Frameworks

App and Engine

Assignment

Algorithm

Manual

QuantumLattices.Essentials.Frameworks.Algorithm — Type

Algorithm(  name::String, engine::Engine;
            din::String=".",
            dout::String=".",
            parameters::Union{Parameters, Nothing}=nothing,
            map::Function=identity,
            assignments::Tuple{Vararg{Assignment}}=(),
            kwargs...
            )

An algorithm associated with an engine.

QuantumLattices.Essentials.Frameworks.App — Type

App

Abstract type for all apps.

QuantumLattices.Essentials.Frameworks.Assignment — Type

Assignment( id::Symbol, app::App, parameters::Parameters;
            map::Function=identity,
            dependences::Tuple{Vararg{Symbol}}=(),
            data::Any=nothing,
            savedata::Bool=true,
            virgin::Bool=true,
            kwargs...
            )

An assignment associated with an app.

QuantumLattices.Essentials.Frameworks.Engine — Type

Engine

Abstract type for all engines.

QuantumLattices.Essentials.Frameworks.dependences — Function

dependences(alg::Algorithm, assign::Assignment, ::Tuple{}=()) -> Tuple{Vararg{Symbol}}
dependences(alg::Algorithm, assign::Assignment, mask::Tuple{Vararg{Symbol}}) -> Tuple{Vararg{Symbol}}

Get the dependences of an assignment and return their ids.

QuantumLattices.Essentials.Frameworks.rundependences! — Function

rundependences!(alg::Algorithm, assign::Assignment, mask::Tuple{Vararg{Symbol}}=()) -> Algorithm

Run the dependences of an assignment. Optionally, some dependences can be jumped by specifying the mask argument.

Base.:== — Method

==(app1::App, app2::App) -> Bool

Judge whether two apps are equivalent to each other.

Base.:== — Method

==(assign1::Assignment, assign2::Assignment) -> Bool

Judge whether two assignments are equivalent to each other.

Base.:== — Method

==(engine1::Engine, engine2::Engine) -> Bool

Judge whether two engines are equivalent to each other.

Base.get — Method

get(alg::Algorithm, id::Symbol) -> Assignment
get(alg::Algorithm, ::Val{id}) where id -> Assignment

Find the assignment registered on a algorithm by its id.

Base.isequal — Method

isequal(app1::App, app2::App) -> Bool

Judge whether two apps are equivalent to each other.

Base.isequal — Method

isequal(assign1::Assignment, assign2::Assignment) -> Bool

Judge whether two assignments are equivalent to each other.

Base.isequal — Method

isequal(engine1::Engine, engine2::Engine) -> Bool

Judge whether two engines are equivalent to each other.

Base.repr — Function

repr(alg::Algorithm, mask::Tuple{Vararg{Symbol}}=(); ndecimal::Int=10) -> String

Get the repr representation of an algorithm.

Optionally, some parameters of the algorithm can be masked. Besides, the maximum number of decimals of the parameters can also be specified.

Base.show — Method

show(io::IO, alg::Algorithm)

Show an algorithm.

Base.summary — Method

summary(alg::Algorithm)

Provide a summary of an algorithm.

Base.valtype — Method

valtype(assign::Assignment)
valtype(::Type{<:Assignment{<:App, <:Parameters, <:Function, <:Tuple{Vararg{Symbol}}, R}}) where R

The type of the data(result) of an assignment.

QuantumLattices.Interfaces.add! — Method

add!(alg::Algorithm, id::Symbol, app::App; kwargs...) -> Algorithm

Add an assignment on a algorithm by providing the contents of the assignment.

The difference between add! and register! is that the add! function does not run the newly added assignment but the register! function does.

QuantumLattices.Interfaces.id — Method

id(assign::Assignment) -> Symbol
id(::Type{<:Assignment{<:App, <:Parameters, <:Function, <:Tuple{Vararg{Symbol}}, <:Any, I}}) where I -> Symbol

The id of an assignment.

QuantumLattices.Interfaces.prepare! — Method

prepare!(alg::Algorithm, assign::Assignment) -> Nothing

Prepare an assignment registered on a algorithm.

QuantumLattices.Interfaces.register! — Method

register!(alg::Algorithm, id::Symbol, app::App; kwargs...) -> Algorithm

Add an assignment on a algorithm by providing the contents of the assignment, and run this assignment.

QuantumLattices.Interfaces.run! — Function

run!(alg::Algorithm, id::Symbol, timing::Bool=true) -> Algorithm
run!(alg::Algorithm, ::Val{id}, timing::Bool=true) where id -> Algorithm

Run an assignment with the given id registered on an algorithm. Optionally, the run process can be timed by setting the timing argument to be true.

QuantumLattices.Interfaces.run! — Method

run!(alg::Algorithm, assign::Assignment) -> Nothing

Run an assignment registered on a algorithm.

QuantumLattices.Interfaces.update! — Method

update!(alg::Algorithm; kwargs...) -> Algorithm

Update the parameters of an algorithm and its associated engine.

QuantumLattices.Interfaces.update! — Method

update!(app::App; kwargs...) -> App

Update the status of an app.

QuantumLattices.Interfaces.update! — Method

update!(assign::Assignment; kwargs...) -> Assignment

Update the parameters of an assignment and the status of its associated app.

QuantumLattices.Interfaces.update! — Method

update!(engine::Engine; kwargs...) -> Engine

Update the status of an engine.