Quantum operators

Quantum operators form an algebra over a field, which are vector spaces with a bilinear operation (often called the "multiplication") between vectors defined.

With the help of the structure constants of the algebra, the result of the bilinear operation between any arbitrary two vectors can be expressed by a sum of individual ones. Therefore, in principle, an algebra can be represented by the complete basis set of its corresponding vector space and a rank-3 tensor encapsulating its structure constants. It is noted that the "bilinear operation" is not restricted to the usual multiplication. For example, it is the commutator, which is a composition of the usual multiplication and subtraction (for any A and B, the commutator [A, B] is defined as [A, B]≝AB-BA) that serves as the bilinear operator for Lie algebras.

In general, there are three basic operations on quantum operators, i.e. the scalar multiplication between a scalar and a quantum operator, the usual addition and the usual multiplication between quantum operators. Other complicated operations can be composed from these basic ones. These basic operations are implemented in this module.

SingularID and ID

SingularID is the building block of the id system of quantum operators, which specifies the basis of the vector space of the corresponding algebra. On the other hand, ID is the direct product (⊗) of several SingularIDs, which could specify the product of the basis quantum operators. This defines an over complete id system for the quantum operators because it does not reflects the structure constants of the corresponding algebra, which would reduce the number of independent basis quantum operators. Extra mechanisms should be provided to kill this redundancy, which goes beyond the current module. Users should define them themselves.

OperatorProd and OperatorSum

OperatorProd defines the product operator as an entity of basis quantum operators while OperatorSum defines the summation as an entity of OperatorProds. Both of them are subtypes of QuantumOperator, which is the abstract type for all quantum operators.

An OperatorProd must have two predefined contents:

value::Number: the coefficient of the quantum operator

id::ID: the id of the quantum operator

Arithmetic operations (+, -, *, /) between a scalar, an OperatorProd or an OperatorSum is defined. See Manual for details.

Manual

QuantumLattices.Essentials.QuantumOperators.ID — Type

ID{I<:SingularID, N}

The (composite) id system of quantum operators.

Type alias for NTuple{N, I} where {N, I<:SingularID}.

QuantumLattices.Essentials.QuantumOperators.ID — Method

ID(ids::SingularID...)
ID(ids::NTuple{N, SingularID}) where N

Get the composite id from singular ids.

QuantumLattices.Essentials.QuantumOperators.ID — Method

ID(::Type{SID}, attrs::Vararg{NTuple{N}, M}) where {SID<:SingularID, N, M}

Get the composite id from the components of singular ids.

QuantumLattices.Essentials.QuantumOperators.Identity — Type

Identity <: Transformation

The identity transformation.

QuantumLattices.Essentials.QuantumOperators.MatrixRepresentation — Type

MatrixRepresentation <: Transformation

The matrix representation.

QuantumLattices.Essentials.QuantumOperators.Numericalization — Type

Numericalization{T<:Number} <: Transformation

The numericalization transformation, which converts the value of a quantum operator to a number of type T.

QuantumLattices.Essentials.QuantumOperators.OperatorProd — Type

OperatorProd{V, I<:ID{SingularID}} <: QuantumOperator{I, V}

The entity that represent the product of several quantum operators.

Basically, a concrete subtype should contain two predefined contents:

value::V: the coefficient of the product

id::I: the total id of the product

QuantumLattices.Essentials.QuantumOperators.OperatorSum — Type

OperatorSum{I<:ID{SingularID}, M<:OperatorProd} <: QuantumOperator{I, M}

The sum of OperatorProds.

Similar items are automatically merged with the aid of the id system.

QuantumLattices.Essentials.QuantumOperators.OperatorSum — Method

OperatorSum(ms::OperatorProd...)
OperatorSum{I, M}(ms::OperatorProd...) where {I<:ID{SingularID}, M<:OperatorProd}

Get the sum of OperatorProds with similar items merged.

QuantumLattices.Essentials.QuantumOperators.QuantumOperator — Type

QuantumOperator{I, V} <:CompositeDict{I, V}

The abstract type of any quantum operator.

QuantumLattices.Essentials.QuantumOperators.Scalar — Type

Scalar{V}

Scalar quantum operator.

QuantumLattices.Essentials.QuantumOperators.SingularID — Type

SingularID <: NamedVector

A singular id is the building block of the id system of quantum operators.

QuantumLattices.Essentials.QuantumOperators.Transformation — Type

Transformation <: Function

Abstract transformation that could transform the quantum operators.

QuantumLattices.Essentials.QuantumOperators.Transformation — Method

(transformation::Transformation)(ms::OperatorSum) -> OperatorSum

Get the transformed quantum operators.

Base.:* — Method

*(factor::Number, m::OperatorProd) -> OperatorProd
*(m::OperatorProd, factor::Number) -> OperatorProd
*(m₁::OperatorProd, m₂::OperatorProd) -> OperatorProd
*(factor::Number, ms::OperatorSum) -> OperatorSum
*(ms::OperatorSum, factor::Number) -> OperatorSum
*(m::OperatorProd, ms::OperatorSum) -> OperatorSum
*(ms::OperatorSum, m::OperatorProd) -> OperatorSum
*(ms₁::OperatorSum, ms₂::OperatorSum) -> OperatorSum

Overloaded * between quantum operators or a quantum operator and a scalar.

Base.:* — Method

*(sid₁::SingularID, sid₂::SingularID) -> ID{SingularID}
*(sid::SingularID, cid::ID{SingularID}) -> ID{SingularID}
*(cid::ID{SingularID}, sid::SingularID) -> ID{SingularID}
*(cid₁::ID{SingularID}, cid₂::ID{SingularID}) -> ID{SingularID}

Get the product of the id system.

Base.:+ — Method

+(m::OperatorProd) -> typeof(m)
+(m::OperatorProd, factor::Number) -> OperatorSum
+(factor::Number, m::OperatorProd) -> OperatorSum
+(m₁::OperatorProd, m₂::OperatorProd) -> OperatorSum
+(ms::OperatorSum) -> typeof(ms)
+(ms::OperatorSum, factor::Number) -> OperatorSum
+(factor::Number, ms::OperatorSum) -> OperatorSum
+(ms::OperatorSum, m::OperatorProd) -> OperatorSum
+(m::OperatorProd, ms::OperatorSum) -> OperatorSum
+(ms₁::OperatorSum, ms₂::OperatorSum) -> OperatorSum

Overloaded + between quantum operators.

Base.:- — Method

-(m::OperatorProd) -> typeof(m)
-(m::OperatorProd, factor::Number) -> OperatorSum
-(factor::Number, m::OperatorProd) -> OperatorSum
-(m₁::OperatorProd, m₂::OperatorProd) -> OperatorSum
-(ms::OperatorSum) -> typeof(ms)
-(ms::OperatorSum, factor::Number) -> OperatorSum
-(factor::Number, ms::OperatorSum) -> OperatorSum
-(m::OperatorProd, ms::OperatorSum) -> OperatorSum
-(ms::OperatorSum, m::OperatorProd) -> OperatorSum
-(ms₁::OperatorSum, ms₂::OperatorSum) -> OperatorSum

Overloaded - between quantum operators.

Base.:/ — Method

/(m::OperatorProd, factor::Number) -> OperatorProd
/(m::OperatorProd, factor::Scalar) -> OperatorProd
/(ms::OperatorSum, factor::Number) -> OperatorSum
/(ms::OperatorSum, factor::Scalar) -> OperatorSum

Overloaded / between a quantum operator and a scalar.

Base.:// — Method

//(m::OperatorProd, factor::Integer) -> OperatorProd
//(m::OperatorProd, factor::Scalar) -> OperatorProd
//(ms::OperatorSum, factor::Integer) ->  OperatorSum
//(ms::OperatorSum, factor::Scalar) -> OperatorSum

Overloaded // between a quantum operator and a scalar.

Base.:^ — Method

^(m::OperatorProd, n::Integer) -> OperatorProd
^(ms::OperatorSum, n::Integer) -> OperatorSum

Overloaded ^ between a quantum operator and an integer.

Base.convert — Method

convert(::Type{M}, m::Scalar) where {M<:Scalar}
convert(::Type{M}, m::Number) where {M<:Scalar}
convert(::Type{M}, m::OperatorProd) where {M<:OperatorProd}

Convert a scalar quantum operator from one type to another;

Convert a scalar to a scalar quantum operator;

Convert a quantum operator from one type to another.

Base.getindex — Method

getindex(m::OperatorProd, i) -> OperatorProd

Overloaded [].

Base.getproperty — Method

getproperty(cid::ID{SingularID}, name::Symbol)

Get the property of a composite id.

Base.isapprox — Method

isapprox(m₁::OperatorProd, m₂::OperatorProd; atol::Real=atol, rtol::Real=rtol) -> Bool

Compare two OperatorProds and judge whether they are approximate to each other.

Base.isapprox — Method

isapprox(ms₁::OperatorSum, ms₂::OperatorSum; atol::Real=atol, rtol::Real=rtol) -> Bool

Compare two OperatorSums and judge whether they are approximate to each other.

Base.isless — Method

isless(::Type{<:SingularID}, cid₁::ID{SingularID}, cid₂::ID{SingularID}) -> Bool

Compare two ids and judge whether the first is less than the second.

The comparison rule are as follows:

ids with smaller ranks are always less than those with higher ranks;

if two ids are of the same rank, the comparison goes just like that between tuples.

Base.length — Method

length(m::OperatorProd) -> Int

Get the length of an OperatorProd.

Base.one — Method

one(::Type{M}) where M<:OperatorProd
one(m::OperatorProd)

Get the identity quantum operator.

Base.permute! — Method

permute!(result::OperatorSum, m::OperatorProd, table) -> OperatorSum
permute!(result::OperatorSum, ms::OperatorSum, table) -> OperatorSum

Permute the singular ids of a quantum operator to the descending order according to a table, and store the permuted quantum operators in result.

Note

To use this function, the user must implement a method of permute, which computes the result of the permutation of two ids and takes the following interface:

permute(id₁::SingularID, id₂::SingularID) -> Union{OperatorProd, OperatorSum}

Here, id₁ and id₂ are two arbitrary singular ids contained in id(m).

Base.promote_rule — Method

promote_rule(::Type{M₁}, ::Type{M₂}) where {M₁<:OperatorProd, M₂<:OperatorProd}

Define the promote rule for OperatorProd types.

Base.promote_type — Method

promote_type(::Type{Tuple{}}, I::Type{<:ID{SingularID, N}}) where N
promote_type(I::Type{<:ID{SingularID, N}}, ::Type{Tuple{}}) where N

Define the promote rule for ID types.

Base.promote_type — Method

promote_type(::Type{M}, ::Type{V}, ::Val{:*}) where {M<:OperatorProd, V<:Number}
promote_type(::Type{V}, ::Type{M}, ::Val{:*}) where {M<:OperatorProd, V<:Number}

Define the promote rule for the multiplication between an OperatorProd and a scalar.

Base.propertynames — Method

propertynames(::Type{I}) where I<:ID{SingularID} -> Tuple{Vararg{Symbol}}

Get the property names of a composite id.

Base.replace — Method

replace(m::OperatorProd, v::Number) -> OperatorProd

Replace the value of an OperatorProd.

Base.replace — Method

replace(m::OperatorProd, pairs::Pair{<:SingularID, <:Union{OperatorProd, OperatorSum}}...) -> OperatorProd/OperatorSum
replace(ms::OperatorSum, pairs::Pair{<:SingularID, <:Union{OperatorProd, OperatorSum}}...) -> OperatorSum

Replace the rank-1 components of a quantum operator with new quantum operators.

Base.replace — Method

replace(m::OperatorProd; kwargs...) -> typeof(m)

Return a copy of a concrete OperatorProd with some of the field values replaced by the keyword arguments.

Base.repr — Method

repr(ms::OperatorSum) -> String

Get the repr representation of a sum of OperatorProds.

Base.split — Method

split(m::OperatorProd) -> Tuple{Any, Vararg{OperatorProd}}

Split an OperatorProd into the coefficient and a sequence of rank-1 OperatorProds.

Base.valtype — Method

valtype(m::QuantumOperator)
valtype(::Type{T}) where {T<:QuantumOperator}

Get the type of the value of a QuantumOperator.

LinearAlgebra.:⋅ — Method

⋅(m::OperatorProd, ms::OperatorSum) -> OperatorSum
⋅(ms::OperatorSum, m::OperatorProd) -> OperatorSum
⋅(ms₁::OperatorSum, ms₂::OperatorSum) -> OperatorSum

Overloaded ⋅ between quantum operators.

LinearAlgebra.mul! — Method

mul!(ms::OperatorSum, factor::Scalar) -> OperatorSum
mul!(ms::OperatorSum, factor::Number) -> OperatorSum

Get the in-place multiplication of an OperatorSum with a scalar.

LinearAlgebra.rank — Method

rank(m::QuantumOperator) -> Int
rank(::Type{M}) where M<:QuantumOperator -> Int

Get the rank of a QuantumOperator.

LinearAlgebra.rank — Method

rank(id::ID{SingularID}) -> Int
rank(::Type{<:ID{SingularID}}) -> Any
rank(::Type{<:ID{SingularID, N}}) where N -> Int

Get the rank of a composite id.

QuantumLattices.Essentials.QuantumOperators.idtype — Method

idtype(m::QuantumOperator)
idtype(::Type{T}) where {T<:QuantumOperator}

The type of the id of a QuantumOperator.

QuantumLattices.Essentials.QuantumOperators.matrix — Function

matrix

Generic matrix representation.

QuantumLattices.Essentials.QuantumOperators.matrix! — Function

matrix!

Generic matrix representation.

QuantumLattices.Essentials.QuantumOperators.sequence — Method

sequence(m::OperatorProd, table) -> NTuple{rank(m), Int}

Get the sequence of the id of a quantum operator according to a table.

QuantumLattices.Interfaces.:⊗ — Method

⊗(m::OperatorProd, ms::OperatorSum) -> OperatorSum
⊗(ms::OperatorSum, m::OperatorProd) -> OperatorSum
⊗(ms₁::OperatorSum, ms₂::OperatorSum) -> OperatorSum

Overloaded ⊗ between quantum operators.

QuantumLattices.Interfaces.add! — Method

add!(ms::OperatorSum) -> typeof(ms)
add!(ms::OperatorSum, m::Number) -> typeof(ms)
add!(ms::OperatorSum, m::OperatorProd) -> typeof(ms)
add!(ms::OperatorSum, mms::OperatorSum) -> typeof(ms)

Get the in-place addition of quantum operators.

QuantumLattices.Interfaces.div! — Method

div!(ms::OperatorSum, factor::Scalar) -> OperatorSum
div!(ms::OperatorSum, factor::Number) -> OperatorSum

Get the in-place division of an OperatorSum with a scalar.

QuantumLattices.Interfaces.id — Method

id(m::OperatorProd) -> idtype(m)

Get the id of an OperatorProd.

QuantumLattices.Interfaces.permute — Method

permute(m::OperatorProd, table) -> OperatorSum
permute(ms::OperatorSum, table) -> OperatorSum

Permute the singular ids of a quantum operator to the descending order according to a table.

QuantumLattices.Interfaces.permute — Method

permute(id₁::SingularID, id₂::SingularID) -> Tuple{Vararg{OperatorProd}}

Permutation rule of two singular ids.

QuantumLattices.Interfaces.sub! — Method

sub!(ms::OperatorSum) -> typeof(ms)
sub!(ms::OperatorSum, m::Number) -> typeof(ms)
sub!(ms::OperatorSum, m::OperatorProd) -> typeof(ms)
sub!(ms::OperatorSum, mms::OperatorSum) -> typeof(ms)

Get the in-place subtraction of quantum operators.

QuantumLattices.Interfaces.value — Method

value(m::OperatorProd) -> valtype(m)

Get the value of an OperatorProd.