Interfaces

This module contains the generic functions that are extended by the package.

Due to the multi-dispatch feature of Julia, generic functions can be extended by local methods for different types. However, a local definition of a method also claims a new generic function if the generic function is not imported to the current scope, thus ruins the definitions in other modules. Therefore, it is quite necessary to predefine the common generic functions in a separate module, so that other modules can extend them with their own by a simple import.

Manual

LinearAlgebra.:⋅ — Function

Dot product.

LinearAlgebra.mul! — Function

In-place multiplication.

LinearAlgebra.rank — Function

Rank.

QuantumLattices.Interfaces.:⊕ — Function

Direct sum.

QuantumLattices.Interfaces.:⊗ — Function

Direct product.

QuantumLattices.Interfaces.add! — Function

In-place addition.

QuantumLattices.Interfaces.decompose — Function

Decompose.

QuantumLattices.Interfaces.decompose! — Function

In place decomposition.

QuantumLattices.Interfaces.dimension — Function

Dimension.

QuantumLattices.Interfaces.div! — Function

In-place division.

QuantumLattices.Interfaces.expand — Function

Get the expansion.

QuantumLattices.Interfaces.expand! — Function

In place expansion.

QuantumLattices.Interfaces.id — Function

ID.

QuantumLattices.Interfaces.permute — Function

Get the permutation.

QuantumLattices.Interfaces.sub! — Function

In-place subtraction.

QuantumLattices.Interfaces.value — Function

Value.