Combinatorics

Combinatorics

This module implements the combinations and permutations of an indexable object, with duplicate elements allowed or not. Compared to another Julia package Combinatorics, the iterators return tuples instead of vectors, which greatly decreases the momory allocation times and improves the code efficiency.

AbstractCombinatorics

AbstractCombinatorics{M,C} is the abstract type of all combinatoric algorithms. It has two type parameters:

M: the number of elements to be taken

C: the type of the collection of candidate elements

To avoid momery allocation, the iteration of a concrete combinatoric algorithm returns a tuple, whose length is M and eltype is eltype(C).

Combinations and DulCombinations

Combinations{M,C} and DulCombinations generate all the combinations of M elements from an indexable collection whose type is C, with the differences being that the former forbids duplicate elements in the combinations while the latter allows.

Permutations and DulPermutations

Permutations{M,C} and DulPermutations generate all the permutations of M elements from an indexable collection whose type is C, with the differences being that the former forbids duplicate elements in the permutations while the latter allows.

Manul

QuantumLattices.Mathematics.Combinatorics.AbstractCombinatorics — Type.

AbstractCombinatorics{M,C}

Abstract combinatoric algorithms.

source

QuantumLattices.Mathematics.Combinatorics.Combinations — Type.

Combinations{M}(contents::C) where {M,C}

Combinations of M elements from contents. Duplicates are not allowed.

source

QuantumLattices.Mathematics.Combinatorics.DulCombinations — Type.

DulCombinations{M}(contents::C) where {M,C}

Combinations of M elements from contents. Duplicates are allowed.

source

QuantumLattices.Mathematics.Combinatorics.DulPermutations — Type.

DulPermutations{M}(contents::C) where {M,C}

Permutations of M elements from contents. Duplicates are not allowed.

source

QuantumLattices.Mathematics.Combinatorics.Permutations — Type.

Permutations{M}(contents::C) where {M,C}

Permutations of M elements from contents. Duplicates are allowed.

source