Vector spaces

Vector spaces

A vector space is a linear space, in which the addition of vectors and multiplication of a vector by a scalar are defined.

Vector spaces are frequently encountered in physics, e.g. the Hilbert space in quantum mechanics. In this submodule, we only implement those with finite dimensions. We want to remark that in our implementation, a vector space is a subtype of an abstract vector, therefore, the bases always possess a order, which means, two vector spaces are not considered to be equal to each other even if their corresponding actual mathmatical spaces are the same but the the orders of the bases are different.

AbstractVectorSpace

AbstractVectorSpace{B} is the abstaction of a vector space, which has only one type parameter:

B<:Any: the type of the bases of the vector space

Basically, a subtype should implement the following 3 methods:

dimension(vs::AbstractVectorSpace) -> Int
Get the dimension of a vector space

Base.getindex(vs::AbstractVectorSpace{B},i::Int)  where B -> B
Get the ith basis of a vector space

Base.searchsortedfirst(vs::AbstractVectorSpace{B},basis::B) where B -> Int
Search the index of a basis in a vector space

However, we provide several interfaces, including type traits and methods to deal with common situations:

A vector space whose bases are stored in a table under the attribute name :table can be ascribed to the HasTable trait and the TableSorted trait. Specifically, the first trait must be implemented as
HasTable(::Type{SubType})=HasTable(true)
While, the second should be implemented as
TableSorted(::Type{SubType})=TableSorted(false)
if the table is unsorted, and as
TableSorted(::Type{SubType})=TableSorted(true)
if the table is sorted.

A vector space whose bases may be represented by a multiindex (Cartesian index) can be ascribed to the traits IsMultiIndexable and MultiIndexOrderStyle. Specifically, the first trait must be implemented as
MultiIndexOrderStyle(::Type{SubType})=MultiIndexOrderStyle('C')
While, the second shoule be implemented as
MultiIndexOrderStyle(::Type{SubType})=MultiIndexOrderStyle('C')
if the order style of the multiindex is C/C++ like, and as
MultiIndexOrderStyle(::Type{SubType})=MultiIndexOrderStyle('F')
if the order style is Fortran like. Furthermore, it should implement the following methods

rank(::Type{SubType}) -> Int
Get the rank of a multiindexable vector space.

dims(vs::SubType) -> NTuple{vs|>typeof|>rank,Int}
Get the dimensions along each axes of a multiindexable vector space.

inds(basis,vs::SubType) ->  NTuple{vs|>typeof|>rank,Int}
Get the Cartesian index representation of a basis in a multiindexable vector space.

eltype(SubType).name.wrapper(index::NTuple{N,Int},vs::SubType)
Construct a basis from a tuple of integers and a multiindexable vector space.

Note that a multiindexable vector space can also have a sorted or unsorted table. But then the trait MultiIndexOrderStyle takes no effects and the sequences of its bases will be completely determined by its attribute :table.

If the type taits and methods are defined properly as stated above, the dimension, getindex and searchsortedfirst methods get default implementations. No need to worry about them any more.

Other features include

comparison: == and isequal

iteration: iterate

inquiry: size, findfirst and in

DirectVectorSpace

DirectVectorSpace{S,B,N} is the simplest vector space, whose bases are stored in the attribute :table as an ntuple.

The :table attribute can be sorted or unsorted, which is determined by the type parameter S, with 'T' for sorted and 'F' for unsorted.

AbstractOrderedIndices

AbstractOrderedIndices{N} defines a simplest class of multiindexable vector spaces, whose bases are just tuples of integers.

This class of vector spaces must have the following attribute: dims::NTuple{N,Int}: the dimesnions of the Cartesian indices along all axes

DirectOrderedIndices

DirectOrderedIndices{M,N} is the direct ordered Cartesian indices.

It is worth noting that

It can be C/C++ ordered or Fortran ordered depending on the first type parameter M, with 'C' for the former and 'F' the latter.

For its bases (Cartesian indices), there is no restriction except that they should be in the proper range defined by its dims.

TabledOrderedIndices

TabledOrderedIndices{S,N} defines the tabled ordered Cartesian indices.

Compared to DirectOrderedIndices, the bases of this kind of vector spaces are stored in the attribute :table, which must be a vector of tuple of integers. The :table attribute can be sorted or unsorted, which is determined by the type parameter S, with 'T' for sorted and 'F' for unsorted. This type suits the situations when the Cartesian indices are restricted by extra conditions except that propoesed by the attribute :dims.

Manul

Hamiltonian.Mathematics.VectorSpaces.AbstractOrderedIndices — Type.

AbstractOrderedIndices{N} <: AbstractVectorSpace{NTuple{N,Int}}

A simplest class of multiindexable vector spaces, whose bases are just tuples of integers.

This class of vector spaces must have the following attribute: dims::NTuple{N,Int}: the dimesnions of the Cartesian indices along all axes

source

Hamiltonian.Mathematics.VectorSpaces.AbstractVectorSpace — Type.

AbstractVectorSpace{B} <: AbstractVector{B}

Abstract vector space.

source

Hamiltonian.Mathematics.VectorSpaces.DirectOrderedIndices — Type.

DirectOrderedIndices{M}(dims::NTuple{N,Int}) where {M,N}

Direct ordered Cartesian indices.

Note

It can be C/C++ ordered or Fortran ordered depending on the first type parameter M, with 'C' for the former and 'F' the latter.

For its bases (Cartesian indices), there is no restriction except that they should be in the proper range defined by its dims.

source

Hamiltonian.Mathematics.VectorSpaces.DirectVectorSpace — Type.

DirectVectorSpace{S}(table::NTuple{N,B}) where {S,B,N}
DirectVectorSpace{S}(table...) where S

Simplest vector space, whose bases are stored in the attribute :table as an ntuple.

The :table attribute can be sorted or unsorted, which is determined by the type parameter S, with 'T' for sorted and 'F' for unsorted.

source

Hamiltonian.Mathematics.VectorSpaces.HasTable — Type.

HasTable(B::Bool)
HasTable(::Type{<:AbstractVectorSpace})

Trait of whether a subtype of AbstractVectorSpace has the attribute :table.

Only two instances are allowed, the first of which is the default for a subtype:

HasTable(false): indication of not having the attribute :table

HasTable(true): indication of having the attribute :table

source

Hamiltonian.Mathematics.VectorSpaces.IsMultiIndexable — Type.

IsMultiIndexable(B::Bool)
IsMultiIndexable(::Type{<:AbstractVectorSpace})

Trait of whether the bases of a subtype of AbstractVectorSpace can be represented by multiindices (Cartesian indices).

Only two instances are allowed, the first of which is the default for a subtype:

IsMultiIndexable(false): indication of irrepresentability by Cartesian indices

IsMultiIndexable(true): indication of representability by Cartesian indices

source

Hamiltonian.Mathematics.VectorSpaces.MultiIndexOrderStyle — Type.

MultiIndexOrderStyle(M::Char)
MultiIndexOrderStyle(::Type{<:AbstractVectorSpace})

Trait of the order style of the Cartesian-index representation of the bases of a subtype of AbstractVectorSpace.

Only two instances are allowed, the first of which is the default for a subtype:

MultiIndexOrderStyle(C): indication of C/C++ order style

MultiIndexOrderStyle(F): indication of Fortran order style

source

Hamiltonian.Mathematics.VectorSpaces.TableSorted — Type.

TableSorted(B::Bool)
TableSorted(::Type{<:AbstractVectorSpace})

Trait of whether the attribute :table of a subtype of AbstractVectorSpace is sorted.

Only two instances are allowed, the first of which is the default for a subtype:

TableSorted(false): indication of unsorted attribute :table

TableSorted(true): indication of sorted attribute :table

source

Hamiltonian.Mathematics.VectorSpaces.TabledOrderedIndices — Type.

TabledOrderedIndices{S}(dims::NTuple{N,Int},table::Vector{NTuple{N,Int}}) where {S,N}
TabledOrderedIndices{N}(::Type{M},len::Int) where {N,M<:AbstractCombinatorics}

Tabled ordered Cartesian indices.

Compared to DirectOrderedIndices, the bases of this kind of vector spaces are stored in the attribute :table, which must be a vector of tuple of integers. The :table attribute can be sorted or unsorted, which is determined by the type parameter S, with 'T' for sorted and 'F' for unsorted. This type suits the situations when the Cartesian indices are restricted by extra conditions except that propoesed by the attribute :dims.

source

Hamiltonian.Prerequisites.Interfaces.:⊕ — Method.

⊕(basis1::B,basis2::B) where B
⊕(basis::B,vs::DirectVectorSpace{<:Any,B}) where B
⊕(vs::DirectVectorSpace{<:Any,B},basis::B) where B
⊕(vs1::DirectVectorSpace{<:Any,B},vs2::DirectVectorSpace{<:Any,B}) where B

Get the direct sum between bases or direct vector spaces.

source

Hamiltonian.Prerequisites.Interfaces.dimension — Method.

dimension(vs::AbstractVectorSpace)

The dimension of a vector space.

source

Base.Sort.searchsortedfirst — Method.

searchsortedfirst(vs::AbstractVectorSpace{B},basis::B) where B -> Int
searchsortedfirst(vs,basis) -> Int

Search the index of a basis in a vector space.

source

Base.findfirst — Method.

findfirst(basis::B,vs::AbstractVectorSpace{B}) where B -> Int
findfirst(bases,vs::AbstractVectorSpace) -> NTuple{length(bases),Int}

Get the index of a basis or the indices of a couple of bases in a vector space.

source

Base.getindex — Method.

getindex(vs::AbstractVectorSpace,i::Int)

Get the ith basis of a vector space by the [] operator.

source

Base.in — Method.

in(basis::B,vs::AbstractVectorSpace{B}) where B

Judge whether a basis belongs to a vector space.

source