Factories

The aim of Factories is to provide tools to hack into Julia codes without knowing the details of their abstract syntax trees and regularize the mechanism to "escape" variables in Expr expressions, so that users can manipulate the existing codes, modify them and generate new ones in macros. In particular, a factory in this module means the representation of certain blocks of Julia codes by a usual Julia struct. This representation is much easier to comprehend than the canonical Expr representation and makes it far more convenient to define macros. In general, we propose the following requirements that any factory must satisfy:

DECOMPOSITION - An Expr expression can be decomposed into its corresponding factory by the factory's constructor.

COMPOSITION - A factory can compose its corresponding Expr expression by calling itself.

ESCAPE - A variable should be or not be escaped in the composed Expr expression by a factory depends on predefined escape mechanisms.

These three requirements also define the basic interfaces to interact with factories. In practice, we combine the second and the third in a single interface, i.e. by passing an instance of certain concrete EscapeMechanism as the only argument of calling a factory, the needed Expr expression with variables correctly escaped can be obtained.

Escape mechanisms

We adopt Julia structs to denote escape mechanisms so that we can utilize Julia's multi-dispatch to implement different mechanisms whereas keeping the same interface.

EscapeMechanism

EscapeMechanism is the abstract type for all concrete escape mechanisms.

Escaped

Escaped has only one attribute:

names::Tuple{Vararg{Symbol}}: the names of variables to be escaped

Apparently, a variable should be escaped if its name is in the names of an Escaped. This mechanism suits a factory whose variables should be unescaped by default.

UnEscaped

UnEscaped also has only on attribute:

names::Tuple{Vararg{Symbol}}: the names of variables not to be escaped

Obviously, on the contrary to Escaped, a variable should be escaped if its name is not in the names of an UnEscaped. This mechanism suits a factory whose variables should be escaped by default.

MixEscaped

MixEscaped has two attributes:

escaped::Escaped: the escaped part of the mixed mechanism

unescaped::UnEscaped: the UnEscaped part of the mixed mechanism

This mechanism suits complex factories that parts of it suit the "escaped" mechanism while others suit the "unescaped" mechanism.

RawExpr

RawExpr has no attributes and it means "raw expression without any variable escaped". This mechanism is used for the print of all factories by default.

Concrete factories

Out of practical purposes, we implement 7 kinds of factories:

Inference: a type inference,

Argument: a function argument,

Parameter: a method or type parameter,

Field: a struct field,

Block: a begin ... end block,

FunctionFactory: a function itself,

TypeFactory: a struct itself.

Some of the basic methods making the above three requirements fulfilled with these types are based on the powerful functions defined in MacroTools.

We want to give a remark that although the types and functions provided in this module helps a lot for the definition of macros, macros should not be abused. On the one hand, some macros may change the language specifications, which makes it hard to understand the codes, and even splits the community; on the one hand, macros usually increases the precompiling/jit time, which means enormous uses of macros in a module may lead to an extremely long load time. Besides, due to the limited ability of the author, the codes in this module are not optimal, which adds to the jit overhead. Any promotion that keeps the interfaces unchanged is welcomed on GitHub issues.

Inference

An Inference has 3 attributes:

head::Union{Symbol, Nothing}: the head of the type inference, which must be one of (nothing, :(<:), :curly)

name::Union{Symbol, Nothing}: the name of the type inference

params::Union{Inference, Vector{Inference}, Nothing}: the parameters of the type inference

All valid expressions representing type inferences can be passed to the constructor.

A type variable represented by a Symbol:

Inference(:T)

Inference(
  head:   nothing
  name:   T
  params: nothing
)

A range of types represented by a UnionAll:

Inference(:(<:Number))

Inference(
  head:   <:
  name:   nothing
  params: Number
)

A type with formal type parameters:

Inference(:(Vector{T}))

Inference(
  head:   curly
  name:   Vector
  params: Inference[T]
)

A definite parametric type:

Inference(:(Vector{Tuple{String, Int}}))

Inference(
  head:   curly
  name:   Vector
  params: Inference[Tuple{String, Int}]
)

A type with type parameters being UnionAlls:

Inference(:(Type{<:Number}))

Inference(
  head:   curly
  name:   Type
  params: Inference[<:Number]
)

On the other hand, you can use the macro @inference to construct an Inference directly from a type inference:

@inference Vector{Tuple{String, Int}}

Inference(
  head:   curly
  name:   Vector
  params: Inference[Tuple{String, Int}]
)

Note

Inference is a recursive struct, i.e. it recursively decomposes a type inference until the final type inference is just a Symbol.

When the input expression is a Symbol, the head and params attributes of the resulting Inference is nothing. Otherwise, its head is the same with that of the input expression, and the args of the input expression will be further decomposed, whose result will be stored in params.

When the head of the input expression is :(<:), the params is an Inference whereas when the head of the input expression is :curly, the params is a Vector{Inference}.

Inference uses the UnEscaped mechanism to escape variables, i.e. you have to specify the names that are not to be escaped.

Empty UnEscaped() escapes everything:

Inference(:(Vector{T}))(UnEscaped())

:(($(Expr(:escape, :Vector))){$(Expr(:escape, :T))})

Unescape formal type parameters:

Inference(:(Vector{T}))(UnEscaped(:T))

:(($(Expr(:escape, :Vector))){T})

Unescape both type names and type parameters:

Inference(:(Vector{T}))(UnEscaped(:Vector, :T))

:(Vector{T})

Argument

An Argument has 4 attributes:

name::Union{Symbol, Nothing}: the name of the argument

type::Inference: the type inference of the argument

slurp::Bool: whether the argument should be expanded by ...

default::Any: the default value of the argument, nothing for those with no default values

All valid expressions representing the arguments of functions can be passed to the constructor.

A formal argument represented by a Symbol:

Argument(:arg)

Argument(
  name:    arg
  type:    Any
  slurp:   false
  default: nothing
)

A formal argument with its type specified:

Argument(:(arg::ArgType))

Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: nothing
)

A slurped formal argument:

Argument(:(arg::ArgType...))

Argument(
  name:    arg
  type:    ArgType
  slurp:   true
  default: nothing
)

A formal argument with a default value

Argument(:(arg::ArgType=default))

Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: default
)

Or you can use the macro @argument for a direct construction from an argument declaration:

@argument arg::ArgType=default

Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: default
)

The construction from such expressions is based on the MacroTools.splitarg function.

Argument uses the MixEscaped mechanism to escape variables, with the UnEscaped mechanism for type and Escaped mechanism for default, e.g.

Argument(:(arg::Real=zero(Int)))(MixEscaped(UnEscaped(), Escaped(:zero, :Int)))

:($(Expr(:kw, :(arg::$(Expr(:escape, :Real))), :(($(Expr(:escape, :zero)))($(Expr(:escape, :Int)))))))

It can be seen that the name of an argument will never be escaped, which is obvious since the name of a function argument is always local. By the way, the composition of an Argument expression is based on the MacroTools.combinearg function.

Parameter

A Parameter has 3 attributes:

head::Union{Symbol, Nothing}: the head of the parameter, must be either :(<:) or :(::) if not nothing

name::Union{Symbol, Nothing}: the name of the parameter

type::Union{Inference, Nothing}: the type inference of the parameter

All expressions that represent type parameters or method parameters are allowed to be passed to the constructor.

A formal type/function parameter represented by a Symbol:

Parameter(:T)

Parameter(
  head: nothing
  name: T
  type: nothing
)

A formal type/function parameter represented by a UnionAll with the name omitted:

Parameter(:(<:Number))

Parameter(
  head: <:
  name: nothing
  type: Number
)

A formal type/function parameter represented by a UnionAll with the name specified:

Parameter(:(T<:Number))

Parameter(
  head: <:
  name: T
  type: Number
)

The macro @parameter completes the construction directly from a parameter declaration:

@parameter T<:Number

Parameter(
  head: <:
  name: T
  type: Number
)

Note

We use nothing to denote a missing name or type.

Two subtle situations of type/method parameters, e.g. MyType{T} and MyType{Int}, should be distinguished by Parameter(:T) and Parameter(:(::Int)).

Parameter uses the UnEscaped mechanism to escape variables, too, e.g.

Parameter(:(N<:Vector{T}))(UnEscaped(:T))

:(N <: ($(Expr(:escape, :Vector))){T})

As is similar to Argument, the name of a method/type parameter will never be escaped because of its local scope.

Field

A Field has 2 attributes:

name::Symbol: the name of the field

type::Inference: the type inference of the field

Legal expressions can be used to construct a Field instance by its constructor.

A field name specified by a Symbol:

Field(:field)

Field(
  name: field
  type: Any
)

A field with both the name and type specified:

Field(:(field::FieldType))

Field(
  name: field
  type: FieldType
)

A field with the type being parametric:

Field(:(field::ParametricType{T}))

Field(
  name: field
  type: ParametricType{T}
)

The macro @field is also provided to help the construction directly from a field declaration:

@field field::FieldType

Field(
  name: field
  type: FieldType
)

The construction from these expressions is based on the MacroTools.splitarg function.

Field uses the UnEscaped mechanism to escape variables as well, e.g.

Field(:(field::Dict{N, D}))(UnEscaped(:N, :D))

:(field::($(Expr(:escape, :Dict))){N, D})

The name of a struct will never be escaped either because it is a local variable tightly binding to a struct. It is noted that the composition of field expressions is based on the MacroTools.combinefield function.

Block

A Block has only one attribute:

body::Vector{Any}: the body of the begin ... end block

Any expression can be passed to the constructor of Block.

A single expression:

Block(:(x = 1))

Block(
  body: Any[:(x = 1)]
)

A semicolon expression:

Block(:(x = 1; y = 2))

Block(
  body: Any[:(x = 1), :(#= none:1 =#), :(y = 2)]
)

A begin ... end block:

Block(:(begin x = 1 end))

Block(
  body: Any[:(#= none:1 =#), :(x = 1)]
)

A quote block:

Block(quote
        x = 1
        y = 2
    end)

Block(
  body: Any[:(#= none:2 =#), :(x = 1), :(#= none:3 =#), :(y = 2)]
)

Or you can construct a Block instance directly from any code by the macro @block:

@block x=1 y=2

Block(
  body: Any[:(x = 1), :(y = 2)]
)

The body of a block can also be extended by the push! function or the @push! macro.

Note

The body of a Block is somewhat "flattened", i.e. it contains no begin ... end blocks. During the initialization, any such input block will be unblocked and added to the body part by part. So is the push! and @push! procedures.

All LineNumberNodes generated by the input codes will also be included in the block's body. However, you can use rmlines! or @rmlines! to remove them from the body of an existing Block, or use rmlines or @rmlines to get a copy with them removed in the body.

Block uses the Escaped mechanism to escape variables. This is because variables in a block are often local ones and should not be escaped. Therefore, only those defined in other modules should be noted and escaped, which usually constitute the minority. For example,

Block(:(x = 1; y = 2; z = Int[1, 2, 3]))(Escaped(:Int))

quote
    x = 1
    #= none:1 =#
    y = 2
    #= none:1 =#
    z = ($(Expr(:escape, :Int)))[1, 2, 3]
end

FunctionFactory

A FunctionFactory has 7 attributes:

name::Union{Symbol, Expr}: the name of the function

params::Vector{Inference}: the method parameters of the function

args::Vector{Argument}: the positional arguments of the function

kwargs::Vector{Argument}: the keyword arguments of the function

rtype::Inference: the return type of the function

whereparams::Vector{Parameter}: the method parameters specified by the where keyword

body::Block: the body of the function

All expressions that represent functions are allowed to be passed to the constructor.

A function with no arguments:

FunctionFactory(:(f() = nothing))

FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[]
  kwargs:      Argument[]
  rtype:       Any
  whereparams: Parameter[]
  body:        begin
                   #= none:1 =#
                   nothing
               end
)

A function with arguments:

FunctionFactory(:(f(x) = x))

FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::Any]
  kwargs:      Argument[]
  rtype:       Any
  whereparams: Parameter[]
  body:        begin
                   #= none:1 =#
                   x
               end
)

A function with arguments that have default values:

FunctionFactory(:(f(x::Int, y::Int; choice::Function=sum) = choice(x, y)))

FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::Int, y::Int]
  kwargs:      Argument[$(Expr(:kw, :(choice::Function), :sum))]
  rtype:       Any
  whereparams: Parameter[]
  body:        begin
                   #= none:1 =#
                   choice(x, y)
               end
)

A function with arguments whose type are specified by where keyword:

FunctionFactory(:(f(x::T, y::T; choice::Function=sum) where {T<:Number} = choice(x, y)))

FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[$(Expr(:kw, :(choice::Function), :sum))]
  rtype:       Any
  whereparams: Parameter[T <: Number]
  body:        begin
                   #= none:1 =#
                   choice(x, y)
               end
)

A function with return type:

FunctionFactory(:((f(x::T, y::T; choice::Function=sum)::T) where {T<:Number} = choice(x, y)))

FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[$(Expr(:kw, :(choice::Function), :sum))]
  rtype:       T
  whereparams: Parameter[T <: Number]
  body:        begin
                   #= none:1 =#
                   choice(x, y)
               end
)

A multiline function:

FunctionFactory(:(
    function (f(x::T, y::T; choice::Function=sum)::T) where T<:Number
        choice(x, y)
    end
))

FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[$(Expr(:kw, :(choice::Function), :sum))]
  rtype:       T
  whereparams: Parameter[T <: Number]
  body:        begin
                   #= none:2 =#
                   #= none:3 =#
                   choice(x, y)
               end
)

A function in a quote block:

FunctionFactory(
    quote
        function (f(x::T, y::T; choice::Function=sum)::T) where T<:Number
            choice(x, y)
        end
    end
)

FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[$(Expr(:kw, :(choice::Function), :sum))]
  rtype:       T
  whereparams: Parameter[T <: Number]
  body:        begin
                   #= none:3 =#
                   #= none:4 =#
                   choice(x, y)
               end
)

Similarly, an instance can also be constructed from the macro @functionfactory:

@functionfactory (f(x::T, y::T; choice::Function=sum)::T) where {T<:Number} = choice(x, y)

FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[$(Expr(:kw, :(choice::Function), :sum))]
  rtype:       T
  whereparams: Parameter[T <: Number]
  body:        begin
                   #= none:1 =#
                   choice(x, y)
               end
)

The construction from such expressions are based on the MacroTools.splitdef function.

Note

Since Julia 0.7, the form MyType{D}(data::D) where D only appears in struct constructors, therefore, the attribute :params of a function factory is nonempty only when this factory aims to represent a struct constructor.

Usually, the name of a function factory is a Symbol. However, if the factory aims to extend some methods of a function defined in another module, e.g., Base.eltype, the name will be an Expr.

FunctionFactory adopts the MixEscaped mechanism to escape variables, with UnEscaped for params, args, kwargs, rtype and whereparams while Escaped for name and body. It is worth to emphasize that the name of a function factory belongs to the Escaped part. Therefore, when it is an Expr, it will never be escaped because an Expr cannot be a element of a Tuple{Vararg{Symbol}}. See following examples.

Escape the function name:

FunctionFactory(:(
    (f(x::T, y::T; choice::Function=sum)::T) where {T<:Number} = max(x, y, choice(x, y))
    ))(MixEscaped(UnEscaped(:T), Escaped(:f, :max)))

:(function (($(Expr(:escape, :f)))(x::T, y::T; choice::$(Expr(:escape, :Function)) = sum)::T) where T <: $(Expr(:escape, :Number))
      #= none:2 =#
      ($(Expr(:escape, :max)))(x, y, choice(x, y))
  end)

Do not escape the function name:

FunctionFactory(:(
    (f(x::T, y::T; choice::Function=sum)::T) where {T<:Number} = max(x, y, choice(x, y))
    ))(MixEscaped(UnEscaped(:T), Escaped(:max)))

:(function (f(x::T, y::T; choice::$(Expr(:escape, :Function)) = sum)::T) where T <: $(Expr(:escape, :Number))
      #= none:2 =#
      ($(Expr(:escape, :max)))(x, y, choice(x, y))
  end)

The compositions of function expressions are based on the MacroTools.combinedef function.

Other features include:

Positional arguments can be added by addargs! or @addargs!

Keyword arguments can be added by addkwargs! or @addkwargs!

Where parameters can be added by addwhereparams! or @addwhereparams!

Body can be extended by extendbody! or @extendbody!

TypeFactory

A TypeFactory has 6 attributes:

name::Symbol: the name of the struct

mutable::Bool: whether or not the struct is mutable

params::Vector{Parameter}: the type parameters of the struct

supertype::Inference: the supertype of the struct

fields::Vector{Field}: the fields of the struct

constructors::Vector{FunctionFactory}: the inner constructors of the struct

Any expression representing valid struct definitions can be passed to the constructor.

A struct with no fields:

TypeFactory(:(struct StructName end))

TypeFactory(
  name:         StructName
  mutable:      false
  params:       Parameter[]
  supertype:    Any
  fields:       Field[]
  constructors: FunctionFactory[]
)

A parametric struct with no fields:

TypeFactory(:(struct StructName{T} end))

TypeFactory(
  name:         StructName
  mutable:      false
  params:       Parameter[T]
  supertype:    Any
  fields:       Field[]
  constructors: FunctionFactory[]
)

An empty struct with a supertype:

TypeFactory(:(struct Child{T} <: Parent{T} end))

TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T]
  supertype:    Parent{T}
  fields:       Field[]
  constructors: FunctionFactory[]
)

A fielded struct with a supertype

TypeFactory(:(
    struct Child{T<:Number} <: Parent{T}
        field1::T
        field2::T
    end
))

TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T <: Number]
  supertype:    Parent{T}
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[]
)

A struct in a quote block:

TypeFactory(
    quote
        struct Child{T<:Number} <: Parent{T}
            field1::T
            field2::T
        end
    end
)

TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T <: Number]
  supertype:    Parent{T}
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[]
)

Also, the macro @typefactory supports the construction directly from a type definition:

@typefactory struct Child{T<:Number} <: Parent{T}
                field1::T
                field2::T
                Child(field1::T, field2::T=zero(T)) where {T} = new{T}(field1, field2)
            end

TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T <: Number]
  supertype:    Parent{T}
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[
                  function (Child(field1::T, field2::T = zero(T); )::Any) where T
                      new{T}(field1, field2)
                  end
                  ]
)

The construction from these expressions is based on the MacroTools.splitstructdef function.

TypeFactory also uses the MixEscaped mechanism to escape variables, with the UnEscaped part for params, supertype and fields, the Escaped part for name, and both for constructors. For example,

@typefactory(struct Child{T<:Number} <: Parent{T}
    field::T
    Child(field::T) where {T} = new{T}(field)
end)(MixEscaped(UnEscaped(:T), Escaped(:Child)))

:(struct ($(Expr(:escape, :Child))){T <: $(Expr(:escape, :Number))} <: ($(Expr(:escape, :Parent))){T}
      #= /home/travis/.julia/packages/MacroTools/X77lQ/src/structdef.jl:57 =#
      field::T
      #= /home/travis/.julia/packages/MacroTools/X77lQ/src/structdef.jl:58 =#
      function (($(Expr(:escape, :Child)))(field::T; )::$(Expr(:escape, :Any))) where T
          new{T}(field)
      end
  end)

The composition of a type expression is based on the MacroTools.combinestructdef function.

Other features include:

Fields can be added by addfields! or @addfields!

Type parameters can be added by addparams! or @addparams!

Inner constructors can be added by addconstructors! or @addconstructors!

Manual

QuantumLattices.Prerequisites.Factories.FExpr — Constant

Factory expression types, which is defined as Union{Symbol, Expr}.

QuantumLattices.Prerequisites.Factories.rawexpr — Constant

rawexpr

Indicate that no variable in a factory should be escaped.

QuantumLattices.Prerequisites.Factories.AbstractFactory — Type

AbstractFactory

Abstract type for all concrete factories.

QuantumLattices.Prerequisites.Factories.Argument — Type

Argument(name::Union{Symbol, Nothing}, type::Inference, slurp::Bool, default::Any)
Argument(; name::Union{Symbol, Nothing}=nothing, type::Inference=Inference(:Any), slurp::Bool=false, default::Any=nothing)
Argument(expr::FExpr)

The struct to describe a argument of a function.

QuantumLattices.Prerequisites.Factories.Argument — Method

(a::Argument)(em::RawExpr) -> Expr
(a::Argument)(em::MixEscaped) -> Expr

Convert an Argument to the Expr representation of the argument it describes.

QuantumLattices.Prerequisites.Factories.Block — Type

Block(parts::FExpr...)

The struct to describe a begin ... end block.

QuantumLattices.Prerequisites.Factories.Block — Method

(b::Block)(em::RawExpr) -> Expr
(b::Block)(em::Escaped) -> Expr
(b::Block)(em::MixEscaped) -> Expr

Convert a Block to the Expr representation of the begin ... end block it describes.

QuantumLattices.Prerequisites.Factories.EscapeMechanism — Type

Abstract escape mechanism.

QuantumLattices.Prerequisites.Factories.Escaped — Type

Escaped(names::Symbol...)

Indicate that symbols of a factory should be escaped if they are in names.

QuantumLattices.Prerequisites.Factories.Field — Type

Field(name::Symbol, type::Inference)
Field(; name::Symbol, type::FExpr=Inference(:Any))
Field(expr::FExpr)

The struct to describe a field of a struct.

QuantumLattices.Prerequisites.Factories.Field — Method

(f::Field)(em::RawExpr) -> Expr
(f::Field)(em::UnEscaped) -> Expr
(f::Field)(em::MixEscaped) -> Expr

Convert a Field to the Expr representation of the field it describes.

QuantumLattices.Prerequisites.Factories.FunctionFactory — Type

FunctionFactory(name::FExpr,
                params::Vector{Inference},
                args::Vector{Argument},
                kwargs::Vector{Argument},
                rtype::Inference,
                whereparams::Vector{Parameter},
                body::Block
                )
FunctionFactory(;   name::FExpr,
                    params::Vector{Inference}=Inference[],
                    args::Vector{Argument}=Argument[],
                    kwargs::Vector{Argument}=Argument[],
                    rtype::Inference=Inference(:Any),
                    whereparams::Vector{Parameter}=Parameter[],
                    body::Block=Block()
                    )
FunctionFactory(expr::Expr)

The struct to describe a function.

QuantumLattices.Prerequisites.Factories.FunctionFactory — Method

(ff::FunctionFactory)(em::RawExpr) -> Expr
(ff::FunctionFactory)(em::MixEscaped) -> Expr

Convert a FunctionFactory to the Expr representation of the function it describes.

QuantumLattices.Prerequisites.Factories.Inference — Type

Inference(head::Union{Symbol, Nothing}, name::Union{Symbol, Nothing}, params::Union{Inference, Vector{Inference}, Nothing})
Inference(;
        head::Union{Symbol, Nothing}=nothing,
        name::Union{Symbol, Nothing}=nothing,
        params::Union{Inference, Vector{Inference}, Nothing}=nothing,
        )
Inference(expr::FExpr)

The struct to describe a type inference.

QuantumLattices.Prerequisites.Factories.Inference — Method

(i::Inference)(em::RawExpr) -> FExpr
(i::Inference)(em::UnEscaped) -> FExpr
(i::Inference)(em::MixEscaped) -> FExpr

Convert a Inference to the Expr representation of the type inference it describes.

QuantumLattices.Prerequisites.Factories.MixEscaped — Type

MixEscaped(escaped::Escaped)
MixEscaped(unescaped::UnEscaped)
MixEscaped(escaped::Escaped, unescaped::UnEscaped)
MixEscaped(unescaped::UnEscaped, escaped::Escaped)

Indicate that some parts of a factory use the Escaped mechanism while other parts use the UnEscaped mechanism.

QuantumLattices.Prerequisites.Factories.Parameter — Type

Parameter(head::Union{Symbol, Nothing}, name::Union{Symbol, Nothing}, type::Union{Inference, Nothing})
Parameter(; head::Union{Symbol, Nothing}=nothing, name::Union{Symbol, Nothing}=nothing, type::Union{Inference, Nothing}=nothing)
Parameter(expr::FExpr)

The struct to describe a parameter of a function or a type.

QuantumLattices.Prerequisites.Factories.Parameter — Method

(p::Parameter)(em::RawExpr) -> FExpr
(p::Parameter)(em::UnEscaped) -> FExpr
(p::Parameter)(em::MixEscaped) -> FExpr

Convert a Parameter to the Expr representation of the parameter it describes.

QuantumLattices.Prerequisites.Factories.RawExpr — Type

Raw expression without any variable escaped.

QuantumLattices.Prerequisites.Factories.TypeFactory — Type

TypeFactory(name::Symbol, mutable::Bool, params::Vector{Parameter}, supertype::Inference, fields::Vector{Field}, constructors::Vector{FunctionFactory})
TypeFactory(;   name::Symbol,
                mutable::Bool=false,
                params::Vector{Parameter}=Parameter[],
                supertype::Inference=Inference(:Any),
                fields::Vector{Field}=Field[],
                constructors::Vector{FunctionFactory}=FunctionFactory[]
                )
TypeFactory(expr::Expr)

The struct to describe a struct.

QuantumLattices.Prerequisites.Factories.TypeFactory — Method

(tf::TypeFactory)(em::RawExpr) -> Expr
(tf::TypeFactory)(em::MixEscaped) -> Expr

Convert a TypeFactory to the Expr representation of the struct it describes.

QuantumLattices.Prerequisites.Factories.UnEscaped — Type

UnEscaped(names::Symbol...)

IIndicate that symbols of a factory should be escaped if they are not in names.

QuantumLattices.Prerequisites.Factories.@addargs! — Macro

@addargs! ff args::FExpr...

Add a couple of positional arguments to a function factory.

QuantumLattices.Prerequisites.Factories.@addconstructors! — Macro

@addconstructors! tf constructors::Expr...

Add a couple of constructors to a type factory.

QuantumLattices.Prerequisites.Factories.@addfields! — Macro

@addfields! tf fields::FExpr...

Add a couple of fields to a type factory.

QuantumLattices.Prerequisites.Factories.@addkwargs! — Macro

@addkwargs! ff kwargs::FExpr...

Add a couple of keyword arguments to a function factory.

QuantumLattices.Prerequisites.Factories.@addparams! — Macro

@addparams! f params::FExpr...

Add a couple of method parameters to a function factory or a type factory.

QuantumLattices.Prerequisites.Factories.@addwhereparams! — Macro

@addwhereparams! ff whereparams::FExpr...

Add a couple of method parameters to a function factory or a type factory.

QuantumLattices.Prerequisites.Factories.@argument — Macro

@argument expr::FExpr

Construct an Argument directly from an argument statement.

QuantumLattices.Prerequisites.Factories.@block — Macro

@block parts::FExpr...

Construct a Block directly from a begin ... end block definition.

QuantumLattices.Prerequisites.Factories.@extendbody! — Macro

@extendbody! ff parts::FExpr...

Extend the body of a function factory.

QuantumLattices.Prerequisites.Factories.@field — Macro

@field expr::FExpr

Construct a Field directly from a field statement.

QuantumLattices.Prerequisites.Factories.@functionfactory — Macro

@functionfactory expr::FExpr

Construct a FunctionFactory directly from a function definition.

QuantumLattices.Prerequisites.Factories.@inference — Macro

@inference expr::FExpr

Construct an Inference directly from a type inference.

QuantumLattices.Prerequisites.Factories.@parameter — Macro

@parameter expr::FExpr

Construct a Parameter directly from an parameter statement.

QuantumLattices.Prerequisites.Factories.@push! — Macro

@push! b parts::FExpr...

Push other parts into the body of a block.

QuantumLattices.Prerequisites.Factories.@rmlines! — Macro

@rmlines! b::Expr

Remove line number nodes in the body of a block.

QuantumLattices.Prerequisites.Factories.@rmlines — Macro

@rmlines b::Expr

Return a copy of a block with the line number nodes removed.

QuantumLattices.Prerequisites.Factories.@typefactory — Macro

@typefactory expr::Expr

Construct a TypeFactory directly from a type definition.

Base.:== — Method

==(f1::F, f2::F) where F<:AbstractFactory -> Bool

Overloaded equivalent operator.

Base.isequal — Method

isequal(f1::F, f2::F) where F<:AbstractFactory -> Bool

Overloaded equivalent operator.

Base.names — Method

names(expr::Symbol) -> Vector{Symbol}
names(expr::Expr) -> Vector{Symbol}

Get all the symbols in an expression.

Base.push! — Method

push!(b::Block, parts::FExpr...) -> Block
push!(b::Block, parts::Block...) -> Block

Push other parts into the body of a block.

Base.replace — Method

replace(f::AbstractFactory; kwargs...) -> typeof(f)

Return a copy of a concrete AbstractFactory with some of the field values replaced by the keyword arguments.

Base.show — Method

show(io::IO, f::AbstractFactory)

Show a concrete AbstractFactory.

MacroTools.rmlines — Method

rmlines(b::Block) -> Block

Return a copy of a block with the line number nodes removed.

QuantumLattices.Prerequisites.Factories.addargs! — Method

addargs!(ff::FunctionFactory, args::Argument...) -> FunctionFactory
addargs!(ff::FunctionFactory, args::FExpr...) -> FunctionFactory

Add a couple of positional arguments to a function factory.

QuantumLattices.Prerequisites.Factories.addconstructors! — Method

addconstructors!(tf::TypeFactory, constructors::FunctionFactory...) -> TypeFactory
addconstructors!(tf::TypeFactory, constructors::Expr...) -> TypeFactory

Add a couple of constructors to a type factory.

QuantumLattices.Prerequisites.Factories.addfields! — Method

addfields!(tf::TypeFactory, fields::Field...) -> TypeFactory
addfields!(tf::TypeFactory, fields::FExpr...) -> TypeFactory

Add a couple of fields to a type factory.

QuantumLattices.Prerequisites.Factories.addkwargs! — Method

addkwargs!(ff::FunctionFactory, kwargs::Argument...) -> FunctionFactory
addkwargs!(ff::FunctionFactory, kwargs::FExpr...) -> FunctionFactory

Add a couple of keyword arguments to a function factory.

QuantumLattices.Prerequisites.Factories.addparams! — Method

addparams!(f::FunctionFactory, params::Inference...) ->FunctionFactory
addparams!(f::FunctionFactory, params::FExpr...) -> FunctionFactory
addparams!(f::TypeFactory, params::Parameter...) -> TypeFactory
addparams!(f::TypeFactory, params::FExpr...) -> TypeFactory

Add a couple of parameters to a function factory or a type factory.

QuantumLattices.Prerequisites.Factories.addwhereparams! — Method

addwhereparams!(ff::FunctionFactory, whereparams::Parameter...) -> FunctionFactory
addwhereparams!(ff::FunctionFactory, whereparams::FExpr...) -> FunctionFactory

Add a couple of method where parameters to a function factory or a type factory.

QuantumLattices.Prerequisites.Factories.escape — Method

escape(expr, ::RawExpr) -> Any
escape(expr, ::Escaped) -> Any
escape(expr, ::UnEscaped) -> Any
escape(expr::Symbol, em::Escaped) -> FExpr
escape(expr::Expr, em::Escaped) -> Expr
escape(expr::Symbol, em::UnEscaped) -> FExpr
escape(expr::Expr, em::UnEscaped) -> Expr

Escape the variables in the input expression.

QuantumLattices.Prerequisites.Factories.extendbody! — Method

extendbody!(ff::FunctionFactory, parts::FExpr...) -> FunctionFactory
extendbody!(ff::FunctionFactory, parts::Block...) -> FunctionFactory

Extend the body of a function factory.

QuantumLattices.Prerequisites.Factories.rmlines! — Method

rmlines!(b::Block) -> Block

Remove line number nodes in the body of a block.