Interfaces

Interfaces

This submodule contains the generic functions that are extended by the package.

Due to the multidispatch feature of Julia, generic functions can be extended by local methods for different types. However, a local definition of a method also claims a new generic function if the generic function is not imported to the current scope, thus ruins the definitions in other modules. Therefore, it is quite necessary to prefine the common generic functions in a separate module, so that other modules can extend them with their own by a simple import.

Manual

Hamiltonian.Prerequisites.Interfaces.:⊕ — Function.

Direct sum.

source

Hamiltonian.Prerequisites.Interfaces.:⊗ — Function.

Direct product.

source

Hamiltonian.Prerequisites.Interfaces.add! — Function.

Inplace addition.

source

Hamiltonian.Prerequisites.Interfaces.decompose — Function.

Decompose.

source

Hamiltonian.Prerequisites.Interfaces.degree — Function.

Degree

source

Hamiltonian.Prerequisites.Interfaces.dimension — Function.

Dimension.

source

Hamiltonian.Prerequisites.Interfaces.dims — Function.

Dimensions.

source

Hamiltonian.Prerequisites.Interfaces.div! — Function.

Inplace division.

source

Hamiltonian.Prerequisites.Interfaces.expand — Function.

Get the expansion.

source

Hamiltonian.Prerequisites.Interfaces.index — Function.

Index.

source

Hamiltonian.Prerequisites.Interfaces.inds — Function.

Indices.

source

Hamiltonian.Prerequisites.Interfaces.matrix — Function.

Matrix representation.

source

Hamiltonian.Prerequisites.Interfaces.mul! — Function.

Inplace multiplication.

source

Hamiltonian.Prerequisites.Interfaces.permute — Function.

Get the permutation.

source

Hamiltonian.Prerequisites.Interfaces.rank — Function.

Rank.

source

Hamiltonian.Prerequisites.Interfaces.sub! — Function.

Inplace subtraction.

source

Hamiltonian.Prerequisites.Interfaces.vector — Function.

Vector representation.

source

Hamiltonian.Prerequisites.Interfaces.update! — Function.

Inplace Update.

source