Simple trees

Simple trees

The aim of this module is to represent the standard tree structure in efficiency-non-sensitive cases. Please note that the default implementation of tree methods are far from optimal in efficiency. Therefore, please DO NOT use it if you need an efficient tree for addition, deletion, sort and inquiry. This module of codes apply only when the structure of tree matters but not the efficiency.

AbstractSimpleTree

AbstractSimpleTree{N,D} is the abstract type for all concrete trees. By design, it has two type parameters:

N: the type of the tree's node

D: the type of the tree's data

To fully utilize the methods designed for a tree structure, in our protocol, a concrete subtype must implement the following methods:

inquiry related methods

root(tree::AbstractSimpleTree{N}) where N -> Union{N,Nothing}
Get a tree's root node (nothing for empty trees)

haskey(tree::AbstractSimpleTree{N},node::N) where N -> Bool
Check whether a node is in a tree.

length(tree::AbstractSimpleTree) -> Int
Get the number of a tree's nodes.

parent(tree::AbstractSimpleTree{N},
       node::N,
       superparent::Union{N,Nothing}=nothing
       ) where N -> Union{N,Nothing}
Get the parent of a tree's node or return superparent when the input node is the tree's root.

children(tree::AbstractSimpleTree{N},node::N) where N -> Vector{N}
Get the children of a tree's node.

structure modification related methods

addnode!(tree::AbstractSimpleTree{N},
         parent::Union{N,Nothing},
         node::N
         ) where N
Update the structure of a tree by adding a node. When the parent is nothing, the input tree must be empty and the input node becomes the tree's root.

deletenode!(tree::AbstractSimpleTree{N},node::N) where N
Update the structure of a tree by deleting a node.

index related methods

getindex(tree::AbstractSimpleTree{N,D},node::N) where {N,D} -> D
Get the data of a tree's node

setindex!(tree::AbstractSimpleTree{N,D},node::N,data::D) where {N,D}
Set the data of a tree's node.

Based on these methods, we implement several generic functions for inquiries and manipulations

inquiry for type parameters: keytype, valtype, eltype

expansion over nodes/data-records: keys, values, pairs

inquiry for info of nodes: isleaf, level

inquiry for nodes: ancestor, descendants, siblings, leaves

modification: push!, append!, delete!, empty!

And optionally, when a subtype implement the following method,

empty(tree::AbstractSimpleTree) -> typeof(tree)

which constructs an empty tree of the same type with the input one, two more more methods are supported:

subtree: Get a subtree starting from a node.

move!: Move a subtree to a new position.

SimpleTreeCore and SimpleTree

To implement all the prerequisites listed above costs a bit efforts. We provide two lazy ways to get over this:

Inheritance AbstractSimpleTree with TREECORE::SimpleTreeCore as the last attribute

Inclusion an attribute which is an instance of SimpleTree

SimpleTreeCore

SimpleTreeCore{N,D}, as the literal meaning indicates, is the core of a tree. It encapsulates all the data structures needed by the default implementation, which constains 4 attributes:

root::N: the tree's root node

contents::Dict{N,D}: the tree's (node,data) pairs

parent::Dict{N,N}: records of the parent of each of the tree's nodes

children::Dict{N,Vector{N}}: records of the children of each of the tree's nodes

As above, the first lazy way is to include this struct with the special name :TREECORE in your concrete subtype as the last attribute. This process can be even lazier, in that we provide a macro @simpletree to decorate your "raw" struct automatically, e.g.

julia> @simpletree(struct SimpleSubTree end)

julia> @simpletree(struct SubTreeWithTreeParameters end,
             {N<:AbstractString,D<:Number}
             )

julia> @simpletree(struct SubTreeWithCertainTreeParameters end,
             {<:String,<:Int}
             )

julia> @simpletree(struct SubTreeWithFields info::Vector{Int} end,
             {N<:AbstractString,D<:Number}
             )

julia> @simpletree(struct SubTreeWithParametricFields{T} info::Vector{T} end,
             {N<:AbstractString,D<:Number}
             )

julia> @simpletree(struct SubTreeWithOverlappedParametricFields{N} info::Vector{N} end,
             {N<:AbstractString,D<:Number}
             )

SimpleTree

SimpleTree{N,D} is the minimum struct that implements all the default tree methods. You can include an instance of it as an attribute in your own type to utilize all the tree methods.

Manual

QuantumLattices.Prerequisites.SimpleTrees.simpletreedepth — Constant.

simpletreedepth

Indicate that the iteration over a tree is depth-first.

source

QuantumLattices.Prerequisites.SimpleTrees.simpletreewidth — Constant.

simpletreewidth

Indicate that the iteration over a tree is width-first.

source

QuantumLattices.Prerequisites.SimpleTrees.AbstractSimpleTree — Type.

AbstractSimpleTree{Node,Data}

Abstract type for all concrete trees.

source

QuantumLattices.Prerequisites.SimpleTrees.SimpleTree — Type.

SimpleTree{N,D}() where {N,D}

The minimum tree structure that implements all the default tree methods.

source

QuantumLattices.Prerequisites.SimpleTrees.SimpleTreeCore — Type.

SimpleTreeCore()

The core of a tree.

source

QuantumLattices.Prerequisites.SimpleTrees.@simpletree — Macro.

@simpletree structdef treeparams::Union{Expr,Nothing}=nothing

Decorate a "raw" struct to be a subtype of AbstractSimpleTree.

Note

A "raw" struct means:

It has no explicit supertype;

It has no inner constructor;

It has no attribute :TREECORE.

The keytype and valtype can be assigned by the argument treeparams in the form {keytype,valtype}.

When the formal argument names of keytype and valtype are not assigned, they can be automatically generated by the functioin gensym. For example, all of the structs after the decration by the following codes
@simpletree(struct TreeWithWrongTypeParameterNames{N} info::Vector{N} end)
@simpletree(struct TreeWithWrongTypeParameterNames{N} info::Vector{N} end,
      {::String,::Int}
      )
@simpletree(struct TreeWithWrongTypeParameterNames{N} info::Vector{N} end,
      {<:AbstractString,<:Number}
      )
will have three type parameters.

When the formal argument names of keytype and valtype overlap with those of the raw struct type parameters, the duplicates will be considered as the same. For example, the decorated struct SubTreeWithOverlappedParametricFields by the following code
@simpletree(struct TreeWithOverlappedParametricFields{N} info::Vector{N} end,
      {N<:AbstractString,D<:Number}
      )
only has two type parameters N<:AbstractString and D<:Number, where the N in the info::Vector{N} is the same N with that in the decorated attribute TREECORE::SimpleTreeCore{N,D}.

When the formal argument names of keytype and valtype have no intersection with those of the raw struct type parameters, the type parameters of the decorated struct will be just extended by keytype and valtype. For example, the decorated struct SubTreeWithParametricFields by the following code
@simpletree(struct TreeWithParametricFields{T} info::Vector{T} end,
      {N<:AbstractString,D<:Number}
      )
have 3 type parameters, T, N<:AbstractString and D<:Number.

source

QuantumLattices.Prerequisites.SimpleTrees.addnode! — Method.

addnode!(tree::AbstractSimpleTree{N},node::N) where N} -> typeof(tree)
addnode!(tree::AbstractSimpleTree{N},::Nothing,node::N) where N -> typeof(tree)
addnode!(tree::AbstractSimpleTree{N},parent::N,node::N) where N -> typeof(tree)

Update the structure of a tree by adding a node. When the parent is nothing, the input tree must be empty and the input node becomes the tree's root.

source

QuantumLattices.Prerequisites.SimpleTrees.ancestor — Method.

ancestor(tree::AbstractSimpleTree{N},node::N,generation::Int=1) where N -> N

Get the ancestor of a tree's node of the n-th generation.

source

QuantumLattices.Prerequisites.SimpleTrees.children — Method.

children(tree::AbstractSimpleTree) -> Vector{keytype(tree)}
children(tree::AbstractSimpleTree,::Nothing) -> Vector{keytype(tree)}
children(tree::AbstractSimpleTree{N},node::N) where N -> Vector{N}

Get the children of a tree's node.

source

QuantumLattices.Prerequisites.SimpleTrees.deletenode! — Method.

deletenode!(tree::AbstractSimpleTree{N},node::N) where N -> typeof(tree)

Update the structure of a tree by deleting a node.

source

QuantumLattices.Prerequisites.SimpleTrees.descendants — Method.

descendants(tree::AbstractSimpleTree{N},node::N,generation::Int=1) where N -> Vector{N}

Get the descendants of a tree's node of the n-th generation.

source

QuantumLattices.Prerequisites.SimpleTrees.isleaf — Method.

isleaf(tree::AbstractSimpleTree{N},node::N) where N -> Bool

Judge whether a tree's node is a leaf (a node without children) or not.

source

QuantumLattices.Prerequisites.SimpleTrees.leaves — Method.

leaves(tree::AbstractSimpleTree) -> Vector{keytype(tree)}

Get a tree's leaves.

source

QuantumLattices.Prerequisites.SimpleTrees.level — Method.

level(tree::AbstractSimpleTree{N},node::N) where N -> Int

Get the level of tree's node.

source

QuantumLattices.Prerequisites.SimpleTrees.move! — Method.

move!(tree::AbstractSimpleTree{N},node::N,parent::N) where N -> typeof(tree)

Move a subtree to a new position.

source

QuantumLattices.Prerequisites.SimpleTrees.parent — Method.

parent(tree::AbstractSimpleTree{N},node::N,superparent::Union{N,Nothing}=nothing) where N -> Union{N,Nothing}

Get the parent of a tree's node. When node is the tree's root, return superparent.

source

QuantumLattices.Prerequisites.SimpleTrees.root — Method.

root(tree::AbstractSimpleTree) -> Union{keytype(tree),Nothing}

Get a tree's root node.

source

QuantumLattices.Prerequisites.SimpleTrees.siblings — Method.

siblings(tree::AbstractSimpleTree{N},node::N) where N -> Vector{N}

Get the siblings (other nodes sharing the same parent) of a tree's node.

source

QuantumLattices.Prerequisites.SimpleTrees.subtree — Method.

subtree(tree::AbstractSimpleTree{N},node::N) where N -> typeof(tree)

Get a subtree whose root is node.

source

Base.:== — Method.

==(tc1::TC,tc2::TC) where TC<:SimpleTreeCore -> Bool
isequal(tc1::TC,tc2::TC) where TC<:SimpleTreeCore -> Bool

Overloaded equivalent operator.

source

Base.:== — Method.

==(t1::T,t2::T) where T<:AbstractSimpleTree -> Bool
isequal(t1::T,t2::T) where T<:AbstractSimpleTree -> Bool

Overloaded equivalent operator.

source

Base.append! — Method.

append!(tree::AbstractSimpleTree{N,D},subtree::AbstractSimpleTree{N,D}) where {N,D} -> typeof(tree)
append!(tree::AbstractSimpleTree{N,D},node::Union{N,Nothing},subtree::AbstractSimpleTree{N,D}) where {N,D} -> typeof(tree)

Append a subtree to a tree.

source

Base.delete! — Method.

delete!(tree::AbstractSimpleTree{N},node::N) where N -> typeof(tree)

Delete a node and all its descendants from a tree.

source

Base.eltype — Method.

eltype(tree::AbstractSimpleTree)
eltype(::Type{<:AbstractSimpleTree{N,D}}) where {N,D}

Get the eltype of a tree.

source

Base.empty! — Method.

empty!(tree::AbstractSimpleTree) -> typeof(tree)

Empty a tree.

source

Base.empty — Method.

empty(tree::AbstractSimpleTree)

Construct an empty tree of the same type with the input one.

source

Base.getindex — Method.

getindex(tree::AbstractSimpleTree{N},node::N) where N -> N

Get the data of a tree's node.

source

Base.haskey — Method.

haskey(tree::AbstractSimpleTree{N},node::N) where N -> Bool

Check whether a node is in a tree.

source

Base.keys — Method.

keys(tree::AbstractSimpleTree{N},::SimpleTreeDepth,node::Union{N,Nothing}=tree|>root) where N
keys(tree::AbstractSimpleTree{N},::SimpleTreeWidth,node::Union{N,Nothing}=tree|>root) where N

Iterate over a tree's nodes starting from a certain node by depth first search or width first search.

source

Base.keytype — Method.

keytype(tree::AbstractSimpleTree)
keytype(::Type{<:AbstractSimpleTree{N}}) where N

Get a tree's node type.

source

Base.length — Method.

length(tree::AbstractSimpleTree) -> Int

Get the number of a tree's nodes.

source

Base.pairs — Method.

pairs(tree::AbstractSimpleTree{N},::SimpleTreeDepth,node::Union{N,Nothing}=tree|>root) where N
pairs(tree::AbstractSimpleTree{N},::SimpleTreeWidth,node::Union{N,Nothing}=tree|>root) where N

Iterate over a tree's (node,data) pairs starting from a certain node by depth first search or width first search.

source

Base.push! — Method.

push!(tree::AbstractSimpleTree{N,D},node::N,data::D) where {N,D} -> typeof(tree)
push!(tree::AbstractSimpleTree{N,D},parent::Union{N,Nothing},node::N,data::D) where {N,D} -> typeof(tree)

Push a new node to a tree. When parent is nothing, this function set the root node of an empty tree.

source

Base.setindex! — Method.

setindex!(tree::AbstractSimpleTree{N,D},data::D,node::N) where {N,D}

Set the data of a tree's node.

source

Base.valtype — Method.

valtype(tree::AbstractSimpleTree)
valtype(::Type{<:AbstractSimpleTree{N,D} where N}) where D=D

Get a tree's data type.

source

Base.values — Method.

values(tree::AbstractSimpleTree{N},::SimpleTreeDepth,node::Union{N,Nothing}=tree|>root) where N
values(tree::AbstractSimpleTree{N},::SimpleTreeWidth,node::Union{N,Nothing}=tree|>root) where N

Iterate over a tree's data starting from a certain node by depth first search or width first search.

source