Trees

Trees

The aim of this module is to represent the standard tree structure in efficiency-non-sensitive cases. Please note that the default implementation of tree methods are far from optimal in efficiency. Therefore, please DO NOT use it if you need an efficient tree for addition, deletion, sort and inquiry. This module of codes apply only when the structure of tree matters but not the efficiency.

AbstractTree

AbstractTree{N,D} is the abstract type for all concrete trees. By design, it has two type parameters:

N: the type of the tree's node

D: the type of the tree's data

To fully utilize the methods designed for a tree structure, in our protocol, a concrete subtype must implement the following methods:

inquiry related methods

root(tree::AbstractTree{N,D}) where {N,D} -> Union{N,Nothing}
Get a tree's root node (nothing for empty trees)

haskey(tree::AbstractTree{N,D},node::N) where {N,D} -> Bool
Check whether a node is in a tree.

length(tree::AbstractTree) -> Int
Get the number of a tree's nodes.

parent(tree::AbstractTree{N,D},
       node::N,
       superparent::Union{N,Nothing}=nothing
       ) where {N,D} -> Union{N,Nothing}
Get the parent of a tree's node or return superparent when the input node is the tree's root.

children(tree::AbstractTree{N,D},node::N) where {N,D} -> Vector{N}
Get the children of a tree's node.

structure modification related methods

addnode!(tree::AbstractTree{N,D},
         parent::Union{N,Nothing},
         node::N
         ) where {N,D}
Update the structure of a tree by adding a node. When the parent is nothing, the input tree must be empty and the input node becomes the tree's root.

deletenode!(tree::AbstractTree{N,D},node::N) where {N,D}
Update the structure of a tree by deleting a node.

index related methods

getindex(tree::AbstractTree{N,D},node::N) where {N,D} -> D
Get the data of a tree's node

setindex!(tree::AbstractTree{N,D},node::N,data::D) where {N,D}
Set the data of a tree's node.

Based on these methods, we implement several generic functions for inquiries and manipulations

inquiry for type parameters: keytype, valtype, eltype

expansion over nodes/data-records: keys, values, pairs

inquiry for info of nodes: isleaf, level

inquiry for nodes: ancestor, descendants, siblings, leaves

modification: push!, append!, delete!, empty!

And optionally, when a subtype implement the following method,

empty(tree::AbstractTree) -> typeof(tree)

which constructs an empty tree of the same type with the input one, two more more methods are supported:

subtree: Get a subtree starting from a node.

move!: Move a subtree to a new position.

TreeCore and SimpleTree

To implement all the prerequisites listed above costs a bit efforts. We provide two lazy ways to get over this:

Inheritance AbstractTree with TREECORE::TreeCore as the last attribute

Inclusion an attribute which is an instance of SimpleTree

TreeCore

TreeCore{N,D}, as the literal meaning indicates, is the core of a tree. It encapsulates all the data structures needed by the default implementation, which constains 4 attributes:

root::N: the tree's root node

contents::Dict{N,D}: the tree's (node,data) pairs

parent::Dict{N,N}: records of the parent of each of the tree's nodes

children::Dict{N,Vector{N}}: records of the children of each of the tree's nodes

As above, the first lazy way is to include this struct with the special name :TREECORE in your concrete subtype as the last attribute. This process can be even lazier, in that we provide a macro @tree to decorate your "raw" struct automatically, e.g.

julia> @tree(struct SimpleSubTree end)

julia> @tree(struct SubTreeWithTreeParameters end,
             {N<:AbstractString,D<:Number}
             )

julia> @tree(struct SubTreeWithCertainTreeParameters end,
             {<:String,<:Int}
             )

julia> @tree(struct SubTreeWithFields info::Vector{Int} end,
             {N<:AbstractString,D<:Number}
             )

julia> @tree(struct SubTreeWithParametricFields{T} info::Vector{T} end,
             {N<:AbstractString,D<:Number}
             )

julia> @tree(struct SubTreeWithOverlappedParametricFields{N} info::Vector{N} end,
             {N<:AbstractString,D<:Number}
             )

SimpleTree

SimpleTree{N,D} is the minimum struct that implements all the default tree methods. You can include an instance of it as an attribute in your own type to utilize all the tree methods.

Manual

Hamiltonian.Prerequisites.Trees.treedepth — Constant.

treedepth

Indicate that the iteration over a tree is depth-first.

source

Hamiltonian.Prerequisites.Trees.treewidth — Constant.

treewidth

Indicate that the iteration over a tree is width-first.

source

Hamiltonian.Prerequisites.Trees.AbstractTree — Type.

AbstractTree{Node,Data}

Abstract type for all concrete trees.

source

Hamiltonian.Prerequisites.Trees.SimpleTree — Type.

SimpleTree{N,D}() where {N,D}

The minimum tree structure that implements all the default tree methods.

source

Hamiltonian.Prerequisites.Trees.TreeCore — Type.

TreeCore()

The core of a tree.

source

Hamiltonian.Prerequisites.Trees.@tree — Macro.

@tree structdef treeparams::Union{Expr,Nothing}=nothing

Decorate a "raw" struct to be a subtype of AbstractTree.

Note

A "raw" struct means:

It has no explicit supertype;

It has no inner constructor;

It has no attribute :TREECORE.

The keytype and valtype can be assigned by the argument treeparams in the form {keytype,valtype}.

When the formal argument names of keytype and valtype are not assigned, they can be automatically generated by the functioin gensym. For example, all of the structs after the decration by the following codes
@tree(struct TreeWithWrongTypeParameterNames{N} info::Vector{N} end)
@tree(struct TreeWithWrongTypeParameterNames{N} info::Vector{N} end,
      {::String,::Int}
      )
@tree(struct TreeWithWrongTypeParameterNames{N} info::Vector{N} end,
      {<:AbstractString,<:Number}
      )
will have three type parameters.

When the formal argument names of keytype and valtype overlap with those of the raw struct type parameters, the duplicates will be considered as the same. For example, the decorated struct SubTreeWithOverlappedParametricFields by the following code
@tree(struct TreeWithOverlappedParametricFields{N} info::Vector{N} end,
      {N<:AbstractString,D<:Number}
      )
only has two type parameters N<:AbstractString and D<:Number, where the N in the info::Vector{N} is the same N with that in the decorated attribute TREECORE::TreeCore{N,D}.

When the formal argument names of keytype and valtype have no intersection with those of the raw struct type parameters, the type parameters of the decorated struct will be just extended by keytype and valtype. For example, the decorated struct SubTreeWithParametricFields by the following code
@tree(struct TreeWithParametricFields{T} info::Vector{T} end,
      {N<:AbstractString,D<:Number}
      )
have 3 type parameters, T, N<:AbstractString and D<:Number.

source

Hamiltonian.Prerequisites.Trees.addnode! — Method.

addnode!(tree::AbstractTree{N,D},node::N) where {N,D} -> typeof(tree)
addnode!(tree::AbstractTree{N,D},::Nothing,node::N) where {N,D} -> typeof(tree)
addnode!(tree::AbstractTree{N,D},parent::N,node::N) where {N,D} -> typeof(tree)

Update the structure of a tree by adding a node. When the parent is nothing, the input tree must be empty and the input node becomes the tree's root.

source

Hamiltonian.Prerequisites.Trees.ancestor — Method.

ancestor(tree::AbstractTree{N,D},node::N,generation::Int=1) where {N,D} -> N

Get the ancestor of a tree's node of the n-th generation.

source

Hamiltonian.Prerequisites.Trees.children — Method.

children(tree::AbstractTree) -> Vector{keytype(tree)}
children(tree::AbstractTree,::Nothing) -> Vector{keytype(tree)}
children(tree::AbstractTree{N,D},node::N) where {N,D} -> Vector{N}

Get the children of a tree's node.

source

Hamiltonian.Prerequisites.Trees.deletenode! — Method.

deletenode!(tree::AbstractTree{N,D},node::N) where {N,D} -> typeof(tree)

Update the structure of a tree by deleting a node.

source

Hamiltonian.Prerequisites.Trees.descendants — Method.

descendants(tree::AbstractTree{N,D},node::N,generation::Int=1) where {N,D} -> Vector{N}

Get the descendants of a tree's node of the n-th generation.

source

Hamiltonian.Prerequisites.Trees.isleaf — Method.

isleaf(tree::AbstractTree{N,D},node::N) where{N,D} -> Bool

Judge whether a tree's node is a leaf (a node without children) or not.

source

Hamiltonian.Prerequisites.Trees.leaves — Method.

leaves(tree::AbstractTree) -> Vector{keytype(tree)}

Get a tree's leaves.

source

Hamiltonian.Prerequisites.Trees.level — Method.

level(tree::AbstractTree{N,D},node::N) where {N,D} -> Int

Get the level of tree's node.

source

Hamiltonian.Prerequisites.Trees.move! — Method.

move!(tree::AbstractTree{N,D},node::N,parent::N) where {N,D} -> typeof(tree)

Move a subtree to a new position.

source

Hamiltonian.Prerequisites.Trees.parent — Method.

parent(tree::AbstractTree{N,D},node::N,superparent::Union{N,Nothing}=nothing) where {N,D} -> Union{N,Nothing}

Get the parent of a tree's node. When node is the tree's root, return superparent.

source

Hamiltonian.Prerequisites.Trees.root — Method.

root(tree::AbstractTree) -> Union{keytype(tree),Nothing}

Get a tree's root node.

source

Hamiltonian.Prerequisites.Trees.siblings — Method.

siblings(tree::AbstractTree{N,D},node::N) where{N,D} -> Vector{N}

Get the siblings (other nodes sharing the same parent) of a tree's node.

source

Hamiltonian.Prerequisites.Trees.subtree — Method.

subtree(tree::AbstractTree{N,D},node::N) where{N,D} -> typeof(tree)

Get a subtree whose root is node.

source

Base.:== — Method.

==(tc1::TC,tc2::TC) where TC<:TreeCore -> Bool
isequal(tc1::TC,tc2::TC) where TC<:TreeCore -> Bool

Overloaded equivalent operator.

source

Base.:== — Method.

==(t1::T,t2::T) where T<:AbstractTree -> Bool
isequal(t1::T,t2::T) where T<:AbstractTree -> Bool

Overloaded equivalent operator.

source

Base.append! — Method.

append!(tree::AbstractTree{N,D},subtree::AbstractTree{N,D}) where {N,D} -> typeof(tree)
append!(tree::AbstractTree{N,D},node::Union{N,Nothing},subtree::AbstractTree{N,D}) where {N,D} -> typeof(tree)

Append a subtree to a tree.

source

Base.delete! — Method.

delete!(tree::AbstractTree{N,D},node::N) where {N,D} -> typeof(tree)

Delete a node and all its descendants from a tree.

source

Base.eltype — Method.

eltype(tree::AbstractTree)
eltype(::Type{<:AbstractTree{N,D}}) where {N,D}

Get the eltype of a tree.

source

Base.empty! — Method.

empty!(tree::AbstractTree) -> typeof(tree)

Empty a tree.

source

Base.empty — Method.

empty(tree::AbstractTree)

Construct an empty tree of the same type with the input one.

source

Base.getindex — Method.

getindex(tree::AbstractTree{N,D},node::N) where {N,D} -> N

Get the data of a tree's node.

source

Base.haskey — Method.

haskey(tree::AbstractTree{N,D},node::N) where {N,D} -> Bool

Check whether a node is in a tree.

source

Base.keys — Method.

keys(tree::AbstractTree{N,D},::TreeDepth,node::Union{N,Nothing}=tree|>root) where {N,D}
keys(tree::AbstractTree{N,D},::TreeWidth,node::Union{N,Nothing}=tree|>root) where {N,D}

Iterate over a tree's nodes starting from a certain node by depth first search or width first search.

source

Base.keytype — Method.

keytype(tree::AbstractTree)
keytype(::Type{<:AbstractTree{N,D}}) where {N,D}

Get a tree's node type.

source

Base.length — Method.

length(tree::AbstractTree) -> Int

Get the number of a tree's nodes.

source

Base.pairs — Method.

pairs(tree::AbstractTree,::TreeDepth,node::Union{N,Nothing}=tree|>root) where {N,D}
pairs(tree::AbstractTree,::TreeWidth,node::Union{N,Nothing}=tree|>root) where {N,D}

Iterate over a tree's (node,data) pairs starting from a certain node by depth first search or width first search.

source

Base.push! — Method.

push!(tree::AbstractTree{N,D},node::N,data::D) where {N,D} -> typeof(tree)
push!(tree::AbstractTree{N,D},parent::Union{N,Nothing},node::N,data::D) where {N,D} -> typeof(tree)

Push a new node to a tree. When parent is nothing, this function set the root node of an empty tree.

source

Base.setindex! — Method.

setindex!(tree::AbstractTree{N,D},data::D,node::N) where {N,D}

Set the data of a tree's node.

source

Base.valtype — Method.

valtype(tree::AbstractTree)
valtype(::Type{<:AbstractTree{N,D}}) where {N,D}

Get a tree's data type.

source

Base.values — Method.

values(tree::AbstractTree,::TreeDepth,node::Union{N,Nothing}=tree|>root) where {N,D}
values(tree::AbstractTree,::TreeWidth,node::Union{N,Nothing}=tree|>root) where {N,D}

Iterate over a tree's data starting from a certain node by depth first search or width first search.

source