Vector spaces

A vector space is a linear space, in which the addition of vectors and multiplication of a vector by a scalar are defined.

Vector spaces are frequently encountered in physics, e.g. the Hilbert space in quantum mechanics. In this submodule, we only implement those with finite dimensions. We want to remark that in our implementation, a vector space is a subtype of an abstract vector, therefore, the bases always possess a order, which means, two vector spaces are not considered to be equal to each other even if their corresponding actual mathematical spaces are the same but the orders of the bases are different.

VectorSpace

VectorSpace{B} is the abstraction of a vector space, which has only one type parameter:

B<:Any: the type of the bases of the vector space

Basically, a subtype should implement the following 3 methods:

dimension(vs::VectorSpace) -> Int
Get the dimension of a vector space

Base.getindex(vs::VectorSpace{B}, i::Int)  where B -> B
Get the ith basis of a vector space

Base.searchsortedfirst(vs::VectorSpace{B}, basis::B) where B -> Int
Search the index of a basis in a vector space

Other features include

comparison: == and isequal

iteration: iterate

inquiry: size, findfirst and in

Manual

QuantumLattices.Prerequisites.VectorSpaces.DirectProductedNamedVectorSpace — Type

DirectProductedNamedVectorSpace{T<:Tuple{Vararg{NamedVectorSpace}}} <: NamedVectorSpace{Any}

Zipped named vector space.

QuantumLattices.Prerequisites.VectorSpaces.NamedVectorSpace — Type

NamedVectorSpace{B} <: VectorSpace{B}

Abstract named vector space.

QuantumLattices.Prerequisites.VectorSpaces.ParameterSpace — Type

ParameterSpace{N, T, B} <: SimpleNamedVectorSpace{N, B}

Parameter space.

QuantumLattices.Prerequisites.VectorSpaces.SimpleNamedVectorSpace — Type

SimpleNamedVectorSpace{N, B} <: NamedVectorSpace{B}

Abstract simple named vector space.

The subtypes must have the following predefined content:

:content::T: the content of the named vector space.

QuantumLattices.Prerequisites.VectorSpaces.VectorSpace — Type

VectorSpace{B} <: AbstractVector{B}

Abstract vector space.

QuantumLattices.Prerequisites.VectorSpaces.VectorSpaceCartesian — Type

VectorSpaceCartesian <: VectorSpaceStyle

Cartesian vector space style, which indicates that every basis in it could be represented by a Cartesian index.

QuantumLattices.Prerequisites.VectorSpaces.VectorSpaceDirectProducted — Type

VectorSpaceDirectProducted <: VectorSpaceStyle

Vector space style which indicates that a vector space is the direct product of its sub-components.

QuantumLattices.Prerequisites.VectorSpaces.VectorSpaceDirectSummed — Type

VectorSpaceDirectSummed <: VectorSpaceStyle

Vector space style which indicates that a vector space is the direct sum of its sub-components.

QuantumLattices.Prerequisites.VectorSpaces.VectorSpaceEnumerative — Type

VectorSpaceEnumerative <: VectorSpaceStyle

Enumerative vector space style, which indicates that the vector space has a predefined content named :table that contains all its bases.

QuantumLattices.Prerequisites.VectorSpaces.VectorSpaceStyle — Type

VectorSpaceStyle

The style of a concrete type of vector space.

QuantumLattices.Prerequisites.VectorSpaces.VectorSpaceZipped — Type

VectorSpaceZipped <: VectorSpaceStyle

Vector space style which indicates that a vector space is the zip of its sub-components.

QuantumLattices.Prerequisites.VectorSpaces.ZippedNamedVectorSpace — Type

ZippedNamedVectorSpace{T<:Tuple{Vararg{NamedVectorSpace}}} <: NamedVectorSpace{Any}

Zipped named vector space.

QuantumLattices.Interfaces.:⊕ — Method

⊕(vs₁::SimpleNamedVectorSpace, vs₂::SimpleNamedVectorSpace) -> ZipNamedVectorSpace
⊕(vs₁::SimpleNamedVectorSpace, vs₂::ZippedNamedVectorSpace) -> ZipNamedVectorSpace
⊕(vs₁::ZippedNamedVectorSpace, vs₂::SimpleNamedVectorSpace) -> ZipNamedVectorSpace
⊕(vs₁::ZippedNamedVectorSpace, vs₂::ZippedNamedVectorSpace) -> ZipNamedVectorSpace

The zip of named vector spaces.

QuantumLattices.Interfaces.:⊗ — Method

⊗(vs₁::SimpleNamedVectorSpace, vs₂::SimpleNamedVectorSpace) -> DirectProductedNamedVectorSpace
⊗(vs₁::SimpleNamedVectorSpace, vs₂::DirectProductedNamedVectorSpace) -> DirectProductedNamedVectorSpace
⊗(vs₁::DirectProductedNamedVectorSpace, vs₂::SimpleNamedVectorSpace) -> DirectProductedNamedVectorSpace
⊗(vs₁::DirectProductedNamedVectorSpace, vs₂::DirectProductedNamedVectorSpace) -> DirectProductedNamedVectorSpace

The direct product of named vector spaces.