Factory

Factory

The aim of Factory is to provide tools to hack into Julia codes without knowing the details of their abstract syntax trees, so that users can somewhat manipulate the existing codes, modify them and generate new ones. In particular, Factory in this module means the representation of certain blocks of Julia codes by a usual Julia struct. This representation is much easier to comprehend than the canonical Expr representation. In general, we propose two basic requirements that any factory must satisfy:

Besides the default constructor, a concrete factory can be constructed from any legal Expr expression that represents the block of codes it aims to represent;

The canonical Expr expression of the block of codes that a concrete factory represents can be obtained by "calling" the factory itself.

These two requirements also defines the basic interfaces to interact with factories.

Out of practical purposes, we only implemente 6 kinds of factories, i.e. Argument, Parameter, Field, Block, FunctionFactory and TypeFactory, which represent a function argument, a method or type parameter, a struct field, a begin ... end block, a function itself and a struct itself, respectively. Some of the basic methods making the above requirements fulfilled with these types are based on the powerful functions defined in MacroTools.

Argument

An Argument has 4 attributes:

name::Symbol: the name of the argument

type::Union{Symbol,Expr}: the type of the argument

slurp::Bool: whether the argument should be expanded by ...

default::Any: the default value of the argument, nothing for those with no default values

All valid expressions representing the arguments of functions can be passed to the constructor:

julia> Argument(:arg)
Argument(
  name:    arg
  type:    Any
  slurp:   false
  default: nothing
)

julia> Argument(:(arg::ArgType))
Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: nothing
)

julia> Argument(:(arg::ArgType...))
Argument(
  name:    arg
  type:    ArgType
  slurp:   true
  default: nothing
)

julia> Argument(:(arg::ArgType=default))
Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: default
)

Or you can use the macro @argument for construction directly from an argument declaration:

julia> @argument arg::ArgType=default
Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: default
)

The construction from such expressions is based on the the MacroTools.splitarg function. On the other hand, calling an instance of Argument will get the corresponding Expr expression, e.g.,

julia> Argument(:arg,:ArgType,false,:default)()
:(arg::ArgType=default)

This feature is based on the MacroTools.combinearg function.

Parameter

A Parameter has 2 attributes:

name::Symbol: the name of the parameter

type::Union{Symbol,Expr}: the type of the parameter

All expressions that represent a type parameter or a method parameter are allowed to be passed to the constructor:

julia> Parameter(:Int)
Parameter(
  name: Int
  type: nothing
)

julia> Parameter(:T)
Parameter(
  name: T
  type: nothing
)

julia> Parameter(:(<:Number))
Parameter(
  name: nothing
  type: Number
)

julia> Parameter(:(T<:Number))
Parameter(
  name: T
  type: Number
)

The macro @parameter completes the construction directly from a parameter declaration:

julia> @parameter T<:Number
Parameter(
  name: T
  type: Number
)

Field

A Field has 2 attributes:

name::Symbol: the name of the field

type::Union{Symbol,Expr}: the type of the field

Legal expressions can be used to construct a Field instance by its constructor:

julia> Field(:field)
Field(
  name: field
  type: Any
)

julia> Field(:(field::FieldType))
Field(
  name: field
  type: FieldType
)

julia> Field(:(field::ParametricType{Parameter}))
Field(
  name: field
  type: ParametricType{Parameter}
)

The macro @field is also provided to help the construction directly from a field declaration:

julia> @field field::FieldType
Field(
  name: field
  type: FieldType
)

The construction from these expressions is based on the MacroTools.splitarg function and the convertion to these expressions is based on the MacroTools.combinefield function.

Block

A Block has only one attribute:

body::Vector{Any}: the body of the begin ... end block

Any expression can be passed to the constructor of Block:

julia> Block(:(x=1))
Block(
  body: Any[:(x = 1)]
)

julia> Block(:(x=1;y=2))
Block(
  body: Any[:(x = 1), :(#= none:1 =#), :(y = 2)]
)

julia> Block(:(begin x=1 end))
Block(
  body: Any[:(#= none:1 =#), :(x = 1)]
)

julia> Block(quote
               x=1
               y=2
           end)
Block(
  body: Any[:(#= none:2 =#), :(x = 1), :(#= none:3 =#), :(y = 2)]
)

Or you can construct a Block instance directly from any code by the macro @block:

julia> @block x=1 y=2
Block(
  body: Any[:(x = 1), :(y = 2)]
)

The body of a block can also be extended by the push! function or the @push! macro.

Note

The body of a Block is somewhat "flattened", i.e. it contains no begin ... end blocks. During the initialization, any such input block will be unblocked and added to the body part by part. So is the push! and @push! processes.

All LineNumberNodes generated by the input codes will also be included in the block's body. However, you can use rmlines! or @rmlines! to remove them from the body of an existing Block, or use rmlines or @rmlines to get a copy with them removed in the body.

FunctionFactory

A FunctionFactory has 6 attributes:

name::Symbol: the name of the function

args::Vector{Argument}: the positional arguments of the function

kwargs::Vector{Argument}: the keyword arguments of the function

rtype::Any: the return type of the function

params::Vector{Parameter}: the method parameters specified by the where keyword

body::Expr: the body of the function

All expressions that represent functions are allowed to be passed to the constructor:

julia> FunctionFactory(:(f()=nothing))
FunctionFactory(
  name:   f
  args:   Argument[]
  kwargs: Argument[]
  rtype:  Any
  params: Parameter[]
  body:   begin
              nothing
          end
)

julia> FunctionFactory(:(f(x)=x))
FunctionFactory(
  name:   f
  args:   Argument[x::Any]
  kwargs: Argument[]
  rtype:  Any
  params: Parameter[]
  body:   begin
              x
          end
)

julia> FunctionFactory(:(f(x::Int,y::Int;choice::Function=sum)=choice(x,y)))
FunctionFactory(
  name:   f
  args:   Argument[x::Int, y::Int]
  kwargs: Argument[choice::Function=sum]
  rtype:  Any
  params: Parameter[]
  body:   begin
              choice(x, y)
          end
)

julia> FunctionFactory(:(f(x::T,y::T;choice::Function=sum) where T<:Number=choice(x,y)))
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::Function=sum]
  rtype:  Any
  params: Parameter[T <: Number]
  body:   begin
              choice(x, y)
          end
)

julia> FunctionFactory(:((f(x::T,y::T;choice::Function=sum)::T) where T<:Number=choice(x,y)))
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::Function=sum]
  rtype:  T
  params: Parameter[T <: Number]
  body:   begin
              choice(x, y)
          end
)

julia> FunctionFactory(:(
           function (f(x::T,y::T;choice::Function=sum)::T) where T<:Number
               choice(x,y)
           end
       ))
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::Function=sum]
  rtype:  T
  params: Parameter[T <: Number]
  body:   begin
              #= none:3 =#
              choice(x, y)
          end
)

julia> FunctionFactory(quote
           function (f(x::T,y::T;choice::Function=sum)::T) where T<:Number
               choice(x,y)
           end
       end)
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::Function=sum]
  rtype:  T
  params: Parameter[T <: Number]
  body:   begin
              #= none:3 =#
              choice(x, y)
          end
)

Similarly, an instance can also be constructed from the macro @functionfactory:

julia> @functionfactory (f(x::T,y::T;choice::Function=sum)::T) where T<:Number=choice(x,y)
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::Function=sum]
  rtype:  T
  params: Parameter[T <: Number]
  body:   begin
              choice(x, y)
          end
)

The construction from and the convertion to such expressions are based on the MacroTools.splitdef and MacroTools.combinedef functions, respectively.

Note

Because the form f{T}(x::T,y::T;choice::Function=sum) has no longer been supported since Julia 0.7, the entry :params in the returned dict by MacroTools.splitarg is always missing. Therefore, we abandon its corresponding field in FunctionFactory but use the attribute :params to denote the :whereparams entry.

Other features include:

Positional arguments can be added by addargs! or @addargs!

Keyword arguments can be added by addkwargs! or @addkwargs!

Method parameters can be added by addparams! or @addparams!

Body can be extended by extendbody! or @extendbody!

TypeFactory

A TypeFactory has 6 attributes:

name::Symbol: the name of the struct

mutable::Bool: whether or not the struct is mutable

params::Vector{Parameter}: the type parameters of the struct

supertype::Any: the supertype of the struct

fields::Vector{Field}: the fields of the struct

constructors::Vector{FunctionFactory}: the inner constructors of the struct

Any expression representing valid struct definitions can be passed to the constructor:

julia> TypeFactory(:(struct StructName end))
TypeFactory(
  name:         StructName
  mutable:      false
  params:       Parameter[]
  supertype:    Any
  fields:       Field[]
  constructors: FunctionFactory[]
)

julia> TypeFactory(:(struct StructName{T} end))
TypeFactory(
  name:         StructName
  mutable:      false
  params:       Parameter[T]
  supertype:    Any
  fields:       Field[]
  constructors: FunctionFactory[]
)

julia> TypeFactory(:(struct Child{T} <: Parent{T} where T end))
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T]
  supertype:    Parent{T} where T
  fields:       Field[]
  constructors: FunctionFactory[]
)

julia> TypeFactory(:(
           struct Child{T} <: Parent{T} where T<:TypeParameter
               field1::T
               field2::T
           end
       ))
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T]
  supertype:    Parent{T} where T <: TypeParameter
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[]
)

julia> TypeFactory(quote
           struct Child{T} <: Parent{T} where T<:TypeParameter
               field1::T
               field2::T
           end
       end)
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T]
  supertype:    Parent{T} where T <: TypeParameter
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[]
)

Also, the macro @typefactory supports the construction directly from a type definition:

julia> @typefactory struct Child{T} <: Parent{T} where T<:TypeParameter
                       field1::T
                       field2::T
                       Child(field1::T,field2::T=zero(T)) where T=new{T}(field1,field2)
                   end
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T]
  supertype:    Parent{T} where T <: TypeParameter
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[
                  function (Child(field1::T, field2::T=zero(T); )::Any) where T
                      #= /home/travis/.julia/packages/MacroTools/4AjBS/src/utils.jl:302 =#
                      begin
                          new{T}(field1, field2)
                      end
                  end
                  ]
)

The construction from these expressions is based on the MacroTools.splitstructdef function. Meanwhile, the convertion to the corresponding expression from a TypeFactory is based on the MacroTools.combinestructdef function.

Other features include:

Fields can be added by addfields! or @addfields!

Type parameters can be added by addparams! or @addparams!

Inner constructors can be added by addconstructors! or @addconstructors!

Manual

Hamiltonian.Utilities.Factory.Argument — Type.

Argument(name::Symbol,type::FExpr,slurp::Bool,default::Any)
Argument(name::Symbol;type::FExpr=:Any,slurp::Bool=false,default::Any=nothing)
Argument(expr::Expr)

The struct to describe a argument of a function.

source

Hamiltonian.Utilities.Factory.Argument — Method.

(a::Argument)()

Convert an Argument to the Expr representation of the argument it describes.

source

Hamiltonian.Utilities.Factory.Block — Type.

Block(parts::FExpr...)

The struct to describe a begin ... end block.

source

Hamiltonian.Utilities.Factory.Block — Method.

(b::Block)()

Convert a Block to the Expr representation of the begin ... end block it describes.

source

Hamiltonian.Utilities.Factory.Field — Type.

Field(name::Symbol,type::FExpr)
Field(name::Symbol;type::FExpr=:Any)
Field(expr::Expr)

The struct to describe a field of a struct.

source

Hamiltonian.Utilities.Factory.Field — Method.

(f::Field)()

Convert a Field to the Expr representation of the field it describes.

source

Hamiltonian.Utilities.Factory.FunctionFactory — Type.

FunctionFactory(name::Symbol,args::Vector{Argument},kwargs::Vector{Argument},rtype::FExpr,params::Vector{Parameter},body::Block)
FunctionFactory(    name::Symbol;
                    args::Vector{Argument}=Argument[],
                    kwargs::Vector{Argument}=Argument[],
                    rtype::FExpr=:Any,
                    params::Vector{Parameter}=Parameter[],
                    body::Block=Block()
                    )
FunctionFactory(expr::Expr)

The struct to describe a function.

source

Hamiltonian.Utilities.Factory.FunctionFactory — Method.

(ff::FunctionFactory)()

Convert a FunctionFactory to the Expr representation of the function it describes.

source

Hamiltonian.Utilities.Factory.Parameter — Type.

Parameter(name::Symbol,type::FExpr)
Parameter(name::Symbol;type::FExpr=:Any)
Parameter(expr::FExpr)

The struct to describe a parameter of a function or a type.

source

Hamiltonian.Utilities.Factory.Parameter — Method.

(p::Parameter)()

Convert a Parameter to the Expr representation of the parameter it describes.

source

Hamiltonian.Utilities.Factory.TypeFactory — Type.

TypeFactory(name::Symbol,mutable::Bool,params::Vector{Parameter},supertype::FExpr,fields::Vector{Field},constructors::Vector{FunctionFactory})
TypeFactory(    name::Symbol;
                mutable::Bool=false,
                params::Vector{Parameter}=Parameter[],
                supertype::FExpr=:Any,
                fields::Vector{Field}=Field[],
                constructors::Vector{FunctionFactory}=FunctionFactory[]
                )
TypeFactory(expr::Expr)

The struct to describe a struct.

source

Hamiltonian.Utilities.Factory.TypeFactory — Method.

(tf::TypeFactory)()

Convert a TypeFactory to the Expr representation of the struct it describes.

source

Hamiltonian.Utilities.Factory.@addargs! — Macro.

@addargs! ff args::FExpr...

Add a couple of positional arguments to a function factory.

source

Hamiltonian.Utilities.Factory.@addconstructors! — Macro.

@addconstructors! tf constructors::Expr...

Add a couple of constructors to a type factory.

source

Hamiltonian.Utilities.Factory.@addfields! — Macro.

@addfields! tf fields::FExpr...

Add a couple of fields to a type factory.

source

Hamiltonian.Utilities.Factory.@addkwargs! — Macro.

@addkwargs! ff kwargs::FExpr...

Add a couple of keyword arguments to a function factory.

source

Hamiltonian.Utilities.Factory.@addparams! — Macro.

@addparams! f params::FExpr...

Add a couple of method parameters to a function factory or a type factory.

source

Hamiltonian.Utilities.Factory.@argument — Macro.

@argument expr::FExpr

Construct an Argument directly from an argument statement.

source

Hamiltonian.Utilities.Factory.@block — Macro.

@block parts::FExpr...

Construct a Block directly from a begin ... end block definition.

source

Hamiltonian.Utilities.Factory.@extendbody! — Macro.

@extendbody! ff parts::FExpr...

Extend the body of a function factory.

source

Hamiltonian.Utilities.Factory.@field — Macro.

@field expr::FExpr

Construct a Field directly from a field statement.

source

Hamiltonian.Utilities.Factory.@functionfactory — Macro.

@functionfactory expr::Expr

Construct a FunctionFactory directly from a function definition.

source

Hamiltonian.Utilities.Factory.@parameter — Macro.

@parameter expr::FExpr

Construct a Parameter directly from an parameter statement.

source

Hamiltonian.Utilities.Factory.@push! — Macro.

@push! b parts::FExpr...

Push other parts into the body of a block.

source

Hamiltonian.Utilities.Factory.@rmlines! — Macro.

@rmlines! b::Expr

Remove line number nodes in the body of a block.

source

Hamiltonian.Utilities.Factory.@rmlines — Macro.

@rmlines b::Expr

Return a copy of a block with the line number nodes removed.

source

Hamiltonian.Utilities.Factory.@typefactory — Macro.

@typefactory expr::Expr

Construct a TypeFactory directly from a type definition.

source

Hamiltonian.Utilities.Factory.addargs! — Method.

addargs!(ff::FunctionFactory,args::FExpr...)
addargs!(ff::FunctionFactory,args::Argument...)

Add a couple of positional arguments to a function factory.

source

Hamiltonian.Utilities.Factory.addconstructors! — Method.

addconstructors!(tf::TypeFactory,constructors::Expr...)
addconstructors!(tf::TypeFactory,constructors::FunctionFactory...)

Add a couple of constructors to a type factory.

source

Hamiltonian.Utilities.Factory.addfields! — Method.

addfields!(tf::TypeFactory,fields::FExpr...)
addfields!(tf::TypeFactory,fields::Field...)

Add a couple of fields to a type factory.

source

Hamiltonian.Utilities.Factory.addkwargs! — Method.

addkwargs!(ff::FunctionFactory,kwargs::FExpr...)
addkwargs!(ff::FunctionFactory,kwargs::Argument...)

Add a couple of keyword arguments to a function factory.

source

Hamiltonian.Utilities.Factory.addparams! — Method.

addparams!(f::Union{FunctionFactory,TypeFactory},params::FExpr...)
addparams!(f::Union{FunctionFactory,TypeFactory},params::Parameter...)

Add a couple of method parameters to a function factory or a type factory.

source

Hamiltonian.Utilities.Factory.extendbody! — Method.

extendbody!(ff::FunctionFactory,parts::FExpr...)
extendbody!(ff::FunctionFactory,parts::Block...)

Extend the body of a function factory.

source

Hamiltonian.Utilities.Factory.rmlines! — Method.

rmlines!(b::Block)

Remove line number nodes in the body of a block.

source

MacroTools.rmlines — Method.

rmlines(b::Block)

Return a copy of a block with the line number nodes removed.

source

Hamiltonian.Utilities.Factory.FExpr — Constant.

Factory expression types.

source

Hamiltonian.Utilities.Factory.AbstractFactory — Type.

AbstractFactory

Abstract type for all concrete factories.

source

Base.:== — Method.

==(f1::F,f2::F) where F<:AbstractFactory

Overloaded == operator.

source

Base.push! — Method.

push!(b::Block,parts::FExpr...)
push!(b::Block,parts::Block...)

Push other parts into the body of a block.

source

Base.replace — Method.

replace(f::AbstractFactory;kwargs...)

Return a copy of a concrete AbstractFactory with some of the field values replaced by the keyword arguments.

source

Base.show — Method.

show(io::IO,f::AbstractFactory)

Show a concrete AbstractFactory.

source