Good quantum numbers

Good quantum numbers

Good qunatum numbers can be considered as the conserved labels for the bases of a Hilbert space when a quantum system hosts some symmetries. Here we only implement Abelian good quantum numbers because non-Abelian ones are far more complicated yet much less used. In practice, good quantum numbers can be integers or half integers, therefore, we use real numbers to denote them in this module for simplicity. Independent quantum numbers, such as the particle number and the spin z-component, can coexist at the same time. We use type QuantumNumber to represent the complete set of independent ones for a single basis of a Hilbert space, and type QuantumNumbers to represent the whole quantum numbers for the total bases.

QuantumNumber

The abstract type for the complete set of independent good quantum numbers for a single basis.

Main features include:

function fieldnames: get the names of the quantum numbers

function periods: get the periods of the quantum numbers

arithmetic operations: +,-,*,⊕

hashable: concrete instances can be used as keys for a dict or a set

iterable: concrete instances are iterable over their values

In particular, QuantumNumber <: AbstractNamedVector{Float64}, all features supported by AbstractNamedVector are also available for QuantumNumber. See also AbstractNamedVector.

For convenience, 4 kinds of good quantum numbers are predefined in this module, i.e.

SQN: for spin z-component reserved systems

PQN: for particle number reserved systems

SPQN: for both particle number and spin-z component reserved systems

Z2QN: for systems with a $Z_2$ conservation quantum number

Users who want to define their own $Z_N$-like quantum numbers must handle the periodicities in the construction function, otherwise, wrong results will be get when arithmetic operations, such as + or -, are involved. It is highly recommended to use the macro @quantumnumber to define your own concrete QuantumNumbers.

QuantumNumbers

The whole quantum numbers for the total bases.

To achieve high efficiency:

The quantum numbers are stored in a compressed form similiar to that of a CSC/CSR sparse matrix.

The contents of a QuantumNumbers are not an array of some kind of concrete QuantumNumbers, but an 2d array of floats with the columns being the values of those concrete QuantumNumbers.

Main features include:

function eltype: get the concrete type of the quantum numbers it contains

arithmetic operations: +,-,*,^,⊗,⊕

iteration: concrete QuantumNumbers it contains will be constructed and returned

Manual

Hamiltonian.Utilities.GoodQuantumNumber.PQN — Type.

PQN(N::Real)

The concrete QuantumNumber of a quantum system with particle number N conserved.

source

Hamiltonian.Utilities.GoodQuantumNumber.QNSProtocol — Type.

Protocol used for the initilization of QuantumNumbers.

qnscounts: initilization by counts

qnsindptr: initilization by indptr

source

Hamiltonian.Utilities.GoodQuantumNumber.QuantumNumber — Type.

Abstract type for all concrete quantum numbers for a single basis.

source

Hamiltonian.Utilities.GoodQuantumNumber.QuantumNumbers — Type.

QuantumNumbers(form::Char,::Type{QN},contents::Array{Float64,2},info::Array{Int,1},protocol::QNSProtocol=qnscounts,regularize::Bool=true) where QN<:QuantumNumber
QuantumNumbers(form::Char,contents::Array{QN,1},info::Array{Int,1},protocol::QNSProtocol=qnscounts) where QN<:QuantumNumber
QuantumNumbers(qn::QuantumNumber,count::Int=1)

The whole quantum numbers of the total bases of a Hilbert space.

source

Hamiltonian.Utilities.GoodQuantumNumber.SPQN — Type.

SPQN(N::Real,Sz::Real)

The concrete QuantumNumber of a quantum system with both particle number N and spin z-component Sz conserved.

source

Hamiltonian.Utilities.GoodQuantumNumber.SQN — Type.

SQN(Sz::Real)

The concrete QuantumNumber of a quantum system with spin z-component Sz conserved.

source

Hamiltonian.Utilities.GoodQuantumNumber.Z2QN — Type.

Z2QN(N::Real)

The concrete QuantumNumber of a quantum system with a Z₂-like conserved quantity.

source

Hamiltonian.Utilities.GoodQuantumNumber.@quantumnumber — Macro.

@quantumnumber typename fieldnames fieldperiods

Construct a concrete QuantumNumber with the type name being typename, fieldnames specified by fieldnames and periods specified by fieldperiods.

source

Hamiltonian.Utilities.GoodQuantumNumber.:⊕ — Function.

⊕(qns::QN...;signs::Union{Array{Int,1},Nothing}=nothing) where QN<:QuantumNumber
⊕(qnses::QuantumNumbers{QN}...;signs::Union{Array{Int,1},Nothing}=nothing) where QN

Get the direct sum of some QuantumNumbers or QuantumNumberss.

source

Hamiltonian.Utilities.GoodQuantumNumber.:⊗ — Function.

⊗(::Type{QN},qn1::QuantumNumber,qn2::QuantumNumber) where QN<:QuantumNumber
⊗(qnses::QuantumNumbers{QN}...;signs::Union{Array{Int,1},Nothing}=nothing) where QN

Get the direct product of some QuantumNumbers or QuantumNumberss.

source

Hamiltonian.Utilities.GoodQuantumNumber.regularization — Method.

regularization(::Type{QN},array::AbstractArray{Float64,N}) where {QN<:QuantumNumber,N}

Get the regularized array of the array representation of a QuantumNumber/QuantumNumbers by the periods.

source

Hamiltonian.Utilities.GoodQuantumNumber.regularize! — Function.

regularize!(::Type{QN},array::AbstractVector{Float64}) where QN<:QuantumNumber
regularize!(::Type{QN},array::AbstractArray{Float64,2}) where QN<:QuantumNumber

Regularize an array representation of a QuantumNumber/QuantumNumbers by the concrete QuantumNumber's periods.

source

Base.:* — Function.

*(qn::QuantumNumber,factor::Integer)
*(factor::Integer,qn::QuantumNumber)
*(qns::QuantumNumbers,factor::Integer)
*(factor::Integer,qns::QuantumNumbers)

Overloaded * operator for QuantumNumber and QuantumNumbers..

source

Base.:+ — Function.

+(qn::QuantumNumber)
+(qn1::QN,qn2::QN) where QN<:QuantumNumber
+(qn1::QN,qn2::QN,qns::QN...) where QN<:QuantumNumber
+(qns::QuantumNumbers)
+(qn::QN,qns::QuantumNumbers{QN}) where QN
+(qns::QuantumNumbers{QN},qn::QN) where QN

Overloaded + operator for QuantumNumber and QuantumNumbers.

source

Base.:- — Function.

-(qn::QuantumNumber)
-(qn1::QN,qn2::QN) where QN<:QuantumNumber
-(qns::QuantumNumbers)
-(qn::QN,qns::QuantumNumbers{QN}) where QN
-(qns::QuantumNumbers{QN},qn::QN) where QN

Overloaded - operator for QuantumNumber and QuantumNumbers.

source

Base.:^ — Function.

^(qn::QuantumNumber,factor::Integer)
^(qns::QuantumNumbers,factor::Integer)

Overloaded ^ operator for QuantumNumber and QuantumNumbers.

source

Base.eltype — Method.

eltype(::Type{<:QuantumNumbers{QN}}) where QN

Get the type of the concrete QuantumNumber contained in QuantumNumbers.

source

Base.getindex — Method.

getindex(qns::QuantumNumbers,index::Int)

Overloaded [] operator.

source

Base.iterate — Function.

iterate(qns::QuantumNumbers,state::Int=1)

Iterate over the concrete QuantumNumbers the QuantumNumbers contains.

source

Base.keys — Method.

keys(qns::QuantumNumbers)

Iterate over the concrete QuantumNumbers the QuantumNumbers contains.

source

Base.length — Function.

length(qns::QuantumNumbers,duplicate::Bool=true)

Get the number of qunatum numbers in the QuantumNumbers.

duplicate==true: the duplicate quantum numbers are counted duplicately. Then the result equals the dimension of the QuantumNumbers.

duplicate==false: only unduplicate quantum numbers are counted. Then the result equals the number of columns of the QuantumNumbers' contents.

source

Base.pairs — Function.

pairs(qns::QuantumNumbers,protocol::QNSProtocol=qnsindptr)

Iterate over the QuantumNumber=>slice or QuantumNumber=>count pairs.

source

Base.show — Method.

show(io::IO,qns::QuantumNumbers)

Show a QuantumNumbers.

source

Base.string — Method.

string(qns::QuantumNumbers)

Convert a QuantumNumbers to string.

source

Base.values — Function.

values(qns::QuantumNumbers,protocol::QNSProtocol=qnsindptr)

Iterate over the slices/counts of the QuantumNumbers.

source