Introduction

Introduction

This module contains the utilities of the Hamiltonian package.

The constants, types, macros, functions or submodules defined in this module will NOT be exported by the package. Instead, they serve as the prerequisites and fundamentals. The range of the utilities' contents are quite wide, but basically, they fall into four categories:

Global constants and miscellaneous tiny useful functions;

Generic functions that are extended by other parts of the package;

Basic data structures as supplements to the Julia.Base and other common packages;

Basic mathematical algorithms or structures that are needed by or specific to this package.

The first category is contained in the main body of this module, while the others come in separate submodules.

Constants and functions

All the following constants and functions in this section are defined in the main body and are exported by this module.

Hamiltonian.Utilities.atol — Constant.

Absolute tolerance for float numbers.

source

Hamiltonian.Utilities.rtol — Constant.

Relative tolerance for float numbers.

source

Hamiltonian.Utilities.Float — Type.

Default float type.

source

Hamiltonian.Utilities.decimaltostr — Function.

decimaltostr(number::Integer,n::Int=5)
decimaltostr(number::Rational,n::Int=5)
decimaltostr(number::AbstractFloat,n::Int=5)
decimaltostr(number::Complex,n::Int=5)

Convert a number to a string with at most n decimal places.

source

Hamiltonian.Utilities.ordinal — Function.

ordinal(number::Interger)

Convert a positive number to its corresponding ordinal.

source

Hamiltonian.Utilities.delta — Function.

delta(i,j) -> Int

Kronecker delta function.

source

Generic interfaces

For details, please refer to the page Interfaces:

Interfaces

Manual

Basic structures

Here lists the table of contents of the basic data structures that are supplements to the Julia.Base and other common packages:

Type traits

EfficientOperations

MemoryOrder

Manual

Factories

Escape mechanisms

Concrete factories

Manual

Composite structures

CompositeNTuple

CompositeVector

CompositeDict

Manul

Trees

AbstractTree

TreeCore and SimpleTree

Manual

Named vectors

AbstractNamedVector

HomoNamedVector

Manual

Mathematical supplements

The following lists the submodules containing the mathematical algorithms or structures that are needed by or specific to this package:

Combinatorics

AbstractCombinatorics

Manul

Algebra over fields

ID

Element and Elements

Manual

Quantum numbers

AbelianNumber

AbelianNumbers

Manual