NamedVector

NamedVector

A named vector is similiar to a named tuple, which associate each of its values with a name. Although the names of a named vector cannot be changed, the values can be modified if needed. In contrast to the predefined NamedTuple in Julia, which employs the names as type parameters, we just implement a named vector as a composite struct equipped with the getindex and setindex! functions, with the fieldnames being its names. This simple implementation makes it possible to define your own concrete named vector with any of your preferred type names, and ensures that all instances of a certain concrete named vector share the same names. Therefore, if you are familiar with Python, you will find that our named vector is more qualified to be the counterpart of the namedtuple in Python than the default Julia implementation. Specifically, we also define a macro @namedvector as the type factory to help users to define their own concrete named vectors. Last but not least important, it is also worth noted that a named vector is not a vector, as is similar to that a named tuple is not a tuple in Julia. This results from our basic expectation that a named vector should be more like a tuple other than a vector so that not all operations valid to vectors are also valid to named vectors.

AbstractNamedVector

AbstractNamedVector defines the abstract type for all concrete named vectors.

Main features include:

Values can be accessed or modified either by the . operator or by the [] operator.

Comparisons, such as ≡, ≢, ==, ≠, >, <, ≥, ≤ are supported. Therefore a vector of named vectors can be sorted by the default sort function.

Hash is supported by hash. Therefore, a named vector can be used as the key of a dict or set.

Iteration over its fieldnames is supported by keys, over its values is supported by values, over its field-value pairs is supported by pairs. A reverse iteration is also supported.

To subtype it, please note:

A concrete type can be either mutable or immutable as you need, but all its fields should be of the same type. A recommended template for the subtype is
[mutable] struct YourNamedVector{T} <: AbstractNamedVector{T}
    filedname1::T
    filedname2::T
    ...
end

It is recommended to overload the Base.fieldnames function for concrete subtypes to ensure type stability and improve efficiency, which though is not a necessity. A template for such an overloading is
Base.fieldnames(Type{YourNamedVector})=(:fieldname1,:fieldname2,...)

For all concrete subtypes, if inner constructors are defined, the one which has the same interface with the default one must be implemented. Otherwise, some functionalities will not work.

Arithmetic operations, such as +, -, *, /, %, ÷, etc. are NOT supported. However, an efficient map function is implemented, which can help users do the overloadings of these operations.

Manual

Hamiltonian.Utilities.NamedVector.AbstractNamedVector — Type.

AbstractNamedVector{T}

Abstract type for all concrete named vectors.

source

Hamiltonian.Utilities.NamedVector.@namedvector — Macro.

@namedvector mutableornot::Bool typename fieldnames dtype::Union{Expr,Symbol}=:nothing supertypename=:AbstractNamedVector

Construct a mutable or immutable concrete named vector with the type name being typename and the fieldnames specified by fieldnames, and optionally, the type parameters specified by dtype and the supertype specified by supertypename.

source

Core.Type — Method.

(::Type{NV})(values::NTuple{N,T}) where {NV<:AbstractNamedVector,N,T}

Construct a concrete named vector by a tuple.

source

Base.:< — Method.

<(nv1:NV,nv2:NV) where NV<:AbstractNamedVector -> Bool

Overloaded < operator.

source

Base.:== — Method.

==(nv1::AbstractNamedVector,nv2::AbstractNamedVector) -> Bool

Overloaded == operator. Two named vector are equal to each other if and only if their keys as well as their values are equal to each other.

Note

It is not necessary for two named vectors to be of the same concrete type to be equal to each other.

source

Base.convert — Method.

convert(::Type{Tuple},nv::AbstractNamedVector) -> NTuple{nv|>length,nv|>eltype}
convert(::Type{NTuple},nv::AbstractNamedVector) -> NTuple{nv|>length,nv|>eltype}
convert(::Type{NTuple{N,T}},nv::AbstractNamedVector{T}) where {N,T} -> NTuple{nv|>length,nv|>eltype}

Convert a named vector to tuple.

source

Base.eltype — Method.

eltype(::Type{NV}) where NV<:AbstractNamedVector{T} where T
eltype(nv::AbstractNamedVector)

Get the type parameter of a concrete AbstractNamedVector.

source

Base.getindex — Method.

getindex(nv::AbstractNamedVector,index::Int)

Get the value by the [] syntax.

source

Base.hash — Method.

hash(nv::AbstractNamedVector,h::UInt)

Hash a concrete AbstractNamedVector.

source

Base.isless — Method.

isless(nv1::NV,nv2::NV) where NV<:AbstractNamedVector -> Bool

Overloaded isless function.

source

Base.iterate — Function.

iterate(nv::AbstractNamedVector,state=1)
iterate(rv::Iterators.Reverse{<:AbstractNamedVector},state=length(rv.itr))

Iterate or reversely iterate over the values of a concrete AbstractNamedVector.

source

Base.keys — Method.

keys(nv::AbstractNamedVector) -> NTuple(nv|>length,Symbol)

Iterate over the names.

source

Base.length — Method.

length(::Type{NV}) where NV<:AbstractNamedVector -> Int
length(nv::AbstractNamedVector) -> Int

Get the length of a concrete AbstractNamedVector.

source

Base.map — Method.

map(f,nvs::NV...) where NV<:AbstractNamedVector -> NV

Apply function f elementwise on the input named vectors.

source

Base.pairs — Method.

pairs(nv::AbstractNamedVector)

Iterate over the name=>value pairs.

source

Base.replace — Method.

replace(nv::AbstractNamedVector;kwargs...) -> typeof(nv)

Return a copy of a concrete AbstractNamedVector with some of the filed values replaced by the keyword arguments.

source

Base.setindex! — Method.

setindex!(nv::AbstractNamedVector,value,index::Int)

Set the value by the [] syntax if mutable.

source

Base.show — Method.

show(io::IO,nv::AbstractNamedVector)

Show a concrete AbstractNamedVector.

source

Base.values — Method.

values(nv::AbstractNamedVector) -> NTuple{nv|>length,nv|>eltype}

Iterate over the values.

source

Base.zero — Method.

zero(::Type{NV}) where NV<:AbstractNamedVector
zero(nv::AbstractNamedVector)

Get a concrete AbstractNamedVector with all values being zero.

source