Type trait

Type trait

EfficientOperations

Hamiltonian.Utilities.TypeTrait.efficientoperations — Constant.

efficientoperations

Indicate that the efficient operations, i.e. "=="/"isequal", "<"/"isless" or "replace", will be used.

source

Base.:< — Method.

<(::EfficientOperations,o1,o2) -> Bool
isless(::EfficientOperations,o1,o2) -> Bool

Compare two objects and judge whether the first is less than the second.

source

Base.:== — Method.

==(::EfficientOperations,o1,o2) -> Bool
isequal(::EfficientOperations,o1,o2) -> Bool

Compare two objects and judge whether they are eqaul to each other.

source

Base.replace — Method.

replace(::EfficientOperations,o;kwargs...) -> typeof(o)

Return a copy of the input object with some of the field values replaced by the keyword arguments.

source