Type traits

Type traits

This module defines generic type traits that are useful to the package.

Julia does not support multi-inheritance, which is sometimes not convenient. A way around this is to use traits, i.e. by utilizing the dispatch on certain singleton types known as traits to simulate multi-inheritance. Although this method cannot aviod small repetitive codes, it suits methods well that are complicated and lengthy.

EfficientOperations

EfficientOperations defines efficient operations such as ==/isequal, </isless, replace, etc, that ensure type stability.

Type stability is the key of Julia to improve the code efficiency. However, it cannot be ensured in some unexpected cases, especially where an iterator is involved. For example, the following codes appears type unstable:

function Base.:(==)(o1::AbstractType,o2::AbstractType)
    n1,n2=o1|>typeof|>fieldcount,o2|>typeof|>fieldcount
    n1==n2 ? all(getfield(o1,i)==getfield(o2,i) for i=1:n1) : false
end

Methods like above are common when we design abstract types, but they are not type stable. To get rid of it, the generated function trick can be used:

@generated function Base.:(==)(o1::AbstractType,o2::AbstractType)
    n1,n2=o1|>typeof|>fieldcount,o2|>typeof|>fieldcount
    if n1==n2
        expr=:(getfield(o1,1)==getfield(o2,1))
        for i=2:fcount
            expr=Expr(:&&,expr,:(getfield(o1,$i)==getfield(o2,$i)))
        end
        return expr
    else
        return :(false)
    end
end

Then type stability can be ensured. We use this trick to implement the methods such as ==/isequal, </isless, replace, etc, with the trait efficientoperations::EfficientOperations. Other types can resort to these methods by passing efficientoperations as the first argument.

Manual

Hamiltonian.Utilities.TypeTraits.efficientoperations — Constant.

efficientoperations

Indicate that the efficient operations, i.e. "=="/"isequal", "<"/"isless" or "replace", will be used.

source

Base.:< — Method.

<(::EfficientOperations,o1,o2) -> Bool
isless(::EfficientOperations,o1,o2) -> Bool

Compare two objects and judge whether the first is less than the second.

source

Base.:== — Method.

==(::EfficientOperations,o1,o2) -> Bool
isequal(::EfficientOperations,o1,o2) -> Bool

Compare two objects and judge whether they are eqaul to each other.

source

Base.replace — Method.

replace(::EfficientOperations,o;kwargs...) -> typeof(o)

Return a copy of the input object with some of the field values replaced by the keyword arguments.

source