DegreeOfFreedom

DegreeOfFreedom

Table

Internal and Index

IDFConfig

There are two purposes with IDFConfig

provide a complete set of internal degrees of freedom on a lattice

offer the ordering of the internal degrees of freedom on a lattice

IndexPack and IndexPacks

Manul

Hamiltonian.Essentials.DegreeOfFreedom.Coupling — Type.

Coupling

The coupling intra/inter interanl degrees of freedom at different lattice points.

source

Hamiltonian.Essentials.DegreeOfFreedom.Couplings — Type.

Couplings{I,C<:Coupling} <: AbstractDict{I,C}

A pack of couplings intra/inter interanl degrees of freedom at different lattice points.

Alias for Elements{I,C}.

source

Hamiltonian.Essentials.DegreeOfFreedom.DirectIndexToTuple — Type.

DirectIndexToTuple

Direct index to tuple.

source

Hamiltonian.Essentials.DegreeOfFreedom.DirectIndexToTuple — Method.

(indextotuple::DirectIndexToTuple)(index::Index) -> Tuple
(indextotuple::DirectIndexToTuple)(index::Index,mask::Symbol...) -> Tuple

Convert an index to tuple directly.

When mask is nonempty, those attributes of Index whose names are in mask will be omitted during the conversion.

source

Hamiltonian.Essentials.DegreeOfFreedom.IDFConfig — Type.

IDFConfig{I}(indextotuple::IndexToTuple,map::Function,pids::AbstractVector{<:PID}=[]) where I<:Internal

Configuration of the internal degrees of freedom at a lattice.

map maps a PID to an Internal, while indextotuple maps an Index to a tuple.

source

Hamiltonian.Essentials.DegreeOfFreedom.IID — Type.

IID

The id of an internal degree of freedom.

source

Hamiltonian.Essentials.DegreeOfFreedom.Index — Type.

Index{P,I}

The complete index of a degree of freedom, which consist of the spatial part and the internal part.

source

Hamiltonian.Essentials.DegreeOfFreedom.IndexToTuple — Type.

IndexToTuple

The rules for converting an index to a tuple.

As a function, every instance should accept two positional arguments

index::Index: the index to be converted to a tuple

mask::NTuple{N,Symbol}: the names of the attributes of an index to be omitted during the conversion

source

Hamiltonian.Essentials.DegreeOfFreedom.Internal — Type.

Internal

The whole internal degrees of freedom at a single point.

source

Hamiltonian.Essentials.DegreeOfFreedom.Table — Type.

Table{I<:Index} <: AbstractDict{I,Int}

Index-sequence table. Alias for Dict{I<:Index,Int}.

source

Hamiltonian.Essentials.DegreeOfFreedom.Table — Method.

Table(indices::AbstractVector{<:Index};by::IndexToTuple=directindextotuple,mask::NTuple{N,Symbol}=()) where N -> Table

Convert an sequence of indices to the corresponding index-sequence table.

The input indices will be converted to tuples by the by function along with the mask parameter, which contains the names of omitted attributes of the indices during this conversion. Then duplicates are removed and the resulting unique tuples are sorted, which determines the sequence of the input indices. Note that two indices have the same sequence if their converted tupels are equal to each other.

source

Hamiltonian.Essentials.DegreeOfFreedom.Table — Method.

Table(config::IDFConfig;mask::NTuple{N,Symbol}=()) where N -> Table

Get the index-sequence table of the whole internal Hilbert spaces at a lattice.

When mask is nonempty, it contains the names of the attributes that will be omitted during the conversion from an index to tuple.

source

Hamiltonian.Essentials.DegreeOfFreedom.directindextotuple — Function.

directindextotuple

Indicate that the conversion from an index to a tuple is direct.

source

Hamiltonian.Essentials.DegreeOfFreedom.iid — Method.

iid(index::Index) -> IID

The internal part of an index.

source

Hamiltonian.Essentials.DegreeOfFreedom.iidtype — Method.

iidtype(index::Index)
iidtype(::Type{<:Index{P,I}}) where {P,I}

The type of the internal part of an index.

source

Hamiltonian.Essentials.DegreeOfFreedom.pid — Method.

pid(index::Index) -> PID

The spatial part of an index.

source

Hamiltonian.Essentials.DegreeOfFreedom.pidtype — Method.

pidtype(index::Index)
pidtype(::Type{<:Index{P,I}}) where {P,I}

The type of the spatial part of an index.

source

Core.Type — Method.

(INDEX::Type{<:Index})(pid::PID,iid::IID) -> INDEX

Get the corresponding index from a pid and an iid.

source

Base.:== — Method.

==(i1::I,i2::I) where I<:Internal
isequal(i1::I,i2::I) where I<:Internal

Compare two internals and judge whether they are equal to each other.

source

Base.convert — Method.

convert(::Type{Tuple},index::Index;by::IndexToTuple=directindextotuple,mask::NTuple{N,Symbol}=()) where N -> Tuple

Convert an index to tuple.

by specifies the algorithm to convert an index to tuple, while mask contains the names of the omitted attributes of an index during the conversion.

source

Base.eltype — Method.

eltype(internal::Internal)
eltype(::Type{<:Internal{I}}) where I

Get the type of the IIDs that an Internal contains.

source

Base.replace! — Method.

replace!(config::IDFConfig,pids::PID...) -> IDFConfig

Reset the idfconfig with new pids.

source

Base.reverse — Method.

reverse(table::Table) -> Dict{Int,Set{<:Index}}

Convert an index-sequence table to a sequence-indices table.

Since different indices may correspond to the same sequence, the reverse is a one-to-many map.

source

Base.show — Method.

show(io::IO,i::Internal)

Show an internal.

source

Base.union — Method.

union(::Type{P},::Type{I}) where {P<:PID,I<:IID}

Combine a concrete PID type and a concrete IID type to a concrete Index type.

source

Base.union — Method.

union(tables::Table...;by::IndexToTuple=directindextotuple,mask::NTuple{N,Symbol}=()) where N -> Table

Unite several index-sequence tables.

See Table for more details.

source