Degrees of freedom

Degrees of freedom

IID and Internal

Index

IDFConfig

Table

OID and Operator

Operators

Manul

Hamiltonian.Essentials.DegreesOfFreedom.DirectIndexToTuple — Type.

DirectIndexToTuple

Direct index to tuple.

source

Hamiltonian.Essentials.DegreesOfFreedom.DirectIndexToTuple — Method.

(indextotuple::DirectIndexToTuple)(index::Index) -> Tuple

Convert an index to tuple directly.

source

Hamiltonian.Essentials.DegreesOfFreedom.FilteredAttributes — Type.

FilteredAttributes(::Type{I}) where I<:Index

A method that converts an arbitary index to a tuple, by iterating over the selected attributes in a specific order.

source

Hamiltonian.Essentials.DegreesOfFreedom.FilteredAttributes — Method.

(indextotuple::FilteredAttributes)(index::Index) -> Tuple

Convert an index to tuple by the "filtered attributes" method.

source

Hamiltonian.Essentials.DegreesOfFreedom.IDFConfig — Type.

IDFConfig{I}(map::Function,pids::Union{AbstractVector{<:PID},Tuple{}}=()) where I<:Internal

Configuration of the internal degrees of freedom at a lattice.

map maps a PID to an Internal.

source

Hamiltonian.Essentials.DegreesOfFreedom.IID — Type.

IID

The id of an internal degree of freedom.

source

Hamiltonian.Essentials.DegreesOfFreedom.Index — Type.

Index{P,I}

The complete index of a degree of freedom, which consist of the spatial part and the internal part.

source

Hamiltonian.Essentials.DegreesOfFreedom.IndexToTuple — Type.

IndexToTuple

The rules for converting an index to a tuple.

As a function, every instance should accept only one positional argument, i.e. the index to be converted to a tuple.

source

Hamiltonian.Essentials.DegreesOfFreedom.Internal — Type.

Internal

The whole internal degrees of freedom at a single point.

source

Hamiltonian.Essentials.DegreesOfFreedom.OID — Type.

OID(index::Index,::Nothing,::Nothing,seq::Union{Nothing,Int})
OID(index::Index,rcoord::SVector{N,Float},icoord::SVector{N,Float},seq::Union{Nothing,Int}) where N
OID(index::Index,rcoord::Vector{Float},icoord::Vector{Float},seq::Union{Nothing,Int})
OID(index::Index;rcoord::Union{Nothing,SVector,Vector{Float}}=nothing,icoord::Union{Nothing,SVector,Vector{Float}}=nothing,seq::Union{Nothing,Int}=nothing)

Operator id.

source

Hamiltonian.Essentials.DegreesOfFreedom.Operator — Type.

Operator{N,V<:Number,I<:ID{<:NTuple{N,OID}}} <: Element{N,V,I}

Abstract type for an operator.

source

Hamiltonian.Essentials.DegreesOfFreedom.Operators — Type.

Operators(opts::Operator...)

A set of operators.

Type alias of Operators{I<:ID,O<:Operator}=Elements{I,O}.

source

Hamiltonian.Essentials.DegreesOfFreedom.Table — Type.

Table{I<:Index} <: AbstractDict{I,Int}

Index-sequence table. Alias for Dict{I<:Index,Int}.

source

Hamiltonian.Essentials.DegreesOfFreedom.Table — Method.

Table(indices::AbstractVector{<:Index};by::IndexToTuple=directindextotuple) -> Table

Convert an sequence of indices to the corresponding index-sequence table.

The input indices will be converted to tuples by the by function with the duplicates removed. The resulting unique tuples are sorted, which determines the sequence of the input indices. Note that two indices have the same sequence if their converted tupels are equal to each other.

source

Hamiltonian.Essentials.DegreesOfFreedom.Table — Method.

Table(config::IDFConfig;by::IndexToTuple=directindextotuple) -> Table

Get the index-sequence table of the whole internal Hilbert spaces at a lattice.

source

Hamiltonian.Essentials.DegreesOfFreedom.directindextotuple — Function.

directindextotuple

Indicate that the conversion from an index to a tuple is direct.

source

Hamiltonian.Essentials.DegreesOfFreedom.iid — Method.

iid(index::Index) -> IID

Get the internal part of an index.

source

Hamiltonian.Essentials.DegreesOfFreedom.iidtype — Method.

iidtype(index::Index)
iidtype(::Type{<:Index{P,I}}) where {P,I}

Get the type of the internal part of an index.

source

Hamiltonian.Essentials.DegreesOfFreedom.isHermitian — Method.

isHermitian(opts::Operators) -> Bool

Judge whether a set of operators as a whole is Hermitian.

source

Hamiltonian.Essentials.DegreesOfFreedom.isHermitian — Method.

isHermitian(opt::Operator) -> Bool

Judge whether an operator is Hermitian.

source

Hamiltonian.Essentials.DegreesOfFreedom.isHermitian — Method.

isHermitian(oid::ID{<:NTuple{N,OID}}) where N -> Bool

Judge whether an operator id is Hermitian.

source

Hamiltonian.Essentials.DegreesOfFreedom.oidtype — Function.

oidtype

Get the compatible oid type.

source

Hamiltonian.Essentials.DegreesOfFreedom.otype — Function.

otype

Get the compatible operator type from a term type, a bond type and a table type.

source

Hamiltonian.Essentials.DegreesOfFreedom.pid — Method.

pid(index::Index) -> PID

Get the spatial part of an index.

source

Hamiltonian.Essentials.Spatials.icoord — Method.

icoord(opt::Operator{1}) -> SVector
icoord(opt::Operator{2}) -> SVector

Get the whole icoord of an operator.

source

Hamiltonian.Essentials.Spatials.pidtype — Method.

pidtype(index::Index)
pidtype(::Type{<:Index{P,I}}) where {P,I}

Get the type of the spatial part of an index.

source

Hamiltonian.Essentials.Spatials.rcoord — Method.

rcoord(opt::Operator{1}) -> SVector
rcoord(opt::Operator{2}) -> SVector

Get the whole rcoord of an operator.

source

Core.Type — Method.

(INDEX::Type{<:Index})(pid::PID,iid::IID) -> INDEX

Get the corresponding index from a pid and an iid.

source

Base.adjoint — Method.

adjoint(opts::Operators) -> Operators

Get the adjoint of a set of operators.

source

Base.adjoint — Method.

adjoint(index::Index) -> typeof(index)

Get the adjoint of an index.

source

Base.adjoint — Method.

adjoint(oid::OID) -> typeof(oid)
adjoint(oid::ID{<:NTuple{N,OID}}) where N -> typeof(oid)

Get the adjoint of an operator id.

source

Base.adjoint — Method.

adjoint(opt::Operator{N}) where N -> Operator

Get the adjoint of an operator.

source

Base.filter — Method.

filter(f::Function,indextotuple::FilteredAttributes)

Filter the attributes of a "filtered attributes" method.

source

Base.length — Method.

length(indextotuple::FilteredAttributes) -> Int
length(::Type{<:FilteredAttributes{N}}) where N -> Int

Get the length of the filtered attributes.

source

Base.replace! — Method.

replace!(config::IDFConfig,pids::PID...) -> IDFConfig

Reset the idfconfig with new pids.

source

Base.reverse — Method.

reverse(table::Table) -> Dict{Int,Set{<:Index}}

Convert an index-sequence table to a sequence-indices table.

Since different indices may correspond to the same sequence, the reverse is a one-to-many map.

source

Base.show — Method.

show(io::IO,i::Internal)

Show an internal.

source

Base.show — Method.

show(io::IO,oid::OID)

Show an operator id.

source

Base.show — Method.

show(io::IO,opt::Operator)

Show an operator.

source

Base.union — Method.

union(tables::Table...;by::IndexToTuple=directindextotuple) -> Table

Unite several index-sequence tables.

See Table for more details.

source

Base.union — Method.

union(::Type{P},::Type{I}) where {P<:PID,I<:IID}

Combine a concrete PID type and a concrete IID type to a concrete Index type.

source