Degrees of freedom

Internal, IID, Index and OID

IIDSpace

Hilbert

Subscript and Subscripts

Coupling and Couplings

Metric and Table

Term

Boundary

Manual

QuantumLattices.Essentials.DegreesOfFreedom.AbstractCoupling — Type

AbstractCoupling{V, I<:ID{OperatorUnit}} <: OperatorProd{V, I}

The abstract coupling intra/inter internal degrees of freedom at different lattice points.

QuantumLattices.Essentials.DegreesOfFreedom.AbstractOID — Type

AbstractOID <: OperatorUnit

Abstract type of operator id.

QuantumLattices.Essentials.DegreesOfFreedom.CompositeIID — Type

CompositeIID{T<:Tuple{Vararg{SimpleIID}}} <: IID

The composition of several single internal ids.

QuantumLattices.Essentials.DegreesOfFreedom.CompositeIID — Method

CompositeIID(contents::SimpleIID...)

Construct a composite iid from a set of simple iids.

QuantumLattices.Essentials.DegreesOfFreedom.CompositeInternal — Type

CompositeInternal{T<:Tuple{Vararg{SimpleInternal}}} <: Internal{CompositeIID}

The composition of several single internal spaces.

QuantumLattices.Essentials.DegreesOfFreedom.CompositeInternal — Method

CompositeInternal(contents::SimpleInternal...)

Construct a composite internal space from a set of simple internal spaces.

QuantumLattices.Essentials.DegreesOfFreedom.CompositeOID — Type

CompositeOID{I<:Index} <: AbstractOID

The abstract type of composite operator id.

QuantumLattices.Essentials.DegreesOfFreedom.Coupling — Type

Coupling{V, I<:ID{SimpleIID}, C<:Subscripts, CI<:SubscriptsID} <: AbstractCoupling{V, Tuple{CompositeIID{I}, CI}}

The coupling intra/inter internal degrees of freedom at different lattice points.

QuantumLattices.Essentials.DegreesOfFreedom.Couplings — Type

Couplings(cps::AbstractCoupling...)

A pack of couplings intra/inter internal degrees of freedom at different lattice points.

Alias for OperatorSum{<:AbstractCoupling, <:ID{OperatorUnit}}.

QuantumLattices.Essentials.DegreesOfFreedom.Hilbert — Type

Hilbert{I<:Internal, P<:AbstractPID, M<:Function} <: CompositeDict{P, I}

Hilbert space at a lattice.

QuantumLattices.Essentials.DegreesOfFreedom.Hilbert — Method

Hilbert(map::Function, pids::AbstractVector{<:AbstractPID})

Construct a Hilbert space from a function and a set of point ids.

Here, map maps a AbstractPID to an Internal.

QuantumLattices.Essentials.DegreesOfFreedom.Hilbert — Method

Hilbert(ps::Pair...)
Hilbert(kv)

Construct a Hilbert space the same way as a Dict.

QuantumLattices.Essentials.DegreesOfFreedom.Hilbert — Method

Hilbert{I}(map::Function, pids::AbstractVector{<:AbstractPID}) where {I<:Internal}

Construct a Hilbert space from a function and a set of point ids.

Here, map maps a AbstractPID to an Internal.

QuantumLattices.Essentials.DegreesOfFreedom.IID — Type

IID <: OperatorUnit

The id of an internal degree of freedom.

QuantumLattices.Essentials.DegreesOfFreedom.IIDSpace — Type

IIDSpace{I<:IID, V<:Internal, Kind} <: CartesianVectorSpace{IID}

The space expanded by a "labeled" iid.

QuantumLattices.Essentials.DegreesOfFreedom.Index — Type

Index{P<:AbstractPID, I<:SimpleIID} <: AbstractOID

The index of a degree of freedom, which consist of the spatial part and the internal part.

QuantumLattices.Essentials.DegreesOfFreedom.Internal — Type

Internal{I<:IID} <: CartesianVectorSpace{I}

The whole internal degrees of freedom at a single point.

QuantumLattices.Essentials.DegreesOfFreedom.Metric — Type

Metric <: Function

The rules for measuring a concrete oid so that oids can be compared.

As a function, every instance should accept only one positional argument, i.e. the concrete oid to be measured.

QuantumLattices.Essentials.DegreesOfFreedom.OID — Type

OID{I<:Index, V<:SVector} <: CompositeOID{I}

Operator id.

QuantumLattices.Essentials.DegreesOfFreedom.OID — Method

OID(index::Index, rcoord, icoord)
OID(index::Index; rcoord, icoord)

Construct an operator id.

QuantumLattices.Essentials.DegreesOfFreedom.OIDToTuple — Type

OIDToTuple{Fields} <: Metric

A rule that converts an oid to a tuple by iterating over a set of selected fields in a specific order.

QuantumLattices.Essentials.DegreesOfFreedom.OIDToTuple — Method

(oidtotuple::OIDToTuple)(index::Index) -> Tuple

Convert a concrete oid to a tuple.

QuantumLattices.Essentials.DegreesOfFreedom.OIDToTuple — Method

OIDToTuple(::Type{I}) where {I<:Index}

Construct the conversion rule from the information of subtypes of AbstractOID.

QuantumLattices.Essentials.DegreesOfFreedom.SimpleIID — Type

SimpleIID <: IID

The id of a simple internal degree of freedom.

QuantumLattices.Essentials.DegreesOfFreedom.SimpleInternal — Type

SimpleInternal{I<:SimpleIID} <: Internal{I}

The simple internal degrees of freedom at a single point.

QuantumLattices.Essentials.DegreesOfFreedom.Subscript — Type

Subscript{P<:Tuple, C<:Function} <: CompositeTuple{P}

The subscript representative of a certain internal degree of freedom.

QuantumLattices.Essentials.DegreesOfFreedom.Subscript — Method

Subscript(N::Int)
Subscript(pattern::Tuple, check_constant::Bool=false)

Construct the subscript representative of a certain internal degree of freedom.

QuantumLattices.Essentials.DegreesOfFreedom.Subscripts — Type

Subscripts{T<:Tuple{Vararg{NamedContainer{Subscript}}}} <: CompositeTuple{T}

The complete set of subscripts of the internal degrees of freedom.

QuantumLattices.Essentials.DegreesOfFreedom.Subscripts — Method

Subscripts(contents::NamedContainer{Subscript}...)

Construct the complete set of subscripts.

QuantumLattices.Essentials.DegreesOfFreedom.SubscriptsID — Type

SubscriptsID{T<:Tuple{Vararg{Pair{UnitRange{Int}, <:Tuple{Vararg{String}}}}}} <: OperatorUnit

The id of the subscripts.

QuantumLattices.Essentials.DegreesOfFreedom.SubscriptsID — Method

SubscriptsID(subscripts::Subscripts)

Construct the id of the subscripts.

QuantumLattices.Essentials.DegreesOfFreedom.Table — Type

Table(hilbert::Hilbert, by::Metric=OIDToTuple(typeof(hilbert))) -> Table

Get the oid-sequence table of a Hilbert space.

QuantumLattices.Essentials.DegreesOfFreedom.Table — Type

Table(oids::AbstractVector{<:AbstractOID}, by::Metric=OIDToTuple(eltype(oids)))

Convert a set of concrete oids to the corresponding table of oid-sequence pairs.

The input oids are measured by the input by function with the duplicates removed. The resulting unique values are sorted, which determines the sequence of the input oids. Note that two oids have the same sequence if their converted values are equal to each other.

QuantumLattices.Essentials.DegreesOfFreedom.Table — Type

Table{I, B<:Metric} <: CompositeDict{I, Int}

The table of oid-sequence pairs.

QuantumLattices.Essentials.DegreesOfFreedom.Term — Type

Term{K, I, V, B, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}

A term of a quantum lattice system.

QuantumLattices.Essentials.DegreesOfFreedom.Term — Method

Term{K}(id::Symbol, value, bondkind;
    couplings::Union{Function, Couplings},
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false
    ) where K

Construct a term.

QuantumLattices.Essentials.DegreesOfFreedom.TermAmplitude — Type

TermAmplitude(amplitude::Union{Function, Nothing}=nothing)

The function for the amplitude of a term.

QuantumLattices.Essentials.DegreesOfFreedom.TermCouplings — Type

TermCouplings(couplings::Union{Couplings, Function})

The function for the couplings of a term.

QuantumLattices.Essentials.DegreesOfFreedom.TermFunction — Type

TermFunction <: Function

Abstract type for concrete term functions.

QuantumLattices.Essentials.DegreesOfFreedom.TermModulate — Type

TermModulate(id::Symbol, modulate::Function)
TermModulate(id::Symbol, modulate::Bool)

The function for the modulation of a term.

QuantumLattices.Essentials.DegreesOfFreedom.@couplings — Macro

@couplings cps -> Couplings

Convert an expression/literal to a set of couplings.

QuantumLattices.Essentials.DegreesOfFreedom.@subscript_str — Macro

subscript"..." -> Subscript

Construct the subscript from a literal string.

Base.:* — Method

*(cp₁::Coupling, cp₂::Coupling) -> Coupling

Get the multiplication between two couplings.

Base.:* — Method

*(subscripts₁::Subscripts, subscripts₂::Subscripts) -> Subscripts

Get the combination of two sets of subscripts.

Base.:+ — Method

+(term::Term) -> Term
-(term::Term) -> Term
*(term::Term, factor) -> Term
*(factor, term::Term) -> Term
/(term::Term, factor) -> Term

Allowed arithmetic operations for a term.

Base.adjoint — Method

adjoint(index::Index) -> typeof(index)

Get the adjoint of an index.

Base.adjoint — Method

adjoint(oid::OID) -> typeof(oid)

Get the adjoint of an operator id.

Base.filter — Method

filter(f::Function, oidtotuple::OIDToTuple) -> OIDToTuple

Filter the selected fields.

Base.filter — Method

filter(iid::SimpleIID, ci::CompositeInternal) -> Union{Nothing, SimpleInternal, CompositeInternal}
filter(::Type{I}, ci::CompositeInternal) where {I<:SimpleIID} -> Union{Nothing, SimpleInternal, CompositeInternal}

Filter the composite internal space and select those that matches I or the type of iid.

Base.filter — Method

filter(iid::SimpleIID, i::SimpleInternal) -> Union{Nothing, typeof(i)}
filter(::Type{I}, i::SimpleInternal) where {I<:SimpleIID} -> Union{Nothing, typeof(i)}

Filter a simple internal space with respect to the input iid or type I.

Base.filter — Method

filter(iid::SimpleIID, ::Type{C}) where {C<:CompositeInternal}
filter(::Type{I}, ::Type{C}) where {I<:SimpleIID, C<:CompositeInternal}

Filter the type of a composite internal space and select those that matches I or the type of iid.

Base.filter — Method

filter(iid::SimpleIID, ::Type{T}) where {T<:SimpleInternal}
filter(::Type{I}, ::Type{T}) where {I<:SimpleIID, T<:SimpleInternal}

Filter the type of a simple internal space with respect to the input iid or type I.

Base.getindex — Method

getindex(table::Table, oid::AbstractOID) -> Int

Inquiry the sequence of an oid.

Base.haskey — Method

haskey(table::Table, oid::AbstractOID) -> Bool
haskey(table::Table, id::ID{AbstractOID}) -> Tuple{Vararg{Bool}}

Judge whether a single oid or a set of oids have been assigned with sequences in table.

Base.keys — Method

keys(::OIDToTuple{Fields}) where Fields -> Fields
keys(::Type{<:OIDToTuple{Fields}}) where Fields -> Fields

Get the names of the selected fields.

Base.match — Method

match(iid::SimpleIID, i::SimpleInternal) -> Bool
match(::Type{I}, ::Type{SI}) where {I<:SimpleIID, SI<:SimpleInternal}

Judge whether a simple iid or a simple iid type matches a simple internal space or a simple internal space type.

Here, "match" means that the eltype of the simple internal space has the same type name with the simple iid.

Base.match — Method

match(subscript::Subscript, values::Tuple) -> Bool

Judge whether a set of values matches the pattern specified by the subscript.

Base.match — Method

match(subscripts::Subscripts, iids::Tuple{Vararg{SimpleIID}}) -> Bool
match(subscripts::Subscripts, ciid::CompositeIID) -> Bool

Judge whether a composite iid matches the patterns specified by the subscripts.

Base.one — Method

one(term::Term) -> Term

Get a unit term.

Base.replace — Method

replace(term::Term; kwargs...) -> Term

Replace some attributes of a term with key word arguments.

Base.repr — Method

repr(term::Term, bond::AbstractBond, hilbert::Hilbert) -> String

Get the repr representation of a term on a bond with a given Hilbert space.

Base.union — Method

union(tables::Table...) -> Table

Unite several oid-sequence tables.

Base.valtype — Method

valtype(term::Term)
valtype(::Type{<:Term)

Get the value type of a term.

Base.valtype — Method

valtype(::Type{<:OIDToTuple}, ::Type{<:Index})

Get the valtype of applying an OIDToTuple rule to a subtype of AbstractOID.

Base.zero — Method

zero(term::Term) -> Term

Get a zero term.

LinearAlgebra.ishermitian — Method

ishermitian(term::Term) -> Bool
ishermitian(::Type{<:Term}) -> Bool

LinearAlgebra.rank — Method

rank(ciid::CompositeIID) -> Int
rank(::Type{<:CompositeIID{T}}) where {T<:Tuple{Vararg{SimpleIID}}} -> Int

Get the number of simple iids in a composite iid.

LinearAlgebra.rank — Method

rank(ci::CompositeInternal) -> Int
rank(::Type{<:CompositeInternal{T}}) where {T<:Tuple{Vararg{SimpleInternal}}} -> Int

Get the number of simple internal spaces in a composite internal space.

LinearAlgebra.rank — Method

rank(subscripts::Subscripts, i::Integer) -> Int
rank(::Type{<:Subscripts{T}}, i::Integer) where {T<:Tuple{Vararg{NamedContainer{Subscript}}}} -> Int

Get the rank of the ith homogenous segment of the subscripts.

LinearAlgebra.rank — Method

rank(subscripts::Subscripts) -> Int
rank(::Type{<:Subscripts{T}}) where {T<:Tuple{Vararg{NamedContainer{Subscript}}}} -> Int

Get the rank of the subscripts.

LinearAlgebra.rank — Method

rank(subscript::Subscript) -> Int
rank(::Type{<:Subscript}) -> Int

Get the number of the whole variables of the subscript.

LinearAlgebra.rank — Method

rank(term::Term) -> Int
rank(::Type{<:Term) -> Int

Get the rank of a term.

QuantumLattices.Essentials.DegreesOfFreedom.abbr — Method

abbr(term::Term) -> Symbol
abbr(::Type{<:Term}) -> Symbol

Get the abbreviation of the kind of a term.

QuantumLattices.Essentials.DegreesOfFreedom.couplingcenters — Method

couplingcenters(coupling::AbstractCoupling, bond::AbstractBond, info::Val) -> NTuple{rank(coupling), Int}

Get the acting centers of the coupling on a bond.

QuantumLattices.Essentials.DegreesOfFreedom.couplinginternals — Method

couplinginternals(coupling::AbstractCoupling, bond::AbstractBond, hilbert::Hilbert, info::Val) -> NTuple{rank(coupling), SimpleInternal}

Get the internal spaces where each order of the coupling acts on.

QuantumLattices.Essentials.DegreesOfFreedom.couplingpoints — Method

couplingpoints(coupling::AbstractCoupling, bond::AbstractBond, info::Val) -> NTuple{rank(coupling), eltype(bond)}

Get the points where each order of the coupling acts on.

QuantumLattices.Essentials.DegreesOfFreedom.iidtype — Method

iidtype(ciid::CompositeIID, i::Integer)
iidtype(::Type{<:CompositeIID{T}}, i::Integer) where {T<:Tuple{Vararg{SimpleIID}}}

Get the type of the ith simple iid in a composite iid.

QuantumLattices.Essentials.DegreesOfFreedom.iidtype — Method

iidtype(index::Index)
iidtype(::Type{I}) where {I<:Index}

Get the type of the internal part of an index.

QuantumLattices.Essentials.DegreesOfFreedom.indextype — Method

indextype(::CompositeOID)
indextype(::Type{<:CompositeOID})

Get the index type of a composite operator id.

QuantumLattices.Essentials.DegreesOfFreedom.ismodulatable — Method

ismodulatable(term::Term) -> Bool
ismodulatable(::Type{<:Term}) -> Bool

Judge whether a term could be modulated by its modulate function.

QuantumLattices.Essentials.DegreesOfFreedom.oidtype — Method

oidtype(I::Type{<:SimpleInternal}, P::Type{<:Point}, ::Val)

Get the compatible oid type from the combination of the internal part and the spatial part.

QuantumLattices.Essentials.DegreesOfFreedom.statistics — Method

statistics(index::Index) -> Symbol
statistics(::Type{<:Index{<:AbstractPID, I}}) where {I<:SimpleIID} -> Symbol

Get the statistics of an index.

QuantumLattices.Essentials.DegreesOfFreedom.statistics — Method

statistics(oid::CompositeOID) -> Symbol
statistics(::Type{<:CompositeOID{I}}) where {I<:Index} -> Symbol

Get the statistics of a composite operator id.

QuantumLattices.Essentials.DegreesOfFreedom.statistics — Method

statistics(i::SimpleInternal) -> Symbol
statistics(::Type{<:SimpleInternal{I}}) where {I<:SimpleIID} -> Symbol

Get the statistics of a simple internal space.

QuantumLattices.Essentials.QuantumOperators.idtype — Method

idtype(subscripts::Subscripts)
idtype(::Type{<:Subscripts{T}}) where {T<:Tuple{Vararg{NamedContainer{Subscript}}}}

Get the type of the id of the subscripts.

QuantumLattices.Essentials.QuantumOperators.optype — Method

optype(::Type{T}, ::Type{H}, ::Type{B}) where {T<:Term, H<:Hilbert, B<:AbstractBond}

Get the compatible OperatorProd type from the type of a term, a Hilbert space and a bond.

QuantumLattices.Essentials.QuantumOperators.script — Method

script(::Val{:integralicoord}, oid::OID; vectors, kwargs...)

Get the integral script of the icoord of an oid.

QuantumLattices.Essentials.QuantumOperators.script — Method

script(::Val{:rcoord}, oid::OID; kwargs...) -> String
script(::Val{:icoord}, oid::OID; kwargs...) -> String

Get the :rcoord/:icoord script of an oid.

QuantumLattices.Essentials.QuantumOperators.script — Method

script(::Val{attr}, oid::OID; kwargs...) where attr

Get the attr script of an oid, which is contained in its index.

QuantumLattices.Essentials.Spatials.icoord — Method

icoord(opt::Operator{<:Number, <:ID{OID}}) -> SVector

Get the whole icoord of an operator.

QuantumLattices.Essentials.Spatials.pidtype — Method

pidtype(index::Index)
pidtype(::Type{I}) where {I<:Index}

Get the type of the spatial part of an index.

QuantumLattices.Essentials.Spatials.rcoord — Method

rcoord(opt::Operator{<:Number, <:ID{OID}}) -> SVector

Get the whole rcoord of an operator.

QuantumLattices.Essentials.kind — Method

kind(term::Term) -> Symbol
kind(::Type{<:Term) -> Symbol

Get the kind of a term.

QuantumLattices.Essentials.reset! — Method

reset!(hilbert::Hilbert, pids) -> Hilbert

Reset the Hilbert space with new pids.

QuantumLattices.Essentials.reset! — Method

reset!(table::Table, oids::AbstractVector{<:AbstractOID}) -> Table

Reset a table by a new set of oids.

QuantumLattices.Essentials.reset! — Method

reset!(table::Table, hilbert::Hilbert) -> Table

Reset a table by a Hilbert space.

QuantumLattices.Essentials.update! — Method

update!(term::Term, args...; kwargs...) -> Term

Update the value of a term by its modulate function.

QuantumLattices.Interfaces.:⊗ — Method

⊗(iid₁::SimpleIID, iid₂::SimpleIID) -> CompositeIID
⊗(iid::SimpleIID, ciid::CompositeIID) -> CompositeIID
⊗(ciid::CompositeIID, iid::SimpleIID) -> CompositeIID
⊗(ciid₁::CompositeIID, ciid₂::CompositeIID) -> CompositeIID

Direct product between simple iids and composite iids.

QuantumLattices.Interfaces.:⊗ — Method

⊗(i₁::SimpleInternal, i₂::SimpleInternal) -> CompositeInternal
⊗(i::SimpleInternal, ci::CompositeInternal) -> CompositeInternal
⊗(ci::CompositeInternal, i::SimpleInternal) -> CompositeInternal
⊗(ci₁::CompositeInternal, ci₂::CompositeInternal) -> CompositeInternal

Direct product between simple internal spaces and composite internal spaces.

QuantumLattices.Interfaces.expand! — Method

expand!(operators::Operators, term::Term, bond::AbstractBond, hilbert::Hilbert; half::Bool=false, table::Union{Nothing, Table}=nothing) -> Operators
expand!(operators::Operators, term::Term, bonds, hilbert::Hilbert; half::Bool=false, table::Union{Nothing, Table}=nothing) -> Operators

Expand the operators of a term on a bond/set-of-bonds with a given Hilbert space.

The half parameter determines the behavior of generating operators, which falls into the following two categories

true: "Hermitian half" of the generated operators

false: "Hermitian whole" of the generated operators

QuantumLattices.Interfaces.expand — Method

expand(coupling::AbstractCoupling, bond::AbstractBond, hilbert::Hilbert, info::Val)

Expand a coupling with the given bond and Hilbert space.

QuantumLattices.Interfaces.expand — Method

expand(term::Term, bond::AbstractBond, hilbert::Hilbert; half::Bool=false, table::Union{Nothing, Table}=nothing) -> Operators
expand(term::Term, bonds::Bonds, hilbert::Hilbert; half::Bool=false, table::Union{Nothing, Table}=nothing) -> Operators

Expand the operators of a term on a bond/set-of-bonds with a given Hilbert space.

QuantumLattices.Interfaces.expand — Method

expand(iids::NTuple{N, IID}, internals::NTuple{N, Internal}) where N -> IIDSpace

Get the space expanded by a set of "labeled" iids.

QuantumLattices.Interfaces.id — Method

id(term::Term) -> Symbol
id(::Type{<:Term) -> Symbol

Get the id of a term.