Degrees of freedom

Degrees of freedom

IID and Internal

Index

IDFConfig

Table

OID and Operator

Operators

LaTeX

Boundary

Manul

QuantumLattices.Essentials.DegreesOfFreedom.coordabsent — Constant.

coordabsent

Indicate that the :icoord and :rcoord attributes in an oid should be nothing

source

QuantumLattices.Essentials.DegreesOfFreedom.coordpresent — Constant.

coordpresent

Indicate that the :icoord and :rcoord attributes in an oid should not be nothing.

source

QuantumLattices.Essentials.DegreesOfFreedom.Boundary — Type.

Boundary{Names}(values::AbstractVector{Float},vectors::AbstractVector{<:AbstractVector{Float}}) where Names
Boundary()

Boundary twist of operators.

source

QuantumLattices.Essentials.DegreesOfFreedom.Boundary — Method.

(bound::Boundary)(operator::Operator) -> Operator
(bound::Boundary{()})(operator::Operator) -> Operator

Get the boundary twisted operator.

source

QuantumLattices.Essentials.DegreesOfFreedom.DirectIndexToTuple — Type.

DirectIndexToTuple

Direct index to tuple.

source

QuantumLattices.Essentials.DegreesOfFreedom.DirectIndexToTuple — Method.

(indextotuple::DirectIndexToTuple)(index::Index) -> Tuple

Convert an index to tuple directly.

source

QuantumLattices.Essentials.DegreesOfFreedom.FilteredAttributes — Type.

FilteredAttributes(::Type{I}) where I<:Index

A method that converts an arbitary index to a tuple, by iterating over the selected attributes in a specific order.

source

QuantumLattices.Essentials.DegreesOfFreedom.FilteredAttributes — Method.

(indextotuple::FilteredAttributes)(index::Index) -> Tuple

Convert an index to tuple by the "filtered attributes" method.

source

QuantumLattices.Essentials.DegreesOfFreedom.IDFConfig — Type.

IDFConfig{I}(map::Function,pids::Union{AbstractVector{<:PID},Tuple{}}=()) where I<:Internal

Configuration of the internal degrees of freedom at a lattice.

map maps a PID to an Internal.

source

QuantumLattices.Essentials.DegreesOfFreedom.IID — Type.

IID

The id of an internal degree of freedom.

source

QuantumLattices.Essentials.DegreesOfFreedom.Index — Type.

Index{P,I}

The complete index of a degree of freedom, which consist of the spatial part and the internal part.

source

QuantumLattices.Essentials.DegreesOfFreedom.IndexToTuple — Type.

IndexToTuple

The rules for converting an index to a tuple.

As a function, every instance should accept only one positional argument, i.e. the index to be converted to a tuple.

source

QuantumLattices.Essentials.DegreesOfFreedom.Internal — Type.

Internal

The whole internal degrees of freedom at a single point.

source

QuantumLattices.Essentials.DegreesOfFreedom.LaTeX — Type.

LaTeX{SP,SB}(body) where {SP,SB}

LaTeX string representation.

source

QuantumLattices.Essentials.DegreesOfFreedom.OID — Type.

OID(index::Index,::Nothing,::Nothing,seq::Union{Nothing,Int})
OID(index::Index,rcoord::SVector{N,Float},icoord::SVector{N,Float},seq::Union{Nothing,Int}) where N
OID(index::Index,rcoord::Vector{Float},icoord::Vector{Float},seq::Union{Nothing,Int})
OID(index::Index;rcoord::Union{Nothing,SVector,Vector{Float}}=nothing,icoord::Union{Nothing,SVector,Vector{Float}}=nothing,seq::Union{Nothing,Int}=nothing)

Operator id.

source

QuantumLattices.Essentials.DegreesOfFreedom.Operator — Type.

Operator{V,I<:ID} <: Element{V,I}

Abstract type for an operator.

source

QuantumLattices.Essentials.DegreesOfFreedom.Operators — Type.

Operators(opts::Operator...)

A set of operators.

Type alias for Elements{<:ID,<:Operator}.

source

QuantumLattices.Essentials.DegreesOfFreedom.Table — Type.

Table{I}(by::IndexToTuple) where I<:Index

Index-sequence table.

source

QuantumLattices.Essentials.DegreesOfFreedom.Table — Type.

Table(config::IDFConfig,by::IndexToTuple=directindextotuple) -> Table

Get the index-sequence table of the whole internal Hilbert spaces at a lattice.

source

QuantumLattices.Essentials.DegreesOfFreedom.Table — Type.

Table(indices::AbstractVector{<:Index},by::IndexToTuple=directindextotuple) -> Table

Convert an sequence of indices to the corresponding index-sequence table.

The input indices will be converted to tuples by the by function with the duplicates removed. The resulting unique tuples are sorted, which determines the sequence of the input indices. Note that two indices have the same sequence if their converted tupels are equal to each other.

source

QuantumLattices.Essentials.DegreesOfFreedom.directindextotuple — Function.

directindextotuple

Indicate that the conversion from an index to a tuple is direct.

source

QuantumLattices.Essentials.DegreesOfFreedom.iid — Method.

iid(index::Index) -> IID

Get the internal part of an index.

source

QuantumLattices.Essentials.DegreesOfFreedom.iidtype — Method.

iidtype(index::Index)
iidtype(::Type{<:Index{<:PID,I}}) where {I<:IID}

Get the type of the internal part of an index.

source

QuantumLattices.Essentials.DegreesOfFreedom.isHermitian — Method.

isHermitian(opts::Operators) -> Bool

Judge whether a set of operators as a whole is Hermitian.

source

QuantumLattices.Essentials.DegreesOfFreedom.isHermitian — Method.

isHermitian(opt::Operator) -> Bool

Judge whether an operator is Hermitian.

source

QuantumLattices.Essentials.DegreesOfFreedom.isHermitian — Method.

isHermitian(oid::ID{<:NTuple{N,OID}}) where N -> Bool

Judge whether an operator id is Hermitian.

source

QuantumLattices.Essentials.DegreesOfFreedom.latexformat — Method.

latexformat(T::Type{<:Operator}) -> LaTeX
latexformat(T::Type{<:Operator},l::LaTeX) -> LaTeX

Get/Set the latex format for a subtype of Operator.

source

QuantumLattices.Essentials.DegreesOfFreedom.oidtype — Method.

oidtype(I::Type{<:IID},B::Type{<:AbstractBond},::Type{Nothing},::Val{true})
oidtype(I::Type{<:IID},B::Type{<:AbstractBond},::Type{<:Table},::Val{true})
oidtype(I::Type{<:IID},B::Type{<:AbstractBond},::Type{Nothing},::Val{false})
oidtype(I::Type{<:IID},B::Type{<:AbstractBond},::Type{<:Table},::Val{false})

Get the compatible oid type.

source

QuantumLattices.Essentials.DegreesOfFreedom.otype — Function.

otype

Get the compatible operator type from a term type, a bond type and a table type.

source

QuantumLattices.Essentials.DegreesOfFreedom.pid — Method.

pid(index::Index) -> PID

Get the spatial part of an index.

source

QuantumLattices.Essentials.DegreesOfFreedom.script — Method.

script(oid::OID,l::LaTeX,::Val{:B}) -> Any
script(oid::OID,l::LaTeX,::Val{:SP}) -> Tuple
script(oid::OID,l::LaTeX,::Val{:SB}) -> Tuple

Get the body/superscript/subscript of the latex string representation of an oid.

source

QuantumLattices.Essentials.DegreesOfFreedom.script — Method.

script(oid::OID,::Val{attr}) where attr -> String

Get the :rcoord/:icoord script of an oid.

source

QuantumLattices.Essentials.DegreesOfFreedom.twist — Method.

twist(operator::Operator,vectors::AbstractVector{<:AbstractVector{Float}},values::AbstractVector{Float}) -> Operator

Twist an operator.

source

Core.Type — Method.

(INDEX::Type{<:Index})(pid::PID,iid::IID) -> INDEX

Get the corresponding index from a pid and an iid.

source

Base.:== — Method.

==(bound1::Boundary,bound2::Boundary) -> Bool
isequal(bound1::Boundary,bound2::Boundary) -> Bool

source

Base.adjoint — Method.

adjoint(opts::Operators) -> Operators

Get the adjoint of a set of operators.

source

Base.adjoint — Method.

adjoint(index::Index) -> typeof(index)

Get the adjoint of an index.

source

Base.adjoint — Method.

adjoint(oid::OID) -> typeof(oid)
adjoint(oid::ID{<:NTuple{N,OID}}) where N -> typeof(oid)

Get the adjoint of an operator id.

source

Base.adjoint — Method.

adjoint(opt::Operator{N}) where N -> Operator

Get the adjoint of an operator.

source

Base.angle — Method.

angle(bound::Boundary,operator::Operator) -> Float
angle(bound::Boundary{()},operator::Operator) -> Int

Get the boundary twist phase of an operator.

source

Base.angle — Method.

angle(id::ID{<:Tuple{Vararg{OID}}},vectors::AbstractVector{<:AbstractVector{Float}},values::AbstractVector{Float}) -> Float

Get the total twist phase of an id.

source

Base.filter — Method.

filter(f::Function,indextotuple::FilteredAttributes)

Filter the attributes of a "filtered attributes" method.

source

Base.length — Method.

length(indextotuple::FilteredAttributes) -> Int
length(::Type{<:FilteredAttributes{N}}) where N -> Int

Get the length of the filtered attributes.

source

Base.repr — Function.

repr(opt::Operator,l::Union{LaTeX,Nothing}=nothing) -> String

Get the latex string representation of an operator.

source

Base.repr — Function.

repr(opts::Operators,l::Union{LaTeX,Nothing}=nothing) -> String

Get the latex string representation of a set of operators.

source

Base.repr — Method.

repr(oid::OID,l::LaTeX) -> String

LaTeX string representation of an oid.

source

Base.reverse — Method.

reverse(table::Table) -> Dict{Int,Set{<:Index}}

Convert an index-sequence table to a sequence-indices table.

Since different indices may correspond to the same sequence, the reverse is a one-to-many map.

source

Base.show — Method.

show(io::IO,i::Internal)

Show an internal.

source

Base.show — Method.

show(io::IO,::MIME"text/latex",opts::Operators)

Show latex formed operators.

source

Base.show — Method.

show(io::IO,oid::OID)

Show an operator id.

source

Base.show — Method.

show(io::IO,opt::Operator)
show(io::IO,::MIME"text/latex",opt::Operator)

Show an operator.

source

Base.summary — Method.

summary(io::IO,opts::Operators)

Print a brief description of a set of operators to an io.

source

Base.union — Method.

union(tables::Table...) -> Table

Unite several index-sequence tables.

See Table for more details.

source

Base.union — Method.

union(::Type{P},::Type{I}) where {P<:PID,I<:IID}

Combine a concrete PID type and a concrete IID type to a concrete Index type.

source

QuantumLattices.Essentials.Spatials.icoord — Method.

icoord(opt::Operator) -> SVector

Get the whole icoord of an operator.

source

QuantumLattices.Essentials.Spatials.pidtype — Method.

pidtype(index::Index)
pidtype(::Type{<:Index{P}}) where {P<:PID}

Get the type of the spatial part of an index.

source

QuantumLattices.Essentials.Spatials.rcoord — Method.

rcoord(opt::Operator) -> SVector

Get the whole rcoord of an operator.

source

QuantumLattices.Interfaces.reset! — Method.

reset!(config::IDFConfig,pids) -> IDFConfig

Reset the idfconfig with new pids.

source

QuantumLattices.Interfaces.reset! — Method.

reset!(table::Table,indices::AbstractVector{<:Index}) -> Table
reset!(table::Table,config::IDFConfig) -> Table

Reset a table.

source

QuantumLattices.Interfaces.sequence — Function.

sequence(opt::Operator,table=nothing) -> NTuple{rank(opt),Int}

Get the sequence of the oids of an operator according to a table.

source

QuantumLattices.Interfaces.update! — Method.

update!(bound::Boundary{Names},args...;kwargs...) where Names -> Boundary

Update the values of the boundary twisted phase.

source

QuantumLattices.Mathematics.AlgebraOverFields.rawelement — Method.

rawelement(::Type{<:Operator})

Get the raw name of a type of Operator.

source