Degrees of freedom

IID, Internal and Index

Hilbert

OID

Operator and Operators

Metric and Table

LaTeX

Boundary

Manual

QuantumLattices.Essentials.DegreesOfFreedom.AbstractCompositeOID — Type

AbstractCompositeOID{I<:Index} <: AbstractOID

The abstract type of composite operator id.

QuantumLattices.Essentials.DegreesOfFreedom.AbstractOID — Type

AbstractOID <: SingularID

Abstract type of operator id.

QuantumLattices.Essentials.DegreesOfFreedom.AbstractOperator — Type

AbstractOperator{V<:Number, I<:ID{AbstractOID}} <: OperatorProd{V, I}

Abstract type for an operator.

QuantumLattices.Essentials.DegreesOfFreedom.Boundary — Type

Boundary{Names}(values::AbstractVector{<:Number}, vectors::AbstractVector{<:AbstractVector{<:Number}}) where Names

Boundary twist of operators.

QuantumLattices.Essentials.DegreesOfFreedom.Boundary — Method

(bound::Boundary)(operator::AbstractOperator) -> AbstractOperator

Get the boundary twisted operator.

QuantumLattices.Essentials.DegreesOfFreedom.CompositeIID — Type

CompositeIID{T<:Tuple{Vararg{SimpleIID}}} <: IID

The composition of several single internal ids.

QuantumLattices.Essentials.DegreesOfFreedom.CompositeIID — Method

CompositeIID(contents::SimpleIID...)

Construct a composite iid from a set of simple iids.

QuantumLattices.Essentials.DegreesOfFreedom.CompositeInternal — Type

CompositeInternal{T<:Tuple{Vararg{SimpleInternal}}} <: Internal{CompositeIID}

The composition of several single internal spaces.

QuantumLattices.Essentials.DegreesOfFreedom.CompositeInternal — Method

CompositeInternal(contents::SimpleInternal...)

Construct a composite internal space from a set of simple internal spaces.

QuantumLattices.Essentials.DegreesOfFreedom.Hilbert — Type

Hilbert{I<:Internal, P<:AbstractPID, M<:Function} <: CompositeDict{P, I}

Hilbert space at a lattice.

QuantumLattices.Essentials.DegreesOfFreedom.Hilbert — Method

Hilbert(map::Function, pids::AbstractVector{<:AbstractPID})

Construct a Hilbert space from a function and a set of point ids.

Here, map maps a AbstractPID to an Internal.

QuantumLattices.Essentials.DegreesOfFreedom.Hilbert — Method

Hilbert(ps::Pair...)
Hilbert(kv)

Construct a Hilbert space the same way as a Dict.

QuantumLattices.Essentials.DegreesOfFreedom.Hilbert — Method

Hilbert{I}(map::Function, pids::AbstractVector{<:AbstractPID}) where {I<:Internal}

Construct a Hilbert space from a function and a set of point ids.

Here, map maps a AbstractPID to an Internal.

QuantumLattices.Essentials.DegreesOfFreedom.IID — Type

IID <: SingularID

The id of an internal degree of freedom.

QuantumLattices.Essentials.DegreesOfFreedom.Index — Type

Index{P<:AbstractPID, I<:SimpleIID} <: AbstractOID

The index of a degree of freedom, which consist of the spatial part and the internal part.

QuantumLattices.Essentials.DegreesOfFreedom.Internal — Type

Internal{I<:IID} <: CartesianVectorSpace{I}

The whole internal degrees of freedom at a single point.

QuantumLattices.Essentials.DegreesOfFreedom.InternalIndex — Type

InternalIndex{I}

Index of a simple internal space in the composite internal spaces.

QuantumLattices.Essentials.DegreesOfFreedom.LaTeX — Type

LaTeX{SP, SB}(body, spdelimiter::String=", ", sbdelimiter::String=", "; options...) where {SP, SB}

LaTeX string representation.

QuantumLattices.Essentials.DegreesOfFreedom.Metric — Type

Metric <: Function

The rules for measuring a concrete oid so that oids can be compared.

As a function, every instance should accept only one positional argument, i.e. the concrete oid to be measured.

QuantumLattices.Essentials.DegreesOfFreedom.OID — Type

OID{I<:Index, V<:SVector} <: AbstractCompositeOID{I}

Operator id.

QuantumLattices.Essentials.DegreesOfFreedom.OID — Method

OID(index::Index, rcoord, icoord)
OID(index::Index; rcoord, icoord)

Construct an operator id.

QuantumLattices.Essentials.DegreesOfFreedom.OIDToTuple — Type

OIDToTuple{Fields} <: Metric

A rule that converts an oid to a tuple by iterating over a set of selected fields in a specific order.

QuantumLattices.Essentials.DegreesOfFreedom.OIDToTuple — Method

(oidtotuple::OIDToTuple)(index::Index) -> Tuple

Convert a concrete oid to a tuple.

QuantumLattices.Essentials.DegreesOfFreedom.OIDToTuple — Method

OIDToTuple(::Type{I}) where {I<:Index}

Construct the conversion rule from the information of subtypes of AbstractOID.

QuantumLattices.Essentials.DegreesOfFreedom.Operator — Type

Operator{V<:Number, I<:ID{AbstractOID}} <: AbstractOperator{V, I}

Operator.

QuantumLattices.Essentials.DegreesOfFreedom.Operators — Type

Operators(opts::AbstractOperator...)

A set of operators.

Type alias for OperatorSum{<:ID{AbstractOID}, <:AbstractOperator}.

QuantumLattices.Essentials.DegreesOfFreedom.SimpleIID — Type

SimpleIID <: IID

The id of a simple internal degree of freedom.

QuantumLattices.Essentials.DegreesOfFreedom.SimpleInternal — Type

SimpleInternal{I<:SimpleIID} <: Internal{I}

The simple internal degrees of freedom at a single point.

QuantumLattices.Essentials.DegreesOfFreedom.Table — Type

Table{I, B<:Metric} <: CompositeDict{I, Int}

The table of oid-sequence pairs.

QuantumLattices.Essentials.DegreesOfFreedom.Table — Type

Table(oids::AbstractVector{<:AbstractOID}, by::Metric=OIDToTuple(eltype(oids)))

Convert a set of concrete oids to the corresponding table of oid-sequence pairs.

The input oids are measured by the input by function with the duplicates removed. The resulting unique values are sorted, which determines the sequence of the input oids. Note that two oids have the same sequence if their converted values are equal to each other.

QuantumLattices.Essentials.DegreesOfFreedom.Table — Type

Table(hilbert::Hilbert, by::Metric=OIDToTuple(typeof(hilbert))) -> Table

Get the oid-sequence table of a Hilbert space.

Base.adjoint — Method

adjoint(index::Index) -> typeof(index)

Get the adjoint of an index.

Base.adjoint — Method

adjoint(oid::OID) -> typeof(oid)
adjoint(oid::ID{AbstractOID}) -> ID{AbstractOID}

Get the adjoint of an operator id.

Base.adjoint — Method

adjoint(opts::Operators) -> Operators

Get the adjoint of a set of operators.

Base.adjoint — Method

adjoint(opt::Operator) -> Operator

Get the adjoint of an operator.

Base.angle — Method

angle(bound::Boundary, operator::AbstractOperator) -> Number

Get the boundary twist phase of an operator.

Base.angle — Method

angle(id::ID{AbstractOID}, vectors::AbstractVector{<:AbstractVector{<:Number}}, values::AbstractVector{<:Number}) -> Number

Get the total twist phase of an id.

Base.filter — Method

filter(f::Function, oidtotuple::OIDToTuple) -> OIDToTuple

Filter the selected fields.

Base.filter — Method

filter(iid::SimpleIID, ci::CompositeInternal) -> Union{Nothing, SimpleInternal, CompositeInternal}
filter(::Type{I}, ci::CompositeInternal) where {I<:SimpleIID} -> Union{Nothing, SimpleInternal, CompositeInternal}

Filter the composite internal space and select those that matches I or the type of iid.

Base.filter — Method

filter(iid::SimpleIID, i::SimpleInternal) -> Union{Nothing, typeof(i)}
filter(::Type{I}, i::SimpleInternal) where {I<:SimpleIID} -> Union{Nothing, typeof(i)}

Filter a simple internal space with respect to the input iid or type I.

Base.filter — Method

filter(iid::SimpleIID, ::Type{C}) where {C<:CompositeInternal}
filter(::Type{I}, ::Type{C}) where {I<:SimpleIID, C<:CompositeInternal}

Filter the type of a composite internal space and select those that matches I or the type of iid.

Base.filter — Method

filter(iid::SimpleIID, ::Type{T}) where {T<:SimpleInternal}
filter(::Type{I}, ::Type{T}) where {I<:SimpleIID, T<:SimpleInternal}

Filter the type of a simple internal space with respect to the input iid or type I.

Base.getindex — Method

getindex(table::Table, oid::AbstractOID) -> Int

Inquiry the sequence of an oid.

Base.getindex — Method

getindex(ci::CompositeInternal, ::InternalIndex{I}) where I

Get the Ith simple internal space in a composite internal space.

Base.haskey — Method

haskey(table::Table, oid::AbstractOID) -> Bool
haskey(table::Table, id::ID{AbstractOID}) -> Tuple{Vararg{Bool}}

Judge whether a single oid or a set of oids have been assigned with sequences in table.

Base.keys — Method

keys(bound::Boundary) -> Tuple{Vararg{Symbol}}
keys(::Type{<:Boundary{Names}}) where Names -> Names

Get the names of the boundary parameters.

Base.keys — Method

keys(::OIDToTuple{Fields}) where Fields -> Fields
keys(::Type{<:OIDToTuple{Fields}}) where Fields -> Fields

Get the names of the selected fields.

Base.match — Method

match(iid::SimpleIID, i::SimpleInternal) -> Bool
match(::Type{I}, ::Type{SI}) where {I<:SimpleIID, SI<:SimpleInternal}

Judge whether a simple iid or a simple iid type matches a simple internal space or a simple internal space type.

Here, "match" means that the eltype of the simple internal space has the same type name with the simple iid.

Base.repr — Method

repr(opts::Operators) -> String

Get the string representation of a set of operators in the LaTeX format.

Base.repr — Method

repr(oid::AbstractOID) -> String
repr(oid::AbstractOID, l::LaTeX) -> String

LaTeX string representation of an oid.

Base.repr — Method

repr(opt::AbstractOperator) -> String

Get the string representation of an operator in the LaTeX format.

Base.show — Method

show(io::IO, ::MIME"text/latex", opts::Operators)

Show LaTeX formed operators.

Base.show — Method

show(io::IO, ::MIME"text/latex", opt::AbstractOperator)

Show an operator.

Base.summary — Method

summary(io::IO, opts::Operators)

Print a brief description of a set of operators to an io.

Base.union — Method

union(tables::Table...) -> Table

Unite several oid-sequence tables.

Base.valtype — Method

valtype(::Type{<:OIDToTuple}, ::Type{<:Index})

Get the valtype of applying an OIDToTuple rule to a subtype of AbstractOID.

QuantumLattices.Essentials.DegreesOfFreedom.iidtype — Method

iidtype(ciid::CompositeIID, i::Integer)
iidtype(::Type{<:CompositeIID{T}}, i::Integer) where {T<:Tuple{Vararg{SimpleIID}}}

Get the type of the ith simple iid in a composite iid.

QuantumLattices.Essentials.DegreesOfFreedom.iidtype — Method

iidtype(index::Index)
iidtype(::Type{<:Index{<:AbstractPID, I}}) where {I<:SimpleIID}

Get the type of the internal part of an index.

QuantumLattices.Essentials.DegreesOfFreedom.indextype — Method

indextype(::AbstractCompositeOID)
indextype(::Type{<:AbstractCompositeOID})

Get the index type of a composite operator id.

QuantumLattices.Essentials.DegreesOfFreedom.internaltype — Method

internaltype(ci::CompositeInternal, index::Int)
internaltype(::Type{<:CompositeInternal{T}}, index::Int) where {T<:Tuple{Vararg{SimpleInternal}}}

Get the type of the ith simple internal space in a composite internal space.

QuantumLattices.Essentials.DegreesOfFreedom.isHermitian — Method

isHermitian(opts::Operators) -> Bool

Judge whether a set of operators as a whole is Hermitian.

QuantumLattices.Essentials.DegreesOfFreedom.isHermitian — Method

isHermitian(opt::Operator) -> Bool

Judge whether an operator is Hermitian.

QuantumLattices.Essentials.DegreesOfFreedom.isHermitian — Method

isHermitian(id::ID{AbstractOID, N}) where N -> Bool

Judge whether an operator id is Hermitian.

QuantumLattices.Essentials.DegreesOfFreedom.latexformat — Method

latexformat(T::Type{<:AbstractOID}) -> LaTeX
latexformat(T::Type{<:AbstractOID}, l::LaTeX) -> LaTeX

Get/Set the LaTeX format for a subtype of AbstractOID.

QuantumLattices.Essentials.DegreesOfFreedom.latexname — Method

latexname(T::Type{<:AbstractOID}) -> Symbol

Get the name of a type of AbstractOID in the latex format lookups.

QuantumLattices.Essentials.DegreesOfFreedom.oidtype — Method

oidtype(I::Type{<:SimpleInternal}, P::Type{<:Point}, ::Val)

Get the compatible oid type from the combination of the internal part and the spatial part.

QuantumLattices.Essentials.DegreesOfFreedom.plain — Function

plain

Plain boundary condition without any twist.

QuantumLattices.Essentials.DegreesOfFreedom.script — Method

script(::Val{:BD}, oid::AbstractOID, l::LaTeX) -> Any
script(::Val{:SP}, oid::AbstractOID, l::LaTeX) -> Tuple
script(::Val{:SB}, oid::AbstractOID, l::LaTeX) -> Tuple

Get the body/superscript/subscript of the LaTeX string representation of an oid.

QuantumLattices.Essentials.DegreesOfFreedom.script — Method

script(::Val{:integralicoord}, oid::OID; vectors, kwargs...)

Get the integral script of the icoord of an oid.

QuantumLattices.Essentials.DegreesOfFreedom.script — Method

script(::Val{:rcoord}, oid::OID; kwargs...) -> String
script(::Val{:icoord}, oid::OID; kwargs...) -> String

Get the :rcoord/:icoord script of an oid.

QuantumLattices.Essentials.DegreesOfFreedom.script — Method

script(::Val{attr}, oid::OID; kwargs...) where attr

Get the attr script of an oid, which is contained in its index.

QuantumLattices.Essentials.DegreesOfFreedom.statistics — Method

statistics(index::Index) -> Symbol
statistics(::Type{<:Index{<:AbstractPID, I}}) where {I<:SimpleIID} -> Symbol

Get the statistics of an index.

QuantumLattices.Essentials.DegreesOfFreedom.statistics — Method

statistics(oid::AbstractCompositeOID) -> Symbol
statistics(::Type{<:AbstractCompositeOID{I}}) where {I<:Index} -> Symbol

Get the statistics of a composite operator id.

QuantumLattices.Essentials.DegreesOfFreedom.statistics — Method

statistics(i::SimpleInternal) -> Symbol
statistics(::Type{<:SimpleInternal{I}}) where {I<:SimpleIID} -> Symbol

Get the statistics of a simple internal space.

QuantumLattices.Essentials.DegreesOfFreedom.twist — Method

twist(operator::AbstractOperator, vectors::AbstractVector{<:AbstractVector{<:Number}}, values::AbstractVector{<:Number}) -> AbstractOperator

Twist an operator.

QuantumLattices.Essentials.QuantumOperators.sequence — Method

sequence(opt::AbstractOperator, table::AbstractDict) -> NTuple{rank(opt), Int}

Get the sequence of the oids of an operator according to a table.

QuantumLattices.Essentials.Spatials.icoord — Method

icoord(opt::Operator) -> SVector

Get the whole icoord of an operator.

QuantumLattices.Essentials.Spatials.pidtype — Method

pidtype(index::Index)
pidtype(::Type{<:Index{P}}) where {P<:AbstractPID}

Get the type of the spatial part of an index.

QuantumLattices.Essentials.Spatials.rcoord — Method

rcoord(opt::Operator) -> SVector

Get the whole rcoord of an operator.

QuantumLattices.Essentials.reset! — Method

reset!(hilbert::Hilbert, pids) -> Hilbert

Reset the Hilbert space with new pids.

QuantumLattices.Essentials.reset! — Method

reset!(table::Table, oids::AbstractVector{<:AbstractOID}) -> Table

Reset a table by a new set of oids.

QuantumLattices.Essentials.reset! — Method

reset!(table::Table, hilbert::Hilbert) -> Table

Reset a table by a Hilbert space.

QuantumLattices.Essentials.update! — Method

update!(bound::Boundary, args...; kwargs...) -> Boundary

Update the values of the boundary twisted phase.

QuantumLattices.Interfaces.:⊗ — Method

⊗(iid₁::SimpleIID, iid₂::SimpleIID) -> CompositeIID
⊗(iid::SimpleIID, ciid::CompositeIID) -> CompositeIID
⊗(ciid::CompositeIID, iid::SimpleIID) -> CompositeIID
⊗(ciid₁::CompositeIID, ciid₂::CompositeIID) -> CompositeIID

Direct product between simple iids and composite iids.

QuantumLattices.Interfaces.:⊗ — Method

⊗(i₁::SimpleInternal, i₂::SimpleInternal) -> CompositeInternal
⊗(i::SimpleInternal, ci::CompositeInternal) -> CompositeInternal
⊗(ci::CompositeInternal, i::SimpleInternal) -> CompositeInternal
⊗(ci₁::CompositeInternal, ci₂::CompositeInternal) -> CompositeInternal

Direct product between simple internal spaces and composite internal spaces.

QuantumLattices.Interfaces.rank — Method

rank(ciid::CompositeIID) -> Int
rank(::Type{<:CompositeIID{T}}) where {T<:Tuple{Vararg{SimpleIID}}} -> Int

Get the number of simple iids in a composite iid.

QuantumLattices.Interfaces.rank — Method

rank(ci::CompositeInternal) -> Int
rank(::Type{<:CompositeInternal{T}}) where {T<:Tuple{Vararg{SimpleInternal}}} -> Int

Get the number of simple internal spaces in a composite internal space.