Fock package

Fock package

Fock degrees of freedom

Hamiltonian.Essentials.FockPackage.ANNIHILATION — Constant.

ANNIHILATION

Indicate that the nambu index is ANNIHILATION.

source

Hamiltonian.Essentials.FockPackage.CREATION — Constant.

CREATION

Indicate that the nambu index is CREATION.

source

Hamiltonian.Essentials.FockPackage.FCID — Type.

FCID{A,O,S,N} <: SimpleID

The id of a Fock coupling.

source

Hamiltonian.Essentials.FockPackage.FCID — Method.

FCID(;atom=nothing,orbital=nothing,spin=nothing,nambu=nothing)

Construct an id of a Fock coupling.

source

Hamiltonian.Essentials.FockPackage.FID — Type.

FID <: IID

The Fock id.

source

Hamiltonian.Essentials.FockPackage.FID — Method.

FID(;orbital::Int=1,spin::Int=1,nambu::Int=ANNIHILATION)

Create a Fock id.

source

Hamiltonian.Essentials.FockPackage.FIndex — Type.

FIndex{S} <: Index{PID{S},FID}

The Fock index.

source

Hamiltonian.Essentials.FockPackage.Fock — Type.

Fock <: Internal{FID}

The Fock interanl degrees of freedom.

source

Hamiltonian.Essentials.FockPackage.Fock — Method.

Fock(;atom::Int=1,norbital::Int=1,nspin::Int=2,nnambu::Int=2)

Construct a Fock degrees of freedom.

source

Hamiltonian.Essentials.FockPackage.FockCoupling — Type.

FockCoupling{N,V,I<:ID{N,<:FCID}} <: Coupling{V,I,N}

A Fock coupling.

source

Hamiltonian.Essentials.FockPackage.FockCoupling — Method.

FockCoupling{N}(    value::Number;
                    atoms::Union{NTuple{N,Int},Nothing}=nothing,
                    orbitals::Union{NTuple{N,Int},Nothing}=nothing,
                    spins::Union{NTuple{N,Int},Nothing}=nothing,
                    nambus::Union{NTuple{N,Int},Nothing}=nothing) where N

Construct a Fock coupling.

source

Hamiltonian.Essentials.FockPackage.nambufockindextotuple — Function.

nambufockindextotuple

Indicate that the filtered attributes are (:scope,:nambu,:site,:orbital,:spin) when converting a Fock index to tuple.

source

Hamiltonian.Essentials.FockPackage.usualfockindextotuple — Function.

usualfockindextotuple

Indicate that the filtered attributes are (:scope,:site,:orbital,:spin) when converting a Fock index to tuple.

source

Hamiltonian.Essentials.FockPackage.σʸ — Method.

σʸ(mode::String)

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σˣ — Method.

σˣ(mode::String)

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σᶻ — Method.

σᶻ(mode::String)

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σ⁰ — Method.

σ⁰(mode::String)

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σ⁺ — Method.

σ⁺(mode::String)

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σ⁻ — Method.

σ⁻(mode::String)

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Utilities.Interfaces.expand — Method.

expand(fc::FockCoupling,bond::AbstractBond,focks::Fock...) -> FCExpand

Expand a Fock coupling on a bond with the given Fock degrees of freedom.

source

Base.:* — Method.

*(fc1::FockCoupling{2},fc2::FockCoupling{2}) -> FockCoupling{2}

Get the multiplication between two rank-2 Fock couplings.

source

Base.adjoint — Method.

adjoint(fid::FID) -> FID

Get the adjoint of a Fock id.

source

Base.fieldnames — Method.

fieldnames(::Type{<:FCID}) -> NTuple{4,Symbol}

Get the fieldnames of FCID.

source

Base.fieldnames — Method.

fieldnames(::Type{<:FIndex}) -> NTuple{5,Symbol}

Get the fieldnames of a Fock index.

source

Base.fieldnames — Method.

fieldnames(::Type{FID}) -> NTuple{3,Symbol}

Get the fieldnames of FID.

source

Base.iterate — Function.

iterate(fock::Fock,state=1)

Iterate over a Fock degrees of freedom.

source

Base.length — Method.

length(fock::Fock) -> Int

Get the dimension of a Fock degrees of freedom.

source

Base.repr — Method.

repr(fc::FockCoupling) -> String

Get the repr representation of a Fock coupling.

source

Base.show — Method.

show(io::IO,fc::FockCoupling)

Show a Fock coupling.

source

Base.union — Method.

union(::Type{P},::Type{FID})

Get the union type of PID and FID.

source