Fock package

Fock package

Fock degrees of freedom

FID and Fock

FIndex

Subscript and Subscripts

FCID and FockCoupling

σ⁰,σˣ,σʸ,σᶻ,σ⁺,σ⁻

Manual

Hamiltonian.Essentials.FockPackage.ANNIHILATION — Constant.

ANNIHILATION

Indicate that the nambu index is ANNIHILATION.

source

Hamiltonian.Essentials.FockPackage.CREATION — Constant.

CREATION

Indicate that the nambu index is CREATION.

source

Hamiltonian.Essentials.FockPackage.FCID — Type.

FCID(;center=wildcard,atom=wildcard,orbital=wildcard,spin=wildcard,nambu=wildcard,obsub=wildcard,spsub=wildcard)

The id of a Fock coupling.

source

Hamiltonian.Essentials.FockPackage.FID — Type.

FID <: IID

The Fock id.

source

Hamiltonian.Essentials.FockPackage.FID — Method.

FID(;orbital::Int=1,spin::Int=1,nambu::Int=ANNIHILATION)

Create a Fock id.

source

Hamiltonian.Essentials.FockPackage.FIndex — Type.

FIndex{S} <: Index{PID{S},FID}

The Fock index.

source

Hamiltonian.Essentials.FockPackage.Fock — Type.

Fock <: Internal{FID}

The Fock interanl degrees of freedom.

source

Hamiltonian.Essentials.FockPackage.Fock — Method.

Fock(;atom::Int=1,norbital::Int=1,nspin::Int=2,nnambu::Int=2)

Construct a Fock degrees of freedom.

source

Hamiltonian.Essentials.FockPackage.FockCoupling — Type.

FockCoupling(value::Number,id::ID{N,I},obsubscripts::Subscripts,spsubscripts::Subscripts) where {N,I<:FCID}
FockCoupling{N}(value::Number;
                centers::Union{NTuple{N,Int},Nothing}=nothing,
                atoms::Union{NTuple{N,Int},Nothing}=nothing,
                orbitals::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                spins::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                nambus::Union{NTuple{N,Int},Nothing}=nothing) where N

Fock coupling.

source

Hamiltonian.Essentials.FockPackage.Subscript — Type.

Subscript(  ipattern::NTuple{N1,Any},
            opattern::NTuple{N2,Any},
            mapping::Union{Function,Nothing}=nothing,
            constrain::Union{Function,Nothing}=nothing,
            identifier::Union{Symbol,Char}=wildcard
            ) where {N1,N2}
Subscript{N}() where N
Subscript(opattern::NTuple{N,Int}) where N

The subscripts of some orbital/spin degrees of freedom.

source

Hamiltonian.Essentials.FockPackage.Subscript — Method.

(subscript::Subscript{N})(::Val{'M'},values::Vararg{Int,N}) where N -> NTuple{dimension(subscript),Int}
(subscript::Subscript{N})(::Val{'C'},values::Vararg{Int,N}) where N -> Bool

Construct the subscripts from a set of independent variables.

Judge whether a set of independent variables are valid to construct the subscripts.

source

Hamiltonian.Essentials.FockPackage.Subscripts — Type.

Subscripts(contents::Subscript...)

A complete set of all the independent subscripts of the orbital/spin degrees of freedom.

source

Hamiltonian.Essentials.FockPackage.Subscripts — Method.

(subscripts::Subscripts)(::Val{'M'},values::NTuple{N,Int}) where N -> NTuple{dimension(subscripts),Int}
(subscripts::Subscripts)(::Val{'C'},values::NTuple{N,Int}) where N -> Bool

Construct the complete set of subscripts from a complete set of independent variables.

Judge whether a complete set of independent variables are valid to construct the complete subscripts.

source

Hamiltonian.Essentials.FockPackage.@subscript — Macro.

@subscript expr::Expr with constrain::Expr -> Subscript

Construct a subscript from a map and optionally with a constrain.

source

Hamiltonian.Essentials.FockPackage.nambufockindextotuple — Function.

nambufockindextotuple

Indicate that the filtered attributes are (:scope,:nambu,:site,:orbital,:spin) when converting a Fock index to tuple.

source

Hamiltonian.Essentials.FockPackage.usualfockindextotuple — Function.

usualfockindextotuple

Indicate that the filtered attributes are (:scope,:site,:orbital,:spin) when converting a Fock index to tuple.

source

Hamiltonian.Essentials.FockPackage.σʸ — Method.

σʸ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σˣ — Method.

σˣ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σᶻ — Method.

σᶻ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σ⁰ — Method.

σ⁰(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σ⁺ — Method.

σ⁺(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σ⁻ — Method.

σ⁻(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Prerequisites.Interfaces.:⊗ — Method.

⊗(fc1::FockCoupling{N},fc2::FockCoupling{N}) where N -> FockCoupling

Get the direct product between two Fock couplings.

source

Hamiltonian.Prerequisites.Interfaces.dimension — Method.

dimension(subscripts::Subscripts,i::Int) -> Int
dimension(::Type{<:Subscripts{T}},i::Int) where T -> Int

Get the total number of the whole variables of a component of the complete subscript set.

source

Hamiltonian.Prerequisites.Interfaces.dimension — Method.

dimension(subscripts::Subscripts) -> Int
dimension(::Type{S}) where S<:Subscripts -> Int

Get the total number of the whole variables of the complete subscript set.

source

Hamiltonian.Prerequisites.Interfaces.dimension — Method.

dimension(subscript::Subscript) -> Int
dimension(::Type{<:Subscript{N1,N2}}) where {N1,N2} -> Int

Get the number of the whole variables that are used to describe the subscripts of some orbital/spin degrees of freedom.

source

Hamiltonian.Prerequisites.Interfaces.expand — Method.

expand(fc::FockCoupling,pid::PID,fock::Fock) -> FCExpand
expand(fc::FockCoupling,pids::NTuple{R,PID},focks::NTuple{R,Fock}) where R -> FCExpand

Expand a Fock coupling with the given set of point ids and Fock degrees of freedom.

source

Hamiltonian.Prerequisites.Interfaces.expand — Method.

expand(subscripts::Subscripts,dimensions::NTuple{N,Int}) where N -> SbExpand

Expand a complete set of subscripts with a given set of variable ranges.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(subscripts::Subscripts,i::Int) -> Int
rank(::Type{<:Subscripts{T}},i::Int) where T -> Int

Get the number of the independent variables of a component of the complete subscript set.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(subscripts::Subscripts) -> Int
rank(::Type{S}) where S<:Subscripts -> Int

Get the total number of the independent variables of the complete subscript set.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(subscript::Subscript) -> Int
rank(::Type{<:Subscript{N}}) where N -> Int

Get the number of the independent variables that are used to describe the subscripts of some orbital/spin degrees of freedom.

source

Base.:* — Method.

*(fc1::FockCoupling,fc2::FockCoupling) -> FockCoupling

Get the multiplication between two Fock couplings.

source

Base.:* — Method.

*(sub1::Subscript,sub2::Subscript) -> Subscripts
*(subs::Subscripts,sub::Subscript) -> Subscripts
*(sub::Subscript,subs::Subscripts) -> Subscripts
*(subs1::Subscripts,subs2::Subscripts) -> Subscripts

Get the multiplication between subscripts or complete sets of subscripts.

source

Base.:== — Method.

==(sub1::Subscript,sub2::Subscript) -> Bool
isequal(sub1::Subscript,sub2::Subscript) -> Bool

Judge whether two subscripts are equivalent to each other.

source

Base.adjoint — Method.

adjoint(fid::FID) -> FID

Get the adjoint of a Fock id.

source

Base.repr — Method.

repr(fc::FockCoupling) -> String

Get the repr representation of a Fock coupling.

source

Base.show — Method.

show(io::IO,fc::FockCoupling)

Show a Fock coupling.

source

Base.show — Method.

show(io::IO,subscript::Subscript)

Show a subscript.

source

Base.union — Method.

union(::Type{P},::Type{FID})

Get the union type of PID and FID.

source