Fock package

Fock package

Fock degrees of freedom

FID and Fock

FIndex

Predefined Fock operators

Fock terms

FCID and FockCoupling

Predefined Fock couplings

Predefined Fock terms

Manual

QuantumLattices.Essentials.FockPackage.ANNIHILATION — Constant.

ANNIHILATION

Indicate that the nambu index is ANNIHILATION.

source

QuantumLattices.Essentials.FockPackage.CREATION — Constant.

CREATION

Indicate that the nambu index is CREATION.

source

QuantumLattices.Essentials.FockPackage.BOperator — Type.

BOperator(value::Number,id::ID{<:NTuple{N,OID}}) where N

Bosonic Fock operator.

source

QuantumLattices.Essentials.FockPackage.Coulomb — Type.

Coulomb{ST}(    id::Symbol,value::Number;
                neighbor::Int=1,
                couplings::Union{Tuple{<:Tuple{Vararg{Couplings}},<:Function},Coupling,Couplings,Nothing}=nothing,
                amplitude::Union{Function,Nothing}=nothing,
                modulate::Union{Function,Bool}=false,
                factor::Number=1
                ) where {ST}

Coulomb term.

Type alias for Term{Statistics,:Coulomb,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.FCID — Type.

FCID(;center=wildcard,atom=wildcard,orbital=wildcard,spin=wildcard,nambu=wildcard,obsub=wildcard,spsub=wildcard)

The id of a Fock coupling.

source

QuantumLattices.Essentials.FockPackage.FID — Type.

FID <: IID

The Fock id.

source

QuantumLattices.Essentials.FockPackage.FID — Method.

FID(;orbital::Int=1,spin::Int=1,nambu::Int=ANNIHILATION)

Create a Fock id.

source

QuantumLattices.Essentials.FockPackage.FIndex — Type.

FIndex{S} <: Index{PID{S},FID}

The Fock index.

source

QuantumLattices.Essentials.FockPackage.FOperator — Type.

FOperator(value::Number,id::ID{<:NTuple{N,OID}}) where N

Fermionic Fock operator.

source

QuantumLattices.Essentials.FockPackage.Fock — Type.

Fock <: Internal{FID}

The Fock interanl degrees of freedom.

source

QuantumLattices.Essentials.FockPackage.Fock — Method.

Fock(;atom::Int=1,norbital::Int=1,nspin::Int=2,nnambu::Int=2)

Construct a Fock degrees of freedom.

source

QuantumLattices.Essentials.FockPackage.FockCoupling — Type.

FockCoupling(value::Number,id::ID{<:NTuple{N,FCID}},obsubscripts::Subscripts,spsubscripts::Subscripts) where N
FockCoupling{N}(value::Number=1;
                centers::Union{NTuple{N,Int},Nothing}=nothing,
                atoms::Union{NTuple{N,Int},Nothing}=nothing,
                orbitals::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                spins::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                nambus::Union{NTuple{N,Int},Nothing}=nothing) where N

Fock coupling.

source

QuantumLattices.Essentials.FockPackage.FockOperator — Type.

FockOperator{N,V<:Number,I<:ID{<:NTuple{N,OID}}} <: Operator{N,V,I}

Abstract type for all Fock operators.

source

QuantumLattices.Essentials.FockPackage.Hopping — Type.

Hopping{ST}(id::Symbol,value::Number;
            neighbor::Int=1,
            couplings::Union{Tuple{<:Tuple{Vararg{Couplings}},<:Function},Coupling,Couplings,Nothing}=nothing,
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            factor::Number=1
            ) where {ST}

Hopping term.

Type alias for Term{Statistics,:Hopping,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.Hubbard — Type.

Hubbard{ST}(id::Symbol,value::Real;
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            factor::Number=1
            ) where {ST}

Hubbard term.

Type alias for Term{Statistics,:Hubbard,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.InterOrbitalInterSpin — Type.

InterOrbitalInterSpin{ST}(  id::Symbol,value::Real;
                            amplitude::Union{Function,Nothing}=nothing,
                            modulate::Union{Function,Bool}=false,
                            factor::Number=1
                            ) where {ST}

Interorbital-interspin term.

Type alias for Term{Statistics,:InterOrbitalInterSpin,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.InterOrbitalIntraSpin — Type.

InterOrbitalIntraSpin{ST}(  id::Symbol,value::Real;
                            amplitude::Union{Function,Nothing}=nothing,
                            modulate::Union{Function,Bool}=false,
                            factor::Number=1
                            ) where {ST}

Interorbital-intraspin term.

Type alias for Term{Statistics,:InterOrbitalIntraSpin,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.Onsite — Type.

Onsite{ST}( id::Symbol,value::Number;
            couplings::Union{Tuple{<:Tuple{Vararg{Couplings}},<:Function},Coupling,Couplings,Nothing}=nothing,
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            factor::Number=1
            ) where {ST}

Onsite term.

Type alias for Term{Statistics,:Onsite,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.PairHopping — Type.

PairHopping{ST}(    id::Symbol,value::Real;
                    amplitude::Union{Function,Nothing}=nothing,
                    modulate::Union{Function,Bool}=false,
                    factor::Number=1
                    ) where {ST}

Pair-hopping term.

Type alias for Term{Statistics,:PairHopping,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.Pairing — Type.

Pairing{ST}(id::Symbol,value::Number;
            neighbor::Int=0,
            couplings::Union{Tuple{<:Tuple{Vararg{Couplings}},<:Function},Coupling,Couplings,Nothing}=nothing,
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            factor::Number=1
            ) where {ST}

Pairing term.

Type alias for Term{Statistics,:Pairing,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.SpinFlip — Type.

SpinFlip{ST}(   id::Symbol,value::Real;
                amplitude::Union{Function,Nothing}=nothing,
                modulate::Union{Function,Bool}=false,
                factor::Number=1
                ) where {ST}

Spin-flip term.

Type alias for Term{Statistics,:SpinFlip,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.isnormalordered — Method.

isnormalordered(opt::FockOperator) -> Bool

Judge whether a FockOperator is normal ordered.

source

QuantumLattices.Essentials.FockPackage.nambufockindextotuple — Function.

nambufockindextotuple

Indicate that the filtered attributes are (:scope,:nambu,:site,:orbital,:spin) when converting a Fock index to tuple.

source

QuantumLattices.Essentials.FockPackage.usualfockindextotuple — Function.

usualfockindextotuple

Indicate that the filtered attributes are (:scope,:site,:orbital,:spin) when converting a Fock index to tuple.

source

QuantumLattices.Essentials.FockPackage.σʸ — Method.

σʸ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{<:NTuple{2,FCID}},FockCoupling{2,Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σˣ — Method.

σˣ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{<:NTuple{2,FCID}},FockCoupling{2,Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σᶻ — Method.

σᶻ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{<:NTuple{2,FCID}},FockCoupling{2,Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σ⁰ — Method.

σ⁰(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{<:NTuple{2,FCID}},FockCoupling{2,Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σ⁺ — Method.

σ⁺(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{<:NTuple{2,FCID}},FockCoupling{2,Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σ⁻ — Method.

σ⁻(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{<:NTuple{2,FCID}},FockCoupling{2,Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Base.:* — Method.

*(fc1::FockCoupling,fc2::FockCoupling) -> FockCoupling

Get the multiplication between two Fock couplings.

source

Base.adjoint — Method.

adjoint(fid::FID) -> FID

Get the adjoint of a Fock id.

source

Base.repr — Method.

repr(fc::FockCoupling) -> String

Get the repr representation of a Fock coupling.

source

Base.show — Method.

show(io::IO,fc::FockCoupling)

Show a Fock coupling.

source

Base.union — Method.

union(::Type{P},::Type{FID}) where P<:PID

Get the union type of PID and FID.

source

QuantumLattices.Essentials.DegreesOfFreedom.oidtype — Method.

oidtype(::Val{:Fock},B::Type{<:AbstractBond},::Type{Nothing})
oidtype(::Val{:Fock},B::Type{<:AbstractBond},::Type{<:Table})

Get the compatible Fock OID type with an AbstractBond type and a Table/Nothing type.

source

QuantumLattices.Essentials.DegreesOfFreedom.otype — Method.

otype(::Val{:Fock},O::Type{<:Term{'F'}},B::Type{<:AbstractBond},T::Type{<:Union{Nothing,Table}})
otype(::Val{:Fock},O::Type{<:Term{'B'}},B::Type{<:AbstractBond},T::Type{<:Union{Nothing,Table}})

Get the compatible Fock operator type with a Term type, an AbstractBond type and a Table/Nothing type.

source

QuantumLattices.Essentials.DegreesOfFreedom.twist — Method.

twist(id::OID{<:FIndex},vectors::AbstractVector{<:AbstractVector{Float}},values::AbstractVector{Float}) -> Complex{Float}

Get the twist phase corresponding to a Fock oid.

source

QuantumLattices.Essentials.Terms.statistics — Method.

statistics(opt::BOperator)
statistics(::Type{<:BOperator})

Get the statistics of BOperator.

source

QuantumLattices.Essentials.Terms.statistics — Method.

statistics(opt::FOperator) -> Char
statistics(::Type{<:FOperator}) -> Char

Get the statistics of FOperator.

source

QuantumLattices.Interfaces.:⊗ — Method.

⊗(fc1::FockCoupling{N},fc2::FockCoupling{N}) where N -> FockCoupling

Get the direct product between two Fock couplings.

source

QuantumLattices.Interfaces.expand — Method.

expand(fc::FockCoupling,pid::PID,fock::Fock,species::Union{Val{S},Nothing}=nothing) where S -> Union{FCExpand,Tuple{}}
expand(fc::FockCoupling,pids::NTuple{R,PID},focks::NTuple{R,Fock},species::Union{Val{S},Nothing}=nothing) where {R,S} -> Union{FCExpand,Tuple{}}

Expand a Fock coupling with the given set of point ids and Fock degrees of freedom.

source