Fock package

Fock package

Fock degrees of freedom

FID and Fock

FIndex

FCID and FockCoupling

Predefined Fock couplings

Manual

Hamiltonian.Essentials.FockPackage.ANNIHILATION — Constant.

ANNIHILATION

Indicate that the nambu index is ANNIHILATION.

source

Hamiltonian.Essentials.FockPackage.CREATION — Constant.

CREATION

Indicate that the nambu index is CREATION.

source

Hamiltonian.Essentials.FockPackage.FCID — Type.

FCID(;center=wildcard,atom=wildcard,orbital=wildcard,spin=wildcard,nambu=wildcard,obsub=wildcard,spsub=wildcard)

The id of a Fock coupling.

source

Hamiltonian.Essentials.FockPackage.FID — Type.

FID <: IID

The Fock id.

source

Hamiltonian.Essentials.FockPackage.FID — Method.

FID(;orbital::Int=1,spin::Int=1,nambu::Int=ANNIHILATION)

Create a Fock id.

source

Hamiltonian.Essentials.FockPackage.FIndex — Type.

FIndex{S} <: Index{PID{S},FID}

The Fock index.

source

Hamiltonian.Essentials.FockPackage.Fock — Type.

Fock <: Internal{FID}

The Fock interanl degrees of freedom.

source

Hamiltonian.Essentials.FockPackage.Fock — Method.

Fock(;atom::Int=1,norbital::Int=1,nspin::Int=2,nnambu::Int=2)

Construct a Fock degrees of freedom.

source

Hamiltonian.Essentials.FockPackage.FockCoupling — Type.

FockCoupling(value::Number,id::ID{N,I},obsubscripts::Subscripts,spsubscripts::Subscripts) where {N,I<:FCID}
FockCoupling{N}(value::Number=1;
                centers::Union{NTuple{N,Int},Nothing}=nothing,
                atoms::Union{NTuple{N,Int},Nothing}=nothing,
                orbitals::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                spins::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                nambus::Union{NTuple{N,Int},Nothing}=nothing) where N

Fock coupling.

source

Hamiltonian.Essentials.FockPackage.FockCouplings — Method.

FockCouplings(  ::Val{N},value::Number=1;
                centers::Union{NTuple{N,Int},Nothing}=nothing,
                atoms::Union{NTuple{N,Int},Nothing}=nothing,
                orbitals::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                spins::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                nambus::Union{NTuple{N,Int},Nothing}=nothing) where N -> Couplings

Construct an instance of Couplings that contains only one element of FockCoupling.

source

Hamiltonian.Essentials.FockPackage.nambufockindextotuple — Function.

nambufockindextotuple

Indicate that the filtered attributes are (:scope,:nambu,:site,:orbital,:spin) when converting a Fock index to tuple.

source

Hamiltonian.Essentials.FockPackage.usualfockindextotuple — Function.

usualfockindextotuple

Indicate that the filtered attributes are (:scope,:site,:orbital,:spin) when converting a Fock index to tuple.

source

Hamiltonian.Essentials.FockPackage.σʸ — Method.

σʸ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σˣ — Method.

σˣ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σᶻ — Method.

σᶻ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σ⁰ — Method.

σ⁰(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σ⁺ — Method.

σ⁺(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Essentials.FockPackage.σ⁻ — Method.

σ⁻(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID{2,FCID},FockCoupling{2,Int,ID{2,FCID}}}

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Hamiltonian.Prerequisites.Interfaces.:⊗ — Method.

⊗(fc1::FockCoupling{N},fc2::FockCoupling{N}) where N -> FockCoupling

Get the direct product between two Fock couplings.

source

Hamiltonian.Prerequisites.Interfaces.expand — Method.

expand(fc::FockCoupling,pid::PID,fock::Fock) -> FCExpand
expand(fc::FockCoupling,pids::NTuple{R,PID},focks::NTuple{R,Fock}) where R -> FCExpand

Expand a Fock coupling with the given set of point ids and Fock degrees of freedom.

source

Base.:* — Method.

*(fc1::FockCoupling,fc2::FockCoupling) -> FockCoupling

Get the multiplication between two Fock couplings.

source

Base.adjoint — Method.

adjoint(fid::FID) -> FID

Get the adjoint of a Fock id.

source

Base.repr — Method.

repr(fc::FockCoupling) -> String

Get the repr representation of a Fock coupling.

source

Base.show — Method.

show(io::IO,fc::FockCoupling)

Show a Fock coupling.

source

Base.union — Method.

union(::Type{P},::Type{FID}) where P<:PID

Get the union type of PID and FID.

source