Fock package

FID and Fock

FockCoupling

Fock terms

Manual

QuantumLattices.Essentials.FockPackage.ANNIHILATION — Constant

ANNIHILATION

Indicate that the nambu index is ANNIHILATION.

QuantumLattices.Essentials.FockPackage.CREATION — Constant

CREATION

Indicate that the nambu index is CREATION.

QuantumLattices.Essentials.FockPackage.MAJORANA — Constant

MAJORANA

Indicate that the nambu index is MAJORANA.

QuantumLattices.Essentials.FockPackage.bdefaultlatex — Constant

bdefaultlatex

The default LaTeX format for a bosonic oid.

QuantumLattices.Essentials.FockPackage.fdefaultlatex — Constant

fdefaultlatex

The default LaTeX format for a fermionic oid.

QuantumLattices.Essentials.FockPackage.Coulomb — Type

Coulomb(id::Symbol, value, bondkind::Int=1;
    couplings::Union{Function, Couplings, Nothing}=nothing,
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false
    )

Coulomb term.

Type alias for Term{:Coulomb, 4, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.FockPackage.FID — Type

FID{T, O<:Union{Int, Symbol}, S<:Union{Int, Symbol}, N<:Union{Int, Symbol}} <: SimpleIID

The Fock id.

QuantumLattices.Essentials.FockPackage.FID — Method

FID{T}(; orbital::Union{Int, Symbol}=1, spin::Union{Int, Symbol}=1, nambu::Union{Int, Symbol}=ANNIHILATION) where T

Create a Fock id.

QuantumLattices.Essentials.FockPackage.Fock — Type

Fock{T} <: SimpleInternal{FID{T, Int, Int, Int}}

The Fock interanl degrees of freedom.

QuantumLattices.Essentials.FockPackage.Fock — Method

Fock{T}(; norbital::Int=1, nspin::Int=2, nnambu::Int=2) where T

Construct a Fock degrees of freedom.

QuantumLattices.Essentials.FockPackage.FockCoupling — Type

FockCoupling(value::Number,
    orbitals::Subscript{<:NTuple{N, Union{Int, Symbol}}},
    spins::Subscript{<:NTuple{N, Union{Int, Symbol}}},
    nambus::Union{NTuple{N, Int}, NTuple{N, Symbol}}
    ) where N

Fock coupling.

Type alias for Coupling{V, I<:ID{FID}, C<:IIDConstrain, CI<:ConstrainID}.

QuantumLattices.Essentials.FockPackage.FockCoupling — Method

FockCoupling{N}(value::Number=1;
    orbitals::Union{NTuple{N, Int}, Subscript}=Subscript(N),
    spins::Union{NTuple{N, Int}, Subscript}=Subscript(N),
    nambus::Union{NTuple{N, Int}, NTuple{N, Symbol}}=ntuple(i->wildcard, Val(N))
    ) where N

Construct a Fock coupling.

QuantumLattices.Essentials.FockPackage.Hopping — Type

Hopping(id::Symbol, value, bondkind::Int=1;
    couplings::Union{Function, Couplings, Nothing}=nothing,
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Hopping term.

Type alias for Term{:Hopping, 2, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.FockPackage.Hubbard — Type

Hubbard(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Hubbard term.

Type alias for Term{:Hubbard, 4, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.FockPackage.InterOrbitalInterSpin — Type

InterOrbitalInterSpin(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Interorbital-interspin term.

Type alias for Term{:InterOrbitalInterSpin, 4, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.FockPackage.InterOrbitalIntraSpin — Type

InterOrbitalIntraSpin(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Interorbital-intraspin term.

Type alias for Term{:InterOrbitalIntraSpin, 4, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.FockPackage.Onsite — Type

Onsite(id::Symbol, value;
    couplings::Union{Function, Couplings, Nothing}=nothing,
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Onsite term.

Type alias for Term{:Onsite, 2, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.FockPackage.PairHopping — Type

PairHopping(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Pair-hopping term.

Type alias for Term{:PairHopping, 4, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.FockPackage.Pairing — Type

Pairing(id::Symbol, value, bondkind::Int=0;
    couplings::Union{Function, Couplings, Nothing}=nothing,
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Pairing term.

Type alias for Term{:Pairing, 2, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.FockPackage.SpinFlip — Type

SpinFlip(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Spin-flip term.

Type alias for Term{:SpinFlip, 4, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.FockPackage.@fc_str — Macro

fc"..." -> FockCoupling

Construct a FockCoupling from a literal string.

QuantumLattices.Essentials.FockPackage.@σʸ_str — Macro

σʸ"sp" -> Couplings
σʸ"ob" -> Couplings
σʸ"ph" -> Couplings

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σˣ_str — Macro

σˣ"sp" -> Couplings
σˣ"ob" -> Couplings
σˣ"ph" -> Couplings

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σᶻ_str — Macro

σᶻ"sp" -> Couplings
σᶻ"ob" -> Couplings
σᶻ"ph" -> Couplings

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σ⁰_str — Macro

σ⁰"sp" -> Couplings
σ⁰"ob" -> Couplings
σ⁰"ph" -> Couplings

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σ⁺_str — Macro

σ⁺"sp" -> Couplings
σ⁺"ob" -> Couplings
σ⁺"ph" -> Couplings

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σ⁻_str — Macro

σ⁻"sp" -> Couplings
σ⁻"ob" -> Couplings
σ⁻"ph" -> Couplings

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

Base.:* — Method

*(f1::Operator{<:Number, <:ID{AbstractCompositeOID{<:Index{<:AbstractPID, <:FID{:f}}}}}, f2::Operator{<:Number, <:ID{AbstractCompositeOID{<:Index{<:AbstractPID, <:FID{:f}}}}}) -> Union{Nothing, Operator}

Get the multiplication of two fermionic Fock operators.

Base.angle — Method

angle(id::OID{<:Index{<:AbstractPID, <:FID}}, vectors::AbstractVector{<:AbstractVector{<:Number}}, values::AbstractVector{<:Number}) -> Complex{<:Number}

Get the twist phase corresponding to a Fock oid.

QuantumLattices.Essentials.DegreesOfFreedom.script — Method

script(::Val{:site}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> Int
script(::Val{:orbital}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> Int
script(::Val{:spinint}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> Int
script(::Val{:spinsym}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> String
script(::Val{:nambu}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> String

Get the required script of a Fock index.

QuantumLattices.Essentials.FockPackage.isnormalordered — Method

isnormalordered(opt::Operator{<:Number, <:ID{AbstractCompositeOID{<:Index{<:AbstractPID, <:FID}}}}) -> Bool

Judge whether an operator is normal ordered.

QuantumLattices.Essentials.FockPackage.nambufockindextotuple — Function

nambufockindextotuple

Indicate that the choosed fields are (:nambu, :site, :orbital, :spin) when converting a Fock index to tuple.

QuantumLattices.Essentials.FockPackage.usualfockindextotuple — Function

usualfockindextotuple

Indicate that the choosed fields are (:site, :orbital, :spin) when converting a Fock index to tuple.

QuantumLattices.Interfaces.:⊗ — Method

⊗(fc₁::FockCoupling, fc₂::FockCoupling) -> FockCoupling

Get the direct product between two Fock couplings.

QuantumLattices.Interfaces.:⋅ — Method

⋅(fc₁::FockCoupling, fc₂::FockCoupling) -> FockCoupling

Get the dot product of two rank-2 Fock couplings.

A rank-2 FockCoupling can be considered as a matrix acting on the orbital, spin and nambu spaces. Therefore, the dot product here is defined as the multiplication between such matrices.

QuantumLattices.Interfaces.permute — Method

permute(id₁::OID{<:Index{<:AbstractPID, <:FID{:b}}}, id₂::OID{<:Index{<:AbstractPID, <:FID{:b}}}) -> Tuple{Vararg{Operator}}

Permute two bosonic oids and get the result.

QuantumLattices.Interfaces.permute — Method

permute(id₁::OID{<:Index{<:AbstractPID, <:FID{:f}}}, id₂::OID{<:Index{<:AbstractPID, <:FID{:f}}}) -> Tuple{Vararg{Operator}}

Permute two fermionic oids and get the result.