Fock package

Fock package

Fock degrees of freedom

FID and Fock

FIndex

Predefined Fock operators

Fock terms

FCID and FockCoupling

Predefined Fock couplings

Predefined Fock terms

Manual

QuantumLattices.Essentials.FockPackage.ANNIHILATION — Constant.

ANNIHILATION

Indicate that the nambu index is ANNIHILATION.

source

QuantumLattices.Essentials.FockPackage.BFockTerm — Constant.

BFockTerm

Bosonic Fock term types.

source

QuantumLattices.Essentials.FockPackage.CREATION — Constant.

CREATION

Indicate that the nambu index is CREATION.

source

QuantumLattices.Essentials.FockPackage.FFockTerm — Constant.

FFockTerm

Fermionic Fock term types.

source

QuantumLattices.Essentials.FockPackage.MAJORANA — Constant.

MAJORANA

Indicate that the nambu index is MAJORANA.

source

QuantumLattices.Essentials.FockPackage.boptdefaultlatex — Constant.

boptdefaultlatex

The default LaTeX format of the oids of a bosonic operator.

source

QuantumLattices.Essentials.FockPackage.foptdefaultlatex — Constant.

foptdefaultlatex

The default LaTeX format of the oids of a fermionic operator.

source

QuantumLattices.Essentials.FockPackage.BOperator — Type.

BOperator(value,id::ID{<:Tuple{Vararg{OID}}}=ID())

Bosonic Fock operator.

source

QuantumLattices.Essentials.FockPackage.Coulomb — Type.

Coulomb{ST}(    id::Symbol,value::Any,bondkind::Int=1;
                couplings::Union{Function,Coupling,Couplings,Nothing}=nothing,
                amplitude::Union{Function,Nothing}=nothing,
                modulate::Union{Function,Bool}=false,
                ) where {ST}

Coulomb term.

Type alias for Term{statistics,:Coulomb,4,id,V,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.FCID — Type.

FCID(;center=wildcard,atom=wildcard,orbital=wildcard,spin=wildcard,nambu=wildcard,obsub=wildcard,spsub=wildcard)

The id of a Fock coupling.

source

QuantumLattices.Essentials.FockPackage.FID — Type.

FID <: IID

The Fock id.

source

QuantumLattices.Essentials.FockPackage.FID — Method.

FID(;orbital::Int=1,spin::Int=1,nambu::Int=ANNIHILATION)

Create a Fock id.

source

QuantumLattices.Essentials.FockPackage.FIndex — Type.

FIndex{S} <: Index{PID{S},FID}

The Fock index.

source

QuantumLattices.Essentials.FockPackage.FOperator — Type.

FOperator(value,id::ID{<:Tuple{Vararg{OID}}}=ID())

Fermionic Fock operator.

source

QuantumLattices.Essentials.FockPackage.Fock — Type.

Fock <: Internal{FID}

The Fock interanl degrees of freedom.

source

QuantumLattices.Essentials.FockPackage.Fock — Method.

Fock(;atom::Int=1,norbital::Int=1,nspin::Int=2,nnambu::Int=2)

Construct a Fock degrees of freedom.

source

QuantumLattices.Essentials.FockPackage.FockCoupling — Type.

FockCoupling(value,id::ID{<:Tuple{Vararg{FCID}}},obsubscripts::Subscripts,spsubscripts::Subscripts)
FockCoupling{N}(value1;
                centers::Union{NTuple{N,Int},Nothing}=nothing,
                atoms::Union{NTuple{N,Int},Nothing}=nothing,
                orbitals::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                spins::Union{NTuple{N,Int},Subscript,Nothing}=nothing,
                nambus::Union{NTuple{N,Int},Nothing}=nothing) where N

Fock coupling.

source

QuantumLattices.Essentials.FockPackage.FockOperator — Type.

FockOperator{V,I<:ID} <: Operator{V,I}

Abstract type for all Fock operators.

source

QuantumLattices.Essentials.FockPackage.Hopping — Type.

Hopping{ST}(id::Symbol,value::Any,bondkind::Int=1;
            couplings::Union{Function,Coupling,Couplings,Nothing}=nothing,
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            ) where {ST}

Hopping term.

Type alias for Term{statistics,:Hopping,2,id,V,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.Hubbard — Type.

Hubbard{ST}(id::Symbol,value::Any;
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            ) where {ST}

Hubbard term.

Type alias for Term{statistics,:Hubbard,4,id,V,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.InterOrbitalInterSpin — Type.

InterOrbitalInterSpin{ST}(  id::Symbol,value::Any;
                            amplitude::Union{Function,Nothing}=nothing,
                            modulate::Union{Function,Bool}=false,
                            ) where {ST}

Interorbital-interspin term.

Type alias for Term{statistics,:InterOrbitalInterSpin,4,id,V,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.InterOrbitalIntraSpin — Type.

InterOrbitalIntraSpin{ST}(  id::Symbol,value::Any;
                            amplitude::Union{Function,Nothing}=nothing,
                            modulate::Union{Function,Bool}=false,
                            ) where {ST}

Interorbital-intraspin term.

Type alias for Term{statistics,:InterOrbitalIntraSpin,4,id,V,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.Onsite — Type.

Onsite{ST}( id::Symbol,value::Any;
            couplings::Union{Function,Coupling,Couplings,Nothing}=nothing,
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            ) where {ST}

Onsite term.

Type alias for Term{statistics,:Onsite,2,id,V,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.PairHopping — Type.

PairHopping{ST}(    id::Symbol,value::Any;
                    amplitude::Union{Function,Nothing}=nothing,
                    modulate::Union{Function,Bool}=false,
                    ) where {ST}

Pair-hopping term.

Type alias for Term{statistics,:PairHopping,4,id,V,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.Pairing — Type.

Pairing{ST}(id::Symbol,value::Any,bondkind::Int=0;
            couplings::Union{Function,Coupling,Couplings,Nothing}=nothing,
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            ) where {ST}

Pairing term.

Type alias for Term{statistics,:Pairing,2,id,V,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.SpinFlip — Type.

SpinFlip{ST}(   id::Symbol,value::Any;
                amplitude::Union{Function,Nothing}=nothing,
                modulate::Union{Function,Bool}=false,
                ) where {ST}

Spin-flip term.

Type alias for Term{statistics,:SpinFlip,4,id,V,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.FockPackage.@fc_str — Macro.

fc"..." -> FockCoupling

Construct a FockCoupling from a literal string.

source

QuantumLattices.Essentials.FockPackage.@σʸ_str — Macro.

σʸ"sp"/σʸ"sp@(c₁,c₂)" -> Couplings
σʸ"ob"/σʸ"ob@(c₁,c₂)" -> Couplings
σʸ"sl"/σʸ"sl@(c₁,c₂)" -> Couplings
σʸ"ph"/σʸ"ph@(c₁,c₂)" -> Couplings

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.@σˣ_str — Macro.

σˣ"sp"/σˣ"sp@(c₁,c₂)" -> Couplings
σˣ"ob"/σˣ"ob@(c₁,c₂)" -> Couplings
σˣ"sl"/σˣ"sl@(c₁,c₂)" -> Couplings
σˣ"ph"/σˣ"ph@(c₁,c₂)" -> Couplings

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.@σᶻ_str — Macro.

σᶻ"sp"/σᶻ"sp@(c₁,c₂)" -> Couplings
σᶻ"ob"/σᶻ"ob@(c₁,c₂)" -> Couplings
σᶻ"sl"/σᶻ"sl@(c₁,c₂)" -> Couplings
σᶻ"ph"/σᶻ"ph@(c₁,c₂)" -> Couplings

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.@σ⁰_str — Macro.

σ⁰"sp"/σ⁰"sp@(c₁,c₂)" -> Couplings
σ⁰"ob"/σ⁰"ob@(c₁,c₂)" -> Couplings
σ⁰"sl"/σ⁰"sl@(c₁,c₂)" -> Couplings
σ⁰"ph"/σ⁰"ph@(c₁,c₂)" -> Couplings

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.@σ⁺_str — Macro.

σ⁺"sp"/σ⁺"sp@(c₁,c₂)" -> Couplings
σ⁺"ob"/σ⁺"ob@(c₁,c₂)" -> Couplings
σ⁺"sl"/σ⁺"sl@(c₁,c₂)" -> Couplings
σ⁺"ph"/σ⁺"ph@(c₁,c₂)" -> Couplings

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.@σ⁻_str — Macro.

σ⁻"sp"/σ⁻"sp@(c₁,c₂)" -> Couplings
σ⁻"ob"/σ⁻"ob@(c₁,c₂)" -> Couplings
σ⁻"sl"/σ⁻"sl@(c₁,c₂)" -> Couplings
σ⁻"ph"/σ⁻"ph@(c₁,c₂)" -> Couplings

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.isnormalordered — Method.

isnormalordered(opt::FockOperator) -> Bool

Judge whether a FockOperator is normal ordered.

source

QuantumLattices.Essentials.FockPackage.nambufockindextotuple — Function.

nambufockindextotuple

Indicate that the filtered attributes are (:scope,:nambu,:site,:orbital,:spin) when converting a Fock index to tuple.

source

QuantumLattices.Essentials.FockPackage.usualfockindextotuple — Function.

usualfockindextotuple

Indicate that the filtered attributes are (:scope,:site,:orbital,:spin) when converting a Fock index to tuple.

source

QuantumLattices.Essentials.FockPackage.σʸ — Method.

σʸ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID,FockCoupling{Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σˣ — Method.

σˣ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID,FockCoupling{Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σᶻ — Method.

σᶻ(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID,FockCoupling{Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σ⁰ — Method.

σ⁰(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID,FockCoupling{Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σ⁺ — Method.

σ⁺(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID,FockCoupling{Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

QuantumLattices.Essentials.FockPackage.σ⁻ — Method.

σ⁻(mode::String;centers::Union{NTuple{2,Int},Nothing}=nothing) -> Couplings{ID,FockCoupling{Int,ID{<:NTuple{2,FCID}}}}

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

source

Base.:* — Method.

*(f1::FOperator,f2::FOperator) -> Union{Nothing,FOperator}

Get the multiplication of two fermionic Fock operators.

source

Base.:* — Method.

*(fc1::FockCoupling,fc2::FockCoupling) -> FockCoupling

Get the multiplication between two Fock couplings.

source

Base.adjoint — Method.

adjoint(fid::FID) -> FID

Get the adjoint of a Fock id.

source

Base.angle — Method.

angle(id::OID{<:FIndex},vectors::AbstractVector{<:AbstractVector{Float}},values::AbstractVector{Float}) -> Complex{Float}

Get the twist phase corresponding to a Fock oid.

source

Base.repr — Method.

repr(fc::FockCoupling) -> String

Get the repr representation of a Fock coupling.

source

Base.show — Method.

show(io::IO,fc::FockCoupling)

Show a Fock coupling.

source

Base.union — Method.

union(::Type{P},::Type{FID}) where P<:PID

Get the union type of PID and FID.

source

QuantumLattices.Essentials.DegreesOfFreedom.otype — Method.

otype(T::Type{<:FFockTerm},I::Type{<:OID})
otype(T::Type{<:BFockTerm},I::Type{<:OID})

Get the operator type of a Fock term.

source

QuantumLattices.Essentials.DegreesOfFreedom.script — Method.

script(oid::OID{<:FIndex},::Val{:site}) -> Int
script(oid::OID{<:FIndex},::Val{:orbital}) -> Int
script(oid::OID{<:FIndex},::Val{:spinint}) -> Int
script(oid::OID{<:FIndex},::Val{:spinsym}) -> String
script(oid::OID{<:FIndex},::Val{:nambu}) -> String

Get the required script of an Fock oid.

source

QuantumLattices.Essentials.Terms.statistics — Method.

statistics(opt::BOperator)
statistics(::Type{<:BOperator})

Get the statistics of BOperator.

source

QuantumLattices.Essentials.Terms.statistics — Method.

statistics(opt::FOperator) -> Char
statistics(::Type{<:FOperator}) -> Char

Get the statistics of FOperator.

source

QuantumLattices.Interfaces.:⊗ — Method.

⊗(fc1::FockCoupling,fc2::FockCoupling) -> FockCoupling

Get the direct product between two Fock couplings.

source

QuantumLattices.Interfaces.:⋅ — Method.

⋅(fc1::FockCoupling,fc2::FockCoupling) -> FockCoupling

Get the dot product of two rank-2 Fock couplings.

A rank-2 FockCoupling can be considered as a matrix acting on the sublattice, orbital, spin and nambu spaces. The dot product here is defined as the multiplication between such matrices.

source

QuantumLattices.Interfaces.expand — Method.

expand(fc::FockCoupling,pid::PID,fock::Fock,kind::Union{Val{K},Nothing}=nothing) where K -> Union{FCExpand,Tuple{}}
expand(fc::FockCoupling,pids::NTuple{R,PID},focks::NTuple{R,Fock},kind::Union{Val{K},Nothing}=nothing) where {R,K} -> Union{FCExpand,Tuple{}}

Expand a Fock coupling with the given set of point ids and Fock degrees of freedom.

source

QuantumLattices.Interfaces.permute — Function.

permute(::Type{<:FOperator},id1::OID{<:FIndex},id2::OID{<:FIndex},::Any=nothing) -> Tuple{Vararg{FOperator}}

Permute two fermionic oid and get the result.

source

QuantumLattices.Interfaces.permute — Function.

permute(::Type{<:BOperator},id1::OID{<:FIndex},id2::OID{<:FIndex},::Any=nothing) -> Tuple{Vararg{BOperator}}

Permute two bosonic oid and get the result.

source

QuantumLattices.Mathematics.AlgebraOverFields.rawelement — Method.

rawelement(::Type{<:BOperator})

Get the raw name of a type of BOperator.

source

QuantumLattices.Mathematics.AlgebraOverFields.rawelement — Method.

rawelement(::Type{<:FOperator})

Get the raw name of a type of FOperator.

source

QuantumLattices.Mathematics.AlgebraOverFields.rawelement — Method.

rawelement(::Type{<:FockCoupling})

Get the raw name of a type of FockCoupling.

source