Fock package

FID and Fock

FockCoupling

Fock terms

Manual

QuantumLattices.Essentials.FockPackage.ANNIHILATION — Constant

ANNIHILATION

Indicate that the nambu index is ANNIHILATION.

QuantumLattices.Essentials.FockPackage.CREATION — Constant

CREATION

Indicate that the nambu index is CREATION.

QuantumLattices.Essentials.FockPackage.MAJORANA — Constant

MAJORANA

Indicate that the nambu index is MAJORANA.

QuantumLattices.Essentials.FockPackage.bdefaultlatex — Constant

bdefaultlatex

The default LaTeX format for a bosonic oid.

QuantumLattices.Essentials.FockPackage.fdefaultlatex — Constant

fdefaultlatex

The default LaTeX format for a fermionic oid.

QuantumLattices.Essentials.FockPackage.Coulomb — Type

Coulomb(id::Symbol, value::Any, bondkind::Int=1;
    couplings::Union{Function, Coupling, Couplings, Nothing}=nothing,
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false
    )

Coulomb term.

Type alias for Term{:Coulomb, 4, id, V, Int, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.FockPackage.FCID — Type

FCID{A<:Tuple, N<:Tuple} <: SimpleID

The id of the atoms and nambus part of a Fock coupling.

QuantumLattices.Essentials.FockPackage.FID — Type

FID <: IID

The Fock id.

QuantumLattices.Essentials.FockPackage.FID — Method

FID{ST}(; orbital::Int=1, spin::Int=1, nambu::Int=ANNIHILATION) where ST

Create a Fock id.

QuantumLattices.Essentials.FockPackage.Fock — Type

Fock{ST} <: Internal{FID{ST}}

The Fock interanl degrees of freedom.

QuantumLattices.Essentials.FockPackage.Fock — Method

Fock{ST}(; atom::Int=1, norbital::Int=1, nspin::Int=2, nnambu::Int=2) where ST

Construct a Fock degrees of freedom.

QuantumLattices.Essentials.FockPackage.FockCoupling — Type

FockCoupling{V, A<:Tuple, N<:Tuple, O<:Subscripts, S<:Subscripts, I<:Tuple{FCID, SubID, SubID}} <: Coupling{V, I}

Fock coupling.

QuantumLattices.Essentials.FockPackage.FockCoupling — Method

FockCoupling{N}(value::Number=1;
    atoms::Union{NTuple{N, Int}, Nothing}=nothing,
    nambus::Union{NTuple{N, Int}, Nothing}=nothing,
    orbitals::Union{NTuple{N, Int}, Subscripts, Nothing}=nothing,
    spins::Union{NTuple{N, Int}, Subscripts, Nothing}=nothing
    ) where N

QuantumLattices.Essentials.FockPackage.Hopping — Type

Hopping(id::Symbol, value::Any, bondkind::Int=1;
    couplings::Union{Function, Coupling, Couplings, Nothing}=nothing,
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Hopping term.

Type alias for Term{:Hopping, 2, id, V, Int, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.FockPackage.Hubbard — Type

Hubbard(id::Symbol, value::Any; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Hubbard term.

Type alias for Term{:Hubbard, 4, id, V, Int, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.FockPackage.InterOrbitalInterSpin — Type

InterOrbitalInterSpin(id::Symbol, value::Any; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Interorbital-interspin term.

Type alias for Term{:InterOrbitalInterSpin, 4, id, V, Int, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.FockPackage.InterOrbitalIntraSpin — Type

InterOrbitalIntraSpin(id::Symbol, value::Any; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Interorbital-intraspin term.

Type alias for Term{:InterOrbitalIntraSpin, 4, id, V, Int, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.FockPackage.Onsite — Type

Onsite(id::Symbol, value::Any;
    couplings::Union{Function, Coupling, Couplings, Nothing}=nothing,
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Onsite term.

Type alias for Term{:Onsite, 2, id, V, Int, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.FockPackage.PairHopping — Type

PairHopping(id::Symbol, value::Any; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Pair-hopping term.

Type alias for Term{:PairHopping, 4, id, V, Int, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.FockPackage.Pairing — Type

Pairing(id::Symbol, value::Any, bondkind::Int=0;
    couplings::Union{Function, Coupling, Couplings, Nothing}=nothing,
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Pairing term.

Type alias for Term{:Pairing, 2, id, V, Int, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.FockPackage.SpinFlip — Type

SpinFlip(id::Symbol, value::Any; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Spin-flip term.

Type alias for Term{:SpinFlip, 4, id, V, Int, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.FockPackage.@fc_str — Macro

fc"..." -> FockCoupling

Construct a FockCoupling from a literal string.

QuantumLattices.Essentials.FockPackage.@σʸ_str — Macro

σʸ"sp" -> Couplings
σʸ"ob" -> Couplings
σʸ"sl" -> Couplings
σʸ"ph" -> Couplings

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σˣ_str — Macro

σˣ"sp" -> Couplings
σˣ"ob" -> Couplings
σˣ"sl" -> Couplings
σˣ"ph" -> Couplings

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σᶻ_str — Macro

σᶻ"sp" -> Couplings
σᶻ"ob" -> Couplings
σᶻ"sl" -> Couplings
σᶻ"ph" -> Couplings

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σ⁰_str — Macro

σ⁰"sp" -> Couplings
σ⁰"ob" -> Couplings
σ⁰"sl" -> Couplings
σ⁰"ph" -> Couplings

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σ⁺_str — Macro

σ⁺"sp" -> Couplings
σ⁺"ob" -> Couplings
σ⁺"sl" -> Couplings
σ⁺"ph" -> Couplings

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.@σ⁻_str — Macro

σ⁻"sp" -> Couplings
σ⁻"ob" -> Couplings
σ⁻"sl" -> Couplings
σ⁻"ph" -> Couplings

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

Base.:* — Method

*(fc₁::FockCoupling, fc₂::FockCoupling) -> FockCoupling

Get the multiplication between two Fock couplings.

Base.:* — Method

*(f1::Operator{<:Number, <:ID{AbstractCompositeOID{<:Index{<:PID, FID{:f}}}}}, f2::Operator{<:Number, <:ID{AbstractCompositeOID{<:Index{<:PID, FID{:f}}}}}) -> Union{Nothing, Operator}

Get the multiplication of two fermionic Fock operators.

Base.angle — Method

angle(id::OID{<:Index{<:PID, <:FID}}, vectors::AbstractVector{<:AbstractVector{<:Number}}, values::AbstractVector{<:Number}) -> Complex{<:Number}

Get the twist phase corresponding to a Fock oid.

Base.repr — Method

repr(fc::FockCoupling) -> String

Get the repr representation of a Fock coupling.

Base.show — Method

show(io::IO, fc::FockCoupling)

Show a Fock coupling.

QuantumLattices.Essentials.DegreesOfFreedom.script — Method

script(::Val{:site}, index::Index{<:PID, <:FID}; kwargs...) -> Int
script(::Val{:orbital}, index::Index{<:PID, <:FID}; kwargs...) -> Int
script(::Val{:spinint}, index::Index{<:PID, <:FID}; kwargs...) -> Int
script(::Val{:spinsym}, index::Index{<:PID, <:FID}; kwargs...) -> String
script(::Val{:nambu}, index::Index{<:PID, <:FID}; kwargs...) -> String

Get the required script of a Fock index.

QuantumLattices.Essentials.FockPackage.isnormalordered — Method

isnormalordered(opt::Operator{<:Number, <:ID{AbstractCompositeOID{<:Index{<:PID, <:FID}}}}) -> Bool

Judge whether an operator is normal ordered.

QuantumLattices.Essentials.FockPackage.nambufockindextotuple — Function

nambufockindextotuple

Indicate that the choosed fields are (:scope, :nambu, :site, :orbital, :spin) when converting a Fock index to tuple.

QuantumLattices.Essentials.FockPackage.usualfockindextotuple — Function

usualfockindextotuple

Indicate that the choosed fields are (:scope, :site, :orbital, :spin) when converting a Fock index to tuple.

QuantumLattices.Essentials.FockPackage.σʸ — Method

σʸ(mode::String) -> Couplings

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.σˣ — Method

σˣ(mode::String) -> Couplings

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.σᶻ — Method

σᶻ(mode::String) -> Couplings

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.σ⁰ — Method

σ⁰(mode::String) -> Couplings

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.σ⁺ — Method

σ⁺(mode::String) -> Couplings

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Essentials.FockPackage.σ⁻ — Method

σ⁻(mode::String) -> Couplings

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob"), sublattices("sl") or particle-holes("ph").

QuantumLattices.Interfaces.:⊗ — Method

⊗(fc₁::FockCoupling, fc₂::FockCoupling) -> FockCoupling

Get the direct product between two Fock couplings.

QuantumLattices.Interfaces.:⋅ — Method

⋅(fc₁::FockCoupling, fc₂::FockCoupling) -> FockCoupling

Get the dot product of two rank-2 Fock couplings.

A rank-2 FockCoupling can be considered as a matrix acting on the sublattice, orbital, spin and nambu spaces. Therefore, the dot product here is defined as the multiplication between such matrices.

QuantumLattices.Interfaces.expand — Method

expand(fc::FockCoupling, points::NTuple{R, Point}, focks::NTuple{R, Fock}, info::Val) where R -> Union{FCExpand, Tuple{}}

Expand a Fock coupling with the given set of points and Fock degrees of freedom.

QuantumLattices.Interfaces.permute — Method

permute(id₁::OID{<:Index{<:PID, FID{:b}}}, id₂::OID{<:Index{<:PID, FID{:b}}}) -> Tuple{Vararg{Operator}}

Permute two bosonic oid and get the result.

QuantumLattices.Interfaces.permute — Method

permute(id₁::OID{<:Index{<:PID, FID{:f}}}, id₂::OID{<:Index{<:PID, FID{:f}}}) -> Tuple{Vararg{Operator}}

Permute two fermionic oid and get the result.