Frameworks

Parameters

Entry

Engine

Generator

SimplifiedGenerator

Action

Assignment

Algorithm

Manual

QuantumLattices.Essentials.Frameworks.AbstractGenerator — Type

AbstractGenerator{E<:Entry, T<:Union{Table, Nothing}, B<:Boundary} <: Engine

Abstract generator for quantum operators.

By protocol, a concrete generator should have the following predefined contents:

operators::E: the entry for the generated quantum operators

table::T: the index-sequence table if it is not nothing

boundary::B: boundary twist for the generated operators

QuantumLattices.Essentials.Frameworks.Action — Type

Action <: Transformation

Abstract type for all actions.

QuantumLattices.Essentials.Frameworks.Algorithm — Type

Algorithm{E<:Engine, P<:Parameters, M<:Function, S<:NamedContainer{Assignment}}

An algorithm associated with an engine.

QuantumLattices.Essentials.Frameworks.Algorithm — Method

Algorithm(name::String, engine::Engine;
    din::String=".",
    dout::String=".",
    parameters::Union{Parameters, Nothing}=nothing,
    map::Function=identity,
    assignments::Tuple{Vararg{Assignment}}=(),
    kwargs...
    )

Construct an algorithm.

QuantumLattices.Essentials.Frameworks.Assignment — Type

Assignment{A<:Action, P<:Parameters, M<:Function, D<:Tuple{Vararg{Symbol}}, R<:Any, I}

An assignment associated with an action.

QuantumLattices.Essentials.Frameworks.Assignment — Method

Assignment(id::Symbol, action::Action, parameters::Parameters;
    map::Function=identity,
    dependences::Tuple{Vararg{Symbol}}=(),
    data::Any=nothing,
    savedata::Bool=true,
    virgin::Bool=true,
    kwargs...
    )

Construct an assignment.

QuantumLattices.Essentials.Frameworks.Engine — Type

Engine

Abstract type for all engines.

QuantumLattices.Essentials.Frameworks.Entry — Type

Entry{C, A<:NamedTuple, B<:NamedTuple}

An entry of quantum operators related to (part of) a quantum lattice system.

QuantumLattices.Essentials.Frameworks.Entry — Method

Entry(terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert;
    half::Bool=false,
    table::Union{Nothing, Table}=nothing
    )

Construct an entry of operators based on the input terms, bonds and Hilbert space.

QuantumLattices.Essentials.Frameworks.Generator — Type

Generator{E<:Entry, TS<:NamedContainer{Term}, BS<:Bonds, H<:Hilbert, T<:Union{Table, Nothing}, B<:Boundary} <: AbstractGenerator{E, T, B}

A generator of operators based on terms, bonds, Hilbert space, and boundary twist.

QuantumLattices.Essentials.Frameworks.Generator — Method

Generator(terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert;
    half::Bool=false, table::Union{Table,Nothing}=nothing, boundary::Boundary=plain
    )

Construct a generator of operators.

QuantumLattices.Essentials.Frameworks.Parameters — Type

Parameters{Names}(values::Number...) where Names

A NamedTuple that contain the key-value pairs.

QuantumLattices.Essentials.Frameworks.Parameters — Method

Parameters(gen::SimplifiedGenerator) -> Parameters

Get the parameters of the terms of a generator.

QuantumLattices.Essentials.Frameworks.Parameters — Method

Parameters(gen::Generator) -> Parameters

Get the parameters of the terms of a generator.

QuantumLattices.Essentials.Frameworks.SimplifiedGenerator — Type

SimplifiedGenerator{P<:Parameters, T<:Union{Table, Nothing}, B<:Boundary, E<:Entry} <: AbstractGenerator{T, B, E}

The simplified generator for the entry of quantum operators.

QuantumLattices.Essentials.Frameworks.SimplifiedGenerator — Method

SimplifiedGenerator(parameters::Parameters, operators::Entry; table::Union{Table,Nothing}=nothing, boundary::Boundary=plain)

Construct an simplified generator of operators.

QuantumLattices.Essentials.QuantumOperators.Transformation — Method

(transformation::Transformation)(gen::AbstractGenerator;
    table::Union{Table, Nothing}=getcontent(gen, :table),
    boundary::Boundary=getcontent(gen, :boundary)
    ) -> SimplifiedGenerator

Get the result of a transformation on the generator for the entry of quantum operators.

QuantumLattices.Essentials.QuantumOperators.Transformation — Method

(transformation::Transformation)(entry::Entry) -> Entry

Get the transformed entry of quantum operators.

Base.eltype — Method

eltype(entry::Entry)
eltype(::Type{<:Entry})

Get the eltype of an entry of quantum operators.

Base.empty! — Method

empty!(gen::Generator) -> Generator

Empty the :bonds, :hilbert, :table and :operators of a generator.

Base.empty! — Method

empty!(entry::Entry) -> Entry

Empty an entry of quantum operators.

Base.empty! — Method

empty!(gen::SimplifiedGenerator) -> SimplifiedGenerator

Empty the operators of an simplified generator.

Base.empty — Method

empty(gen::Generator) -> Generator

Get an empty copy of a generator.

Base.empty — Method

empty(entry::Entry) -> Entry

Get an empty entry of quantum operators.

Base.empty — Method

empty(gen::SimplifiedGenerator) -> SimplifiedGenerator

Get an empty copy of an simplified generator.

Base.get — Method

get(alg::Algorithm, id::Symbol) -> Assignment
get(alg::Algorithm, ::Val{id}) where id -> Assignment

Find the assignment registered on a algorithm by its id.

Base.match — Function

match(params₁::Parameters, params₂::Parameters, atol=atol, rtol=rtol) -> Bool

Judge whether the second set of parameters matches the first.

Base.merge! — Method

merge!(entry::Entry, another::Entry) -> Entry

Merge the entry of quantum operators by another one.

Base.merge — Method

merge(entry::Entry, another::Entry) -> Entry

Get a new entry of quantum operators by merging two entries.

Base.repr — Function

repr(alg::Algorithm, mask::Tuple{Vararg{Symbol}}=(); ndecimal::Int=10) -> String

Get the repr representation of an algorithm.

Optionally, some parameters of the algorithm can be masked. Besides, the maximum number of decimals of the parameters can also be specified.

Base.summary — Method

summary(alg::Algorithm)

Provide a summary of an algorithm.

Base.valtype — Method

valtype(assign::Assignment)
valtype(::Type{<:Assignment})

The type of the data(result) of an assignment.

QuantumLattices.Essentials.Frameworks.dependences — Function

dependences(alg::Algorithm, assign::Assignment, ::Tuple{}=()) -> Tuple{Vararg{Symbol}}
dependences(alg::Algorithm, assign::Assignment, mask::Tuple{Vararg{Symbol}}) -> Tuple{Vararg{Symbol}}

Get the dependences of an assignment and return their ids.

QuantumLattices.Essentials.Frameworks.prepare! — Method

prepare!(alg::Algorithm, assign::Assignment) -> Nothing

Prepare an assignment registered on a algorithm.

QuantumLattices.Essentials.Frameworks.register! — Method

register!(alg::Algorithm, id::Symbol, action::Action; kwargs...) -> Algorithm

Add an assignment on a algorithm by providing the contents of the assignment, and run this assignment.

QuantumLattices.Essentials.Frameworks.run! — Function

run!(alg::Algorithm, id::Symbol, timing::Bool=true) -> Algorithm
run!(alg::Algorithm, ::Val{id}, timing::Bool=true) where id -> Algorithm

Run an assignment with the given id registered on an algorithm. Optionally, the run process can be timed by setting the timing argument to be true.

QuantumLattices.Essentials.Frameworks.run! — Method

run!(alg::Algorithm, assign::Assignment) -> Nothing

Run an assignment registered on a algorithm.

QuantumLattices.Essentials.Frameworks.rundependences! — Function

rundependences!(alg::Algorithm, assign::Assignment, mask::Tuple{Vararg{Symbol}}=()) -> Algorithm

Run the dependences of an assignment. Optionally, some dependences can be jumped by specifying the mask argument.

QuantumLattices.Essentials.reset! — Method

reset!(gen::Generator, lattice::AbstractLattice) -> Generator

Reset a generator by a new lattice.

QuantumLattices.Essentials.reset! — Method

reset!(entry::Entry{<:Operators, <:NamedContainer{Operators}, <:NamedContainer{Operators}},
    terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert;
    half::Bool=false,
    table::Union{Nothing, Table}=nothing
    ) -> Entry

Reset an entry of operators by the new terms, bonds and Hilbert space.

QuantumLattices.Essentials.reset! — Method

reset!(gen::SimplifiedGenerator, operators::Entry; table::Union{Table, Nothing}=nothing) -> SimplifiedGenerator

Reset a generator by a new lattice.

QuantumLattices.Essentials.update! — Method

update!(action::Action; kwargs...) -> Action

Update the status of an action.

QuantumLattices.Essentials.update! — Method

update!(alg::Algorithm; kwargs...) -> Algorithm

Update the parameters of an algorithm and its associated engine.

QuantumLattices.Essentials.update! — Method

update!(assign::Assignment; kwargs...) -> Assignment

Update the parameters of an assignment and the status of its associated action.

QuantumLattices.Essentials.update! — Method

update!(engine::Engine; kwargs...) -> Engine

Update the status of an engine.

QuantumLattices.Essentials.update! — Method

update!(gen::SimplifiedGenerator; kwargs...) -> typeof(gen)

Update the parameters of a generator.

QuantumLattices.Essentials.update! — Method

update!(gen::Generator; kwargs...) -> typeof(gen)

Update the coefficients of the terms in a generator.

QuantumLattices.Interfaces.add! — Method

add!(alg::Algorithm, id::Symbol, action::Action; kwargs...) -> Algorithm

Add an assignment on a algorithm by providing the contents of the assignment.

The difference between add! and register! is that the add! function does not run the newly added assignment but the register! function does.

QuantumLattices.Interfaces.expand! — Method

expand!(operators::Operators,
    entry::Entry{<:Operators, <:NamedContainer{Operators}, <:NamedContainer{Operators}},
    boundary::Boundary;
    kwargs...
    ) -> Operators

Expand an entry of operators with the given boundary twist and term coefficients.

QuantumLattices.Interfaces.expand! — Method

expand!(operators::Operators, gen::AbstractGenerator{E}) where {E<:Entry{<:Operators, <:NamedContainer{Operators}, <:NamedContainer{Operators}}}-> Operators

Expand the generator.

QuantumLattices.Interfaces.expand — Method

expand(gen::Generator, name::Symbol) -> Operators
expand(gen::Generator, i::Int) -> Operators
expand(gen::Generator, name::Symbol, i::Int) -> Operators

Expand the operators of a generator:

the operators of a specific term;

the operators on a specific bond;

the operators of a specific term on a specific bond.

QuantumLattices.Interfaces.expand — Method

expand(gen::AbstractGenerator{E}) where {E<:Entry{<:Operators, <:NamedContainer{Operators}, <:NamedContainer{Operators}}} -> Operators

Expand the generator.

QuantumLattices.Interfaces.id — Method

id(assign::Assignment) -> Symbol
id(::Type{<:Assignment}) -> Symbol

The id of an assignment.