Frameworks

Parameters

Entry

Engine

Generator

SimplifiedGenerator

Action

Assignment

Algorithm

Manual

QuantumLattices.Essentials.Frameworks.AbstractGenerator — Type

AbstractGenerator{E<:Entry, T<:Union{Table, Nothing}, B<:Boundary} <: Engine

Abstract generator for quantum operators.

By protocol, a concrete generator should have the following predefined contents:

operators::E: the entry for the generated quantum operators

table::T: the index-sequence table if it is not nothing

boundary::B: boundary twist for the generated operators

QuantumLattices.Essentials.Frameworks.Action — Type

Action

Abstract type for all actions.

QuantumLattices.Essentials.Frameworks.Algorithm — Type

Algorithm{E<:Engine, P<:Parameters, M<:Function} <: Function

An algorithm associated with an engine.

QuantumLattices.Essentials.Frameworks.Algorithm — Method

(alg::Algorithm)(assign::Assignment) -> Tuple{Algorithm, Assignment}
(assign::Assignment)(alg::Algorithm) -> Tuple{Algorithm, Assignment}

Run an assignment based on an algorithm.

The difference between these two methods is that the first uses the parameters of assign as the current parameters while the second uses those of alg.

QuantumLattices.Essentials.Frameworks.Algorithm — Method

(alg::Algorithm)(id::Symbol, action::Action; info::Bool=true, parameters::Parameters=Parameters{()}(), kwargs...) -> Tuple{Algorithm, Assignment}

Add an assignment on a algorithm by providing the contents of the assignment, and run this assignment.

QuantumLattices.Essentials.Frameworks.Algorithm — Method

Algorithm(name::Symbol, engine::Engine; din::String=".", dout::String=".", parameters::Parameters=Parameters(engine), map::Function=identity)

Construct an algorithm.

QuantumLattices.Essentials.Frameworks.Algorithm — Method

(alg::Algorithm)(id::Symbol; info::Bool=true, parameters::Parameters=Parameters{()}()) -> Tuple{Algorithm, Assignment}

Run an assignment specified by its id.

Optionally, the time of the run process can be informed by setting the info argument to be true.

QuantumLattices.Essentials.Frameworks.Assignment — Type

Assignment{A<:Action, P<:Parameters, M<:Function, D<:Tuple{Vararg{Symbol}}, R} <: Function

An assignment associated with an action.

QuantumLattices.Essentials.Frameworks.Engine — Type

Engine

Abstract type for all engines.

QuantumLattices.Essentials.Frameworks.Entry — Type

Entry{C, A<:NamedTuple, B<:NamedTuple}

An entry of quantum operators related to (part of) a quantum lattice system.

QuantumLattices.Essentials.Frameworks.Entry — Method

Entry(terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert; half::Bool=false, table::Union{Nothing, Table}=nothing)

Construct an entry of operators based on the input terms, bonds and Hilbert space.

QuantumLattices.Essentials.Frameworks.Generator — Type

Generator{E<:Entry, TS<:Tuple{Vararg{Term}}, BS<:Bonds, H<:Hilbert, T<:Union{Table, Nothing}, B<:Boundary} <: AbstractGenerator{E, T, B}

A generator of operators based on terms, bonds, Hilbert space, and boundary twist.

QuantumLattices.Essentials.Frameworks.Generator — Method

Generator(terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert;
    half::Bool=false, table::Union{Table,Nothing}=nothing, boundary::Boundary=plain
    )

Construct a generator of operators.

QuantumLattices.Essentials.Frameworks.Parameters — Type

Parameters{Names}(values::Number...) where Names

A NamedTuple that contain the key-value pairs.

QuantumLattices.Essentials.Frameworks.Parameters — Method

Parameters(gen::Generator) -> Parameters

Get the parameters of the terms of a generator.

QuantumLattices.Essentials.Frameworks.Parameters — Method

Parameters(gen::SimplifiedGenerator) -> Parameters

Get the parameters of the terms of a generator.

QuantumLattices.Essentials.Frameworks.SimplifiedGenerator — Type

SimplifiedGenerator{P<:Parameters, T<:Union{Table, Nothing}, B<:Boundary, E<:Entry} <: AbstractGenerator{T, B, E}

The simplified generator for the entry of quantum operators.

QuantumLattices.Essentials.Frameworks.SimplifiedGenerator — Method

SimplifiedGenerator(parameters::Parameters, operators::Entry; table::Union{Table,Nothing}=nothing, boundary::Boundary=plain)

Construct an simplified generator of operators.

QuantumLattices.Essentials.QuantumOperators.Transformation — Method

(transformation::Transformation)(gen::AbstractGenerator;
    table::Union{Table, Nothing}=getcontent(gen, :table),
    boundary::Boundary=getcontent(gen, :boundary),
    kwargs...
    ) -> SimplifiedGenerator

Get the result of a transformation on the generator for the entry of quantum operators.

QuantumLattices.Essentials.QuantumOperators.Transformation — Method

(transformation::Transformation)(entry::Entry; kwargs...) -> Entry

Get the transformed entry of quantum operators.

Base.empty! — Method

empty!(gen::Generator) -> Generator

Empty the :bonds, :hilbert, :table and :operators of a generator.

Base.empty! — Method

empty!(entry::Entry) -> Entry

Empty an entry of quantum operators.

Base.empty! — Method

empty!(gen::SimplifiedGenerator) -> SimplifiedGenerator

Empty the operators of an simplified generator.

Base.empty — Method

empty(gen::Generator) -> Generator

Get an empty copy of a generator.

Base.empty — Method

empty(entry::Entry) -> Entry

Get an empty entry of quantum operators.

Base.empty — Method

empty(gen::SimplifiedGenerator) -> SimplifiedGenerator

Get an empty copy of an simplified generator.

Base.match — Method

match(params₁::Parameters, params₂::Parameters; atol=atol, rtol=rtol) -> Bool

Judge whether the second set of parameters matches the first.

Base.merge! — Method

merge!(entry::Entry, another::Entry) -> Entry

Merge the entry of quantum operators by another one.

Base.merge — Method

merge(entry::Entry, another::Entry) -> Entry

Get a new entry of quantum operators by merging two entries.

Base.nameof — Method

nameof(alg::Algorithm, assign::Assignment) -> String

Get the name of the combination of an algorithm and an assignment.

Base.repr — Function

repr(alg::Algorithm, f::Function=param->true; ndecimal::Int=10) -> String

Get the repr representation of an algorithm.

Optionally, some parameters of the algorithm can be filtered by specifying the f function. Besides, the maximum number of decimals of the parameters can also be specified.

Base.summary — Method

summary(alg::Algorithm)

Provide a summary of an algorithm.

Base.valtype — Method

valtype(assign::Assignment)
valtype(::Type{<:Assignment})

The type of the data(result) of an assignment.

Base.valtype — Method

valtype(entry::Entry)
valtype(::Type{<:Entry})

Get the valtype of an entry of quantum operators.

QuantumLattices.Essentials.Frameworks.rundependences! — Function

rundependences!(alg::Algorithm, assign::Assignment, f::Function=assign->true) -> Tuple{Algorithm, Assignment}

Run the dependences of an assignment.

Optionally, some dependences can be filtered by specifying the f function.

QuantumLattices.Essentials.Frameworks.save — Method

save(alg::Algorithm, assign::Assignment) -> Tuple{Algorithm, Assignment}

Save the data of an assignment registered on an algorithm.

QuantumLattices.Essentials.reset! — Method

reset!(gen::Generator, lattice::AbstractLattice) -> Generator

Reset a generator by a new lattice.

QuantumLattices.Essentials.reset! — Method

reset!(entry::Entry{<:Operators}, terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert; half::Bool=false, table::Union{Nothing, Table}=nothing) -> Entry

Reset an entry of operators by the new terms, bonds and Hilbert space.

QuantumLattices.Essentials.reset! — Method

reset!(gen::SimplifiedGenerator, operators::Entry; table::Union{Table, Nothing}=nothing) -> SimplifiedGenerator

Reset a generator by a new lattice.

QuantumLattices.Essentials.update! — Method

update!(alg::Algorithm; kwargs...) -> Algorithm

Update the parameters of an algorithm and its associated engine.

QuantumLattices.Essentials.update! — Method

update!(assign::Assignment; kwargs...) -> Assignment

Update the parameters of an assignment and the status of its associated action.

QuantumLattices.Essentials.update! — Method

update!(gen::Generator; kwargs...) -> typeof(gen)

Update the coefficients of the terms in a generator.

QuantumLattices.Essentials.update! — Method

update!(gen::SimplifiedGenerator; kwargs...) -> typeof(gen)

Update the parameters of a generator.

QuantumLattices.Interfaces.add! — Method

add!(alg::Algorithm, id::Symbol, action::Action; parameters::Parameters=Parameters{()}(), kwargs...) -> Tuple{Algorithm, Assignment}

Add an assignment on a algorithm by providing the contents of the assignment without the execution of it.

QuantumLattices.Interfaces.expand! — Method

expand!(result, gen::AbstractGenerator) -> typeof(result)

Expand the generator.

QuantumLattices.Interfaces.expand! — Method

expand!(operators::Operators, entry::Entry, boundary::Boundary; kwargs...) -> Operators

Expand an entry of operators with the given boundary twist and term coefficients.

QuantumLattices.Interfaces.expand — Method

expand(gen::Generator, name::Symbol) -> Operators
expand(gen::Generator, i::Int) -> Operators
expand(gen::Generator, name::Symbol, i::Int) -> Operators

Expand the operators of a generator:

the operators of a specific term;

the operators on a specific bond;

the operators of a specific term on a specific bond.

QuantumLattices.Interfaces.expand — Method

expand(gen::AbstractGenerator) -> valtype(gen)

Expand the generator.

RecipesBase.apply_recipe — Method

@recipe plot(pack::Tuple{Algorithm, Assignment})

Define the recipe for the visualization of an assignment of an algorithm.