Frameworks

Parameters

Entry

Engine

Generator

SimplifiedGenerator

Action

Assignment

Algorithm

Manual

QuantumLattices.Essentials.Frameworks.AbstractGenerator — Type

AbstractGenerator <: Engine

Abstract type of the generator of (representations of) the quantum operators of a quantum lattice system.

QuantumLattices.Essentials.Frameworks.Action — Type

Action

Abstract type for all actions.

QuantumLattices.Essentials.Frameworks.Algorithm — Type

Algorithm{E<:Engine, P<:Parameters, M<:Function} <: Function

An algorithm associated with an engine.

QuantumLattices.Essentials.Frameworks.Algorithm — Method

(alg::Algorithm)(assign::Assignment) -> Tuple{Algorithm, Assignment}
(assign::Assignment)(alg::Algorithm) -> Tuple{Algorithm, Assignment}

Run an assignment based on an algorithm.

The difference between these two methods is that the first uses the parameters of assign as the current parameters while the second uses those of alg.

QuantumLattices.Essentials.Frameworks.Algorithm — Method

(alg::Algorithm)(id::Symbol, action::Action; info::Bool=true, parameters::Parameters=Parameters{()}(), kwargs...) -> Tuple{Algorithm, Assignment}

Add an assignment on a algorithm by providing the contents of the assignment, and run this assignment.

QuantumLattices.Essentials.Frameworks.Algorithm — Method

Algorithm(name::Symbol, engine::Engine; din::String=".", dout::String=".", parameters::Parameters=Parameters(engine), map::Function=identity)

Construct an algorithm.

QuantumLattices.Essentials.Frameworks.Algorithm — Method

(alg::Algorithm)(id::Symbol; info::Bool=true, parameters::Parameters=Parameters{()}()) -> Tuple{Algorithm, Assignment}

Run an assignment specified by its id.

Optionally, the time of the run process can be informed by setting the info argument to be true.

QuantumLattices.Essentials.Frameworks.Assignment — Type

Assignment{A<:Action, P<:Parameters, M<:Function, D<:Tuple{Vararg{Symbol}}, R} <: Function

An assignment associated with an action.

QuantumLattices.Essentials.Frameworks.Boundary — Type

Boundary{Names}(values::AbstractVector{<:Number}, vectors::AbstractVector{<:AbstractVector{<:Number}}) where Names

Boundary twist of operators.

QuantumLattices.Essentials.Frameworks.Boundary — Method

(bound::Boundary)(operator::Operator; origin::Union{AbstractVector, Nothing}=nothing) -> Operator

Get the boundary twisted operator.

QuantumLattices.Essentials.Frameworks.CompositeGenerator — Type

CompositeGenerator{E<:Entry, T<:Union{Table, Nothing}}

Abstract type for composite generator of (representations of) quantum operators.

By protocol, it must have the following predefined contents:

operators::E: the entry for the generated (representations of) quantum operators

table::T: the index-sequence table if it is not nothing

QuantumLattices.Essentials.Frameworks.Engine — Type

Engine

Abstract type for all engines.

An engine is the core to generate/update/manipulate (representations of) the quantum operators of a quantum lattice system.

QuantumLattices.Essentials.Frameworks.Entry — Type

Entry{C, A<:NamedTuple, B<:NamedTuple, P<:Parameters, D<:Boundary} <: AbstractGenerator

An entry of quantum operators (or representations of quantum operators) related to (part of) a quantum lattice system.

QuantumLattices.Essentials.Frameworks.Entry — Method

Entry(terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert;
    half::Bool=false,
    table::Union{Nothing, Table}=nothing,
    boundary::Boundary=plain
    )

Construct an entry of quantum operators based on the input terms, bonds, Hilbert space and (twisted) boundary condition.

QuantumLattices.Essentials.Frameworks.Formulation — Type

Formulation{F<:Function, P<:Parameters} <: AbstractGenerator

Generator of (representations of the) quantum operators of a quantum lattice system by analytical expressions.

QuantumLattices.Essentials.Frameworks.Generator — Type

Generator{E<:Entry{<:Operators}, TS<:Tuple{Vararg{Term}}, BS<:Bonds, H<:Hilbert, T<:Union{Table, Nothing}} <: CompositeGenerator{E, T}

A generator of operators based on terms, bonds and Hilbert space.

QuantumLattices.Essentials.Frameworks.Generator — Method

Generator(terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert;
    half::Bool=false,
    table::Union{Table,Nothing}=nothing,
    boundary::Boundary=plain
    )

Construct a generator of operators.

QuantumLattices.Essentials.Frameworks.Image — Type

Image{E<:Entry, H<:Transformation, T<:Union{Table, Nothing}} <: CompositeGenerator{E, T}

The image of a transformation on an generator of (representation of) quantum operators.

QuantumLattices.Essentials.Frameworks.Parameters — Type

Parameters{Names}(values::Number...) where Names

A NamedTuple that contain the key-value pairs.

QuantumLattices.Essentials.Frameworks.Parameters — Method

Parameters(bound::Boundary)

Get the parameters of the twisted boundary condition.

QuantumLattices.Essentials.Frameworks.Parameters — Method

Parameters(entry::Entry)

Get the complete set of parameters of an entry of (representations of) quantum operators.

QuantumLattices.Essentials.QuantumOperators.Transformation — Method

(transformation::Transformation)(entry::Entry; kwargs...) -> Entry

Get the transformed entry of (representations of) quantum operators.

QuantumLattices.Essentials.QuantumOperators.Transformation — Method

(transformation::Transformation)(gen::AbstractGenerator; table::Union{Table, Nothing}=nothing, kwargs...) -> Image

Get the image of a transformation on an instance of AbstractGenerator.

Base.empty — Method

empty(entry::Entry) -> Entry
empty!(entry::Entry) -> Entry

Get an empty copy entry or empty an entry of (representations of) quantum operators.

Base.empty — Method

empty(gen::Generator) -> Generator
empty!(gen::Generator) -> Generator

Get an empty copy of a generator.

Empty the :bonds, :hilbert, :table and :operators attributes of a generator.

Base.empty — Method

empty(gen::Image) -> Image
empty!(gen::Image) -> Image

Get an empty copy of the image of a transformation on a generator.

Empty the :operators and table attributes of the image of a transformation on a generator.

Base.keys — Method

keys(bound::Boundary) -> Tuple{Vararg{Symbol}}
keys(::Type{<:Boundary{Names}}) where Names -> Names

Get the names of the boundary parameters.

Base.match — Method

match(params₁::Parameters, params₂::Parameters; atol=atol, rtol=rtol) -> Bool

Judge whether the second set of parameters matches the first.

Base.merge! — Method

merge!(bound::Boundary, another::Boundary) -> typeof(bound)

Merge the values and vectors of the twisted boundary condition from another one.

Base.merge — Method

merge(entry::Entry, another::Entry) -> Entry
merge!(entry::Entry, another::Entry) -> Entry

Get a new entry of (representations of) quantum operators by merging two.

Merge an entry of (representations of) quantum operators by another one.

Base.nameof — Method

nameof(alg::Algorithm, assign::Assignment) -> String

Get the name of the combination of an algorithm and an assignment.

Base.repr — Function

repr(alg::Algorithm, f::Function=param->true; ndecimal::Int=10) -> String

Get the repr representation of an algorithm.

Optionally, some parameters of the algorithm can be filtered by specifying the f function. Besides, the maximum number of decimals of the parameters can also be specified.

Base.summary — Method

summary(alg::Algorithm)

Provide a summary of an algorithm.

Base.valtype — Method

valtype(assign::Assignment)
valtype(::Type{<:Assignment})

The type of the data(result) of an assignment.

Base.valtype — Method

valtype(::Type{<:Entry})

Get the valtype of an entry of (representations of) quantum operators.

QuantumLattices.Essentials.Frameworks.plain — Function

plain

Plain boundary condition without any twist.

QuantumLattices.Essentials.Frameworks.rundependences! — Function

rundependences!(alg::Algorithm, assign::Assignment, f::Function=assign->true) -> Tuple{Algorithm, Assignment}

Run the dependences of an assignment.

Optionally, some dependences can be filtered by specifying the f function.

QuantumLattices.Essentials.Frameworks.save — Method

save(alg::Algorithm, assign::Assignment) -> Tuple{Algorithm, Assignment}

Save the data of an assignment registered on an algorithm.

QuantumLattices.Essentials.reset! — Function

reset!(gen::Generator, lattice::AbstractLattice, boundary::Boundary=gen.operators.boundary) -> Generator

Reset a generator by a new lattice and a new twisted boundary condition.

QuantumLattices.Essentials.reset! — Method

reset!(entry::Entry, transformation::Transformation, source::Entry{<:Operators}; kwargs...)

Reset the contents of an entry by its source entry of quantum operators and the corresponding transformation.

QuantumLattices.Essentials.reset! — Method

reset!(entry::Entry{<:Operators}, terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert;
    half::Bool=false,
    table::Union{Nothing, Table}=nothing,
    boundary::Boundary=entry.boundary
    ) -> Entry

Reset an entry of quantum operators by the new terms, bonds, Hilbert space and (twisted) boundary condition.

QuantumLattices.Essentials.reset! — Method

reset!(gen::Image, source::AbstractGenerator;
    transformation::Transformation=gen.transformation,
    table::Union{Table, Nothing}=nothing,
    kwargs...
    ) -> Image

Reset the image of a transformation on a generator.

QuantumLattices.Essentials.update! — Method

update!(alg::Algorithm; parameters...) -> Algorithm

Update the parameters of an algorithm and its associated engine.

QuantumLattices.Essentials.update! — Method

update!(assign::Assignment; parameters...) -> Assignment

Update the parameters of an assignment and the status of its associated action.

QuantumLattices.Essentials.update! — Method

update!(bound::Boundary; parameters...) -> Boundary

Update the values of the boundary twisted phase.

QuantumLattices.Essentials.update! — Method

update!(entry::Entry, transformation::Transformation, source::Entry{<:Operators}; kwargs...) -> Entry

Update the contents of an Entry by its source entry of quantum operators and the corresponding transformation.

QuantumLattices.Essentials.update! — Method

update!(entry::Entry{<:Operators}; parameters...) -> Entry

Update the parameters (including the boundary parameters) of an entry of quantum operators.

!!! Note The coefficients of boundops are also updated due to the change of the boundary parameters.

QuantumLattices.Essentials.update! — Method

update!(gen::Generator; parameters...) -> typeof(gen)

Update the coefficients of the terms in a generator.

QuantumLattices.Essentials.update! — Method

update!(gen::Image, source::AbstractGenerator; kwargs...) -> Image

Update the image of an transformation by its updated source generator.

QuantumLattices.Essentials.update! — Method

update!(gen::Image; parameters...) -> typeof(gen)

Update the parameters of the image of a transformation on a generator.

QuantumLattices.Interfaces.add! — Method

add!(alg::Algorithm, id::Symbol, action::Action; parameters::Parameters=Parameters{()}(), kwargs...) -> Tuple{Algorithm, Assignment}

Add an assignment on a algorithm by providing the contents of the assignment without the execution of it.

QuantumLattices.Interfaces.expand! — Method

expand!(result, entry::Entry) -> typeof(result)

Expand the contents of an entry.

QuantumLattices.Interfaces.expand — Method

expand(gen::Generator, name::Symbol) -> Operators
expand(gen::Generator, i::Int) -> Operators
expand(gen::Generator, name::Symbol, i::Int) -> Operators

Expand the operators of a generator:

the operators of a specific term;

the operators on a specific bond;

the operators of a specific term on a specific bond.

QuantumLattices.Interfaces.expand — Method

expand(gen::AbstractGenerator) -> valtype(gen)
expand!(result, gen::AbstractGenerator) -> typeof(result)

Expand the generator, that is, get the (representations of the) quantum operators of a quantum lattice system (or some part of it).

RecipesBase.apply_recipe — Method

@recipe plot(pack::Tuple{Algorithm, Assignment})

Define the recipe for the visualization of an assignment of an algorithm.