Quantum Systems

Canonical fermionic/bosonic systems

FID and Fock

FockCoupling

Fock terms

SU(2) spin systems

SID and Spin

SpinCoupling

SpinTerm

Phononic systems

PNID and Phonon

PhononCoupling

PhononKinetic and PhononPotential

Manual

QuantumLattices.Essentials.QuantumSystems.annihilation — Constant

annihilation

Indicate that the nambu index is annihilation.

QuantumLattices.Essentials.QuantumSystems.blatex — Constant

blatex

The default LaTeX format for a bosonic oid.

QuantumLattices.Essentials.QuantumSystems.creation — Constant

creation

Indicate that the nambu index is creation.

QuantumLattices.Essentials.QuantumSystems.flatex — Constant

flatex

The default LaTeX format for a fermionic oid.

QuantumLattices.Essentials.QuantumSystems.majorana — Constant

majorana

Indicate that the nambu index is majorana.

QuantumLattices.Essentials.QuantumSystems.nlatex — Constant

nlatex

The default LaTeX format for a phonon oid.

QuantumLattices.Essentials.QuantumSystems.slatex — Constant

slatex

The default LaTeX format for a spin oid.

Base.Matrix — Method

Matrix(sid::SID{S, <:Union{Int, Symbol}, Char}, dtype::Type{<:Number}=Complex{Float}) where S -> Matrix{dtype}

Get the matrix representation of a sid.

QuantumLattices.Essentials.QuantumSystems.Coulomb — Type

Coulomb(id::Symbol, value, bondkind::Int=1;
    couplings::Union{Function, Couplings}=@couplings(FockCoupling{2}()*FockCoupling{2}()),
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false
    )

Coulomb term.

Type alias for Term{:Coulomb, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.DMPhonon — Type

DMPhonon(id::Symbol, value::Any, bondkind::Int;
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false
    )

The DM Magnon-Phonon coupling term.

Type alias for Term{:DMPhonon, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}

QuantumLattices.Essentials.QuantumSystems.FID — Type

FID{T, O<:Union{Int, Symbol}, S<:Union{Int, Symbol}, N<:Union{Int, Symbol}} <: SimpleIID

The Fock id.

QuantumLattices.Essentials.QuantumSystems.FID — Method

FID{T}(; orbital::Union{Int, Symbol}=1, spin::Union{Int, Symbol}=1, nambu::Union{Int, Symbol}=annihilation) where T

Create a Fock id.

QuantumLattices.Essentials.QuantumSystems.Fock — Type

Fock{T} <: SimpleInternal{FID{T, Int, Int, Int}}

The Fock internal degrees of freedom.

QuantumLattices.Essentials.QuantumSystems.Fock — Method

Fock{T}(; norbital::Int=1, nspin::Int=2, nnambu::Int=2) where T

Construct a Fock degrees of freedom.

QuantumLattices.Essentials.QuantumSystems.FockCoupling — Type

FockCoupling(value::Number,
    orbitals::Subscript{<:NTuple{N, Union{Int, Symbol}}},
    spins::Subscript{<:NTuple{N, Union{Int, Symbol}}},
    nambus::Union{NTuple{N, Int}, NTuple{N, Symbol}}
    ) where N

Fock coupling.

Type alias for Coupling{V, I<:ID{FID}, C<:Subscripts, CI<:SubscriptsID}.

QuantumLattices.Essentials.QuantumSystems.FockCoupling — Method

FockCoupling{N}(value::Number=1;
    orbitals::Union{NTuple{N, Int}, Subscript}=Subscript(N),
    spins::Union{NTuple{N, Int}, Subscript}=Subscript(N),
    nambus::Union{NTuple{N, Int}, NTuple{N, Symbol}}=ntuple(i->wildcard, Val(N))
    ) where N

Construct a Fock coupling.

QuantumLattices.Essentials.QuantumSystems.Hopping — Type

Hopping(id::Symbol, value, bondkind::Int=1;
    couplings::Union{Function, Couplings}=@couplings(FockCoupling{2}()),
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Hopping term.

Type alias for Term{:Hopping, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.Hubbard — Type

Hubbard(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Hubbard term.

Type alias for Term{:Hubbard, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.InterOrbitalInterSpin — Type

InterOrbitalInterSpin(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Interorbital-interspin term.

Type alias for Term{:InterOrbitalInterSpin, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.InterOrbitalIntraSpin — Type

InterOrbitalIntraSpin(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Interorbital-intraspin term.

Type alias for Term{:InterOrbitalIntraSpin, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.NID — Type

NID{D<:Union{Char, Symbol}} <: SimpleIID

The phonon id.

QuantumLattices.Essentials.QuantumSystems.Onsite — Type

Onsite(id::Symbol, value;
    couplings::Union{Function, Couplings}=@couplings(FockCoupling{2}()),
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Onsite term.

Type alias for Term{:Onsite, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.PairHopping — Type

PairHopping(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Pair-hopping term.

Type alias for Term{:PairHopping, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.Pairing — Type

Pairing(id::Symbol, value, bondkind::Int=0;
    couplings::Union{Function, Couplings}=@couplings(FockCoupling{2}()),
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Pairing term.

Type alias for Term{:Pairing, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.Phonon — Type

Phonon <: SimpleInternal{NID{Char}}

The phonon internal degrees of freedom.

QuantumLattices.Essentials.QuantumSystems.PhononCoupling — Type

PhononCoupling(value::Number, tags::NTuple{N, Char}, dirs::Subscript{<:Union{NTuple{N, Char}, NTuple{N, Symbol}}}) where N

Phonon coupling.

Type alias for Coupling{V<:Number, I<:ID{NID}, C<:Subscripts, CI<:SubscriptsID}.

QuantumLattices.Essentials.QuantumSystems.PhononCoupling — Method

PhononCoupling(value::Number, tags::NTuple{N, Char};
    dirs::Union{NTuple{N, Char}, NTuple{N, Symbol}, Subscript}=Subscript(N)
    ) where N

Construct a phonon coupling.

QuantumLattices.Essentials.QuantumSystems.PhononKinetic — Type

PhononKinetic(id::Symbol, value::Any;
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false
    )

Kinetic energy part of phonons.

Type alias for Term{:PhononKinetic, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.PhononPotential — Type

PhononPotential(id::Symbol, value::Any, bondkind::Int;
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false
    )

Potential energy part of phonons.

Type alias for Term{:PhononPotential, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}

QuantumLattices.Essentials.QuantumSystems.SID — Type

SID{S, O<:Union{Int, Symbol}} <: SimpleIID

The spin id.

QuantumLattices.Essentials.QuantumSystems.SID — Method

SID{S}(tag::Union{Char, Symbol}; orbital::Union{Int, Symbol}=1) where S

Create a spin id.

QuantumLattices.Essentials.QuantumSystems.Spin — Type

Spin{S} <: SimpleInternal{SID{S, Int, Char}}

The spin internal degrees of freedom.

QuantumLattices.Essentials.QuantumSystems.Spin — Method

Spin{S}(; norbital::Int=1) where S

Construct a spin degrees of freedom.

QuantumLattices.Essentials.QuantumSystems.SpinCoupling — Type

SpinCoupling(value::Number, tags::NTuple{N, Char}, orbitals::Subscript{<:NTuple{N, Union{Int, Symbol}}}) where N

Spin coupling.

Type alias for Coupling{V, I<:ID{SID}, C<:Subscripts, CI<:SubscriptsID}.

QuantumLattices.Essentials.QuantumSystems.SpinCoupling — Method

SpinCoupling(value::Number, tags::NTuple{N, Char}; orbitals::Union{NTuple{N, Int}, Subscript}=Subscript(N)) where N

Construct a spin coupling.

QuantumLattices.Essentials.QuantumSystems.SpinFlip — Type

SpinFlip(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, modulate::Union{Function, Bool}=false)

Spin-flip term.

Type alias for Term{:SpinFlip, id, V, Int, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.SpinTerm — Type

SpinTerm(id::Symbol, value::Any, bondkind::Any, couplings::Union{Function, Coupling, Couplings},
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    )

Spin term.

Type alias for Term{:SpinTerm, id, V, B<:Any, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}.

QuantumLattices.Essentials.QuantumSystems.@dm_str — Macro

dm"x ob[o₁ o₂]" -> Couplings
dm"y ob[o₁ o₂]" -> Couplings
dm"z ob[o₁ o₂]" -> Couplings

The DM couplings.

QuantumLattices.Essentials.QuantumSystems.@dmphonon_str — Macro

dmphonon"" -> Coupling

The DM magnon-phonon couplings.

QuantumLattices.Essentials.QuantumSystems.@fc_str — Macro

fc"..." -> FockCoupling

Construct a FockCoupling from a literal string.

QuantumLattices.Essentials.QuantumSystems.@gamma_str — Macro

gamma"x ob[o₁ o₂]" -> Couplings
gamma"y ob[o₁ o₂]" -> Couplings
gamma"z ob[o₁ o₂]" -> Couplings

The Gamma couplings.

QuantumLattices.Essentials.QuantumSystems.@heisenberg_str — Macro

heisenberg"ob[o₁ o₂]" -> Couplings
heisenberg"xyz ob[o₁ o₂]" -> Couplings
heisenberg"+-z ob[o₁ o₂]" -> Couplings

The Heisenberg couplings.

QuantumLattices.Essentials.QuantumSystems.@ising_str — Macro

ising"x ob[o₁ o₂]" -> Couplings
ising"y ob[o₁ o₂])" -> Couplings
ising"z ob[o₁ o₂]" -> Couplings

The Ising couplings.

QuantumLattices.Essentials.QuantumSystems.@kinetic_str — Macro

kinetic"" -> Couplings

The kinetic energy part of phonon couplings.

QuantumLattices.Essentials.QuantumSystems.@potential_str — Macro

potential"" -> Couplings

The potential energy part of phonon couplings.

QuantumLattices.Essentials.QuantumSystems.@sc_str — Macro

sc"..." -> SpinCoupling

Construct a SpinCoupling from a literal string.

QuantumLattices.Essentials.QuantumSystems.@sˣ_str — Macro

sˣ"ob[o]" -> Couplings
sʸ"ob[o]" -> Couplings
sᶻ"ob[o]" -> Couplings

The single Sˣ/Sʸ/Sᶻ coupling.

QuantumLattices.Essentials.QuantumSystems.@σʸ_str — Macro

σʸ"sp" -> Couplings
σʸ"ob" -> Couplings
σʸ"ph" -> Couplings

The Pauli matrix σʸ, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.QuantumSystems.@σˣ_str — Macro

σˣ"sp" -> Couplings
σˣ"ob" -> Couplings
σˣ"ph" -> Couplings

The Pauli matrix σˣ, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.QuantumSystems.@σᶻ_str — Macro

σᶻ"sp" -> Couplings
σᶻ"ob" -> Couplings
σᶻ"ph" -> Couplings

The Pauli matrix σᶻ, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.QuantumSystems.@σ⁰_str — Macro

σ⁰"sp" -> Couplings
σ⁰"ob" -> Couplings
σ⁰"ph" -> Couplings

The Pauli matrix σ⁰, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.QuantumSystems.@σ⁺_str — Macro

σ⁺"sp" -> Couplings
σ⁺"ob" -> Couplings
σ⁺"ph" -> Couplings

The Pauli matrix σ⁺, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

QuantumLattices.Essentials.QuantumSystems.@σ⁻_str — Macro

σ⁻"sp" -> Couplings
σ⁻"ob" -> Couplings
σ⁻"ph" -> Couplings

The Pauli matrix σ⁻, which can act on the space of spins("sp"), orbitals("ob") or particle-holes("ph").

Base.:* — Method

*(  f1::Operator{<:Number, <:ID{CompositeOID{<:Index{<:AbstractPID, <:FID{:f}}}}},
    f2::Operator{<:Number, <:ID{CompositeOID{<:Index{<:AbstractPID, <:FID{:f}}}}}
    ) -> Union{Nothing, Operator}

Get the multiplication of two fermionic Fock operators.

Base.angle — Method

angle(id::OID{<:Index{<:AbstractPID, <:FID}}, vectors::AbstractVector{<:AbstractVector{<:Number}}, values::AbstractVector{<:Number}) -> Complex{<:Number}

Get the twist phase corresponding to a Fock oid.

LinearAlgebra.:⋅ — Method

⋅(fc₁::FockCoupling, fc₂::FockCoupling) -> FockCoupling

Get the dot product of two rank-2 Fock couplings.

A rank-2 FockCoupling can be considered as a matrix acting on the orbital, spin and nambu spaces. Therefore, the dot product here is defined as the multiplication between such matrices.

QuantumLattices.Essentials.QuantumOperators.script — Method

script(::Val{:BD}, index::Index{<:AbstractPID, <:NID}, l::LaTeX) -> Char
script(::Val{:BD}, oid::CompositeOID{<:Index{<:AbstractPID, <:NID}}, l::LaTeX) -> Char
script(::Val{:site}, index::Index{<:AbstractPID, <:NID}; kwargs...) -> Int
script(::Val{:dir}, index::Index{<:AbstractPID, <:NID}; kwargs...) -> Char

Get the required script of a phonon oid.

QuantumLattices.Essentials.QuantumOperators.script — Method

script(::Val{:site}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> Int
script(::Val{:orbital}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> Int
script(::Val{:spint}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> Int
script(::Val{:spsym}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> String
script(::Val{:nambu}, index::Index{<:AbstractPID, <:FID}; kwargs...) -> String

Get the required script of a Fock index.

QuantumLattices.Essentials.QuantumOperators.script — Method

script(::Val{:site}, index::Index{<:AbstractPID, <:SID}; kwargs...) -> Int
script(::Val{:orbital}, index::Index{<:AbstractPID, <:SID}; kwargs...) -> Int
script(::Val{:tag}, index::Index{<:AbstractPID, <:SID}; kwargs...) -> Char

Get the required script of a spin oid.

QuantumLattices.Essentials.QuantumSystems.isnormalordered — Method

isnormalordered(opt::Operator{<:Number, <:ID{CompositeOID{<:Index{<:AbstractPID, <:FID}}}}) -> Bool

Judge whether an operator is normal ordered.

QuantumLattices.Interfaces.:⊗ — Method

⊗(fc₁::FockCoupling, fc₂::FockCoupling) -> FockCoupling

Get the direct product between two Fock couplings.

QuantumLattices.Interfaces.expand — Method

expand(pnc::PhononCoupling{<:Number, <:ID{NID{Symbol}}}, bond::Bond, hilbert::Hilbert, info::Val{:PhononPotential}) -> PPExpand

Expand the default phonon potential coupling on a given bond.

QuantumLattices.Interfaces.expand — Method

expand(dmp::Coupling{<:Number, <:Tuple{NID{Symbol}, SID{wildcard, Int, Symbol}}}, bond::Bond, hilbert::Hilbert, info::Val{:DMPhonon}) -> DMPExpand

Expand the default DM magnon-phonon coupling on a given bond.

QuantumLattices.Interfaces.permute — Method

permute(id₁::OID{<:Index{<:AbstractPID, <:FID{:b}}}, id₂::OID{<:Index{<:AbstractPID, <:FID{:b}}}) -> Tuple{Vararg{Operator}}

Permute two bosonic oids and get the result.

QuantumLattices.Interfaces.permute — Method

permute(id₁::OID{<:Index{<:AbstractPID, <:FID{:f}}}, id₂::OID{<:Index{<:AbstractPID, <:FID{:f}}}) -> Tuple{Vararg{Operator}}

Permute two fermionic oids and get the result.

QuantumLattices.Interfaces.permute — Method

permute(id₁::OID{<:Index{<:AbstractPID, NID{Char}}}, id₂::OID{<:Index{<:AbstractPID, NID{Char}}}) -> Tuple{Vararg{Operator}}

Permute two phonon oids and get the result.

QuantumLattices.Interfaces.permute — Method

permute(id₁::OID{<:Index{<:AbstractPID, SID{S, Int, Char}}}, id₂::OID{<:Index{<:AbstractPID, SID{S, Int, Char}}}) where S -> Tuple{Vararg{Operator}}

Permute two spin oids and get the result.