Spatial

Spatial

Point

Bond and Link

AbstractLattice

Lattice

SuperLattice

Cylinder

Manul

Hamiltonian.Essentials.Spatial.acrossbonds — Constant.

acrossbonds

Indicate that bonds across the unitcell are inquired, which are in fact those across the periodic boundaries.

source

Hamiltonian.Essentials.Spatial.insidebonds — Constant.

insidebonds

Indicate that bonds inside the unitcell are inquired, which do not contain those across the periodic boundaries.

source

Hamiltonian.Essentials.Spatial.interbonds — Constant.

interbonds

Indicate that bonds inter the sublattices are inquired.

Notes

These bonds do not contain those accorss the periodic boundaries.

source

Hamiltonian.Essentials.Spatial.intrabonds — Constant.

intrabonds

Indicate that bonds intra the sublattices are inquired.

Notes

These bonds do not contain those accorss the periodic boundaries.

source

Hamiltonian.Essentials.Spatial.AbstractLattice — Type.

AbstractLattice{P<:PID,N}

Abstract type for all lattices.

source

Hamiltonian.Essentials.Spatial.AbstractLatticeIndex — Type.

AbstractLatticeIndex{I<:Union{<:PID,Int}}

Abstract index type for a lattice.

source

Hamiltonian.Essentials.Spatial.Bond — Type.

Bond(neighbor::Int,spoint::Point,epoint::Point)

A bond in a lattice.

source

Hamiltonian.Essentials.Spatial.Cylinder — Type.

Cylinder{P}(    name::String,
                block::AbstractMatrix{<:Real},
                translation::AbstractVector{<:Real};
                vector::Union{AbstractVector{<:Real},Nothing}=nothing,
                neighbors::Union{Dict{Int,<:Real},Int}=1,
                ) where P<:PID

Cylinder of 1d and quasi 2d lattices.

source

Hamiltonian.Essentials.Spatial.Cylinder — Method.

(cylinder::Cylinder)(scopes::Any...;coordination::Int=8) -> Lattice

Construct a lattice from a cylinder with the assigned scopes.

source

Hamiltonian.Essentials.Spatial.ICoordIndex — Type.

ICoordIndex(index::Union{<:PID,Int})

Index for getting an icoord of a lattice.

source

Hamiltonian.Essentials.Spatial.Lattice — Type.

Lattice(    name::String,
            pids::Vector{<:PID},
            rcoords::AbstractMatrix{<:Real};
            icoords::AbstractMatrix{<:Real}=SMatrix{0,0,Float}(),
            vectors::AbstractVector{<:AbstractVector{<:Real}}=SVector{0,SVector{size(rcoords,1),Float}}(),
            neighbors::Union{Dict{Int,<:Real},Int}=1,
            coordination::Int=8
        )
Lattice(    name::String,
            points::AbstractVector{<:Point};
            vectors::AbstractVector{<:AbstractVector{<:Real}}=SVector{0,SVector{points|>eltype|>dimension,Float}}(),
            neighbors::Union{Dict{Int,<:Real},Int}=1,
            coordination::Int=8
            )
Lattice(    name::String,
            sublattices::AbstractVector{<:AbstractLattice};
            vectors::AbstractVector{<:AbstractVector{<:Real}}=SVector{0,SVector{sublattices|>eltype|>dimension,Float}}(),
            neighbors::Union{Dict{Int,<:Real},Int}=1,
            coordination::Int=8
            )

Simplest lattice.

A simplest lattice can be construted from its contents, i.e. pids, rcoords and icoords, or from a couple of points, or from a couple of sublattices.

source

Hamiltonian.Essentials.Spatial.Link — Type.

Link(neighbor::Int,sindex::Int,eindex::Int,disp::AbstractVector{<:Real})

A link in a lattice.

source

Hamiltonian.Essentials.Spatial.PID — Type.

PID(scope,site::Int)
PID(;scope="tz",site::Int=1)

The id of a point.

source

Hamiltonian.Essentials.Spatial.Point — Type.

Point(pid::PID,rcoord::SVector{N,<:Real},icoord::SVector{N,<:Real}) where N
Point(pid::PID,rcoord::NTuple{N,<:Real},icoord::NTuple{N,<:Real}=ntuple(i->0.0,N)) where N
Point(pid::PID,rcoord::AbstractVector{<:Real},icoord::AbstractVector{<:Real}=zero(SVector{length(rcoord),Float}))

Labeled point.

source

Hamiltonian.Essentials.Spatial.PointIndex — Type.

PointIndex(index::Union{<:PID,Int})

Index for getting a point of a lattice.

source

Hamiltonian.Essentials.Spatial.RCoordIndex — Type.

RCoordIndex(index::Union{<:PID,Int})

Index for getting a rcoord of a lattice.

source

Hamiltonian.Essentials.Spatial.SuperLattice — Type.

SuperLattice(   name::String,
                sublattices::AbstractVector{<:AbstractLattice};
                vectors::AbstractVector{<:AbstractVector{<:Real}}=SVector{0,SVector{sublattices|>eltype|>dimension,Float}}(),
                neighbors::Dict{Int,<:Real}=Dict{Int,Float}()
                )

SuperLattice that is composed of serveral sublattices.

source

Hamiltonian.Essentials.Spatial.@lattice — Macro.

@lattice structdef::Expr

Decorate a raw struct to be a subtype of AbstractLattice.

Notes

Here, a "raw" struct means:

It has no explicit supertype except Any;

It has none of the required attributes by AbstractLattice, i.e. :name, :pids, :rcoords, :icoords, :vectors, :reciprocals and :neighbors;

It has no inner constructor.

source

Hamiltonian.Essentials.Spatial.azimuth — Method.

azimuth(v::AbstractVector{<:Real}) -> Float

Get the azimuth angle in radians of a vector.

source

Hamiltonian.Essentials.Spatial.azimuthd — Method.

azimuthd(v::AbstractVector{<:Real}) -> Float

Get the azimuth angle in degrees of a vector.

source

Hamiltonian.Essentials.Spatial.bonds — Method.

bonds(lattice::AbstractLattice) -> Vector{Bond}
bonds(lattice::AbstractLattice,::InsideBonds) -> Vector{Bond}
bonds(lattice::AbstractLattice,::AcrossBonds) -> Vector{Bond}

Get the bonds of a lattice.

source

Hamiltonian.Essentials.Spatial.bonds — Method.

bonds(lattice::SuperLattice,::IntraBonds) -> Vector{Bond}
bonds(lattice::SuperLattice,::InterBonds) -> Vector{Bond}

Get the bonds of a superlattice.

source

Hamiltonian.Essentials.Spatial.distance — Method.

distance(p1::AbstractVector{<:Real},p2::AbstractVector{<:Real}) -> Float

Get the distance between two points.

Note

Compared to norm(p1-p2), this function avoids the memory allocation for p1-p2, thus is more efficient.

source

Hamiltonian.Essentials.Spatial.icoord — Method.

icoord(bond::Bond) -> SVector

Get the icoord of the bond.

source

Hamiltonian.Essentials.Spatial.interlinks — Method.

interlinks(cluster1::AbstractMatrix{<:Real},cluster2::AbstractMatrix{<:Real},neighbors::Dict{Int,Float}) -> Vector{Link}

Use kdtree to get the intercluster nearest neighbors.

source

Hamiltonian.Essentials.Spatial.intralinks — Method.

intralinks( cluster::AbstractMatrix{<:Real},
            vectors::AbstractVector{<:AbstractVector{<:Real}},
            neighbors::Dict{Int,Float},
            maxtranslations::NTuple{N,Int}=ntuple(i->length(neighbors)-1,length(vectors))
            ) where N -> Vector{Link}

Use kdtree to get the intracluster nearest neighbors.

As is similar to minimumlengths, when vectors is nonempty, the cluster assumes periodic boundaries. neighbors provides the map between the bond length and the order of nearest neighbors. Note only those with the lengths present in neighbors will be included in the result. maxtranslations determines the maximum number of translations along those directions specified by vectors when the tiled supercluster is construted (See minimumlengths for the explanation of the method for periodic lattices).

source

Hamiltonian.Essentials.Spatial.isintracell — Method.

isintracell(bond::Bond) -> Bool

Judge whether a bond is intra the unit cell of a lattice.

source

Hamiltonian.Essentials.Spatial.isintratriangle — Method.

isintratriangle(p::AbstractVector{<:Real},
                p1::AbstractVector{<:Real},
                p2::AbstractVector{<:Real},
                p3::AbstractVector{<:Real};
                vertexes::NTuple{3,Bool}=(true,true,true),
                edges::NTuple{3,Bool}=(true,true,true),
                atol::Real=atol,
                rtol::Real=rtol
                ) -> Bool

Judge whether a point belongs to the interior of a triangle whose vertexes are p1, 'p2' and p3 with the give tolerance. vertexes and edges define whether the interior should contain the vertexes or edges, respectively.

Notes

The vertexes are in the order (p1,p2,p3) and the edges are in the order (p1p2,p2p3,p3p1).

The edges do not contain the vertexes.

source

Hamiltonian.Essentials.Spatial.isonline — Method.

isonline(p::AbstractVector{<:Real},p1::AbstractVector{<:Real},p2::AbstractVector{<:Real};ends::Tuple{Bool,Bool}=(true,true),atol::Real=atol,rtol::Real=rtol) -> Bool

Judge whether a point is on a line segment whose end points are p1 and p2 with the given tolerance. ends defines whether the line segment should contain its ends.

source

Hamiltonian.Essentials.Spatial.isparallel — Method.

isparallel(v1::AbstractVector{<:Real},v2::AbstractVector{<:Real};atol::Real=atol,rtol::Real=rtol) -> Bool

Judge whether two vectors are parallel to each other with the given tolerance, 0 for not parallel, 1 for parallel and -1 for antiparallel.

source

Hamiltonian.Essentials.Spatial.issubordinate — Method.

issubordinate(rcoord::AbstractVector{<:Real},vectors::AbstractVector{<:AbstractVector{<:Real}};atol::Real=atol,rtol::Real=rtol) -> Bool

Judge whether a coordinate belongs to a lattice defined by vectors with the given tolerance.

source

Hamiltonian.Essentials.Spatial.minimumlengths — Function.

minimumlengths(cluster::AbstractMatrix{<:Real},vectors::AbstractVector{<:AbstractVector{<:Real}},nneighbor::Int=1;coordination::Int=8) -> Vector{Float}

Use kdtree to search the lowest several minimum bond lengths within a lattice translated by a cluster.

When the translation vectors are not empty, the lattice will be considered periodic in the corresponding directions. Otherwise the lattice will be open in all directions. To search for the bonds accorss the periodic boundaries, the cluster will be pretranslated to become a supercluster, which has open boundaries but is large enough to contain all the nearest neighbors within the required order. The coordination parameter sets the average number of each order of nearest neighbors. If it is to small, larger bond lengths may not be searched, and the result will contain Inf. This is a sign that you may need a larger coordination. Another situation that Inf appears in the result occurs when the minimum lengths are searched in open lattices. Indeed, the cluster may be too small so that the required order just goes beyond it. In this case the warning message can be safely ignored.

source

Hamiltonian.Essentials.Spatial.nneighbor — Method.

nneighbor(lattice::AbstractLattice) -> Int

Get the highest order of nearest neighbors.

source

Hamiltonian.Essentials.Spatial.points — Method.

points(lattice::AbstractLattice) -> Vector{Point}

Get the points contained in a lattice.

source

Hamiltonian.Essentials.Spatial.polar — Method.

polar(v::AbstractVector{<:Real}) -> Float

Get the polar angle in radians of a vector.

source

Hamiltonian.Essentials.Spatial.polard — Method.

polard(v::AbstractVector{<:Real}) -> Float

Get the polar angle in degrees of a vector.

source

Hamiltonian.Essentials.Spatial.rcoord — Method.

rcoord(bond::Bond) -> SVector

Get the rcoord of the bond.

source

Hamiltonian.Essentials.Spatial.reciprocals — Method.

reciprocals(vectors::AbstractVector{AbstractVector{<:Real}}) -> Vector{Vector{Float}}

Get the reciprocals dual to the input vectors.

source

Hamiltonian.Essentials.Spatial.rotate — Method.

rotate(cluster::AbstractMatrix{<:Real},angle::Real;axis::Tuple{Union{AbstractVector{<:Real},Nothing},Tuple{<:Real,<:Real}}=(nothing,(0,0))) -> Matrix{Float}

Get a rotated cluster of the original one by a certain angle around an axis.

The axis is determined by a point it gets through (nothing can be used to denote the origin), and its polar as well as azimuth angles in radians. The default axis is the z axis.

Notes

The result is given by the Rodrigues' rotation formula.

Only 2 and 3 dimensional vectors can be rotated.

When the input vectors are 2 dimensional, both the polar and azimuth of the axis must be 0.

source

Hamiltonian.Essentials.Spatial.tile — Method.

tile(cluster::AbstractMatrix{<:Real},vectors::AbstractVector{<:AbstractVector{<:Real}},translations::NTuple{M,NTuple{N,<:Real}}=()) where {N,M} -> Matrix{Float}

Tile a supercluster by translations of the input cluster.

Basically, the final supercluster is composed of several parts, each of which is a translation of the original cluster, with the translation vectors specified by vectors and each set of the translation indices contained in translations. When translation vectors are empty, a copy of the original cluster will be returned.

source

Hamiltonian.Essentials.Spatial.translate — Method.

translate(cluster::AbstractMatrix{<:Real},vector::AbstractVector{<:Real}) -> Matrix{vector|>eltype}

Get the translated cluster of the original one by a vector.

source

Hamiltonian.Essentials.Spatial.volume — Method.

volume(v1::AbstractVector{<:Real},v2::AbstractVector{<:Real},v3::AbstractVector{<:Real}) -> Real

Get the volume spanned by three vectors.

source

Hamiltonian.Utilities.dimension — Method.

dimension(lattice::AbstractLattice) -> Int
dimension(::Type{<:AbstractLattice{P,N}}) where {P,N}-> Int

Get the space dimension of the lattice.

source

Hamiltonian.Utilities.dimension — Method.

dimension(p::Point) -> Int
dimension(::Type{<:Point{P,N}}) where {P,N} -> Int

Get the space dimension of the point.

source

Base.:== — Method.

==(b1::Bond,b2::Bond) -> Bool

Overloaded == operator.

source

Base.:== — Method.

==(l1::Link,l2::Link) -> Bool

Overloaded == operator.

source

Base.:== — Method.

==(p1::Point,p2::Point) -> Bool

Overloaded == operator.

source

Base.getindex — Method.

getindex(lattice::AbstractLattice,i::RCoordIndex) -> SVector
getindex(lattice::AbstractLattice,i::ICoordIndex) -> SVector
getindex(lattice::AbstractLattice,i::PointIndex) -> Point

Get a rcoord, an icoord or a point of a lattice according to the type of the input index.

source

Base.insert! — Method.

insert!(cylinder::Cylinder,ps::S...;cut::Int=length(cylinder)÷2+1,scopes::Union{<:AbstractVector{S},Nothing}=nothing,coordination::Int=9) where S -> Cylinder

Insert a couple of blocks into a cylinder.

The position of the cut of the cylinder is specified by the keyword argument cut, which is the center of the cylinder by default. All pids corresponding to a same newly inserted block share the same scope, which is specified by the parameter ps. Optionally, the scopes of the old pids in the cylinder can be replaced if the parameter scopes is assigned other than nothing. Note the length of ps is equal to the number of newly inserted blocks, while that of scopes should be equal to the old length of the cylinder.

source

Base.keytype — Method.

keytype(lattice::AbstractLattice)
keytype(::Type{<:AbstractLattice{P,N}}) where {P,N}

Get the pid type of the lattice.

source

Base.length — Method.

length(lattice::AbstractLattice) -> Int

Get the number of points contained in a lattice.

source

Base.reverse — Method.

reverse(bond::Bond) -> Bond

Get the reversed bond.

source

Base.show — Method.

show(io::IO,lattice::AbstractLattice)

Show a lattice.

source

Base.show — Method.

show(io::IO,bond::Bond)

Show a bond.

source

Base.show — Method.

show(io::IO,link::Link)

Show a link.

source

Base.show — Method.

show(io::IO,p::Point)

Show a labeled point.

source

Base.valtype — Method.

valtype(lattice::AbstractLattice)
valtype(::Type{<:AbstractLattice{P,N}}) where {P,N}

Get the point type of the lattice.

source