Spatials

Utilities

AbstractBond

Point

Bond

AbstractLattice

Lattice

SuperLattice

Cylinder

Bonds

Manual

QuantumLattices.Essentials.Spatials.acrossbonds — Constant

acrossbonds

Indicate that bonds across the unitcell are inquired, which are in fact those across the periodic boundaries.

QuantumLattices.Essentials.Spatials.allbonds — Constant

allbonds

Indicate that all bonds are inquired.

QuantumLattices.Essentials.Spatials.insidebonds — Constant

insidebonds

Indicate that bonds inside the unitcell are inquired, which do not contain those across the periodic boundaries.

QuantumLattices.Essentials.Spatials.interbonds — Constant

interbonds

Indicate that bonds inter the sublattices are inquired.

Note

These bonds do not contain those accorss the periodic boundaries.

QuantumLattices.Essentials.Spatials.intrabonds — Constant

intrabonds

Indicate that bonds intra the sublattices are inquired.

Note

These bonds do not contain those accorss the periodic boundaries.

QuantumLattices.Essentials.Spatials.zerothbonds — Constant

zerothbonds

Indicate that zeroth bonds, i.e. the points are inquired.

QuantumLattices.Essentials.Spatials.AbstractBond — Type

AbstractBond{N, P<:PID, R}

Abstract bond.

QuantumLattices.Essentials.Spatials.AbstractLattice — Type

AbstractLattice{P<:PID, N}

Abstract type for all lattices.

It should have the following attributes

name::String: the name of the lattice

pids::Vector{P}: the pids of the lattice

rcoords::Matrix{Float}: the rcoords of the lattice

icoords::Matrix{Float}: the icoords of the lattice

vectors::Vector{SVector{N, Float}}: the translation vectors of the lattice

reciprocals::Vector{SVector{N, Float}}: the reciprocals of the lattice

neighbors::Dict{Int, Float}: the order-distance map of the nearest neighbors of the lattice

QuantumLattices.Essentials.Spatials.Bond — Type

Bond(neighbor::Int, spoint::Point, epoint::Point)

A bond in a lattice.

QuantumLattices.Essentials.Spatials.Bonds — Type

Bonds{T, L}(bonds::Tuple{Vararg{Vector{<:AbstractBond}}}) where {T, L<:AbstractLattice}
Bonds(lattice::AbstractLattice, types::LatticeBonds...)

A set of lattice bonds.

Bonds itself is an AbstractVector of AbstractBond. The need for such a struct is to ensure the type stability during the iteration over a set of different concrete bonds. Although the default iterate function does not achieve this goal, users can get it with the generated function trick. Besides, it provides a high level of management of different categories of bonds based on the LatticeBonds system.

QuantumLattices.Essentials.Spatials.Cylinder — Type

Cylinder{P}(name::String,
            block::AbstractMatrix{<:Real},
            translation::SVector{N, <:Real};
            vector::Union{AbstractVector{<:Real}, Nothing}=nothing,
            neighbors::Union{Dict{Int, <:Real}, Int}=1
            ) where {P<:PID, N}

Cylinder of 1d and quasi 2d lattices.

QuantumLattices.Essentials.Spatials.Cylinder — Method

(cylinder::Cylinder)(scopes::Any...; coordination::Int=8) -> Lattice

Construct a lattice from a cylinder with the assigned scopes.

QuantumLattices.Essentials.Spatials.Lattice — Type

Lattice{N}( name::String,
            pids::Vector{<:PID},
            rcoords::AbstractMatrix{<:Real},
            icoords::AbstractMatrix{<:Real},
            vectors::AbstractVector{<:AbstractVector{<:Real}},
            neighbors::Union{Dict{Int, <:Real}, Int}=1;
            coordination::Int=8
            ) where N
Lattice(    name::String,
            points::AbstractVector{<:Point};
            vectors::AbstractVector{<:AbstractVector{<:Real}}=SVector{0, SVector{points|>eltype|>dimension, Float}}(),
            neighbors::Union{Dict{Int, <:Real}, Int}=1,
            coordination::Int=8
            )
Lattice(    name::String,
            sublattices::AbstractVector{<:AbstractLattice};
            vectors::AbstractVector{<:AbstractVector{<:Real}}=SVector{0, SVector{sublattices|>eltype|>dimension, Float}}(),
            neighbors::Union{Dict{Int, <:Real}, Int}=1,
            coordination::Int=8
            )

Simplest lattice.

A simplest lattice can be construted from its contents, i.e. pids, rcoords and icoords, or from a couple of points, or from a couple of sublattices.

QuantumLattices.Essentials.Spatials.LatticeIndex — Type

LatticeIndex{Kind}(index::Union{PID, Int}) where Kind

Lattice index.

Kind must be one of the followings:

'R': for getting the rcoord of a lattice

'I': for getting the icoord of a lattice

'P': for getting the point of a lattice

QuantumLattices.Essentials.Spatials.PID — Type

PID(scope, site::Int)
PID(site::Int)
PID(; scope="tz", site::Int=1)

The id of a point.

QuantumLattices.Essentials.Spatials.Point — Type

Point(pid::PID, rcoord::SVector{N, <:Real}, icoord::SVector{N, <:Real}) where N
Point(pid::PID, rcoord::NTuple{N, <:Real}, icoord::NTuple{N, <:Real}=ntuple(i->0.0, N)) where N
Point(pid::PID, rcoord::AbstractVector{<:Real}, icoord::AbstractVector{<:Real}=zero(SVector{length(rcoord), Float}))
Point{N}(pid::PID, rcoord::AbstractVector{<:Real}, icoord::AbstractVector{<:Real}=zero(SVector{N, Float})) where N

Labeled point.

QuantumLattices.Essentials.Spatials.SuperLattice — Type

SuperLattice(   name::String,
                sublattices::AbstractVector{<:AbstractLattice};
                vectors::AbstractVector{<:AbstractVector{<:Real}}=SVector{0, SVector{sublattices|>eltype|>dimension, Float}}(),
                neighbors::Dict{Int, <:Real}=Dict{Int, Float}()
                )

SuperLattice that is composed of serveral sublattices.

Base.:== — Method

==(lattice1::AbstractLattice, lattice2::AbstractLattice) -> Bool

Overloaded equivalent operator.

Base.:== — Method

==(bonds1::Bonds, bonds2::Bonds) -> Bool

Overloaded equivalent operator.

Base.:== — Method

==(b1::AbstractBond{N, P, R}, b2::AbstractBond{N, P, R}) where {N, P<:PID, R} -> Bool

Overloaded equivalent operator.

Base.eltype — Method

eltype(bond::AbstractBond)
eltype(::Type{<:AbstractBond{N, P}}) where {N, P<:PID}

Get the eltype of a bond.

Base.empty! — Method

empty!(bs::Bonds) -> Bonds

Empty a set of lattice bonds.

Base.empty — Method

empty(bs::Bonds) -> Bonds

Get an empty copy of a set of lattice bonds.

Base.filter — Function

filter(lbs::LatticeBonds, bs::Bonds, choice::Union{Val{:include}, Val{:exclude}}=Val(:include)) -> Bonds
filter(lbs::Tuple{Vararg{LatticeBonds}}, bs::Bonds, choice::Union{Val{:include}, Val{:exclude}}=Val(:include)) -> Bonds

Get a subset of a set of lattice bonds.

When choice=Val(:include), the lattice bonds indicated by lbs will be selected; When choice=Val(:exclude), the lattice bonds not indicated by lbs will be selected.

Base.filter — Method

filter(select::Function, bs::Bonds) -> Bonds

Get a filtered set of bonds by a select function.

Base.getindex — Method

getindex(lattice::AbstractLattice, i::LatticeIndex{'R'}) -> SVector
getindex(lattice::AbstractLattice, i::LatticeIndex{'I'}) -> SVector
getindex(lattice::AbstractLattice, i::LatticeIndex{'P'}) -> Point

Get a rcoord, an icoord or a point of a lattice according to the type of the input index.

Base.getindex — Method

getindex(bs::Bonds, i::Int) -> eltype(bs)

Get the ith bond in the set.

Base.insert! — Method

insert!(cylinder::Cylinder, ps::S...; cut::Int=length(cylinder)÷2+1, scopes::Union{<:AbstractVector{S}, Nothing}=nothing, coordination::Int=9) where S -> Cylinder

Insert a couple of blocks into a cylinder.

The position of the cut of the cylinder is specified by the keyword argument cut, which is the center of the cylinder by default. All pids corresponding to a same newly inserted block share the same scope, which is specified by the parameter ps. Optionally, the scopes of the old pids in the cylinder can be replaced if the parameter scopes is assigned other than nothing. Note the length of ps is equal to the number of newly inserted blocks, while that of scopes should be equal to the old length of the cylinder.

Base.isequal — Method

isequal(lattice1::AbstractLattice, lattice2::AbstractLattice) -> Bool

Overloaded equivalent function.

Base.isequal — Method

isequal(bonds1::Bonds, bonds2::Bonds) -> Bool

Overloaded equivalent function.

Base.isequal — Method

isequal(b1::AbstractBond{N, P, R}, b2::AbstractBond{N, P, R}) where {N, P<:PID, R} -> Bool

Overloaded equivalent function.

Base.iterate — Function

iterate(bs::Bonds, state=(1, 0))

Iterate over the lattice bonds in the set.

Base.iterate — Function

iterate(p::Point, state=1)

Iterate over the point.

Base.iterate — Function

iterate(bond::Bond, state=1)

Iterate over the points in a bond.

Base.keytype — Method

keytype(lattice::AbstractLattice)
keytype(::Type{<:AbstractLattice{N, P} where N}) where {P<:PID}

Get the pid type of the lattice.

Base.length — Method

length(bond::AbstractBond) -> Int
length(::Type{<:AbstractBond{N, <:PID, R} where N}) where R -> Int

Get the number of points of a bond.

Base.length — Method

length(lattice::AbstractLattice) -> Int

Get the number of points contained in a lattice.

Base.reverse — Method

reverse(bond::Bond) -> Bond

Get the reversed bond.

Base.show — Method

show(io::IO, lattice::AbstractLattice)

Show a lattice.

Base.show — Method

show(io::IO, bond::Bond)

Show a bond.

Base.show — Method

show(io::IO, p::Point)

Show a labeled point.

Base.size — Method

size(bs::Bonds) -> Tuple{Int}

Get the size of the set of lattice bonds.

Base.summary — Method

summary(io::IO, bs::Bonds)

Print the brief description of a set of lattice bonds to an io.

Base.valtype — Method

valtype(lattice::AbstractLattice)
valtype(::Type{<:AbstractLattice{N, P}}) where {N, P<:PID}

Get the point type of the lattice.

QuantumLattices.Essentials.Spatials.azimuth — Method

azimuth(v::AbstractVector{<:Real}) -> Float

Get the azimuth angle in radians of a vector.

QuantumLattices.Essentials.Spatials.azimuthd — Method

azimuthd(v::AbstractVector{<:Real}) -> Float

Get the azimuth angle in degrees of a vector.

QuantumLattices.Essentials.Spatials.bonds! — Method

bonds!(bonds::Vector, lattice::AbstractLattice, inquiries::LatticeBonds...) -> Vector
bonds!(bonds::Vector, lattice::AbstractLattice, ::Val{zerothbonds}) -> Vector
bonds!(bonds::Vector, lattice::AbstractLattice, ::Val{insidebonds}) -> Vector
bonds!(bonds::Vector, lattice::AbstractLattice, ::Val{acrossbonds}) -> Vector

Generate the required bonds of a lattice and append them to the input bonds.

QuantumLattices.Essentials.Spatials.bonds! — Method

bonds!(bonds::Vector, lattice::SuperLattice, ::Val{intrabonds}) -> Vector
bonds!(bonds::Vector, lattice::SuperLattice, ::Val{interbonds}) -> Vector

Generate the required bonds of a superlattice and append them to the input bonds.

QuantumLattices.Essentials.Spatials.bonds — Method

bonds(lattice::AbstractLattice, inquiry=allbonds) -> Vector{eltype(lattice|>typeof, inquiry)}
bonds(lattice::AbstractLattice, inquiries...) -> Vector{mapreduce(inquiry->eltype(lattice|>typeof, inquiry), typejoin, inquiries)}

Generate the required bonds of a lattice.

QuantumLattices.Essentials.Spatials.bondtypes — Method

bondtypes(bs::Bonds) -> Tuple{Vararg{LatticeBonds}}
bondtypes(::Type{<:Bonds{T}}) where T -> Tuple{Vararg{LatticeBonds}}

Get the bondtypes of a set of lattice bonds.

QuantumLattices.Essentials.Spatials.distance — Method

distance(p1::AbstractVector{<:Real}, p2::AbstractVector{<:Real}) -> Float

Get the distance between two points.

Note

Compared to norm(p1-p2), this function avoids the memory allocation for p1-p2, thus is more efficient.

QuantumLattices.Essentials.Spatials.icoord — Method

icoord(bond::Bond) -> SVector

Get the icoord of the bond.

QuantumLattices.Essentials.Spatials.interlinks — Method

interlinks(cluster1::AbstractMatrix{<:Real}, cluster2::AbstractMatrix{<:Real}, neighbors::Dict{Int, Float}) -> Vector{Tuple{Int, Int, Int}}

Use kdtree to get the intercluster nearest neighbors.

QuantumLattices.Essentials.Spatials.intralinks — Method

intralinks( cluster::AbstractMatrix{<:Real},
            vectors::AbstractVector{<:AbstractVector{<:Real}},
            neighbors::Dict{Int, Float},
            maxtranslations::NTuple{N, Int}=ntuple(i->length(neighbors), length(vectors))
            ) where N -> Vector{Tuple{Int, Int, Int, SubArray{<:Real, 1}}}

Use kdtree to get the intracluster nearest neighbors.

As is similar to minimumlengths, when vectors is nonempty, the cluster assumes periodic boundaries. neighbors provides the map between the bond length and the order of nearest neighbors. Note only those with the lengths present in neighbors will be included in the result. maxtranslations determines the maximum number of translations along those directions specified by vectors when the tiled supercluster is construted (See minimumlengths for the explanation of the method for periodic lattices).

QuantumLattices.Essentials.Spatials.isintracell — Method

isintracell(bond::Bond) -> Bool

Judge whether a bond is intra the unit cell of a lattice.

QuantumLattices.Essentials.Spatials.isintratriangle — Method

isintratriangle(p::AbstractVector{<:Real},
                p1::AbstractVector{<:Real},
                p2::AbstractVector{<:Real},
                p3::AbstractVector{<:Real};
                vertexes::NTuple{3, Bool}=(true, true, true),
                edges::NTuple{3, Bool}=(true, true, true),
                atol::Real=atol,
                rtol::Real=rtol
                ) -> Bool

Judge whether a point belongs to the interior of a triangle whose vertexes are p1, 'p2' and p3 with the give tolerance. vertexes and edges define whether the interior should contain the vertexes or edges, respectively.

Note

The vertexes are in the order (p1, p2, p3) and the edges are in the order (p1p2, p2p3, p3p1).

The edges do not contain the vertexes.

QuantumLattices.Essentials.Spatials.isonline — Method

isonline(p::AbstractVector{<:Real}, p1::AbstractVector{<:Real}, p2::AbstractVector{<:Real}; ends::Tuple{Bool, Bool}=(true, true), atol::Real=atol, rtol::Real=rtol) -> Bool

Judge whether a point is on a line segment whose end points are p1 and p2 with the given tolerance. ends defines whether the line segment should contain its ends.

QuantumLattices.Essentials.Spatials.isparallel — Method

isparallel(v1::AbstractVector{<:Real}, v2::AbstractVector{<:Real}; atol::Real=atol, rtol::Real=rtol) -> Int

Judge whether two vectors are parallel to each other with the given tolerance, 0 for not parallel, 1 for parallel and -1 for antiparallel.

QuantumLattices.Essentials.Spatials.issubordinate — Method

issubordinate(rcoord::AbstractVector{<:Real}, vectors::AbstractVector{<:AbstractVector{<:Real}}; atol::Real=atol, rtol::Real=rtol) -> Bool

Judge whether a coordinate belongs to a lattice defined by vectors with the given tolerance.

QuantumLattices.Essentials.Spatials.latticebondsstructure — Method

latticebondsstructure(::Type{<:AbstractLattice}) -> SimpleTree{LatticeBonds, Nothing}

The tree structure of the lattice bonds.

QuantumLattices.Essentials.Spatials.latticebondsstructure — Method

latticebondsstructure(::Type{<:SuperLattice}) -> SimpleTree{LatticeBonds, Nothing}

The tree structure of the lattice bonds.

QuantumLattices.Essentials.Spatials.latticetype — Method

latticetype(bs::Bonds)
latticetype(::Type{<:Bonds{T, L} where T}) where {L<:AbstractLattice}

QuantumLattices.Essentials.Spatials.latticetype — Method

latticetype(sl::SuperLattice)
latticetype(::Type{<:SuperLattice{L}}) where {L<:AbstractLattice}

Get the sublattice type of a superlattice.

QuantumLattices.Essentials.Spatials.minimumlengths — Function

minimumlengths(cluster::AbstractMatrix{<:Real}, vectors::AbstractVector{<:AbstractVector{<:Real}}, nneighbor::Int=1; coordination::Int=8) -> Vector{Float}

Use kdtree to search the lowest several minimum bond lengths within a lattice translated by a cluster.

When the translation vectors are not empty, the lattice will be considered periodic in the corresponding directions. Otherwise the lattice will be open in all directions. To search for the bonds accorss the periodic boundaries, the cluster will be pretranslated to become a supercluster, which has open boundaries but is large enough to contain all the nearest neighbors within the required order. The coordination parameter sets the average number of each order of nearest neighbors. If it is to small, larger bond lengths may not be searched, and the result will contain Inf. This is a sign that you may need a larger coordination. Another situation that Inf appears in the result occurs when the minimum lengths are searched in open lattices. Indeed, the cluster may be too small so that the required order just goes beyond it. In this case the warning message can be safely ignored.

QuantumLattices.Essentials.Spatials.nneighbor — Method

nneighbor(lattice::AbstractLattice) -> Int

Get the highest order of nearest neighbors.

QuantumLattices.Essentials.Spatials.pidtype — Method

pidtype(bond::AbstractBond)
pidtype(::Type{<:AbstractBond{N, P} where N}) where {P<:PID}

Get the pid type of a concrete bond.

QuantumLattices.Essentials.Spatials.polar — Method

polar(v::AbstractVector{<:Real}) -> Float

Get the polar angle in radians of a vector.

QuantumLattices.Essentials.Spatials.polard — Method

polard(v::AbstractVector{<:Real}) -> Float

Get the polar angle in degrees of a vector.

QuantumLattices.Essentials.Spatials.rcoord — Method

rcoord(bond::Bond) -> SVector

Get the rcoord of the bond.

QuantumLattices.Essentials.Spatials.reciprocals — Method

reciprocals(vectors::AbstractVector{AbstractVector{<:Real}}) -> Vector{Vector{Float}}

Get the reciprocals dual to the input vectors.

QuantumLattices.Essentials.Spatials.rotate — Method

rotate(cluster::AbstractMatrix{<:Real}, angle::Real; axis::Tuple{Union{AbstractVector{<:Real}, Nothing}, Tuple{<:Real, <:Real}}=(nothing, (0, 0))) -> Matrix{Float}

Get a rotated cluster of the original one by a certain angle around an axis.

The axis is determined by a point it gets through (nothing can be used to denote the origin), and its polar as well as azimuth angles in radians. The default axis is the z axis.

Note

The result is given by the Rodrigues' rotation formula.

Only 2 and 3 dimensional vectors can be rotated.

When the input vectors are 2 dimensional, both the polar and azimuth of the axis must be 0.

QuantumLattices.Essentials.Spatials.tile — Method

tile(cluster::AbstractMatrix{<:Real}, vectors::AbstractVector{<:AbstractVector{<:Real}}, translations::NTuple{M, NTuple{N, <:Real}}=()) where {N, M} -> Matrix{Float}

Tile a supercluster by translations of the input cluster.

Basically, the final supercluster is composed of several parts, each of which is a translation of the original cluster, with the translation vectors specified by vectors and each set of the translation indices contained in translations. When translation vectors are empty, a copy of the original cluster will be returned.

QuantumLattices.Essentials.Spatials.translate — Method

translate(cluster::AbstractMatrix{<:Real}, vector::AbstractVector{<:Real}) -> Matrix{vector|>eltype}

Get the translated cluster of the original one by a vector.

QuantumLattices.Essentials.Spatials.volume — Method

volume(v1::AbstractVector{<:Real}, v2::AbstractVector{<:Real}, v3::AbstractVector{<:Real}) -> Real

Get the volume spanned by three vectors.

QuantumLattices.Interfaces.decompose — Method

decompose(v0::AbstractVector{<:Real}, v1::AbstractVector{<:Real}) -> Tuple{Float}
decompose(v0::AbstractVector{<:Real}, v1::AbstractVector{<:Real}, v2::AbstractVector{<:Real}) -> Tuple{Float, Float}
decompose(v0::AbstractVector{<:Real}, v1::AbstractVector{<:Real}, v2::AbstractVector{<:Real}, v3::AbstractVector{<:Real}) -> Tuple{Float, Float, Float}

Decompose a vector with respect to input basis vectors.

QuantumLattices.Interfaces.dimension — Method

dimension(bond::AbstractBond) -> Int
dimension(::Type{<:AbstractBond{N}}) where N -> Int

Get the space dimension of a concrete bond.

QuantumLattices.Interfaces.dimension — Method

dimension(lattice::AbstractLattice) -> Int
dimension(::Type{<:AbstractLattice{N}}) where N -> Int

Get the space dimension of the lattice.

QuantumLattices.Interfaces.expand — Method

expand(::Type{L}, ::Val{VS}) where {L<:AbstractLattice, VS} -> Tuple

Expand the lattice bond types to the leaf level.

QuantumLattices.Interfaces.kind — Method

kind(bond::Bond) -> Int

Get the bond kind of a bond.

QuantumLattices.Interfaces.kind — Method

kind(::Point) -> 0
kind(::Type{<:Point}) -> 0

Get the bond kind of a point, which is defined to be 0.

QuantumLattices.Interfaces.rank — Method

rank(bond::AbstractBond) -> Int
rank(::Type{<:AbstractBond{N, <:PID, R} where N}) where R -> Int

Get the rank of a bond.

QuantumLattices.Interfaces.rank — Method

rank(bs::Bonds) -> Int
rank(::Type{<:Bonds{T}}) where T -> Int

Get the rank of a set of lattice bonds.

QuantumLattices.Interfaces.reset! — Method

reset!(bs::Bonds, lattice::AbstractLattice) -> Bonds

Reset a set of lattice bonds by a new lattice.