Spatials

Utilities

AbstractBond

Point

Bond

AbstractLattice

Lattice

SuperLattice

Cylinder

Bonds

Manual

QuantumLattices.Essentials.Spatials.acrossbonds — Constant

acrossbonds

Indicate that bonds across the unitcell are inquired, which are in fact those across the periodic boundaries.

QuantumLattices.Essentials.Spatials.allbonds — Constant

allbonds

Indicate that all bonds are inquired.

QuantumLattices.Essentials.Spatials.insidebonds — Constant

insidebonds

Indicate that bonds inside the unitcell are inquired, which do not contain those across the periodic boundaries.

QuantumLattices.Essentials.Spatials.interbonds — Constant

interbonds

Indicate that bonds inter the sublattices are inquired.

Note

These bonds do not contain those across the periodic boundaries.

QuantumLattices.Essentials.Spatials.intrabonds — Constant

intrabonds

Indicate that bonds intra the sublattices are inquired.

Note

These bonds do not contain those across the periodic boundaries.

QuantumLattices.Essentials.Spatials.zerothbonds — Constant

zerothbonds

Indicate that zeroth bonds, i.e. the points are inquired.

QuantumLattices.Essentials.Spatials.AbstractBond — Type

AbstractBond{N, P<:AbstractPID, D<:Number, R}

Abstract bond.

QuantumLattices.Essentials.Spatials.AbstractLattice — Type

AbstractLattice{N, P<:AbstractPID, D<:Number}

Abstract type for all lattices.

It should have the following contents:

name::String: the name of the lattice

pids::Vector{P}: the pids of the lattice

rcoords::Matrix{D}: the rcoords of the lattice

icoords::Matrix{D}: the icoords of the lattice

vectors::Vector{SVector{N, D}}: the translation vectors of the lattice

reciprocals::Vector{SVector{N, D}}: the reciprocals of the lattice

neighbors::Dict{Int, Float}: the order-distance map of the nearest neighbors of the lattice

QuantumLattices.Essentials.Spatials.AbstractPID — Type

AbstractPID <: SingularID

Abstract point id.

QuantumLattices.Essentials.Spatials.Bond — Type

Bond{N, P<:AbstractPID, D<:Number} <: AbstractBond{N, P, D, 2}

A bond in a lattice.

QuantumLattices.Essentials.Spatials.Bonds — Type

Bonds{T, L<:AbstractLattice, BS<:Tuple{Vararg{Vector{<:AbstractBond}}}, B<:AbstractBond} <: AbstractVector{B}

A set of lattice bonds.

Bonds itself is an AbstractVector of AbstractBond. The need for such a struct is to ensure the type stability during the iteration over a set of different concrete bonds. Although the default iterate function does not achieve this goal, users can get it with the generated function trick. Besides, it provides a high level of management of different categories of bonds based on the LatticeBonds system.

QuantumLattices.Essentials.Spatials.Bonds — Method

Bonds{T, L}(bonds::Tuple{Vararg{Vector{<:AbstractBond}}}) where {T, L<:AbstractLattice}
Bonds(lattice::AbstractLattice, types::LatticeBonds...)

Construct a set of lattice bonds.

QuantumLattices.Essentials.Spatials.BrillouinZone — Type

BrillouinZone{K, P<:Momentum, S<:SVector} <: ReciprocalSpace{K, P}

The Brillouin zone of a lattice.

QuantumLattices.Essentials.Spatials.BrillouinZone — Method

BrillouinZone(reciprocals::AbstractVector, nk::Int)
BrillouinZone(::Type{P}, reciprocals::AbstractVector) where {P<:Momentum}
BrillouinZone(reciprocals::AbstractVector, momenta::AbelianNumbers{<:Momentum})

BrillouinZone{K}(reciprocals::AbstractVector, nk::Int) where K
BrillouinZone{K}(::Type{P}, reciprocals::AbstractVector) where {K, P<:Momentum}
BrillouinZone{K}(reciprocals::AbstractVector, momenta::AbelianNumbers{<:Momentum}) where K

Construct a Brillouin zone.

QuantumLattices.Essentials.Spatials.CPID — Type

CPID{S} <: AbstractPID

Composite point id.

QuantumLattices.Essentials.Spatials.CPID — Method

CPID(site::Int)
CPID(; scope='T', site::Int=1)

Construct a composite point id.

QuantumLattices.Essentials.Spatials.Cylinder — Type

Cylinder{P}(name::String, block::AbstractMatrix{<:Real}, translation::SVector{N, <:Real};
    vector::Union{AbstractVector{<:Real}, Nothing}=nothing,
    neighbors::Union{Dict{Int, <:Real}, Int}=1
    ) where {P<:CPID, N}

Cylinder of 1d and quasi 2d lattices.

QuantumLattices.Essentials.Spatials.Cylinder — Method

(cylinder::Cylinder)(scopes::Any...; coordination::Int=8) -> Lattice

Construct a lattice from a cylinder with the assigned scopes.

QuantumLattices.Essentials.Spatials.Lattice — Type

Lattice{N, P<:AbstractPID, D<:Number} <: AbstractLattice{N, P, D}

Simplest lattice.

A simplest lattice can be constructed from its contents, i.e. pids, rcoords and icoords, or from a couple of points, or from a couple of sublattices.

QuantumLattices.Essentials.Spatials.Lattice — Method

Lattice{N}(name::String,
    pids::Vector{<:AbstractPID}, rcoords::AbstractMatrix{<:Number}, icoords::AbstractMatrix{<:Number},
    vectors::AbstractVector{<:AbstractVector{<:Number}},
    neighbors::Union{Dict{Int, <:Real}, Int}=1;
    coordination::Int=8
    ) where N
Lattice(name::String,
    points::AbstractVector{<:Point};
    vectors::AbstractVector{<:AbstractVector{<:Number}}=SVector{0, SVector{points|>eltype|>dimension, points|>eltype|>dtype}}(),
    neighbors::Union{Dict{Int, <:Real}, Int}=1,
    coordination::Int=8
    )
Lattice(name::String,
    sublattices::AbstractVector{<:AbstractLattice};
    vectors::AbstractVector{<:AbstractVector{<:Real}}=SVector{0, SVector{sublattices|>eltype|>dimension, sublattices|>eltype|>dtype}}(),
    neighbors::Union{Dict{Int, <:Real}, Int}=1,
    coordination::Int=8
    )

Construct a lattice.

QuantumLattices.Essentials.Spatials.LatticeIndex — Type

LatticeIndex{Kind}(index::Union{AbstractPID, Int}) where Kind

Lattice index.

Kind must be one of the followings:

'R': for getting the rcoord of a lattice

'I': for getting the icoord of a lattice

'P': for getting the point of a lattice

QuantumLattices.Essentials.Spatials.PID — Type

PID <: AbstractPID

Point id.

QuantumLattices.Essentials.Spatials.Point — Type

Point{N, P<:AbstractPID, D<:Number} <: AbstractBond{N, P, D, 1}

Labeled point.

QuantumLattices.Essentials.Spatials.Point — Method

Point(pid::AbstractPID, rcoord::SVector{N, D}, icoord::SVector{N, D}) where {N, D<:Number}
Point(pid::AbstractPID, rcoord::NTuple{N, <:Number}, icoord::NTuple{N, <:Number}=ntuple(i->0, N)) where N
Point(pid::AbstractPID, rcoord::AbstractVector{<:Number}, icoord::AbstractVector{<:Number}=zero(SVector{length(rcoord), Int}))
Point{N}(pid::AbstractPID, rcoord::AbstractVector{<:Number}, icoord::AbstractVector{<:Number}=zero(SVector{N, Int})) where N

Construct a labeled point.

QuantumLattices.Essentials.Spatials.ReciprocalPath — Type

ReciprocalPath{K, S<:SVector} <: ReciprocalSpace{K, S}

A path in the reciprocal space.

QuantumLattices.Essentials.Spatials.ReciprocalPath — Method

ReciprocalPath(momenta::AbstractVector)
ReciprocalPath(reciprocals::AbstractVector, segments::Segment{<:SVector}...)
ReciprocalPath(reciprocals::AbstractVector, segments::Tuple{Vararg{Segment{<:SVector}}})

ReciprocalPath{K}(momenta::AbstractVector) where K
ReciprocalPath{K}(reciprocals::AbstractVector, segments::Segment{<:SVector}...) where K
ReciprocalPath{K}(reciprocals::AbstractVector, segments::Tuple{Vararg{Segment{<:SVector}}}) where K

Construct a path in the reciprocal space.

QuantumLattices.Essentials.Spatials.ReciprocalPath — Method

ReciprocalPath(reciprocals::AbstractVector, segments::Pair{<:NTuple{N, Number}, <:NTuple{N, Number}}...;
    len::Union{Int, NTuple{M, Int}}=100,
    ends::Union{NTuple{2, Bool}, NTuple{M, NTuple{2, Bool}}}=(true, false)
    ) where {N, M}
ReciprocalPath(reciprocals::AbstractVector, segments::Tuple{Vararg{Pair{<:NTuple{N, Number}, <:NTuple{N, Number}}}};
    len::Union{Int, NTuple{M, Int}}=100,
    ends::Union{NTuple{2, Bool}, NTuple{M, NTuple{2, Bool}}}=(true, false)
    ) where {N, M}

ReciprocalPath{K}(reciprocals::AbstractVector, segments::Pair{<:NTuple{N, Number}, <:NTuple{N, Number}}...;
    len::Union{Int, NTuple{M, Int}}=100,
    ends::Union{NTuple{2, Bool}, NTuple{M, NTuple{2, Bool}}}=(true, false)
    ) where {K, N, M}
ReciprocalPath{K}(reciprocals::AbstractVector, segments::Tuple{Vararg{Pair{<:NTuple{N, Number}, <:NTuple{N, Number}}}};
    len::Union{Int, NTuple{M, Int}}=100,
    ends::Union{NTuple{2, Bool}, NTuple{M, NTuple{2, Bool}}}=(true, false)
    ) where {K, N, M}

Construct a path in the reciprocal space.

QuantumLattices.Essentials.Spatials.ReciprocalSpace — Type

ReciprocalSpace{K, B} <: NamedVectorSpace{:⊗, K, Tuple{B}}

Abstract type of reciprocal spaces.

QuantumLattices.Essentials.Spatials.ReciprocalZone — Type

ReciprocalZone{K, S<:SVector, V<:Number} <: ReciprocalSpace{K, S}

A zone in the reciprocal space.

QuantumLattices.Essentials.Spatials.ReciprocalZone — Method

ReciprocalZone(momenta::AbstractVector, volume::Number)
ReciprocalZone(reciprocals::AbstractVector; len::Int=100)
ReciprocalZone(reciprocals::AbstractVector, bounds::Segment...)
ReciprocalZone(reciprocals::AbstractVector, bounds::Tuple{Vararg{Segment}})

ReciprocalZone{K}(momenta::AbstractVector, volume::Number) where K
ReciprocalZone{K}(reciprocals::AbstractVector; len::Int=100) where K
ReciprocalZone{K}(reciprocals::AbstractVector, bounds::Segment...) where K
ReciprocalZone{K}(reciprocals::AbstractVector, bounds::Tuple{Vararg{Segment}}) where K

Construct a rectangular zone in the reciprocal space.

QuantumLattices.Essentials.Spatials.ReciprocalZone — Method

ReciprocalZone(bz::BrillouinZone)

Construct a reciprocal zone from a Brillouin zone.

QuantumLattices.Essentials.Spatials.Segment — Type

Segment{S} <: AbstractVector{S}

A segment.

QuantumLattices.Essentials.Spatials.Segment — Method

Segment(start::Number, stop::Number, len::Integer; ends::Tuple{Bool, Bool}=(true, false))
Segment(start::AbstractVector, stop::AbstractVector, len::Integer; ends::Tuple{Bool, Bool}=(true, false))
Segment(start::NTuple{N, Number}, stop::NTuple{N, Number}, len::Integer; ends::Tuple{Bool, Bool}=(true, false)) where N

Construct a segment.

QuantumLattices.Essentials.Spatials.SuperLattice — Type

SuperLattice(name::String,
    sublattices::AbstractVector{<:AbstractLattice};
    vectors::AbstractVector{<:AbstractVector{<:Real}}=SVector{0, SVector{sublattices|>eltype|>dimension, sublattices|>eltype|>dtype}}(),
    neighbors::Dict{Int, <:Real}=Dict{Int, Float}()
    )

SuperLattice that is composed of several sublattices.

QuantumLattices.Essentials.Spatials.@hexagon_str — Macro

hexagon"P₁-P₂-P₃-..."

Construct a tuple of start-stop point pairs for the hexagonal reciprocal space.

QuantumLattices.Essentials.Spatials.@line_str — Macro

line"P₁-P₂-P₃-..."

Construct a tuple of start-stop point pairs for the one dimensional reciprocal space.

QuantumLattices.Essentials.Spatials.@rectangle_str — Macro

rectangle"P₁-P₂-P₃-..."

Construct a tuple of start-stop point pairs for the rectangular reciprocal space.

Base.eltype — Method

eltype(bond::AbstractBond)
eltype(::Type{<:AbstractBond{N, P}}) where {N, P<:AbstractPID}

Get the type of the points contained in a bond.

Base.empty! — Method

empty!(bs::Bonds) -> Bonds

Empty a set of lattice bonds.

Base.empty — Method

empty(bs::Bonds) -> Bonds

Get an empty copy of a set of lattice bonds.

Base.filter — Function

filter(lbs::LatticeBonds, bs::Bonds, choice::Union{Val{:include}, Val{:exclude}}=Val(:include)) -> Bonds
filter(lbs::Tuple{Vararg{LatticeBonds}}, bs::Bonds, choice::Union{Val{:include}, Val{:exclude}}=Val(:include)) -> Bonds

Get a subset of a set of lattice bonds.

When choice=Val(:include), the lattice bonds indicated by lbs will be selected; When choice=Val(:exclude), the lattice bonds not indicated by lbs will be selected.

Base.filter — Method

filter(select::Function, bs::Bonds) -> Bonds

Get a filtered set of bonds by a select function.

Base.getindex — Method

getindex(bond::Bond, i::Integer) -> Point

Get the ith point of a bond.

Base.getindex — Method

getindex(p::Point, i::Integer) -> Point

Get the ith point of a point, which by definition i should only be 1 and the result is the point itself.

Base.getindex — Method

getindex(lattice::AbstractLattice, i::LatticeIndex{'R'}) -> SVector
getindex(lattice::AbstractLattice, i::LatticeIndex{'I'}) -> SVector
getindex(lattice::AbstractLattice, i::LatticeIndex{'P'}) -> Point

Get a rcoord, an icoord or a point of a lattice according to the type of the input index.

Base.insert! — Method

insert!(cylinder::Cylinder, ps::S...; cut=length(cylinder)÷2+1, scopes::Union{<:AbstractVector{S}, Nothing}=nothing, coordination=8) where S -> Cylinder

Insert a couple of blocks into a cylinder.

The position of the cut of the cylinder is specified by the keyword argument cut, which is the center of the cylinder by default. All pids corresponding to a same newly inserted block share the same scope, which is specified by the parameter ps. Optionally, the scopes of the old pids in the cylinder can be replaced if the parameter scopes is assigned other than nothing. Note the length of ps is equal to the number of newly inserted blocks, while that of scopes should be equal to the old length of the cylinder.

Base.iterate — Function

iterate(bond::AbstractBond, state=1)

Iterate over the points in a bond.

Base.keytype — Method

keytype(lattice::AbstractLattice)
keytype(::Type{<:AbstractLattice{N, P} where N}) where {P<:AbstractPID}

Get the pid type of the lattice.

Base.length — Method

length(bond::AbstractBond) -> Int
length(::Type{<:AbstractBond{N, <:AbstractPID, <:Number, R} where N}) where R -> Int

Get the number of points of a bond.

Base.length — Method

length(lattice::AbstractLattice) -> Int

Get the number of points contained in a lattice.

Base.reverse — Method

reverse(bond::Bond) -> Bond

Get the reversed bond.

Base.valtype — Method

valtype(lattice::AbstractLattice)
valtype(::Type{<:AbstractLattice{N, P}}) where {N, P<:AbstractPID}

Get the point type of the lattice.

LinearAlgebra.rank — Method

rank(bs::Bonds) -> Int
rank(::Type{<:Bonds{T}}) where T -> Int

Get the rank of a set of lattice bonds.

LinearAlgebra.rank — Method

rank(bond::AbstractBond) -> Int
rank(::Type{<:AbstractBond{N, <:AbstractPID, <:Number, R} where N}) where R -> Int

Get the rank of a bond.

QuantumLattices.Essentials.Spatials.azimuth — Method

azimuth(v::AbstractVector{<:Number}) -> Number

Get the azimuth angle in radians of a vector.

QuantumLattices.Essentials.Spatials.azimuthd — Method

azimuthd(v::AbstractVector{<:Number}) -> Number

Get the azimuth angle in degrees of a vector.

QuantumLattices.Essentials.Spatials.bonds! — Method

bonds!(bonds::Vector, lattice::AbstractLattice, inquiries::LatticeBonds...) -> Vector
bonds!(bonds::Vector, lattice::AbstractLattice, ::Val{zerothbonds}) -> Vector
bonds!(bonds::Vector, lattice::AbstractLattice, ::Val{insidebonds}) -> Vector
bonds!(bonds::Vector, lattice::AbstractLattice, ::Val{acrossbonds}) -> Vector

Generate the required bonds of a lattice and append them to the input bonds.

QuantumLattices.Essentials.Spatials.bonds! — Method

bonds!(bonds::Vector, lattice::SuperLattice, ::Val{intrabonds}) -> Vector
bonds!(bonds::Vector, lattice::SuperLattice, ::Val{interbonds}) -> Vector

Generate the required bonds of a superlattice and append them to the input bonds.

QuantumLattices.Essentials.Spatials.bonds — Method

bonds(lattice::AbstractLattice, inquiry=allbonds) -> Vector{eltype(lattice|>typeof, inquiry)}
bonds(lattice::AbstractLattice, inquiries...) -> Vector{mapreduce(inquiry->eltype(lattice|>typeof, inquiry), typejoin, inquiries)}

Generate the required bonds of a lattice.

QuantumLattices.Essentials.Spatials.bondtypes — Method

bondtypes(bs::Bonds) -> Tuple{Vararg{LatticeBonds}}
bondtypes(::Type{<:Bonds{T}}) where T -> Tuple{Vararg{LatticeBonds}}

Get the bondtypes of a set of lattice bonds.

QuantumLattices.Essentials.Spatials.distance — Method

distance(p₁::AbstractVector{<:Number}, p₂::AbstractVector{<:Number}) -> Number

Get the distance between two points.

Note

Compared to norm(p₁-p₂), this function avoids the memory allocation for p₁-p₂, thus is more efficient.

QuantumLattices.Essentials.Spatials.icoord — Method

icoord(bond::Bond) -> SVector

Get the icoord of the bond.

QuantumLattices.Essentials.Spatials.interlinks — Method

interlinks(cluster₁::AbstractMatrix{<:Number}, cluster₂::AbstractMatrix{<:Number}, neighbors::Dict{Int, <:Number}) -> Vector{Tuple{Int, Int, Int}}

Use kdtree to get the intercluster nearest neighbors.

QuantumLattices.Essentials.Spatials.intralinks — Method

intralinks(cluster::AbstractMatrix{<:Number}, vectors::AbstractVector{<:AbstractVector{<:Number}}, neighbors::Dict{Int, <:Number},
    maxtranslations::NTuple{N, Int}=ntuple(i->length(neighbors), length(vectors))
    ) where N -> Vector{Tuple{Int, Int, Int, Vector{<:Number}}}

Use kdtree to get the intracluster nearest neighbors.

As is similar to minimumlengths, when vectors is nonempty, the cluster assumes periodic boundaries. neighbors provides the map between the bond length and the order of nearest neighbors. Note only those with the lengths present in neighbors will be included in the result. maxtranslations determines the maximum number of translations along those directions specified by vectors when the tiled supercluster is constructed (See minimumlengths for the explanation of the method for periodic lattices).

QuantumLattices.Essentials.Spatials.isintracell — Method

isintracell(bond::Bond) -> Bool

Judge whether a bond is intra the unit cell of a lattice.

QuantumLattices.Essentials.Spatials.isintracell — Method

isintracell(p::Point) -> Bool

Judge whether a point is intra the unitcell.

QuantumLattices.Essentials.Spatials.isintratriangle — Method

isintratriangle(p::AbstractVector{<:Number}, p₁::AbstractVector{<:Number}, p₂::AbstractVector{<:Number}, p₃::AbstractVector{<:Number};
    vertexes::NTuple{3, Bool}=(true, true, true), edges::NTuple{3, Bool}=(true, true, true), atol::Real=atol, rtol::Real=rtol
    ) -> Bool

Judge whether a point belongs to the interior of a triangle whose vertexes are p₁, 'p₂' and p₃ with the give tolerance. vertexes and edges define whether the interior should contain the vertexes or edges, respectively.

Note

The vertexes are in the order (p₁, p₂, p₃) and the edges are in the order (p1p2, p2p3, p3p1).

The edges do not contain the vertexes.

QuantumLattices.Essentials.Spatials.isonline — Method

isonline(p::AbstractVector{<:Number}, p₁::AbstractVector{<:Number}, p₂::AbstractVector{<:Number};
    ends::Tuple{Bool, Bool}=(true, true), atol::Real=atol, rtol::Real=rtol
    ) -> Bool

Judge whether a point is on a line segment whose end points are p₁ and p₂ with the given tolerance. ends defines whether the line segment should contain its ends.

QuantumLattices.Essentials.Spatials.isparallel — Method

isparallel(v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}; atol::Real=atol, rtol::Real=rtol) -> Int

Judge whether two vectors are parallel to each other with the given tolerance, 0 for not parallel, 1 for parallel and -1 for antiparallel.

QuantumLattices.Essentials.Spatials.issubordinate — Method

issubordinate(rcoord::AbstractVector{<:Number}, vectors::AbstractVector{<:AbstractVector{<:Number}}; atol::Real=atol, rtol::Real=rtol) -> Bool

Judge whether a coordinate belongs to a lattice defined by vectors with the given tolerance.

QuantumLattices.Essentials.Spatials.latticebondsstructure — Method

latticebondsstructure(::Type{<:AbstractLattice}) -> SimpleTree{LatticeBonds, Nothing}

The tree structure of the lattice bonds.

QuantumLattices.Essentials.Spatials.latticebondsstructure — Method

latticebondsstructure(::Type{<:SuperLattice}) -> SimpleTree{LatticeBonds, Nothing}

The tree structure of the lattice bonds.

QuantumLattices.Essentials.Spatials.latticetype — Method

latticetype(bs::Bonds)
latticetype(::Type{<:Bonds{T, L} where T}) where {L<:AbstractLattice}

QuantumLattices.Essentials.Spatials.latticetype — Method

latticetype(sl::SuperLattice)
latticetype(::Type{<:SuperLattice{L}}) where {L<:AbstractLattice}

Get the sublattice type of a superlattice.

QuantumLattices.Essentials.Spatials.minimumlengths — Function

minimumlengths(cluster::AbstractMatrix{<:Number}, vectors::AbstractVector{<:AbstractVector{<:Number}}, nneighbor::Int=1; coordination::Int=8) -> Vector{Float}

Use kdtree to search the lowest several minimum bond lengths within a lattice translated by a cluster.

When the translation vectors are not empty, the lattice will be considered periodic in the corresponding directions. Otherwise the lattice will be open in all directions. To search for the bonds across the periodic boundaries, the cluster will be pre-translated to become a supercluster, which has open boundaries but is large enough to contain all the nearest neighbors within the required order. The coordination parameter sets the average number of each order of nearest neighbors. If it is to small, larger bond lengths may not be searched, and the result will contain Inf. This is a sign that you may need a larger coordination. Another situation that Inf appears in the result occurs when the minimum lengths are searched in open lattices. Indeed, the cluster may be too small so that the required order just goes beyond it. In this case the warning message can be safely ignored.

QuantumLattices.Essentials.Spatials.nneighbor — Method

nneighbor(lattice::AbstractLattice) -> Int

Get the highest order of nearest neighbors.

QuantumLattices.Essentials.Spatials.pidtype — Method

pidtype(bond::AbstractBond)
pidtype(::Type{<:AbstractBond{N, P} where N}) where {P<:AbstractPID}

Get the pid type of a concrete bond.

QuantumLattices.Essentials.Spatials.polar — Method

polar(v::AbstractVector{<:Number}) -> Number

Get the polar angle in radians of a vector.

QuantumLattices.Essentials.Spatials.polard — Method

polard(v::AbstractVector{<:Number}) -> Number

Get the polar angle in degrees of a vector.

QuantumLattices.Essentials.Spatials.rcoord — Method

rcoord(bond::Bond) -> SVector

Get the rcoord of the bond.

QuantumLattices.Essentials.Spatials.reciprocals — Method

reciprocals(vectors::AbstractVector{AbstractVector{<:Number}}) -> Vector{Vector{<:Number}}

Get the reciprocals dual to the input vectors.

QuantumLattices.Essentials.Spatials.rotate — Method

rotate(cluster::AbstractMatrix{<:Number}, angle::Number;
    axis::Tuple{Union{AbstractVector{<:Number}, Nothing}, Tuple{<:Number, <:Number}}=(nothing, (0, 0))
    ) -> Matrix{<:Number}

Get a rotated cluster of the original one by a certain angle around an axis.

The axis is determined by a point it gets through (nothing can be used to denote the origin), and its polar as well as azimuth angles in radians. The default axis is the z axis.

Note

The result is given by the Rodrigues' rotation formula.

Only 2 and 3 dimensional vectors can be rotated.

When the input vectors are 2 dimensional, both the polar and azimuth of the axis must be 0.

QuantumLattices.Essentials.Spatials.tile — Method

tile(cluster::AbstractMatrix{<:Number}, vectors::AbstractVector{<:AbstractVector{<:Number}}, translations::NTuple{M, NTuple{N, <:Number}}=()) where {N, M} -> Matrix{<:Number}

Tile a supercluster by translations of the input cluster.

Basically, the final supercluster is composed of several parts, each of which is a translation of the original cluster, with the translation vectors specified by vectors and each set of the translation indices contained in translations. When translation vectors are empty, a copy of the original cluster will be returned.

QuantumLattices.Essentials.Spatials.translate — Method

translate(cluster::AbstractMatrix{<:Number}, vector::AbstractVector{<:Number}) -> Matrix{vector|>eltype}

Get the translated cluster of the original one by a vector.

QuantumLattices.Essentials.Spatials.volume — Method

volume(vectors::AbstractVector{<:SVector}) -> Number
volume(v::AbstractVector{<:Number}) -> Number
volume(v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}) -> Number
volume(v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}, v₃::AbstractVector{<:Number}) -> Number

Get the volume spanned by the input vectors.

QuantumLattices.Essentials.dtype — Method

dtype(bz::BrillouinZone)
dtype(::Type{<:BrillouinZone{K, <:Momentum, S} where K}) where {S<:SVector}

Get the data type of a Brillouin zone.

QuantumLattices.Essentials.dtype — Method

dtype(bond::AbstractBond)
dtype(::Type{<:AbstractBond{N, <:AbstractPID, D} where N}) where {D<:Number}

Get the data type of a bond.

QuantumLattices.Essentials.dtype — Method

dtype(lattice::AbstractLattice)
dtype(::Type{<:AbstractLattice{N, <:AbstractPID, D} where N}) where {D<:Number}

Get the data type of a lattice.

QuantumLattices.Essentials.kind — Method

kind(bond::Bond) -> Int

Get the bond kind of a bond.

QuantumLattices.Essentials.kind — Method

kind(::Point) -> 0
kind(::Type{<:Point}) -> 0

Get the bond kind of a point, which is defined to be 0.

QuantumLattices.Essentials.reset! — Method

reset!(bs::Bonds, lattice::AbstractLattice) -> Bonds

Reset a set of lattice bonds by a new lattice.

QuantumLattices.Interfaces.decompose — Method

decompose(v₀::AbstractVector{<:Number}, v₁::AbstractVector{<:Number}) -> Tuple{Number}
decompose(v₀::AbstractVector{<:Number}, v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}) -> Tuple{Number, Number}
decompose(v₀::AbstractVector{<:Number}, v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}, v₃::AbstractVector{<:Number}) -> Tuple{Number, Number, Number}

Decompose a vector with respect to input basis vectors.

QuantumLattices.Interfaces.dimension — Method

dimension(bond::AbstractBond) -> Int
dimension(::Type{<:AbstractBond{N}}) where N -> Int

Get the space dimension of a concrete bond.

QuantumLattices.Interfaces.dimension — Method

dimension(lattice::AbstractLattice) -> Int
dimension(::Type{<:AbstractLattice{N}}) where N -> Int

Get the space dimension of the lattice.

QuantumLattices.Interfaces.expand — Method

expand(::Type{L}, ::Val{VS}) where {L<:AbstractLattice, VS} -> Tuple

Expand the lattice bond types to the leaf level.