Spin package

Spin package

Spin degrees of freedom

SID and Spin

SIndex

Predefined spin operators

Spin terms

SCID and SpinCoupling

Predefined spin couplings

Predefined spin terms

Manual

Hamiltonian.Essentials.SpinPackage.SCID — Type.

SCID(;center=wildcard,atom=wildcard,orbital=wildcard,tag='i',subscript=wildcard)=SCID(center,atom,orbital,tag,subscript)

The id of a spin coupling.

source

Hamiltonian.Essentials.SpinPackage.SID — Type.

SID <: IID

The spin id.

source

Hamiltonian.Essentials.SpinPackage.SID — Method.

SID(;orbital::Int=1,spin::Real=0.5,tag::Char='i')

Create a spin id.

source

Hamiltonian.Essentials.SpinPackage.SIndex — Type.

SIndex{S} <: Index{PID{S},SID}

The spin index.

source

Hamiltonian.Essentials.SpinPackage.SOperator — Type.

SOperator(value::Number,id::ID{<:NTuple{N,OID}}) where N

Spin operator.

source

Hamiltonian.Essentials.SpinPackage.Spin — Type.

Spin <: Internal{SID}

The spin interanl degrees of freedom.

source

Hamiltonian.Essentials.SpinPackage.Spin — Method.

Spin(;atom::Int=1,norbital::Int=1,spin::Real=0.5)

Construct a spin degrees of freedom.

source

Hamiltonian.Essentials.SpinPackage.SpinCoupling — Type.

SpinCoupling(value::Number,id::ID{<:NTuple{N,SCID}},subscripts::Subscripts) where N
SpinCoupling{N}(    value::Number=1;
                    tags::NTuple{N,Char},
                    centers::Union{NTuple{N,Int},Nothing}=nothing,
                    atoms::Union{NTuple{N,Int},Nothing}=nothing,
                    orbitals::Union{NTuple{N,Int},Subscript,Nothing}=nothing
                    ) where N

Spin coupling.

source

Hamiltonian.Essentials.SpinPackage.SpinTerm — Type.

SpinTerm(   id::Symbol,value::Number;
            neighbor::Int,
            couplings::Union{Tuple{<:Tuple{Vararg{Couplings}},<:Function},Coupling,Couplings},
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            factor::Number=1
            )

Spin term.

Type alias for Term{'B',:SpinTerm,<:Number,Int,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

Hamiltonian.Essentials.DegreesOfFreedom.oidtype — Method.

oidtype(::Val{:Spin},B::Type{<:AbstractBond},::Type{Nothing})
oidtype(::Val{:Spin},B::Type{<:AbstractBond},::Type{<:Table})

Get the compatible spin OID type with an AbstractBond type and a Table/Nothing type.

source

Hamiltonian.Essentials.DegreesOfFreedom.otype — Method.

otype(::Val{:Spin},O::Type{<:Term{'B'}},B::Type{<:AbstractBond},T::Type{<:Union{Nothing,Table}})

Get the compatible spin operator type with a Term type, an AbstractBond type and a Table/Nothing type.

source

Hamiltonian.Essentials.SpinPackage.Heisenberg — Method.

Heisenberg(;centers::Union{NTuple{2,Int},Nothing}=nothing,
            atoms::Union{NTuple{2,Int},Nothing}=nothing,
            orbitals::Union{NTuple{2,Int},Subscript,Nothing}=nothing
            ) -> Couplings{ID{<:NTuple{2,SCID}},SpinCoupling{2,Float,ID{<:NTuple{2,SCID}}}}

The Heisenberg couplings.

source

Hamiltonian.Essentials.SpinPackage.Ising — Method.

Ising(  tag::Char;
        centers::Union{NTuple{2,Int},Nothing}=nothing,
        atoms::Union{NTuple{2,Int},Nothing}=nothing,
        orbitals::Union{NTuple{2,Int},Subscript,Nothing}=nothing
        ) -> Couplings{ID{<:NTuple{2,SCID}},SpinCoupling{2,Float,ID{<:NTuple{2,SCID}}}}

The Ising couplings.

source

Hamiltonian.Essentials.SpinPackage.Sʸ — Method.

Sʸ(;atom::Union{Int,Nothing}=nothing,orbital::Union{Int,Nothing}=nothing) -> Couplings{ID{<:NTuple{2,SCID}},SpinCoupling{2,Float,ID{<:NTuple{2,SCID}}}}

The single Sʸ coupling.

source

Hamiltonian.Essentials.SpinPackage.Sˣ — Method.

Sˣ(;atom::Union{Int,Nothing}=nothing,orbital::Union{Int,Nothing}=nothing) -> Couplings{ID{<:NTuple{2,SCID}},SpinCoupling{2,Float,ID{<:NTuple{2,SCID}}}}

The single Sˣ coupling.

source

Hamiltonian.Essentials.SpinPackage.Sᶻ — Method.

Sᶻ(;atom::Union{Int,Nothing}=nothing,orbital::Union{Int,Nothing}=nothing) -> Couplings{ID{<:NTuple{2,SCID}},SpinCoupling{2,Float,ID{<:NTuple{2,SCID}}}}

The single Sᶻ coupling.

source

Hamiltonian.Essentials.SpinPackage.usualspinindextotuple — Function.

usualspinindextotuple

Indicate that the filtered attributes are (:scope,:site,:orbital) when converting a spin index to tuple.

source

Hamiltonian.Essentials.Terms.statistics — Method.

statistics(opt::SOperator) -> Char
statistics(::Type{<:SOperator}) -> Char

Get the statistics of SOperator.

source

Hamiltonian.Prerequisites.Interfaces.expand — Method.

expand(sc::SpinCoupling,pid::PID,spin::Spin,species::Union{Val{S},Nothing}=nothing) where S -> Union{SCExpand,Tuple{}}
expand(sc::SpinCoupling,pids::NTuple{N,PID},spins::NTuple{N,Spin},species::Union{Val{S},Nothing}=nothing) where {N,S} -> Union{SCExpand,Tuple{}}

Expand a spin coupling with the given set of point ids and spin degrees of freedom.

source

Hamiltonian.Prerequisites.Interfaces.matrix — Function.

matrix(sid::SID,dtype::Type{<:Number}=Complex{Float}) -> Matrix{dtype}

Get the matrix representation of a sid.

source

Base.:* — Method.

*(sc1::SpinCoupling,sc2::SpinCoupling) -> SpinCoupling

Get the multiplication between two spin couplings.

source

Base.adjoint — Method.

adjoint(sid::SID) -> SID

Get the adjoint of a spin id.

source

Base.repr — Method.

repr(sc::SpinCoupling) -> String

Get the repr representation of a spin coupling.

source

Base.show — Method.

show(io::IO,sc::SpinCoupling)

Show a spin coupling.

source

Base.union — Method.

union(::Type{P},::Type{SID}) where {P<:PID}

Get the union type of PID and SID.

source