Spin package

SID and Spin

SpinCoupling

STerm

Manual

QuantumLattices.Essentials.SpinPackage.soptdefaultlatex — Constant

sdefaultlatex

The default LaTeX format for a spin oid.

Base.Matrix — Method

Matrix(sid::SID{S}, dtype::Type{<:Number}=Complex{Float}) where S -> Matrix{dtype}

Get the matrix representation of a sid.

QuantumLattices.Essentials.SpinPackage.SCID — Type

SCID{T<:Tuple{Vararg{Char}}} <: SimpleID

The id of the tags part of a spin coupling.

QuantumLattices.Essentials.SpinPackage.SID — Type

SID{S} <: IID

The spin id.

QuantumLattices.Essentials.SpinPackage.SID — Method

SID{S}(tag::Char; orbital::Int=1) where S

Create a spin id.

QuantumLattices.Essentials.SpinPackage.Spin — Type

Spin{S} <: Internal{SID{S}}

The spin interanl degrees of freedom.

QuantumLattices.Essentials.SpinPackage.Spin — Method

Spin{S}(; norbital::Int=1) where S

Construct a spin degrees of freedom.

QuantumLattices.Essentials.SpinPackage.SpinCoupling — Type

SpinCoupling{V, T<:Tuple, O<:Subscripts, I<:Tuple{SCID, SubID}} <: Coupling{V, I}

Spin coupling.

QuantumLattices.Essentials.SpinPackage.SpinCoupling — Method

SpinCoupling(value::Number,
    tags::NTuple{N, Char};
    orbitals::Union{NTuple{N, Int}, Subscripts}=Subscripts(N)
    ) where N

Spin coupling.

QuantumLattices.Essentials.SpinPackage.SpinTerm — Type

SpinTerm{R}(id::Symbol, value::Any, bondkind::Any;
    couplings::Union{Function, Coupling, Couplings},
    amplitude::Union{Function, Nothing}=nothing,
    modulate::Union{Function, Bool}=false,
    ) where R

Spin term.

Type alias for Term{:SpinTerm, R, id, V, <:Any, <:TermCouplings, <:TermAmplitude, <:TermModulate}.

QuantumLattices.Essentials.SpinPackage.@dm_str — Macro

dm"x ob[o₁ o₂]" -> Couplings
dm"y ob[o₁ o₂]" -> Couplings
dm"z ob[o₁ o₂]" -> Couplings

The DM couplings.

QuantumLattices.Essentials.SpinPackage.@gamma_str — Macro

gamma"x ob[o₁ o₂]" -> Couplings
gamma"y ob[o₁ o₂]" -> Couplings
gamma"z ob[o₁ o₂]" -> Couplings

The Gamma couplings.

QuantumLattices.Essentials.SpinPackage.@heisenberg_str — Macro

heisenberg"ob[o₁ o₂]" -> Couplings
heisenberg"xyz ob[o₁ o₂]" -> Couplings
heisenberg"+-z ob[o₁ o₂]" -> Couplings

The Heisenberg couplings.

QuantumLattices.Essentials.SpinPackage.@ising_str — Macro

ising"x ob[o₁ o₂]" -> Couplings
ising"y ob[o₁ o₂])" -> Couplings
ising"z ob[o₁ o₂]" -> Couplings

The Ising couplings.

QuantumLattices.Essentials.SpinPackage.@sc_str — Macro

sc"..." -> SpinCoupling

Construct a SpinCoupling from a literal string.

QuantumLattices.Essentials.SpinPackage.@sˣ_str — Macro

sˣ"ob[o]" -> Couplings
sʸ"ob[o]" -> Couplings
sᶻ"ob[o]" -> Couplings

The single Sˣ/Sʸ/Sᶻ coupling.

Base.:* — Method

*(sc₁::SpinCoupling, sc₂::SpinCoupling) -> SpinCoupling

Get the multiplication between two spin couplings.

Base.repr — Method

repr(sc::SpinCoupling) -> String

Get the repr representation of a spin coupling.

Base.show — Method

show(io::IO, sc::SpinCoupling)

Show a spin coupling.

QuantumLattices.Essentials.DegreesOfFreedom.script — Method

script(::Val{:site}, index::Index{<:AbstractPID, <:SID}; kwargs...) -> Int
script(::Val{:orbital}, index::Index{<:AbstractPID, <:SID}; kwargs...) -> Int
script(::Val{:tag}, index::Index{<:AbstractPID, <:SID}; kwargs...) -> Char

Get the required script of a spin oid.

QuantumLattices.Essentials.SpinPackage.DM — Method

DM(tag::Char; orbitals::Union{NTuple{2, Int}, Subscripts}=Subscripts(2)) -> Couplings

The DM couplings.

QuantumLattices.Essentials.SpinPackage.Gamma — Method

Gamma(tag::Char; orbitals::Union{NTuple{2, Int}, Subscripts}=Subscripts(2)) -> Couplings

The Gamma couplings.

QuantumLattices.Essentials.SpinPackage.Heisenberg — Function

Heisenberg(mode::String="+-z"; orbitals::Union{NTuple{2, Int}, Subscripts}=Subscripts(2)) -> Couplings

The Heisenberg couplings.

QuantumLattices.Essentials.SpinPackage.Ising — Method

Ising(tag::Char; orbitals::Union{NTuple{2, Int}, Subscripts}=Subscripts(2)) -> Couplings

The Ising couplings.

QuantumLattices.Essentials.SpinPackage.Sʸ — Method

Sʸ(; orbital::Union{Int, Symbol}=wildcard) -> Couplings

The single Sʸ coupling.

QuantumLattices.Essentials.SpinPackage.Sˣ — Method

Sˣ(; orbital::Union{Int, Symbol}=wildcard) -> Couplings

The single Sˣ coupling.

QuantumLattices.Essentials.SpinPackage.Sᶻ — Method

Sᶻ(; orbital::Union{Int, Symbol}=wildcard) -> Couplings

The single Sᶻ coupling.

QuantumLattices.Essentials.SpinPackage.usualspinindextotuple — Function

usualspinindextotuple

Indicate that the choosed fields are (:scope, :site, :orbital) when converting a spin index to tuple.

QuantumLattices.Interfaces.expand — Method

expand(sc::SpinCoupling, bond::AbstractBond, config::Config, info::Val) -> SCExpand

Expand a spin coupling with the given set of points and spin degrees of freedom.

QuantumLattices.Interfaces.permute — Method

permute(id₁::OID{<:Index{<:AbstractPID, <:SID}}, id₂::OID{<:Index{<:AbstractPID, <:SID}}) -> Tuple{Vararg{Operator}}

Permute two spin oid and get the result.