Spin package

Spin package

Spin degrees of freedom

SID and Spin

SIndex

Predefined spin operators

Spin terms

SCID and SpinCoupling

Predefined spin couplings

Predefined spin terms

Manual

QuantumLattices.Essentials.SpinPackage.soptdefaultlatex — Constant.

soptdefaultlatex

The default LaTeX format of the oids of a spin operator.

source

QuantumLattices.Essentials.SpinPackage.SCID — Type.

SCID(;center=wildcard,atom=wildcard,orbital=wildcard,tag='z',subscript=wildcard)=SCID(center,atom,orbital,tag,subscript)

The id of a spin coupling.

source

QuantumLattices.Essentials.SpinPackage.SID — Type.

SID <: IID

The spin id.

source

QuantumLattices.Essentials.SpinPackage.SID — Method.

SID(;orbital::Int=1,spin::Real=0.5,tag::Char='z')

Create a spin id.

source

QuantumLattices.Essentials.SpinPackage.SIndex — Type.

SIndex{S} <: Index{PID{S},SID}

The spin index.

source

QuantumLattices.Essentials.SpinPackage.SOperator — Type.

SOperator(value,id::ID{<:Tuple{Vararg{OID}}}=ID())

Spin operator.

source

QuantumLattices.Essentials.SpinPackage.Spin — Type.

Spin <: Internal{SID}

The spin interanl degrees of freedom.

source

QuantumLattices.Essentials.SpinPackage.Spin — Method.

Spin(;atom::Int=1,norbital::Int=1,spin::Real=0.5)

Construct a spin degrees of freedom.

source

QuantumLattices.Essentials.SpinPackage.SpinCoupling — Type.

SpinCoupling(value,id::ID{<:Tuple{Vararg{SCID}}},subscripts::Subscripts)
SpinCoupling{N}(    value=1;
                    tags::NTuple{N,Char},
                    centers::Union{NTuple{N,Int},Nothing}=nothing,
                    atoms::Union{NTuple{N,Int},Nothing}=nothing,
                    orbitals::Union{NTuple{N,Int},Subscript,Nothing}=nothing
                    ) where N

Spin coupling.

source

QuantumLattices.Essentials.SpinPackage.SpinTerm — Type.

SpinTerm{R}(id::Symbol,value::Any,bondkind::Any;
            couplings::Union{Function,Coupling,Couplings},
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            ) where R

Spin term.

Type alias for Term{'B',:SpinTerm,R,id,V,<:Any,<:TermCouplings,<:TermAmplitude,<:Union{TermModulate,Nothing}}.

source

QuantumLattices.Essentials.SpinPackage.@dm_str — Macro.

dm"x sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings
dm"y sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings
dm"z sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings

The DM couplings.

source

QuantumLattices.Essentials.SpinPackage.@gamma_str — Macro.

gamma"x sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings
gamma"y sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings
gamma"z sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings

The Gamma couplings.

source

QuantumLattices.Essentials.SpinPackage.@heisenberg_str — Macro.

heisenberg"sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings

The Heisenberg couplings.

source

QuantumLattices.Essentials.SpinPackage.@ising_str — Macro.

ising"x sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings
ising"y sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings
ising"z sl(a₁,a₂)⊗ob(o₁,o₂)@(c₁,c₂)" -> Couplings

The Ising couplings.

source

QuantumLattices.Essentials.SpinPackage.@sc_str — Macro.

sc"..." -> SpinCoupling

Construct a SpinCoupling from a literal string.

source

QuantumLattices.Essentials.SpinPackage.@sˣ_str — Macro.

sˣ"sl(a)⊗ob(o)" -> Couplings
sʸ"sl(a)⊗ob(o)" -> Couplings
sᶻ"sl(a)⊗ob(o)" -> Couplings

The single Sˣ/Sʸ/Sᶻ coupling.

source

QuantumLattices.Essentials.SpinPackage.DM — Method.

DM( tag::Char;
    centers::Union{NTuple{2,Int},Nothing}=nothing,
    atoms::Union{NTuple{2,Int},Nothing}=nothing,
    orbitals::Union{NTuple{2,Int},Subscript,Nothing}=nothing
    ) -> Couplings{ID,SpinCoupling{Int,ID{<:NTuple{2,SCID}}}}

source

QuantumLattices.Essentials.SpinPackage.Gamma — Method.

Gamma(  tag::Char;
        centers::Union{NTuple{2,Int},Nothing}=nothing,
        atoms::Union{NTuple{2,Int},Nothing}=nothing,
        orbitals::Union{NTuple{2,Int},Subscript,Nothing}=nothing
        ) -> Couplings{ID,SpinCoupling{Int,ID{<:NTuple{2,SCID}}}}

The Gamma couplings.

source

QuantumLattices.Essentials.SpinPackage.Heisenberg — Function.

Heisenberg( mode::String="+-z";
            centers::Union{NTuple{2,Int},Nothing}=nothing,
            atoms::Union{NTuple{2,Int},Nothing}=nothing,
            orbitals::Union{NTuple{2,Int},Subscript,Nothing}=nothing
            ) -> Couplings{ID,SpinCoupling{Rational{Int}/Int,ID{<:NTuple{2,SCID}}}}

The Heisenberg couplings.

source

QuantumLattices.Essentials.SpinPackage.Ising — Method.

Ising(  tag::Char;
        centers::Union{NTuple{2,Int},Nothing}=nothing,
        atoms::Union{NTuple{2,Int},Nothing}=nothing,
        orbitals::Union{NTuple{2,Int},Subscript,Nothing}=nothing
        ) -> Couplings{ID,SpinCoupling{Int,ID{<:NTuple{2,SCID}}}}

The Ising couplings.

source

QuantumLattices.Essentials.SpinPackage.Sʸ — Method.

Sʸ(;atom::Union{Int,Nothing}=nothing,orbital::Union{Int,Nothing}=nothing) -> Couplings{ID,SpinCoupling{Int,ID{<:Tuple{SCID}}}}

The single Sʸ coupling.

source

QuantumLattices.Essentials.SpinPackage.Sˣ — Method.

Sˣ(;atom::Union{Int,Nothing}=nothing,orbital::Union{Int,Nothing}=nothing) -> Couplings{ID,SpinCoupling{Int,ID{<:Tuple{SCID}}}}

The single Sˣ coupling.

source

QuantumLattices.Essentials.SpinPackage.Sᶻ — Method.

Sᶻ(;atom::Union{Int,Nothing}=nothing,orbital::Union{Int,Nothing}=nothing) -> Couplings{ID,SpinCoupling{Int,ID{<:Tuple{SCID}}}}

The single Sᶻ coupling.

source

QuantumLattices.Essentials.SpinPackage.usualspinindextotuple — Function.

usualspinindextotuple

Indicate that the filtered attributes are (:scope,:site,:orbital) when converting a spin index to tuple.

source

Base.:* — Method.

*(sc1::SpinCoupling,sc2::SpinCoupling) -> SpinCoupling

Get the multiplication between two spin couplings.

source

Base.adjoint — Method.

adjoint(sid::SID) -> SID

Get the adjoint of a spin id.

source

Base.repr — Method.

repr(sc::SpinCoupling) -> String

Get the repr representation of a spin coupling.

source

Base.show — Method.

show(io::IO,sc::SpinCoupling)

Show a spin coupling.

source

Base.union — Method.

union(::Type{P},::Type{SID}) where {P<:PID}

Get the union type of PID and SID.

source

QuantumLattices.Essentials.DegreesOfFreedom.otype — Method.

otype(T::Type{<:SpinTerm},I::Type{<:OID})

Get the operator type of a spin term.

source

QuantumLattices.Essentials.DegreesOfFreedom.script — Method.

script(oid::OID{<:SIndex},::Val{:site}) -> Int
script(oid::OID{<:SIndex},::Val{:orbital}) -> Int
script(oid::OID{<:SIndex},::Val{:spin}) -> Float
script(oid::OID{<:SIndex},::Val{:tag}) -> Char

Get the required script of a spin oid.

source

QuantumLattices.Essentials.Terms.statistics — Method.

statistics(opt::SOperator) -> Char
statistics(::Type{<:SOperator}) -> Char

Get the statistics of SOperator.

source

QuantumLattices.Interfaces.expand — Method.

expand(sc::SpinCoupling,pid::PID,spin::Spin,kind::Union{Val{K},Nothing}=nothing) where K -> Union{SCExpand,Tuple{}}
expand(sc::SpinCoupling,pids::NTuple{N,PID},spins::NTuple{N,Spin},kind::Union{Val{K},Nothing}=nothing) where {N,K} -> Union{SCExpand,Tuple{}}

Expand a spin coupling with the given set of point ids and spin degrees of freedom.

source

QuantumLattices.Interfaces.matrix — Function.

matrix(sid::SID,dtype::Type{<:Number}=Complex{Float}) -> Matrix{dtype}

Get the matrix representation of a sid.

source

QuantumLattices.Interfaces.permute — Method.

permute(::Type{<:SOperator},id1::OID{<:SIndex},id2::OID{<:SIndex},table) -> Tuple{Vararg{SOperator}}

Permute two fermionic oid and get the result.

source

QuantumLattices.Mathematics.AlgebraOverFields.rawelement — Method.

rawelement(::Type{<:SOperator})

Get the raw name of a type of SOperator.

source

QuantumLattices.Mathematics.AlgebraOverFields.rawelement — Method.

rawelement(::Type{<:SpinCoupling})

Get the raw name of a type of SpinCoupling.

source