Terms

Terms

OID

Operator and Operators

Term

Manul

Hamiltonian.Essentials.Terms.OID — Type.

OID(index::Index,rcoord::Union{Nothing,SVector{N,Float}},icoord::Union{Nothing,SVector{N,Float}},seq::Union{Nothing,Int}) where N
OID(index::Index;rcoord::Union{Nothing,SVector}=nothing,icoord::Union{Nothing,SVector}=nothing,seq::Union{Nothing,Int}=nothing)

Operator id.

source

Hamiltonian.Essentials.Terms.Operator — Type.

Operator{N,V<:Number,I<:ID} <: Element{V,I}

Abstract type for an operator.

source

Hamiltonian.Essentials.Terms.Operators — Type.

Operators(opts::Operator...)

A set of operators.

Type alias of Operators{I<:ID,O<:Operator}=Elements{I,O}.

source

Hamiltonian.Essentials.Terms.Term — Type.

Term{ST,SP,RK}(id::Symbol,value::Number,neighbor::Any,couplings::Function,amplitude::Function,modulate::Union{Function,Nothing},factor::Number) where {ST,SP,RK}
Term{ST,SP,RK}( id::Symbol,value::Number,neighbor::Any;
                couplings::Union{Function,Couplings},
                amplitude::Union{Function,Nothing}=nothing,
                modulate::Union{Function,Bool}=false,
                factor::Number=1
                ) where {ST,SP,RK}

A term of a quantum lattice system.

source

Hamiltonian.Essentials.Terms.abbr — Method.

abbr(::Term) -> Symbol

Get the abbreviation of the species of a term.

source

Hamiltonian.Essentials.Terms.isHermitian — Method.

isHermitian(opts::Operators) -> Bool

Judge whether a set of operators as a whole is Hermitian.

source

Hamiltonian.Essentials.Terms.isHermitian — Method.

isHermitian(opt::Operator) -> Bool

Judge whether an operator is Hermitian.

source

Hamiltonian.Essentials.Terms.species — Method.

species(term::Term) -> Symbol
species(::Type{<:Term{ST,SP}}) where {ST,SP} -> Symbol

Get the species of a term.

source

Hamiltonian.Essentials.Terms.statistics — Method.

statistics(term::Term) -> Char
statistics(::Type{<:Term{ST}}) where ST -> Char

Get the statistics of a term.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(term::Term) -> Int
rank(::Type{<:Term{ST,SP,RK}}) where {ST,SP,RK}

Get the rank of a term.

source

Base.:+ — Method.

+(term::Term) -> Term
-(term::Term) -> Term
*(term::Term,factor::Number) -> Term
*(factor::Number,term::Term) -> Term
/(term::Term,factor::Number) -> Term

Allowed arithmetic operations for a term.

source

Base.:== — Method.

==(term1::Term,term2::Term) -> Bool
isequal(term1::Term,term2::Term) -> Bool

Judge whether two terms are equivalent to each other.

source

Base.adjoint — Method.

adjoint(oid::OID) -> OID

Get the adjoint of an operator id.

source

Base.adjoint — Method.

adjoint(opt::Operator{N}) where N -> Operator

Get the adjoint of an operator.

source

Base.one — Method.

one(term::Term) -> Term

Get a unit term.

source

Base.replace — Method.

replace(term::Term{ST,SP,RK};kwargs...) where {ST,SP,RK} -> Term

Replace some attributes of a term with key word arguments.

source

Base.repr — Method.

repr(term::Term,bond::AbstractBond,config::IDFConfig) -> String

Get the repr representation of a term on a bond with a given config.

source

Base.show — Method.

show(io::IO,term::Term)

Show a term.

source

Base.zero — Method.

zero(term::Term) -> Term

Get a zero term.

source