Terms

Terms

OID

Operator and Operators

Term

Manul

Hamiltonian.Essentials.Terms.OID — Type.

OID(index::Index,rcoord::Union{Nothing,SVector{N,Float},Vector{Float}},icoord::Union{Nothing,SVector{N,Float},Vector{Float}},seq::Union{Nothing,Int}) where N
OID(index::Index;rcoord::Union{Nothing,SVector,Vector{Float}}=nothing,icoord::Union{Nothing,SVector,Vector{Float}}=nothing,seq::Union{Nothing,Int}=nothing)

Operator id.

source

Hamiltonian.Essentials.Terms.Operator — Type.

Operator{N,V<:Number,I<:ID{<:NTuple{N,OID}}} <: Element{N,V,I}

Abstract type for an operator.

source

Hamiltonian.Essentials.Terms.Operators — Type.

Operators(opts::Operator...)

A set of operators.

Type alias of Operators{I<:ID,O<:Operator}=Elements{I,O}.

source

Hamiltonian.Essentials.Terms.Term — Type.

Term{ST,SP}(id::Symbol,value::Number,neighbor::Any,couplings::TermCouplings,amplitude::TermAmplitude,modulate::Union{TermModulate,Nothing},factor::Number) where {ST,SP}
Term{ST,SP}(id::Symbol,value::Number,neighbor::Any;
            couplings::Union{Tuple{<:Tuple{Vararg{Couplings}},<:Function},Coupling,Couplings,TermCouplings},
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            factor::Number=1
            ) where {ST,SP}

A term of a quantum lattice system.

source

Hamiltonian.Essentials.Terms.TermAmplitude — Type.

TermAmplitude(amplitude::Union{Function,Nothing}=nothing)

The function for the amplitude of a term.

source

Hamiltonian.Essentials.Terms.TermCouplings — Type.

TermCouplings(candidate::Coupling)
TermCouplings(candidate::Couplings)
TermCouplings(contents::Tuple{<:Tuple{Vararg{Couplings}},<:Function})
TermCouplings(candidates::NTuple{N,<:Couplings},choice::Function) where N

The function for the couplings of a term.

source

Hamiltonian.Essentials.Terms.TermFunction — Type.

TermFunction <: Function

Abstract type for concrete term functions.

source

Hamiltonian.Essentials.Terms.TermModulate — Type.

TermModulate(id::Symbol,modulate::Union{Function,Nothing}=nothing)
TermModulate(id::Symbol,modulate::Bool)

The function for the modulation of a term.

source

Hamiltonian.Essentials.Spatials.icoord — Method.

icoord(opt::Operator{1}) -> SVector
icoord(opt::Operator{2}) -> SVector

Get the whole icoord of an operator.

source

Hamiltonian.Essentials.Spatials.rcoord — Method.

rcoord(opt::Operator{1}) -> SVector
rcoord(opt::Operator{2}) -> SVector

Get the whole rcoord of an operator.

source

Hamiltonian.Essentials.Terms.abbr — Method.

abbr(term::Term) -> Symbol
abbr(::Type{<:Term}) -> Symbol

Get the abbreviation of the species of a term.

source

Hamiltonian.Essentials.Terms.isHermitian — Method.

isHermitian(opts::Operators) -> Bool

Judge whether a set of operators as a whole is Hermitian.

source

Hamiltonian.Essentials.Terms.isHermitian — Method.

isHermitian(opt::Operator) -> Bool

Judge whether an operator is Hermitian.

source

Hamiltonian.Essentials.Terms.isHermitian — Method.

isHermitian(oid::ID{<:NTuple{N,OID}}) where N -> Bool

Judge whether an operator id is Hermitian.

source

Hamiltonian.Essentials.Terms.oidtype — Function.

oidtype

Get the compatible oid type.

source

Hamiltonian.Essentials.Terms.otype — Function.

otype

Get the compatible operator type from a term type, a bond type and a table type.

source

Hamiltonian.Essentials.Terms.species — Method.

species(term::Term) -> Symbol
species(::Type{<:Term{ST,SP}}) where {ST,SP} -> Symbol

Get the species of a term.

source

Hamiltonian.Essentials.Terms.statistics — Method.

statistics(term::Term) -> Char
statistics(::Type{<:Term{ST}}) where ST -> Char

Get the statistics of a term.

source

Hamiltonian.Prerequisites.Interfaces.expand — Function.

expand(otype::Type{<:Operator},term::Term,bond::AbstractBond,config::IDFConfig,table::Union{Table,Nothing}=nothing,half::Union{Bool,Nothing}=nothing)

Expand the operators of a term on a bond with a given config.

The half parameter determines the behavior of generating operators, which falls into the following three categories

false: no extra operations on the generated operators

true: an extra multiplication by 0.5 with the generated operators

nothing: "Hermitian half" of the generated operators

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(tcs::TermCouplings) -> Int
rank(TCS::Type{<:TermCouplings}) -> Int

Get the rank of the couplings it contained.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(term::Term) -> Int
rank(::Type{T}) where T<:Term -> Int

Get the rank of a term.

source

Hamiltonian.Prerequisites.Interfaces.update! — Method.

update!(term::Term,args...;kwargs...) -> Term

Update the value of a term by its modulate function.

source

Base.:+ — Method.

+(term::Term) -> Term
-(term::Term) -> Term
*(term::Term,factor::Number) -> Term
*(factor::Number,term::Term) -> Term
/(term::Term,factor::Number) -> Term

Allowed arithmetic operations for a term.

source

Base.:== — Method.

==(term1::Term,term2::Term) -> Bool
isequal(term1::Term,term2::Term) -> Bool

Judge whether two terms are equivalent to each other.

source

Base.:== — Method.

==(tf1::TermFunction,tf2::TermFunction) -> Bool
isequal(tf1::TermFunction,tf2::TermFunction) -> Bool

Judge whether two concrete term functions are equivalent to each other.

source

Base.adjoint — Method.

adjoint(opts::Operators) -> Operators

Get the adjoint of a set of operators.

source

Base.adjoint — Method.

adjoint(oid::OID) -> typeof(oid)
adjoint(oid::ID{<:NTuple{N,OID}}) where N -> typeof(oid)

Get the adjoint of an operator id.

source

Base.adjoint — Method.

adjoint(opt::Operator{N}) where N -> Operator

Get the adjoint of an operator.

source

Base.one — Method.

one(term::Term) -> Term

Get a unit term.

source

Base.replace — Method.

replace(term::Term{ST,SP};kwargs...) where {ST,SP} -> Term

Replace some attributes of a term with key word arguments.

source

Base.repr — Method.

repr(term::Term,bond::AbstractBond,config::IDFConfig) -> String

Get the repr representation of a term on a bond with a given config.

source

Base.show — Method.

show(io::IO,oid::OID)

Show an operator id.

source

Base.show — Method.

show(io::IO,opt::Operator)

Show an operator.

source

Base.show — Method.

show(io::IO,term::Term)

Show a term.

source

Base.valtype — Method.

valtype(term::Term)
valtype(::Type{<:Term{ST,SP,V}}) where {ST,SP,V<:Number}

Get the value type of a term.

source

Base.zero — Method.

zero(term::Term) -> Term

Get a zero term.

source