Terms

Terms

Subscript and Subscripts

Coupling and Couplings

Term

Manul

Hamiltonian.Essentials.Terms.Coupling — Type.

Coupling{N,V<:Number,I<:ID{<:NTuple{N,SimpleID}}} <: Element{N,V,I}

The coupling intra/inter interanl degrees of freedom at different lattice points.

source

Hamiltonian.Essentials.Terms.Couplings — Type.

Couplings{I<:ID,C<:Coupling} <: AbstractDict{I,C}

A pack of couplings intra/inter interanl degrees of freedom at different lattice points.

Alias for Elements{I,C}.

source

Hamiltonian.Essentials.Terms.Subscript — Type.

Subscript(  ipattern::NTuple{N1,Any},
            opattern::NTuple{N2,Any},
            mapping::Union{Function,Nothing}=nothing,
            constrain::Union{Function,Nothing}=nothing,
            identifier::Union{Symbol,Char}=wildcard
            ) where {N1,N2}
Subscript{N}() where N
Subscript(opattern::NTuple{N,Int}) where N

The subscripts of some orbital/spin degrees of freedom.

source

Hamiltonian.Essentials.Terms.Subscript — Method.

(subscript::Subscript{N})(::Val{'M'},values::Vararg{Int,N}) where N -> NTuple{dimension(subscript),Int}
(subscript::Subscript{N})(::Val{'C'},values::Vararg{Int,N}) where N -> Bool

Construct the subscripts from a set of independent variables.

Judge whether a set of independent variables are valid to construct the subscripts.

source

Hamiltonian.Essentials.Terms.Subscripts — Type.

Subscripts(contents::Subscript...)

A complete set of all the independent subscripts of the orbital/spin degrees of freedom.

source

Hamiltonian.Essentials.Terms.Subscripts — Method.

(subscripts::Subscripts)(::Val{'M'},values::NTuple{N,Int}) where N -> NTuple{dimension(subscripts),Int}
(subscripts::Subscripts)(::Val{'C'},values::NTuple{N,Int}) where N -> Bool

Construct the complete set of subscripts from a complete set of independent variables.

Judge whether a complete set of independent variables are valid to construct the complete subscripts.

source

Hamiltonian.Essentials.Terms.Term — Type.

Term{ST,SP}(id::Symbol,value::Number,neighbor::Any,couplings::TermCouplings,amplitude::TermAmplitude,modulate::Union{TermModulate,Nothing},factor::Number) where {ST,SP}
Term{ST,SP}(id::Symbol,value::Number,neighbor::Any;
            couplings::Union{Tuple{<:Tuple{Vararg{Couplings}},<:Function},Coupling,Couplings,TermCouplings},
            amplitude::Union{Function,Nothing}=nothing,
            modulate::Union{Function,Bool}=false,
            factor::Number=1
            ) where {ST,SP}

A term of a quantum lattice system.

source

Hamiltonian.Essentials.Terms.TermAmplitude — Type.

TermAmplitude(amplitude::Union{Function,Nothing}=nothing)

The function for the amplitude of a term.

source

Hamiltonian.Essentials.Terms.TermCouplings — Type.

TermCouplings(candidate::Coupling)
TermCouplings(candidate::Couplings)
TermCouplings(contents::Tuple{<:Tuple{Vararg{Couplings}},<:Function})
TermCouplings(candidates::NTuple{N,<:Couplings},choice::Function) where N

The function for the couplings of a term.

source

Hamiltonian.Essentials.Terms.TermFunction — Type.

TermFunction <: Function

Abstract type for concrete term functions.

source

Hamiltonian.Essentials.Terms.TermModulate — Type.

TermModulate(id::Symbol,modulate::Union{Function,Nothing}=nothing)
TermModulate(id::Symbol,modulate::Bool)

The function for the modulation of a term.

source

Hamiltonian.Essentials.Terms.@subscript — Macro.

@subscript expr::Expr with constrain::Expr -> Subscript

Construct a subscript from a map and optionally with a constrain.

source

Hamiltonian.Essentials.Terms.abbr — Method.

abbr(term::Term) -> Symbol
abbr(::Type{<:Term}) -> Symbol

Get the abbreviation of the species of a term.

source

Hamiltonian.Essentials.Terms.species — Method.

species(term::Term) -> Symbol
species(::Type{<:Term{ST,SP}}) where {ST,SP} -> Symbol

Get the species of a term.

source

Hamiltonian.Essentials.Terms.statistics — Method.

statistics(term::Term) -> Char
statistics(::Type{<:Term{ST}}) where ST -> Char

Get the statistics of a term.

source

Hamiltonian.Prerequisites.Interfaces.expand — Function.

expand(otype::Type{<:Operator},term::Term,bond::AbstractBond,config::IDFConfig,table::Union{Table,Nothing}=nothing,half::Union{Bool,Nothing}=nothing)

Expand the operators of a term on a bond with a given config.

The half parameter determines the behavior of generating operators, which falls into the following three categories

false: no extra operations on the generated operators

true: an extra multiplication by 0.5 with the generated operators

nothing: "Hermitian half" of the generated operators

source

Hamiltonian.Prerequisites.Interfaces.expand — Method.

expand(subscripts::Subscripts,dimensions::NTuple{N,Int}) where N -> SbExpand

Expand a complete set of subscripts with a given set of variable ranges.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(subscripts::Subscripts,i::Int) -> Int
rank(::Type{<:Subscripts{T}},i::Int) where T -> Int

Get the number of the independent variables of a component of the complete subscript set.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(subscripts::Subscripts) -> Int
rank(::Type{S}) where S<:Subscripts -> Int

Get the total number of the independent variables of the complete subscript set.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(subscript::Subscript) -> Int
rank(::Type{<:Subscript{N}}) where N -> Int

Get the number of the independent variables that are used to describe the subscripts of some orbital/spin degrees of freedom.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(tcs::TermCouplings) -> Int
rank(TCS::Type{<:TermCouplings}) -> Int

Get the rank of the couplings it contained.

source

Hamiltonian.Prerequisites.Interfaces.rank — Method.

rank(term::Term) -> Int
rank(::Type{T}) where T<:Term -> Int

Get the rank of a term.

source

Hamiltonian.Prerequisites.Interfaces.update! — Method.

update!(term::Term,args...;kwargs...) -> Term

Update the value of a term by its modulate function.

source

Base.:* — Method.

*(sub1::Subscript,sub2::Subscript) -> Subscripts
*(subs::Subscripts,sub::Subscript) -> Subscripts
*(sub::Subscript,subs::Subscripts) -> Subscripts
*(subs1::Subscripts,subs2::Subscripts) -> Subscripts

Get the multiplication between subscripts or complete sets of subscripts.

source

Base.:+ — Method.

+(term::Term) -> Term
-(term::Term) -> Term
*(term::Term,factor::Number) -> Term
*(factor::Number,term::Term) -> Term
/(term::Term,factor::Number) -> Term

Allowed arithmetic operations for a term.

source

Base.:== — Method.

==(sub1::Subscript,sub2::Subscript) -> Bool
isequal(sub1::Subscript,sub2::Subscript) -> Bool

Judge whether two subscripts are equivalent to each other.

source

Base.:== — Method.

==(term1::Term,term2::Term) -> Bool
isequal(term1::Term,term2::Term) -> Bool

Judge whether two terms are equivalent to each other.

source

Base.:== — Method.

==(tf1::TermFunction,tf2::TermFunction) -> Bool
isequal(tf1::TermFunction,tf2::TermFunction) -> Bool

Judge whether two concrete term functions are equivalent to each other.

source

Base.one — Method.

one(term::Term) -> Term

Get a unit term.

source

Base.replace — Method.

replace(term::Term{ST,SP};kwargs...) where {ST,SP} -> Term

Replace some attributes of a term with key word arguments.

source

Base.repr — Method.

repr(term::Term,bond::AbstractBond,config::IDFConfig) -> String

Get the repr representation of a term on a bond with a given config.

source

Base.show — Method.

show(io::IO,subscript::Subscript)

Show a subscript.

source

Base.show — Method.

show(io::IO,term::Term)

Show a term.

source

Base.valtype — Method.

valtype(term::Term)
valtype(::Type{<:Term{ST,SP,V}}) where {ST,SP,V<:Number}

Get the value type of a term.

source

Base.zero — Method.

zero(term::Term) -> Term

Get a zero term.

source

Hamiltonian.Prerequisites.Interfaces.dimension — Method.

dimension(subscripts::Subscripts,i::Int) -> Int
dimension(::Type{<:Subscripts{T}},i::Int) where T -> Int

Get the total number of the whole variables of a component of the complete subscript set.

source

Hamiltonian.Prerequisites.Interfaces.dimension — Method.

dimension(subscripts::Subscripts) -> Int
dimension(::Type{S}) where S<:Subscripts -> Int

Get the total number of the whole variables of the complete subscript set.

source

Hamiltonian.Prerequisites.Interfaces.dimension — Method.

dimension(subscript::Subscript) -> Int
dimension(::Type{<:Subscript{N1,N2}}) where {N1,N2} -> Int

Get the number of the whole variables that are used to describe the subscripts of some orbital/spin degrees of freedom.

source