Terms

Subscript and Subscripts

Coupling and Couplings

Term

Generator

Manual

QuantumLattices.Essentials.Terms.AbstractGenerator — Type

AbstractGenerator{coord, TS<:NamedContainer{Term}, BS<:Bonds, C<:Config, T<:Union{Nothing, Table}, B<:Boundary, OS<:GenOperators}

Abstract generator.

By protocol, a concrete generator must have the following attributes:

terms::TS: the terms contained in a generator

bonds::BS: the bonds on which the terms are defined

config::C: the configuration of the interanl degrees of freedom

table::T: the index-sequence table, nothing for not using such a table

half::Bool: true for generating an Hermitian half of the operators and false for generating the whole

boundary::B: boundary twist for the generated operators, nothing for no twist

operators::OS: the generated operators

QuantumLattices.Essentials.Terms.Coupling — Type

Coupling{V, I<:ID{SimpleID}} <: Element{V, I}

The coupling intra/inter interanl degrees of freedom at different lattice points.

QuantumLattices.Essentials.Terms.Couplings — Type

Couplings(cps::Coupling...)

A pack of couplings intra/inter interanl degrees of freedom at different lattice points.

Alias for Elements{<:ID{SimpleID}, <:Coupling}.

QuantumLattices.Essentials.Terms.GenOperators — Type

GenOperators(constops::Operators, alterops::NamedContainer{Operators}, boundops::NamedContainer{Operators})
GenOperators(terms::Tuple{Vararg{Term}}, bonds::Bonds, config::Config, table::Union{Nothing, Table}, half::Bool, ::Val{coord}) where coord

A set of operators. This is the core of AbstractGenerator.

QuantumLattices.Essentials.Terms.Generator — Type

Generator(terms::Tuple{Vararg{Term}}, bonds::Bonds, config::Config, table::Union{Nothing, Table}=nothing, half::Bool=true, boundary::Boundary=Boundary())
Generator{coord}(terms::Tuple{Vararg{Term}}, bonds::Bonds, config::Config, table::Union{Nothing, Table}=nothing, half::Bool=true, boundary::Boundary=Boundary()) where coord

A generator of operators based on terms, configuration of internal degrees of freedom, table of indices and boundary twist.

QuantumLattices.Essentials.Terms.Parameters — Type

Parameters{Names}(values::Number...) where Names

A NamedTuple that contain the key-value pairs.

QuantumLattices.Essentials.Terms.Parameters — Method

Parameters(gen::AbstractGenerator) -> Parameters

Get the parameters of the terms of a generator.

QuantumLattices.Essentials.Terms.Subscript — Type

Subscript(  ipattern::NTuple{N1, Any},
            opattern::NTuple{N2, Any},
            mapping::Union{Function, Nothing}=nothing,
            constrain::Union{Function, Nothing}=nothing,
            identifier::Union{Symbol, Char}=wildcard
            ) where {N1, N2}
Subscript{N}() where N
Subscript(opattern::NTuple{N, Int}) where N

The subscripts of some orbital/spin degrees of freedom.

QuantumLattices.Essentials.Terms.Subscript — Method

(subscript::Subscript{N})(::Val{'M'}, values::Vararg{Int, N}) where N -> NTuple{dimension(subscript), Int}
(subscript::Subscript{N})(::Val{'C'}, values::Vararg{Int, N}) where N -> Bool

Construct the subscripts from a set of independent variables.

Judge whether a set of independent variables are valid to construct the subscripts.

QuantumLattices.Essentials.Terms.Subscripts — Type

Subscripts(contents::Subscript...)

A complete set of all the independent subscripts of the orbital/spin degrees of freedom.

QuantumLattices.Essentials.Terms.Subscripts — Method

(subscripts::Subscripts)(::Val{'M'}, values::NTuple{N, Int}) where N -> NTuple{dimension(subscripts), Int}
(subscripts::Subscripts)(::Val{'C'}, values::NTuple{N, Int}) where N -> Bool

Construct the complete set of subscripts from a complete set of independent variables.

Judge whether a complete set of independent variables are valid to construct the complete subscripts.

QuantumLattices.Essentials.Terms.Term — Type

Term{ST, K, R, I}(value::Any, bondkind::Any, couplings::TermCouplings, amplitude::TermAmplitude, modulate::Union{TermModulate, Nothing}, factor::Any) where {ST, K, R, I}
Term{ST, K, R}( id::Symbol, value::Any, bondkind::Any;
                couplings::Union{Function, Coupling, Couplings, TermCouplings},
                amplitude::Union{Function, Nothing}=nothing,
                modulate::Union{Function, Bool}=false,
                ) where {ST, K, R}

A term of a quantum lattice system.

QuantumLattices.Essentials.Terms.TermAmplitude — Type

TermAmplitude(amplitude::Union{Function, Nothing}=nothing)

The function for the amplitude of a term.

QuantumLattices.Essentials.Terms.TermCouplings — Type

TermCouplings(candidate::Coupling)
TermCouplings(candidate::Couplings)
TermCouplings(contents::Tuple{<:Tuple{Vararg{Couplings}}, <:Function})
TermCouplings(candidates::NTuple{N, <:Couplings}, choice::Function) where N

The function for the couplings of a term.

QuantumLattices.Essentials.Terms.TermFunction — Type

TermFunction <: Function

Abstract type for concrete term functions.

QuantumLattices.Essentials.Terms.TermModulate — Type

TermModulate(id::Symbol, modulate::Union{Function, Nothing}=nothing)
TermModulate(id::Symbol, modulate::Bool)

The function for the modulation of a term.

QuantumLattices.Essentials.Terms.@couplings — Macro

@couplings cps -> Couplings

Convert an expression/literal to a set of couplings.

QuantumLattices.Essentials.Terms.@subscript — Macro

@subscript expr::Expr with constrain::Expr -> Subscript

Construct a subscript from a map and optionally with a constrain.

Base.:* — Method

*(sub1::Subscript, sub2::Subscript) -> Subscripts
*(subs::Subscripts, sub::Subscript) -> Subscripts
*(sub::Subscript, subs::Subscripts) -> Subscripts
*(subs1::Subscripts, subs2::Subscripts) -> Subscripts

Get the multiplication between subscripts or complete sets of subscripts.

Base.:+ — Method

+(term::Term) -> Term
-(term::Term) -> Term
*(term::Term, factor) -> Term
*(factor, term::Term) -> Term
/(term::Term, factor) -> Term

Allowed arithmetic operations for a term.

Base.:== — Method

==(gen1::AbstractGenerator, gen2::AbstractGenerator) -> Bool

Judge whether generators are equivalent to each other.

Base.:== — Method

==(genops1::GenOperators, genops2::GenOperators) -> Bool

Judge whether two sets of operators are equivalent to each other.

Base.:== — Method

==(sub1::Subscript, sub2::Subscript) -> Bool

Judge whether two subscripts are equivalent to each other.

Base.:== — Method

==(term1::Term, term2::Term) -> Bool

Judge whether two terms are equivalent to each other.

Base.:== — Method

==(tf1::TermFunction, tf2::TermFunction) -> Bool

Judge whether two concrete term functions are equivalent to each other.

Base.eltype — Method

eltype(ops::GenOperators)
eltype(::Type{<:GenOperators{S, D, B}}) where {S<:Operators, D<:NamedContainer{Operators}, B<:NamedContainer{Operators}}

Get the eltype of a set of operators, which is defined to be the common operator type of all operators it contains.

Base.empty! — Method

empty!(ops::GenOperators) -> GenOperators

Empty a set of operators.

Base.empty! — Method

empty!(gen::Generator) -> Generator

Empty the :bonds, :config, :table and :operators of a generator.

Base.empty — Method

empty(ops::GenOperators) -> GenOperators

Get an empty copy of a set of operators.

Base.empty — Method

empty(gen::Generator{coord}) where coord -> Generator

Get an empty copy of a generator.

Base.isequal — Method

isequal(gen1::AbstractGenerator, gen2::AbstractGenerator) -> Bool

Judge whether generators are equivalent to each other.

Base.isequal — Method

isequal(genops1::GenOperators, genops2::GenOperators) -> Bool

Judge whether two sets of operators are equivalent to each other.

Base.isequal — Method

isequal(sub1::Subscript, sub2::Subscript) -> Bool

Judge whether two subscripts are equivalent to each other.

Base.isequal — Method

isequal(term1::Term, term2::Term) -> Bool

Judge whether two terms are equivalent to each other.

Base.isequal — Method

isequal(tf1::TermFunction, tf2::TermFunction) -> Bool

Judge whether two concrete term functions are equivalent to each other.

Base.match — Function

match(params1::Parameters, params2::Parameters, atol=atol, rtol=rtol) -> Bool

Judge whether the second set of parameters matches the first.

Base.one — Method

one(term::Term) -> Term

Get a unit term.

Base.replace — Method

replace(term::Term{ST, K, R, I}; kwargs...) where {ST, K, R, I} -> Term

Replace some attributes of a term with key word arguments.

Base.repr — Method

repr(term::Term, bond::AbstractBond, config::Config) -> String

Get the repr representation of a term on a bond with a given config.

Base.show — Method

show(io::IO, subscript::Subscript)

Show a subscript.

Base.show — Method

show(io::IO, term::Term)

Show a term.

Base.valtype — Method

valtype(term::Term)
valtype(::Type{<:Term{ST, K, R, I, V} where {ST, K, R, I}}) where V

Get the value type of a term.

Base.zero — Method

zero(term::Term) -> Term

Get a zero term.

QuantumLattices.Essentials.DegreesOfFreedom.isHermitian — Method

isHermitian(term::Term) -> Bool
isHermitian(::Type{<:Term}) -> Bool

QuantumLattices.Essentials.Terms.abbr — Method

abbr(term::Term) -> Symbol
abbr(::Type{<:Term}) -> Symbol

Get the abbreviation of the kind of a term.

QuantumLattices.Essentials.Terms.statistics — Method

statistics(term::Term) -> Char
statistics(::Type{<:Term{ST}}) where ST -> Char

Get the statistics of a term.

QuantumLattices.Interfaces.dimension — Method

dimension(subscripts::Subscripts, i::Int) -> Int
dimension(::Type{<:Subscripts{T}}, i::Int) where T -> Int

Get the total number of the whole variables of a component of the complete subscript set.

QuantumLattices.Interfaces.dimension — Method

dimension(subscripts::Subscripts) -> Int
dimension(::Type{<:(Subscripts{<:Tuple, R, D} where R)}) where D -> Int

Get the total number of the whole variables of the complete subscript set.

QuantumLattices.Interfaces.dimension — Method

dimension(subscript::Subscript) -> Int
dimension(::Type{<:(Subscript{N1, N2} where N1)}) where N2 -> Int

Get the number of the whole variables that are used to describe the subscripts of some orbital/spin degrees of freedom.

QuantumLattices.Interfaces.expand — Function

expand(term::Term, bond::AbstractBond, config::Config, table::Union{Table, Nothing}=nothing, half::Bool=false, coord::Union{Val{true}, Val{false}}=coordon) -> Operators
expand(term::Term, bonds::Bonds, config::Config, table::Union{Table, Nothing}=nothing, half::Bool=false, coord::Union{Val{true}, Val{false}}=coordon) -> Operators

Expand the operators of a term on a bond/set-of-bonds with a given config.

QuantumLattices.Interfaces.expand! — Function

expand!(operators::Operators, term::Term, bond::AbstractBond, config::Config,
        table::Union{Table, Nothing}=nothing,
        half::Bool=false,
        coord::Union{Val{true}, Val{false}}=coordon
        ) -> Operators
expand!(operators::Operators, term::Term, bonds::Bonds, config::Config, 
        table::Union{Table, Nothing}=nothing,
        half::Bool=false,
        coord::Union{Val{true}, Val{false}}=coordon
        ) -> Operators

Expand the operators of a term on a bond/set-of-bonds with a given config.

The half parameter determines the behavior of generating operators, which falls into the following two categories

true: "Hermitian half" of the generated operators

false: "Hermitian whole" of the generated operators

QuantumLattices.Interfaces.expand! — Method

expand!(operators::Operators, gen::AbstractGenerator) -> Operators

Expand the operators of a generator.

QuantumLattices.Interfaces.expand! — Method

expand!(operators::Operators, ops::GenOperators, boundary::Boundary; kwargs...) -> Operators

Expand the operators with the given boundary twist and term coefficients.

QuantumLattices.Interfaces.expand — Method

expand(gen::AbstractGenerator) -> Operators
expand(gen::AbstractGenerator{coord}, name::Symbol) where coord -> Operators
expand(gen::AbstractGenerator{coord}, i::Int) where coord -> Operators
expand(gen::AbstractGenerator{coord}, name::Symbol, i::Int) where coord -> Operators

Expand the operators of a generator:

the total operators;

the operators of a specific term;

the operators on a specific bond;

the operators of a specific term on a specific bond.

QuantumLattices.Interfaces.expand — Method

expand(subscripts::Subscripts, dimensions::NTuple{N, Int}) where N -> SbExpand

Expand a complete set of subscripts with a given set of variable ranges.

QuantumLattices.Interfaces.id — Method

id(term::Term) -> Symbol
id(::Type{<:Term{ST, K, R, I} where {ST, K, R}}) where I -> Symbol

Get the id of a term.

QuantumLattices.Interfaces.kind — Method

kind(term::Term) -> Symbol
kind(::Type{<:Term{ST, K} where ST}) where K -> Symbol

Get the kind of a term.

QuantumLattices.Interfaces.rank — Method

rank(subscripts::Subscripts, i::Int) -> Int
rank(::Type{<:Subscripts{T}}, i::Int) where T -> Int

Get the number of the independent variables of a component of the complete subscript set.

QuantumLattices.Interfaces.rank — Method

rank(subscripts::Subscripts) -> Int
rank(::Type{<:Subscripts{<:Tuple, R}}) where R -> Int

Get the total number of the independent variables of the complete subscript set.

QuantumLattices.Interfaces.rank — Method

rank(subscript::Subscript) -> Int
rank(::Type{<:Subscript{N}}) where N -> Int

Get the number of the independent variables that are used to describe the subscripts of some orbital/spin degrees of freedom.

QuantumLattices.Interfaces.rank — Method

rank(term::Term) -> Int
rank(::Type{<:Term{ST, K, R} where {ST, K}}) where R -> Int

Get the rank of a term.

QuantumLattices.Interfaces.reset! — Method

reset!(genops::GenOperators, terms::Tuple{Vararg{Term}}, bonds::Bonds, config::Config, table::Union{Nothing, Table}, half::Bool, ::Val{coord}) where coord -> GenOperators

Reset a set of operators by new terms, bonds, config, table, etc..

QuantumLattices.Interfaces.reset! — Method

reset!(gen::Generator{coord}, lattice::AbstractLattice) where coord -> Generator

Reset a generator by a new lattice.

QuantumLattices.Interfaces.update! — Method

update!(gen::AbstractGenerator; kwargs...) -> typeof(gen)

Update the coefficients of the terms in a generator.

QuantumLattices.Interfaces.update! — Method

update!(term::Term, args...; kwargs...) -> Term

Update the value of a term by its modulate function.

QuantumLattices.Mathematics.AlgebraOverFields.idtype — Method

idtype(ops::GenOperators)
idtype(::Type{<:GenOperators{S, D, B}}) where {S<:Operators, D<:NamedContainer{Operators}, B<:NamedContainer{Operators}}

Get the idtype of a set of operators, which is defined to be the common idtype of all operators it contains.