Terms

Coupling and Couplings

Term

Generator

Manual

QuantumLattices.Essentials.Terms.AbstractCoupling — Type

AbstractCoupling{V, I<:ID{SimpleID}} <: Element{V, I}

The abstract coupling intra/inter interanl degrees of freedom at different lattice points.

QuantumLattices.Essentials.Terms.AbstractGenerator — Type

AbstractGenerator{TS<:NamedContainer{Term}, BS<:Bonds, H<:Hilbert, T<:Table, B<:Boundary, OS<:GenOperators}

Abstract generator.

By protocol, a concrete generator should have the following predefined contents:

terms::TS: the terms contained in a generator

bonds::BS: the bonds on which the terms are defined

hilbert::H: the total Hilbert space

half::Bool: true for generating an Hermitian half of the operators and false for generating the whole

table::Table: the index-sequence table

boundary::B: boundary twist for the generated operators, nothing for no twist

operators::OS: the generated operators

QuantumLattices.Essentials.Terms.ConstrainID — Type

ConstrainID{T<:Tuple{Vararg{Pair{UnitRange{Int}, <:Tuple{Vararg{String}}}}}} <: SimpleID

The id of an iid constrain.

QuantumLattices.Essentials.Terms.ConstrainID — Method

ConstrainID(iidc::IIDConstrain)

Construct the id of an iid constrain.

QuantumLattices.Essentials.Terms.Coupling — Type

Coupling{V, I<:ID{SimpleIID}, C<:IIDConstrain, CI<:ConstrainID} <: AbstractCoupling{V, Tuple{CompositeIID{I}, CI}}

The coupling intra/inter interanl degrees of freedom at different lattice points.

QuantumLattices.Essentials.Terms.Couplings — Type

Couplings(cps::AbstractCoupling...)

A pack of couplings intra/inter interanl degrees of freedom at different lattice points.

Alias for Elements{<:ID{SimpleID}, <:AbstractCoupling}.

QuantumLattices.Essentials.Terms.GenOperators — Type

GenOperators{C<:Operators, A<:NamedContainer{Operators}, B<:NamedContainer{Operators}}

A set of operators. This is the core of AbstractGenerator.

QuantumLattices.Essentials.Terms.GenOperators — Method

GenOperators(terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert, half::Bool; table::Union{Nothing, Table}=nothing)

Construct a set of operators.

QuantumLattices.Essentials.Terms.Generator — Type

Generator{TS<:NamedContainer{Term}, BS<:Bonds, H<:Hilbert, T<:Table, B<:Boundary, OS<:GenOperators} <: AbstractGenerator{TS, BS, H, T, B, OS}

A generator of operators based on terms, configuration of internal degrees of freedom, and boundary twist.

QuantumLattices.Essentials.Terms.Generator — Method

Generator(terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert;
    half::Bool=false, table::Union{Table,Nothing}=nothing, boundary::Boundary=plain
    )

Construct a generator of operators.

QuantumLattices.Essentials.Terms.IIDConstrain — Type

IIDConstrain{T<:Tuple{Vararg{NamedContainer{Subscript}}}} <: CompositeTuple{T}

Constrain on a set of simple iids or on a composite iid.

QuantumLattices.Essentials.Terms.IIDConstrain — Method

IIDConstrain(constrain::NamedContainer{Subscript}...)

Construct the constrain.

QuantumLattices.Essentials.Terms.IIDSpace — Type

IIDSpace{I<:IID, V<:Internal, Kind} <: CartesianVectorSpace{IID}

The space expanded by a "labeled" iid.

QuantumLattices.Essentials.Terms.Parameters — Type

Parameters{Names}(values::Number...) where Names

A NamedTuple that contain the key-value pairs.

QuantumLattices.Essentials.Terms.Parameters — Method

Parameters(gen::AbstractGenerator) -> Parameters

Get the parameters of the terms of a generator.

QuantumLattices.Essentials.Terms.Subscript — Type

Subscript{P<:Tuple, C<:Function} <: CompositeTuple{P}

A subscript set of a certain internal degree of freedom.

QuantumLattices.Essentials.Terms.Subscript — Method

Subscript(N::Int)
Subscript(pattern::Tuple, check_constant::Bool=false)

Construct a subscript set of a certain internal degree of freedom.

QuantumLattices.Essentials.Terms.Term — Type

Term{K, R, I, V, B, C<:TermCouplings, A<:TermAmplitude, M<:TermModulate}

A term of a quantum lattice system.

QuantumLattices.Essentials.Terms.Term — Method

Term{K, R}(id::Symbol, value, bondkind;
    couplings::Union{Function, TermCouplings, Couplings},
    amplitude::Union{Function, TermAmplitude, Nothing}=nothing,
    modulate::Union{Function, TermModulate, Bool}=false
    ) where {K, R}

Construct a term.

QuantumLattices.Essentials.Terms.TermAmplitude — Type

TermAmplitude(amplitude::Union{Function, Nothing}=nothing)

The function for the amplitude of a term.

QuantumLattices.Essentials.Terms.TermCouplings — Type

TermCouplings(couplings::Union{Couplings, Function})

The function for the couplings of a term.

QuantumLattices.Essentials.Terms.TermFunction — Type

TermFunction <: Function

Abstract type for concrete term functions.

QuantumLattices.Essentials.Terms.TermModulate — Type

TermModulate(id::Symbol, modulate::Function)
TermModulate(id::Symbol, modulate::Bool)

The function for the modulation of a term.

QuantumLattices.Essentials.Terms.@couplings — Macro

@couplings cps -> Couplings

Convert an expression/literal to a set of couplings.

QuantumLattices.Essentials.Terms.@subscript_str — Macro

subscript"..." -> Subscript

Construct a subscript set from a literal string.

Base.:* — Method

*(cp₁::Coupling, cp₂::Coupling) -> Coupling

Get the multiplication between two couplings.

Base.:* — Method

*(iidc₁::IIDConstrain, iidc₂::IIDConstrain) -> IIDConstrain

Get the combination of two independent constrains on the composition of two iid sets (or two composite iids).

Base.:+ — Method

+(term::Term) -> Term
-(term::Term) -> Term
*(term::Term, factor) -> Term
*(factor, term::Term) -> Term
/(term::Term, factor) -> Term

Allowed arithmetic operations for a term.

Base.eltype — Method

eltype(ops::GenOperators)
eltype(::Type{<:GenOperators})

Get the eltype of a set of operators, which is defined to be the common operator type of all operators it contains.

Base.empty! — Method

empty!(ops::GenOperators) -> GenOperators

Empty a set of operators.

Base.empty! — Method

empty!(gen::Generator) -> Generator

Empty the :bonds, :hilbert, :table and :operators of a generator.

Base.empty — Method

empty(ops::GenOperators) -> GenOperators

Get an empty copy of a set of operators.

Base.empty — Method

empty(gen::Generator) -> Generator

Get an empty copy of a generator.

Base.match — Function

match(params₁::Parameters, params₂::Parameters, atol=atol, rtol=rtol) -> Bool

Judge whether the second set of parameters matches the first.

Base.match — Method

match(iidc::IIDConstrain, iids::Tuple{Vararg{SimpleIID}}) -> Bool
match(iidc::IIDConstrain, ciid::CompositeIID) -> Bool

Judge whether a composite iid matches the constrain.

Base.match — Method

match(subscript::Subscript, values::Tuple) -> Bool

Judge whether a set of values matches the pattern specified by subscript.

Base.one — Method

one(term::Term) -> Term

Get a unit term.

Base.replace — Method

replace(term::Term; kwargs...) -> Term

Replace some attributes of a term with key word arguments.

Base.repr — Method

repr(term::Term, bond::AbstractBond, hilbert::Hilbert) -> String

Get the repr representation of a term on a bond with a given Hilbert space.

Base.valtype — Method

valtype(term::Term)
valtype(::Type{<:Term)

Get the value type of a term.

Base.zero — Method

zero(term::Term) -> Term

Get a zero term.

QuantumLattices.Essentials.DegreesOfFreedom.isHermitian — Method

isHermitian(term::Term) -> Bool
isHermitian(::Type{<:Term}) -> Bool

QuantumLattices.Essentials.QuantumAlgebras.idtype — Method

idtype(ops::GenOperators)
idtype(::Type{<:GenOperators})

Get the idtype of a set of operators, which is defined to be the common idtype of all operators it contains.

QuantumLattices.Essentials.QuantumAlgebras.idtype — Method

idtype(iidc::IIDConstrain)
idtype(::Type{<:IIDConstrain{T}}) where {T<:Tuple{Vararg{NamedContainer{Subscript}}}}

Get the type of the id of an iid constrain.

QuantumLattices.Essentials.Terms.abbr — Method

abbr(term::Term) -> Symbol
abbr(::Type{<:Term}) -> Symbol

Get the abbreviation of the kind of a term.

QuantumLattices.Essentials.Terms.couplingcenters — Method

couplingcenters(coupling::AbstractCoupling, bond::AbstractBond, info::Val) -> NTuple{rank(coupling), Int}

Get the acting centers of the coupling on a bond.

QuantumLattices.Essentials.Terms.couplinginternals — Method

couplinginternals(coupling::AbstractCoupling, bond::AbstractBond, hilbert::Hilbert, info::Val) -> NTuple{rank(coupling), SimpleInternal}

Get the interanl spaces where each order of the coupling acts on.

QuantumLattices.Essentials.Terms.couplingpoints — Method

couplingpoints(coupling::AbstractCoupling, bond::AbstractBond, info::Val) -> NTuple{rank(coupling), eltype(bond)}

Get the points where each order of the coupling acts on.

QuantumLattices.Essentials.Terms.ismodulatable — Method

ismodulatable(term::Term) -> Bool
ismodulatable(::Type{<:Term}) -> Bool

Judge whether a term could be modulated by its modulate function.

QuantumLattices.Essentials.Terms.otype — Method

otype(gen::AbstractGenerator)
otype(::Type{<:AbstractGenerator})

Get the operator type of the generated opeators.

QuantumLattices.Essentials.Terms.otype — Method

otype(T::Type{<:Term}, H::Type{<:Hilbert}, B::Type{<:AbstractBond})

Get the compatible operator type from the type of a term, a Hilbert space and a bond.

QuantumLattices.Essentials.kind — Method

kind(term::Term) -> Symbol
kind(::Type{<:Term) -> Symbol

Get the kind of a term.

QuantumLattices.Essentials.reset! — Method

reset!(genops::GenOperators, terms::Tuple{Vararg{Term}}, bonds::Bonds, hilbert::Hilbert, half::Bool; table::Union{Nothing, Table}=nothing) -> GenOperators

Reset a set of operators by the new terms, bonds and Hilbert space.

QuantumLattices.Essentials.reset! — Method

reset!(gen::Generator, lattice::AbstractLattice) -> Generator

Reset a generator by a new lattice.

QuantumLattices.Essentials.update! — Method

update!(term::Term, args...; kwargs...) -> Term

Update the value of a term by its modulate function.

QuantumLattices.Essentials.update! — Method

update!(gen::AbstractGenerator; kwargs...) -> typeof(gen)

Update the coefficients of the terms in a generator.

QuantumLattices.Interfaces.expand — Function

expand(term::Term, bond::AbstractBond, hilbert::Hilbert, half::Bool=false; table::Union{Nothing, Table}=nothing) -> Operators
expand(term::Term, bonds::Bonds, hilbert::Hilbert, half::Bool=false; table::Union{Nothing, Table}=nothing) -> Operators

Expand the operators of a term on a bond/set-of-bonds with a given Hilbert space.

QuantumLattices.Interfaces.expand! — Function

expand!(operators::Operators, term::Term, bond::AbstractBond, hilbert::Hilbert, half::Bool=false; table::Union{Nothing, Table}=nothing) -> Operators
expand!(operators::Operators, term::Term, bonds::Bonds, hilbert::Hilbert, half::Bool=false; table::Union{Nothing, Table}=nothing) -> Operators

Expand the operators of a term on a bond/set-of-bonds with a given Hilbert space.

The half parameter determines the behavior of generating operators, which falls into the following two categories

true: "Hermitian half" of the generated operators

false: "Hermitian whole" of the generated operators

QuantumLattices.Interfaces.expand! — Method

expand!(operators::Operators, gen::AbstractGenerator) -> Operators

Expand the operators of a generator.

QuantumLattices.Interfaces.expand! — Method

expand!(operators::Operators, ops::GenOperators, boundary::Boundary; kwargs...) -> Operators

Expand the operators with the given boundary twist and term coefficients.

QuantumLattices.Interfaces.expand — Method

expand(coupling::AbstractCoupling, bond::AbstractBond, hilbert::Hilbert, info::Val)

Expand a coupling with the given bond and Hilbert space.

QuantumLattices.Interfaces.expand — Method

expand(gen::AbstractGenerator) -> Operators
expand(gen::AbstractGenerator, name::Symbol) -> Operators
expand(gen::AbstractGenerator, i::Int) -> Operators
expand(gen::AbstractGenerator, name::Symbol, i::Int) -> Operators

Expand the operators of a generator:

the total operators;

the operators of a specific term;

the operators on a specific bond;

the operators of a specific term on a specific bond.

QuantumLattices.Interfaces.expand — Method

expand(iids::NTuple{N, IID}, internals::NTuple{N, Internal}) where N -> IIDSpace

Get the space expanded by a set of "labeled" iids.

QuantumLattices.Interfaces.id — Method

id(term::Term) -> Symbol
id(::Type{<:Term) -> Symbol

Get the id of a term.

QuantumLattices.Interfaces.rank — Method

rank(iidc::IIDConstrain, i::Integer) -> Int
rank(::Type{<:IIDConstrain{T}}, i::Integer) where {T<:Tuple{Vararg{NamedContainer{Subscript}}}} -> Int

Get the rank of the ith homogenous segment of the iid set (or the composite iid) on which the constrain act.

QuantumLattices.Interfaces.rank — Method

rank(iidc::IIDConstrain) -> Int
rank(::Type{<:IIDConstrain{T}}) where {T<:Tuple{Vararg{NamedContainer{Subscript}}}} -> Int

Get the rank of the iid set (or the composite iid) on which the constrain act.

QuantumLattices.Interfaces.rank — Method

rank(subscript::Subscript) -> Int
rank(::Type{<:Subscript}) -> Int

Get the total number of the whole variables of a subscript set.

QuantumLattices.Interfaces.rank — Method

rank(term::Term) -> Int
rank(::Type{<:Term) -> Int

Get the rank of a term.