Frameworks

QuantumLattices.Frameworks.eager — Constant

const eager = Eager()

Singleton instance of Eager.

QuantumLattices.Frameworks.lazy — Constant

const lazy = Lazy()

Singleton instance of Lazy.

QuantumLattices.Frameworks.Action — Type

Action

Abstract type for all actions.

QuantumLattices.Frameworks.Algorithm — Type

Algorithm{F<:Frontend, P<:Parameters, M<:Function} <: Function

An algorithm associated with an frontend.

QuantumLattices.Frameworks.Algorithm — Method

(alg::Algorithm)(assign::Assignment) -> Tuple{Algorithm, Assignment}
(assign::Assignment)(alg::Algorithm) -> Tuple{Algorithm, Assignment}

Run an assignment based on an algorithm.

The difference between these two methods is that the first uses the parameters of assign as the current parameters while the second uses those of alg.

QuantumLattices.Frameworks.Algorithm — Method

(alg::Algorithm)(id::Symbol, action::Action; info::Bool=true, kwargs...) -> Tuple{Algorithm, Assignment}

Add an assignment on a algorithm by providing the contents of the assignment, and run this assignment.

QuantumLattices.Frameworks.Algorithm — Method

Algorithm(name::Symbol, frontend::Frontend; din::String=".", dout::String=".", parameters::Parameters=Parameters(frontend), map::Function=identity)

Construct an algorithm.

QuantumLattices.Frameworks.Algorithm — Method

(alg::Algorithm)(id::Symbol; info::Bool=true, parameters::Parameters=Parameters{()}()) -> Tuple{Algorithm, Assignment}

Run an assignment specified by its id.

Optionally, the time of the run process can be informed by setting the info argument to be true.

QuantumLattices.Frameworks.Assignment — Type

Assignment{A<:Action, P<:Parameters, M<:Function, N, D} <: Function

An assignment associated with an action.

QuantumLattices.Frameworks.CategorizedGenerator — Type

CategorizedGenerator(terms::Tuple{Vararg{Term}}, bonds::Vector{<:Bond}, hilbert::Hilbert, boundary::Boundary=plain, style::ExpansionStyle=eager; half::Bool=false)

Construct a categorized generator of quantum operators based on the input terms, bonds, Hilbert space and (twisted) boundary condition.

When the boundary condition is plain, the boundary operators will be set to be empty for simplicity and efficiency.

QuantumLattices.Frameworks.CategorizedGenerator — Type

CategorizedGenerator{C, A<:NamedTuple, B<:NamedTuple, P<:Parameters, D<:Boundary, S<:ExpansionStyle} <: Generator

Categorized generator that groups the quantum operators in a quantum lattice system into three categories, i.e., the constant, the alterable, and the boundary.

QuantumLattices.Frameworks.CompositeGenerator — Type

CompositeGenerator{CG<:CategorizedGenerator} <: Generator

Abstract type for a composite generator quantum operators in a quantum lattice system.

By protocol, it must have the following predefined content:

operators::CG: the categorized generator of quantum operators

QuantumLattices.Frameworks.Eager — Type

Eager <: ExpansionStyle

Eager expansion style with eager computation so that similar terms are combined in the final result.

QuantumLattices.Frameworks.ExpansionStyle — Type

ExpansionStyle

Expansion style of a generator of quantum operators. It has two singleton subtypes, Eager and Lazy.

QuantumLattices.Frameworks.Formula — Type

Formula{F<:Function, P<:Parameters}

Representation of a quantum lattice system with an explicit analytical formula.

QuantumLattices.Frameworks.Frontend — Type

Frontend

Frontend of algorithms applied to a quantum lattice system.

QuantumLattices.Frameworks.Generator — Type

Generator

Generator of quantum operators in a quantum lattice system.

QuantumLattices.Frameworks.Hamiltonian — Type

Hamiltonian

Abstract type of the Hamiltonian of a quantum lattice system.

QuantumLattices.Frameworks.Hamiltonian — Method

Hamiltonian(representation::Union{OperatorSet, Formula, Generator}) -> SimpleHamiltonian

Construct the Hamiltonian of a quantum lattice system by its representation, which could be an OperatorSet, a Formula, or a Generator.

QuantumLattices.Frameworks.Image — Type

Image{CG<:CategorizedGenerator, H<:Transformation} <: CompositeGenerator{CG}

Generator representing the image of a transformation applied to a generator of quantum operators.

QuantumLattices.Frameworks.Lazy — Type

Lazy <: ExpansionStyle

Lazy expansion style with lazy computation so that similar terms are not combined in the final result.

QuantumLattices.Frameworks.OperatorGenerator — Type

OperatorGenerator(terms::Tuple{Vararg{Term}}, bonds::Vector{<:Bond}, hilbert::Hilbert, boundary::Boundary=plain, style::ExpansionStyle=eager; half::Bool=false)

Construct a generator of operators.

QuantumLattices.Frameworks.OperatorGenerator — Type

OperatorGenerator{CG<:CategorizedGenerator{<:Operators}, TS<:Tuple{Vararg{Term}}, B<:Bond, H<:Hilbert} <: CompositeGenerator{CG}

A generator of operators based on the terms, bonds and Hilbert space of a quantum lattice system.

QuantumLattices.Frameworks.Parameters — Type

Parameters{Names}(values::Number...) where Names

A NamedTuple that contains the key-value pairs.

QuantumLattices.Frameworks.Parameters — Method

Parameters(bound::Boundary)

Get the parameters of the twisted boundary condition.

QuantumLattices.Frameworks.Parameters — Method

Parameters(cat::CategorizedGenerator)

Get the complete set of parameters of a categorized generator of quantum operators.

QuantumLattices.Frameworks.Parameters — Method

Parameters(hamiltonian::SimpleHamiltonian{<:OperatorSet}) -> NamedTuple{(), Tuple{}}
Parameters(hamiltonian::SimpleHamiltonian{<:Union{Formula, Generator}}) -> Parameters

Get the parameters of a Hamiltonian.

QuantumLattices.Frameworks.SimpleHamiltonian — Type

SimpleHamiltonian{R<:Union{OperatorSet, Formula, Generator}} <: Hamiltonian

Hamiltonian of a quantum lattice system by its representation, which could be an OperatorSet, a Formula, or a Generator.

QuantumLattices.QuantumOperators.LinearTransformation — Method

(transformation::LinearTransformation)(cat::CategorizedGenerator; kwargs...) -> CategorizedGenerator

Apply a linear transformation to a categorized generator of quantum operators.

QuantumLattices.QuantumOperators.Transformation — Method

(transformation::Transformation)(gen::CompositeGenerator; kwargs...) -> Image

Get the image of a transformation applied to a generator of quantum operators.

Base.:* — Method

*(cat::CategorizedGenerator, factor) -> CategorizedGenerator
*(factor, cat::CategorizedGenerator) -> CategorizedGenerator

Multiply a categorized generator of quantum operators with a factor.

Base.:+ — Method

+(cat₁::CategorizedGenerator, cat₂::CategorizedGenerator) -> CategorizedGenerator

Addition of two categorized generators of quantum operators.

Base.empty — Method

empty(cat::CategorizedGenerator) -> CategorizedGenerator
empty!(cat::CategorizedGenerator) -> CategorizedGenerator

Get an empty copy of a categorized generator or empty a categorized generator of quantum operators.

Base.empty — Method

empty(gen::Image) -> Image
empty!(gen::Image) -> Image

Get an empty copy of or empty the image of a transformation applied to a generator of quantum operators.

Base.empty — Method

empty(gen::OperatorGenerator) -> OperatorGenerator
empty!(gen::OperatorGenerator) -> OperatorGenerator

Get an empty copy of or empty an operator generator.

Base.isempty — Method

isempty(cat::CategorizedGenerator) -> Bool

Judge whether a categorized generator is empty.

Base.match — Method

match(params₁::Parameters, params₂::Parameters; atol=atol, rtol=rtol) -> Bool

Judge whether the second set of parameters matches the first.

Base.nameof — Method

nameof(alg::Algorithm, assign::Assignment) -> String

Get the name of the combination of an algorithm and an assignment.

Base.show — Method

show(io::IO, alg::Algorithm)

Show an algorithm.

Optionally, some parameters of the algorithm can be filtered by specifying the :select context in io. Besides, the maximum number of decimals of the parameters can also be specified by the :ndecimal context in io.

Base.summary — Method

summary(alg::Algorithm)

Provide a summary of an algorithm.

Base.valtype — Method

valtype(assign::Assignment)
valtype(::Type{<:Assignment})

The type of the data(result) of an assignment.

Base.valtype — Method

valtype(::Type{<:CategorizedGenerator})

Get the valtype of a categorized generator of quantum operators.

QuantumLattices.Frameworks.prepare! — Function

prepare!(alg::Algorithm, assign::Assignment, f::Function=assign->true) -> Tuple{Algorithm, Assignment}

Run the dependencies of an assignment.

Optionally, some dependencies can be filtered by specifying the f function.

QuantumLattices.QuantumOperators.matrix — Method

matrix(hamiltonian::SimpleHamiltonian{<:Formula}; kwargs...) -> AbstractMatrix

Get the matrix representation of a Hamiltonian.

QuantumLattices.Spatials.save — Method

save(alg::Algorithm, assign::Assignment; delimited=false) -> Tuple{Algorithm, Assignment}

Save the data of an assignment registered on an algorithm.

QuantumLattices.add! — Method

add!(alg::Algorithm, id::Symbol, action::Action; parameters::Parameters=Parameters{()}(), map::Function=identity, dependencies::Tuple=(), kwargs...) -> Tuple{Algorithm, Assignment}

Add an assignment on an algorithm by providing the contents of the assignment without the execution of it.

QuantumLattices.dimension — Method

dimension(hamiltonian::SimpleHamiltonian{<:Formula}) -> Int

Get the dimension of the matrix representation of a Hamiltonian.

QuantumLattices.expand! — Method

expand!(result, gen::Generator) -> typeof(result)

Expand the generator to add the quantum operators in a quantum lattice system to result.

QuantumLattices.expand — Method

expand(cat::CategorizedGenerator, ::Lazy)

Expand a categorized generator to get the quantum operators in a quantum lattice system.

QuantumLattices.expand — Method

expand(gen::Generator)
expand(gen::Generator, ::Eager)
expand(gen::Generator, ::Lazy)

Expand the generator to get the quantum operators in a quantum lattice system.

QuantumLattices.expand — Method

expand(gen::OperatorGenerator, name::Symbol) -> Operators
expand(gen::OperatorGenerator, i::Int) -> Operators
expand(gen::OperatorGenerator, name::Symbol, i::Int) -> Operators

Expand an operator generator to get:

the operators of a specific term;

the operators on a specific bond;

the operators of a specific term on a specific bond.

QuantumLattices.reset! — Function

reset!(cat::CategorizedGenerator{<:Operators}, terms::Tuple{Vararg{Term}}, bonds::Vector{<:Bond}, hilbert::Hilbert, boundary::Boundary=cat.boundary; half::Bool=false) -> CategorizedGenerator

Reset a categorized generator of quantum operators by the new terms, bonds, Hilbert space and (twisted) boundary condition.

QuantumLattices.reset! — Method

reset!(cat::CategorizedGenerator, transformation::LinearTransformation, source::CategorizedGenerator{<:Operators}; kwargs...)

Reset a categorized generator by its source categorized generator of quantum operators and the corresponding linear transformation.

QuantumLattices.reset! — Method

reset!(gen::Image, transformation::Transformation, source::CompositeGenerator; kwargs...) -> Image

Reset the image of a transformation applied to a generator of quantum operators.

QuantumLattices.reset! — Method

reset!(gen::OperatorGenerator, lattice::AbstractLattice, hilbert::Hilbert; neighbors=max(map(term->term.bondkind, gen.terms)...)) -> OperatorGenerator

Reset an operator generator by a new lattice and the corresponding new hilbert space.

QuantumLattices.update! — Method

update!(alg::Algorithm; parameters...) -> Algorithm

Update the parameters of an algorithm and its associated frontend.

QuantumLattices.update! — Method

update!(assign::Assignment; parameters...) -> Assignment

Update the parameters of an assignment and the status of its associated action.

QuantumLattices.update! — Method

update!(cat::CategorizedGenerator, transformation::LinearTransformation, source::CategorizedGenerator{<:Operators}; kwargs...) -> CategorizedGenerator

Update the parameters (including the boundary parameters) of a categorized generator based on its source categorized generator of quantum operators and the corresponding linear transformation.

!!! Note The coefficients of boundops are also updated due to the change of the boundary parameters.

QuantumLattices.update! — Method

update!(cat::CategorizedGenerator{<:OperatorSum}; parameters...) -> CategorizedGenerator

Update the parameters (including the boundary parameters) of a categorized generator of quantum operators.

!!! Note The coefficients of boundops are also updated due to the change of the boundary parameters.

QuantumLattices.update! — Method

update!(gen::Image, source::CompositeGenerator; kwargs...) -> Image

Update the parameters of the image based on its source generator of quantum operators.

QuantumLattices.update! — Method

update!(gen::Image; parameters...) -> typeof(gen)

Update the parameters of the image of a transformation applied to a generator of quantum operators.

QuantumLattices.update! — Method

update!(gen::OperatorGenerator; parameters...) -> typeof(gen)

Update the coefficients of the terms in a generator.

QuantumLattices.update! — Method

update!(hamiltonian::SimpleHamiltonian{<:OperatorSet}; parameters...) -> typeof(hamiltonian)
update!(hamiltonian::SimpleHamiltonian{<:Union{Formula, Generator}}; parameters...) -> typeof(hamiltonian)

Update the parameters of a Hamiltonian.

RecipesBase.apply_recipe — Method

@recipe plot(pack::Tuple{Algorithm, Assignment})

Define the recipe for the visualization of an assignment of an algorithm.