Interfaces

This module contains the generic functions that are extended by the package.

Due to the multi-dispatch feature of Julia, generic functions can be extended by local methods for different types. However, a local definition of a method also claims a new generic function if the generic function is not imported to the current scope, thus ruins the definitions in other modules. Therefore, it is quite necessary to predefine the common generic functions in a separate module, so that other modules can extend them with their own by a simple import.

Manual

QuantumLattices.Interfaces.:⊕ — Function

Direct sum.

QuantumLattices.Interfaces.:⊗ — Function

Direct product.

QuantumLattices.Interfaces.:⋅ — Function

Dot product.

QuantumLattices.Interfaces.add! — Function

Inplace addition.

QuantumLattices.Interfaces.decompose — Function

Decompose.

QuantumLattices.Interfaces.decompose! — Function

In place decomposition.

QuantumLattices.Interfaces.dimension — Function

Dimension.

QuantumLattices.Interfaces.dims — Function

Dimensions.

QuantumLattices.Interfaces.div! — Function

Inplace division.

QuantumLattices.Interfaces.expand — Function

Get the expansion.

QuantumLattices.Interfaces.expand! — Function

In place expansion.

QuantumLattices.Interfaces.id — Function

Id.

QuantumLattices.Interfaces.index — Function

Index.

QuantumLattices.Interfaces.inds — Function

Indices.

QuantumLattices.Interfaces.kind — Function

Kind.

QuantumLattices.Interfaces.matrix — Function

Matrix representation.

QuantumLattices.Interfaces.mul! — Function

Inplace multiplication.

QuantumLattices.Interfaces.permute — Function

Get the permutation.

QuantumLattices.Interfaces.prepare! — Function

Inplace prepare.

QuantumLattices.Interfaces.rank — Function

Rank.

QuantumLattices.Interfaces.register! — Function

Inplace register.

QuantumLattices.Interfaces.regularize — Function

Regularize.

QuantumLattices.Interfaces.regularize! — Function

In place regularization.

QuantumLattices.Interfaces.reset! — Function

Inplace reset.

QuantumLattices.Interfaces.run! — Function

Inplace run.

QuantumLattices.Interfaces.sequence — Function

Sequence.

QuantumLattices.Interfaces.sub! — Function

Inplace subtraction.

QuantumLattices.Interfaces.update! — Function

Inplace Update.

QuantumLattices.Interfaces.vector — Function

Vector representation.