Interfaces

Interfaces

This module contains the generic functions that are extended by the package.

Due to the multidispatch feature of Julia, generic functions can be extended by local methods for different types. However, a local definition of a method also claims a new generic function if the generic function is not imported to the current scope, thus ruins the definitions in other modules. Therefore, it is quite necessary to prefine the common generic functions in a separate module, so that other modules can extend them with their own by a simple import.

Manual

QuantumLattices.Interfaces.:⊕ — Function.

Direct sum.

source

QuantumLattices.Interfaces.:⊗ — Function.

Direct product.

source

QuantumLattices.Interfaces.:⋅ — Function.

Dot product.

source

QuantumLattices.Interfaces.add! — Function.

Inplace addition.

source

QuantumLattices.Interfaces.decompose — Function.

Decompose.

source

QuantumLattices.Interfaces.decompose! — Function.

In place decomposition.

source

QuantumLattices.Interfaces.degree — Function.

Degree

source

QuantumLattices.Interfaces.dimension — Function.

Dimension.

source

QuantumLattices.Interfaces.dims — Function.

Dimensions.

source

QuantumLattices.Interfaces.div! — Function.

Inplace division.

source

QuantumLattices.Interfaces.expand — Function.

Get the expansion.

source

QuantumLattices.Interfaces.expand! — Function.

In place expansion.

source

QuantumLattices.Interfaces.id — Function.

Id.

source

QuantumLattices.Interfaces.index — Function.

Index.

source

QuantumLattices.Interfaces.inds — Function.

Indices.

source

QuantumLattices.Interfaces.kind — Function.

Kind.

source

QuantumLattices.Interfaces.matrix — Function.

Matrix representation.

source

QuantumLattices.Interfaces.mul! — Function.

Inplace multiplication.

source

QuantumLattices.Interfaces.permute — Function.

Get the permutation.

source

QuantumLattices.Interfaces.rank — Function.

Rank.

source

QuantumLattices.Interfaces.regularize — Function.

Regularize.

source

QuantumLattices.Interfaces.regularize! — Function.

In place regularization.

source

QuantumLattices.Interfaces.reset! — Function.

Inplace reset.

source

QuantumLattices.Interfaces.sequence — Function.

Sequence.

source

QuantumLattices.Interfaces.sub! — Function.

Inplace subtraction.

source

QuantumLattices.Interfaces.update! — Function.

Inplace Update.

source

QuantumLattices.Interfaces.vector — Function.

Vector representation.

source