Algebra over fields

Algebra over fields

An algebra over a field is a vector space over that field, in which a bilinear operator (often called the "multiplication") between vectors is defined.

With the help of the structure constants of the algebra, the result of the bilinear operation between any arbitary two vectors can be expressed by a sum of individual ones. Therefore, in principle, an algebra can be represented by the complete basis set of its corresponding vector space and a rank-3 tensor encapsulating its structure constants. It is noted that the "bilinear operation" is not restriced to the usual multiplication only. For example, it is the commutator, which is a composition of the usual multiplication and subtraction (for any A and B, the commutator [A,B] is defined as [A,B]≝AB-BA) that serves as the bilinear operator for Lie algebras. In this module, for scalars in the field and elements in the algebra, we only provide the interfaces of the scalar multiplication (including the scalar division) bwteen a sclar and an element, the addition (including the subtraction) and the usual multiplication between two elements. Other complicated operations should be composed from these basic ones.

SimpleID and ID

SimpleID is the building block of the id system of an algebra over a field, while ID defines the specific identifier of an element in that algebra.

Generally, the usual multiplication between two elements of an algebra is not commutable, and the rank of the multiplication is just the add-up before the simplication with the help of the algebra structure. We thus use a simple id to denote a single basis of the corresponding vector space, and an id to denote the identifier of an element. With the help of the direct product (⊗) of two ids, an over complete id system designed for the whole algebra is constructed. This id system is redundant because it does not reflects the structure constants of the algebra, which reduces independent basis elements. Extra mechanisms should be provided to kill this redundancy, which goes beyond the current module. Users should define them themselves.

IdSpace

IdSpace defines the complete vector space that corresponds to an algebra.

An id space uses a set of simple id bases to generated all ranked ids, with the former represented by an instance of DirectVectorSpace and stored in the attribute :sids, while the latter by an instance of GradedTables in the attribute :tables. For the sick of the orderings among different ids, an id will be represented by a multiindex, which are determined by the corresponding sequences of its simple ids in the basis set. The id system can be complete or over complete, depending on whether the attribute :tables contain multiindices that actually represent the same basis of the algebra. Besides, the methods

Base.getindex(idspace::IdSpace,i::Int) -> ID

and

Base.findfirst(id::ID,idspace::IdSpace) -> Int

are provided to get the ith id and to find the index of an id of a idspace, respectively.

Element and Elements

Element defines a single element of an algebra while Elements defines an exprssion composed of several elements from an algebra.

The first and second attributes of an Element must be

value::Number: the coefficient of the element

id::ID: the id of the element

Arithmetic operations (+, -, *, /) bwteen a scalar, an Element or an Elements is defined. See Manual for details.

Manual

QuantumLattices.Mathematics.AlgebraOverFields.Element — Type.

Element{V,I<:ID}

An element of an algebra over a field.

The first and second attributes of an element must be

value: the coefficient of the element

id::ID: the id of the element

source

QuantumLattices.Mathematics.AlgebraOverFields.Elements — Type.

Elements{I<:ID,M<:Element} <: AbstractDict{I,M}

An set of elements of an algebra over a field.

Alias for Dict{I<:ID,M<:Element}. Similar iterms are automatically merged thanks to the id system.

source

QuantumLattices.Mathematics.AlgebraOverFields.Elements — Method.

Elements(ms)
Elements(ms::Pair{I,M}...) where {I<:ID,M<:Element}
Elements(ms::Element...)

Get the set of elements with similar items merged.

source

QuantumLattices.Mathematics.AlgebraOverFields.ID — Type.

ID(ids::NTuple{N,SimpleID}) where N
ID(ids::SimpleID...)
ID(::Type{SID},attrs::Vararg{NTuple{N},M}) where {SID<:SimpleID,N,M}

The id system of an algebra over a field.

source

QuantumLattices.Mathematics.AlgebraOverFields.IdSpace — Type.

IdSpace(sids::DirectVectorSpace,tables::GradedTables)
IdSpace(::Type{M},sids::DirectVectorSpace,gs::Val{GS}) where {M<:AbstractCombinatorics,GS}

The graded id space for an algebra generated by a couple of basic simple ids.

source

QuantumLattices.Mathematics.AlgebraOverFields.ScalarElement — Type.

ScalarElement{V}

Identity element.

source

QuantumLattices.Mathematics.AlgebraOverFields.SimpleID — Type.

SimpleID <: NamedVector

A simple id is the building block of the id system of an algebra over a field.

source

QuantumLattices.Mathematics.AlgebraOverFields.idtype — Method.

idtype(m::Element)
idtype(::Type{<:Element{V,I} where V}) where I<:ID
idtype(::Type{<:Element})

The type of the id of an element.

source

QuantumLattices.Mathematics.AlgebraOverFields.rawelement — Method.

rawelement(::Type{<:Element})

Get the raw name of a type of Element.

source

Base.:* — Method.

*(factor,m::Element) -> Element
*(m::Element,factor) -> Element
*(m1::Element,m2::Element) -> Element
*(m::Element,::Nothing) -> Nothing
*(::Nothing,m::Element) -> Nothing
*(factor,ms::Elements) -> Elements
*(ms::Elements,factor) -> Elements
*(m::Element,ms::Elements) -> Elements
*(ms::Elements,m::Element) -> Elements
*(ms1::Elements,ms2::Elements) -> Elements
*(ms::Elements,::Nothing) -> Nothing
*(::Nothing,ms::Elements) -> Nothing

Overloaded * operator for element-scalar multiplications and element-element multiplications of an algebra over a field.

source

Base.:* — Method.

*(sid1::SimpleID,sid2::SimpleID) -> ID
*(sid::SimpleID,cid::ID) -> ID
*(cid::ID,sid::SimpleID) -> ID
*(cid1::ID,cid2::ID) -> ID

Get the product of the id system.

source

Base.:+ — Method.

+(m::Element) -> typeof(m)
+(ms::Elements) -> typeof(ms)
+(m::Element,::Nothing) -> typeof(m)
+(::Nothing,m::Element) -> typeof(m)
+(m::Element,factor) -> Elements
+(factor,m::Element) -> Elements
+(ms::Elements,::Nothing) -> typeof(ms)
+(::Nothing,ms::Elements) -> typeof(ms)
+(ms::Elements,factor) -> Elements
+(factor,ms::Elements) -> Elements
+(ms::Elements,m::Element) -> Elements
+(m1::Element,m2::Element) -> Elements
+(m::Element,ms::Elements) -> Elements
+(ms1::Elements,ms2::Elements) -> Elements

Overloaded + operator between elements of an algebra over a field.

source

Base.:- — Method.

-(m::Element) -> typeof(m)
-(ms::Elements) -> typeof(ms)
-(m::Element,::Nothing) -> typeof(m)
-(::Nothing,m::Element) -> typeof(m)
-(ms::Elements,::Nothing) -> typeof(ms)
-(::Nothing,ms::Elements) -> typeof(ms)
-(m::Element,factor) -> Elements
-(factor,m::Element) -> Elements
-(ms::Elements,factor) -> Elements
-(factor,ms::Elements) -> Elements
-(m1::Element,m2::Element) -> Elements
-(m::Element,ms::Elements) -> Elements
-(ms::Elements,m::Element) -> Elements
-(ms1::Elements,ms2::Elements) -> Elements

Overloaded - operator between elements of an algebra over a field.

source

Base.:/ — Method.

/(m::Element,factor) -> Element
/(m::Element,factor::ScalarElement) -> Element
/(ms::Elements,factor) -> Elements
/(ms::Elements,factor::ScalarElement) -> Elements

Overloaded / operator for element-scalar division of an algebra over a field.

source

Base.:== — Method.

==(ms::Elements,::Nothing) -> Bool
==(::Nothing,ms::Elements) -> Bool
isequal(ms::Elements,::Nothing) -> Bool
isequal(::Nothing,ms::Elements) -> Bool

source

Base.:== — Method.

(m1::Element,m2::Element) -> Bool
isequal(m1::Element,m2::Element) -> Bool

Compare two elements and judge whether they are equal to each other.

source

Base.:^ — Method.

^(m::Element,n::Int) -> Element
^(ms::Elements,n::Int) -> Elements

Overloaded ^ operator for element-integer power of an algebra over a field.

source

Base.convert — Method.

convert(::Type{M},m::ScalarElement) where {M<:ScalarElement}
convert(::Type{M},m) where {M<:ScalarElement}
convert(::Type{M},m::Element) where {M<:Element}

Convert a scalar element from one type to another;

Convert a scalar to a scalar element;

Convert an element from one type to another.

source

Base.findfirst — Method.

findfirst(id::ID,idspace::IdSpace) -> Int
searchsortedfirst(idspace::IdSpace,id::ID) -> Int

Find the index of an id in a idspace.

source

Base.getindex — Method.

getindex(idspace::IdSpace,i::Int) -> ID

Get the ith id of a idspace.

source

Base.getproperty — Method.

getproperty(cid::ID,name::Symbol)

Get the property of a composite id.

source

Base.isapprox — Method.

isapprox(m1::Element,m2::Element;atol::Real=atol,rtol::Real=rtol) -> Bool

Compare two elements and judge whether they are inexactly equivalent to each other.

source

Base.isless — Method.

isless(cid1::ID,cid2::ID) -> Bool
<(cid1::ID,cid2::ID) -> Bool

Compare two ids and judge whether the first is less than the second.

We assume that ids with smaller ranks are always less than those with higher ranks. If two ids are of the same rank, the comparison goes just like that between tuples.

source

Base.one — Method.

one(::Type{M}) where {M<:Element}

Get the identity operator.

source

Base.permute! — Function.

permute!(result::Elements,m::Element,table=nothing) -> Elements
permute!(result::Elements,ms::Elements,table=nothing) -> Elements

Permute the ids of an-element/a-set-of-elements to the descending order according to a table, and store the permuted elements in result.

source

Base.promote_rule — Method.

promote_rule(::Type{I1},::Type{I2}) where {I1<:ID,I2<:ID}

Define the promote rule for ID types.

source

Base.promote_rule — Method.

promote_rule(::Type{M1},::Type{M2}) where {M1<:Element,M2<:Element}

Define the promote rule for Element types.

source

Base.propertynames — Method.

propertynames(::Type{I},private::Bool=false) where I<:ID -> Tuple

Get the property names of a composite id.

source

Base.replace — Method.

replace(m::Element,pairs::Pair{<:SimpleID,<:Union{Element,Elements}}...) -> Element/Elements
replace(ms::Elements,pairs::Pair{<:SimpleID,<:Union{Element,Elements}}...) -> Elements

Replace the rank-1 components of an element with new element/elements.

source

Base.replace — Method.

replace(m::Element;kwargs...) -> typeof(m)

Return a copy of a concrete Element with some of the field values replaced by the keyword arguments.

source

Base.repr — Method.

repr(ms::Elements) -> String

Get the repr representation of a set of elements.

source

Base.show — Method.

show(io::IO,ms::Elements)

Show a set of elements.

source

Base.show — Method.

show(io::IO,cid::ID)

Show a composite id.

source

Base.split — Method.

split(m::Element) -> Tuple{Any,Vararg{Element}}

Split an element into the coefficient and a sequence of rank-1 elements.

source

Base.valtype — Method.

valtype(m::Element)
valtype(::Type{<:Element})
valtype(::Type{<:Element{V}}) where V

Get the type of the value of an element.

The result is also the type of the field over which the algebra is defined.

source

Base.zero — Method.

zero(ms::Elements) -> Nothing
zero(::Type{<:Elements}) -> Nothing

Get a zero set of elements.

A zero set of elements is defined to be the one with no elements.

source

QuantumLattices.Interfaces.:⊗ — Method.

⊗(m::Element,ms::Elements) -> Elements
⊗(ms::Elements,m::Element) -> Elements
⊗(ms1::Elements,ms2::Elements) -> Elements

Overloaded ⊗ operator for element-element multiplications of an algebra over a field.

source

QuantumLattices.Interfaces.:⋅ — Method.

⋅(m::Element,ms::Elements) -> Elements
⋅(ms::Elements,m::Element) -> Elements
⋅(ms1::Elements,ms2::Elements) -> Elements

Overloaded ⋅ operator for element-element multiplications of an algebra over a field.

source

QuantumLattices.Interfaces.add! — Method.

add!(ms::Elements) -> typeof(ms)
add!(ms::Elements,::Nothing) -> typeof(ms)
add!(ms::Elements,m) -> typeof(ms)
add!(ms::Elements,m::Element) -> typeof(ms)
add!(ms::Elements,mms::Elements) -> typeof(ms)

Get the inplace addition of elements to a set.

source

QuantumLattices.Interfaces.div! — Method.

div!(ms::Elements,factor::ScalarElement) -> Elements
div!(ms::Elements,factor) -> Elements

Get the inplace division of element with a scalar.

source

QuantumLattices.Interfaces.mul! — Method.

mul!(ms::Elements,factor::ScalarElement) -> Elements
mul!(ms::Elements,factor) -> Elements

Get the inplace multiplication of elements with a scalar.

source

QuantumLattices.Interfaces.permute — Function.

permute(m::Element,table=nothing) -> Elements
permute(ms::Elements,table=nothing) -> Elements

Permute the ids of an-element/a-set-of-elements to the descending order according to a table.

source

QuantumLattices.Interfaces.rank — Method.

rank(m::Element) -> Int
rank(::Type{M}) where M<:Element -> Int

Get the rank of an element.

source

QuantumLattices.Interfaces.rank — Method.

rank(id::ID) -> Int
rank(::Type{<:ID}) -> Any
rank(::Type{<:ID{T}}) where T<:Tuple{Vararg{SimpleID}} -> Int

Get the rank of a composite id.

source

QuantumLattices.Interfaces.sequence — Method.

sequence(m::Element,table=nothing) -> NTuple{rank(m),Int}

Get the sequence of the ids of an element according to a table.

source

QuantumLattices.Interfaces.sub! — Method.

sub!(ms::Elements) -> typeof(ms)
sub!(ms::Elements,::Nothing) -> typeof(ms)
add!(ms::Elements,m) -> typeof(ms)
sub!(ms::Elements,m::Element) -> typeof(ms)
sub!(ms::Elements,mms::Elements) -> typeof(ms)

Get the inplace subtraction of elements from a set.

source