Quantum numbers

Quantum numbers can be considered as the conserved labels for the bases of a Hilbert space when a quantum system hosts some symmetries. Here we only implement Abelian quantum numbers because non-Abelian ones are far more complicated yet much less used. In practice, quantum numbers can be integers or half integers, therefore, we use real numbers to denote them in this module for simplicity. Independent quantum numbers, such as the particle number and the spin z-component, can coexist at the same time. We use type AbelianNumber to represent the complete set of independent ones for a single basis of a Hilbert space, and type AbelianNumbers to represent the whole quantum numbers for the total bases.

AbelianNumber

The abstract type for the complete set of independent quantum numbers for a single basis.

Main features include:

function fieldnames: get the names of the quantum numbers

function periods: get the periods of the quantum numbers

arithmetic operations: +, -, *, ^, ⊕, ⊗

hashable: concrete instances can be used as keys for a dict or a set

iterable: concrete instances are iterable over their values

comparable: two concrete instances can be compared

In particular, AbelianNumber <: HomoNamedVector{Float}, all features supported by HomoNamedVector are also available for HomoNamedVector. See also HomoNamedVector.

For convenience, 4 kinds of quantum numbers are predefined in this module, i.e.

SQN: for spin z-component reserved systems

PQN: for particle number reserved systems

SPQN: for both particle number and spin-z component reserved systems

Z2QN: for systems with a $Z_2$ conservation quantum number

Users who want to define their own $Z_N$-like quantum numbers must handle the periodicities in the construction function, otherwise, wrong results will be get when arithmetic operations, such as + or -, are involved. It is highly recommended to use the macro @abeliannumber to define your own concrete AbelianNumbers.

AbelianNumbers

The whole quantum numbers for the total bases.

By design, a AbelianNumbers{QN} has one type parameter:

QN<:AbelianNumber: the type of the quantum numbers contained in it

And 3 attributes:

form::Char: Its form, whose value must be one of the followings

'G': the general form, which has no restriction for its contents

'U': the unitary form, which requires no duplicates in its contents

'C': the canonical form, which requires both no duplicates and ascending-order in its contents

Usually, G-formed and U-formed AbelianNumberses can be transformed to the corresponding C-formed ones by the sort function.

contents::Vector{QN}: The quantum numbers contained in it. To achieve high efficiency, it is required to be an homogenous array of a certain kind of concrete AbelianNumber.

indptr::Vector{Int}: The indptr of the quantum numbers contained in it, which is similar to the colptr attribute of a CSC sparse matrix and records the compression info of its contents.

Main features include:

function eltype: get the concrete type of the quantum numbers it contains

index access: get the contents directly by the getindex function

arithmetic operations: +, -, *, ^, ⊗, ⊕

iterable: various iteration supports, including functions such as iterate, keys, values and pairs

For a complete summation of its features, please refer to the manual.

For convenience, 5 functions are predefined to generate the AbelianNumbers of common physical systems, i.e.

SQNS: a single spin

PQNS: a single-particle state with at most N identical particles

SzPQNS: a single-particle state with at most one particle whose spin-z component is Sz

SPQNS: a single site with internal degrees of freedom that can be ascribed to a spin

Z2QNS: any $Z_2$ Hilbert space

Manual

QuantumLattices.Mathematics.QuantumNumbers.qnbruteforce — Constant

qnbruteforce

Indicate that decompose uses the brute force method.

QuantumLattices.Mathematics.QuantumNumbers.qncompression — Constant

qncompression

Indicate that findall and permute use the compressed contents.

QuantumLattices.Mathematics.QuantumNumbers.qncontents — Constant

qncontents

Indicate that expand uses the compressed/expanded contents.

QuantumLattices.Mathematics.QuantumNumbers.qncounts — Constant

qncounts

Indicate that methods with AbelianNumbers use the count number of the compressed contents.

QuantumLattices.Mathematics.QuantumNumbers.qnexpansion — Constant

qnexpansion

Indicate that findall and permute use the expanded contents.

QuantumLattices.Mathematics.QuantumNumbers.qnindices — Constant

qnindices

Indicate that expand uses the indices of the compressed/expanded contents.

QuantumLattices.Mathematics.QuantumNumbers.qnindptr — Constant

qnindptr

Indicate that methods with AbelianNumbers use the index pointer of the compressed contents.

QuantumLattices.Mathematics.QuantumNumbers.qnmontecarlo — Constant

qnmontecarlo

Indicate that decompose uses the Monte Carlo method.

QuantumLattices.Mathematics.QuantumNumbers.AbelianNumber — Type

Abstract type for all concrete quantum numbers for a single basis.

QuantumLattices.Mathematics.QuantumNumbers.AbelianNumbers — Type

AbelianNumbers(form::Char, contents::Vector{<:AbelianNumber}, counts::Vector{Int}, ::QNCounts)
AbelianNumbers(form::Char, contents::Vector{<:AbelianNumber}, indptr::Vector{Int}, ::QNIndptr)

The whole quantum numbers of the total bases of a Hilbert space.

The default constructors construct an AbelianNumbers from a vector of concrete quantum numbers and an vector containing their counts or indptr.

QuantumLattices.Mathematics.QuantumNumbers.AbelianNumbers — Type

AbelianNumbers(qn::AbelianNumber, count::Int=1)

Construct an AbelianNumbers with one unique quantum number which occurs count times.

QuantumLattices.Mathematics.QuantumNumbers.AbelianNumbers — Method

AbelianNumbers(od::OrderedDict{<:AbelianNumber, <:UnitRange{<:Integer}})

Construct an AbelianNumbers from an ordered dict containing concrete quantum numbers and their slices.

QuantumLattices.Mathematics.QuantumNumbers.AbelianNumbers — Method

AbelianNumbers(od::OrderedDict{<:AbelianNumber, <:Integer})

Construct an AbelianNumbers from an ordered dict containing concrete quantum numbers and their counts.

QuantumLattices.Mathematics.QuantumNumbers.PQN — Type

PQN(N::Real)

The concrete AbelianNumber of a quantum system with particle number N conserved.

QuantumLattices.Mathematics.QuantumNumbers.SPQN — Type

SPQN(N::Real, Sz::Real)

The concrete AbelianNumber of a quantum system with both particle number N and spin z-component Sz conserved.

QuantumLattices.Mathematics.QuantumNumbers.SQN — Type

SQN(Sz::Real)

The concrete AbelianNumber of a quantum system with spin z-component Sz conserved.

QuantumLattices.Mathematics.QuantumNumbers.Z2QN — Type

Z2QN(N::Real)

The concrete AbelianNumber of a quantum system with a Z₂-like conserved quantity.

QuantumLattices.Mathematics.QuantumNumbers.@abeliannumber — Macro

@abeliannumber typename fieldnames fieldperiods

Construct a concrete AbelianNumber with the type name being typename, fieldnames specified by fieldnames and periods specified by fieldperiods.

Base.:* — Method

*(qn::AbelianNumber, factor::Integer) -> typeof(qn)
*(factor::Integer, qn::AbelianNumber) -> typeof(qn)
*(qns::AbelianNumbers, factor::Integer) -> AbelianNumbers
*(factor::Integer, qns::AbelianNumbers) -> AbelianNumbers

Overloaded * operator for the multiplication between an integer and an AbelianNumber or an AbelianNumbers.

Base.:+ — Method

+(qn::AbelianNumber) -> typeof(qn)
+(qn::QN, qns::QN...) where QN<:AbelianNumber -> QN
+(qns::AbelianNumbers) -> AbelianNumbers
+(qn::QN, qns::AbelianNumbers{QN}) where QN<:AbelianNumber -> AbelianNumbers{QN}
+(qns::AbelianNumbers{QN}, qn::QN) where QN<:AbelianNumber -> AbelianNumbers{QN}

Overloaded + operator for AbelianNumber and AbelianNumbers.

Note

The addition between an AbelianNumbers and an AbelianNumber is just a global shift of the contents of the AbelianNumbers by the AbelianNumber, therefore, the result is an AbelianNumbers.

+ cannot be used between two AbelianNumbers because the result is ambiguous. Instead, use ⊕ for direct sum and ⊗ for direct product.

To ensure type stability, two AbelianNumber can be added together if and only if they are of the same type.

Similarly, an AbelianNumber and an AbelianNumbers can be added together if and only if the former's type is the same with the latter's eltype.

Base.:- — Method

-(qn::AbelianNumber) -> typeof(qn)
-(qn1::QN, qn2::QN) where QN<:AbelianNumber -> QN
-(qns::AbelianNumbers) -> AbelianNumbers
-(qn::QN, qns::AbelianNumbers{QN}) where QN<:AbelianNumber -> AbelianNumbers{QN}
-(qns::AbelianNumbers{QN}, qn::QN) where QN<:AbelianNumber -> AbelianNumbers{QN}

Overloaded - operator for AbelianNumber and AbelianNumbers.

Note

The subtraction between an AbelianNumbers and an AbelianNumber is just a global shift of the contents of the AbelianNumbers by the AbelianNumber, therefore, the result is an AbelianNumbers.

- cannot be used between two AbelianNumbers because the result is ambiguous. Instead, use ⊕ with signs for direct sum and ⊗ with signs for direct product.

To ensure type stability, an AbelianNumber can be subtracted by another AbelianNumber if and only if they are of the same type.

Similarly, an AbelianNumber can be subtracted by an AbelianNumbers or vice versa if and only if the former's type is the same with the latter's eltype.

Base.:== — Method

==(qns1::AbelianNumbers, qns2::AbelianNumbers) -> Bool

Overloaded equivalent operator. Two AbelianNumberses are equal to each other if and only if both their contentses and indptrs are elementwise equal to each other.

Note

It is not necessary for two AbelianNumberses to have the same eltype nor the same form to be equal to each other.

Base.:^ — Method

^(qn::AbelianNumber, factor::Integer) -> typeof(qn)
^(qns::AbelianNumbers, factor::Integer) -> AbelianNumbers

Overloaded ^ operator for AbelianNumber and AbelianNumbers. This operation translates into the direct product of factor copies of qn or qns.

Base.eltype — Method

eltype(::Type{<:AbelianNumbers{QN}}) where {QN<:AbelianNumber}
eltype(qns::AbelianNumbers)

Get the type of the concrete AbelianNumber contained in an AbelianNumbers.

Base.filter — Method

filter(target::QN, qns::AbelianNumbers{QN}) where QN<:AbelianNumber -> AbelianNumbers{QN}
filter(targets::NTuple{N, QN}, qns::AbelianNumbers{QN}) where {N, QN<:AbelianNumber} -> AbelianNumbers{QN}

Find a subset of an AbelianNumbers by picking out the quantum numbers in targets.

Base.findall — Method

findall(target::QN, qns::AbelianNumbers{QN}, ::QNCompression) where QN<:AbelianNumber -> Vector{Int}
findall(target::QN, qns::AbelianNumbers{QN}, ::QNExpansion) where QN<:AbelianNumber -> Vector{Int}
findall(targets::NTuple{N, QN}, qns::AbelianNumbers{QN}, ::QNCompression) where {N, QN<:AbelianNumber} -> Vector{Int}
findall(targets::NTuple{N, QN}, qns::AbelianNumbers{QN}, ::QNExpansion) where {N, QN<:AbelianNumber} -> Vector{Int}

Find all the indices of the target quantum numbers in the contents (qncompression case) or the expansion (qnexpansion case) of an AbelianNumbers.

Base.getindex — Method

getindex(qns::AbelianNumbers, index::Int) -> eltype(qns)
getindex(qns::AbelianNumbers, slice::UnitRange{Int}) -> AbelianNumbers
getindex(qns::AbelianNumbers, indices::Vector{Int}) -> AbelianNumbers

Overloaded [] operator.

Note

For an AbelianNumbers, all these getindex functions act on its contents, i.e. its compressed data, but not on its expansion, i.e. the uncompressed data. This definition is consistent with the length function.

When the index is an integer, the result is an AbelianNumber, while when the index is a unit range or a vector of intgers, the result is an AbelianNumbers. The logic is quite reasonable because such behaviors are much alike to those of a vector container.

Base.isequal — Method

isequal(qns1::AbelianNumbers, qns2::AbelianNumbers) -> Bool

Overloaded equivalent operator. Two AbelianNumberses are equal to each other if and only if both their contentses and indptrs are elementwise equal to each other.

Note

It is not necessary for two AbelianNumberses to have the same eltype nor the same form to be equal to each other.

Base.iterate — Function

iterate(qns::AbelianNumbers, state::Int=1)
iterate(rv::Iterators.Reverse{<:AbelianNumbers}, state::Int=length(rv.itr, false))

Iterate or reversely iterate over the concrete AbelianNumbers contained in an AbelianNumbers.

Base.keys — Method

keys(qns::AbelianNumbers) -> Vector{qns|>eltype}

Iterate over the concrete AbelianNumbers contained in an AbelianNumbers.

Base.kron — Method

kron(::Type{QN}, qn1::AbelianNumber, qn2::AbelianNumber) where QN<:AbelianNumber -> QN
kron(qns::Vararg{<:AbelianNumber, N}; signs::NTuple{N, Int}=ntuple(i->1, Val(N))) where N -> eltype(qns)
kron(qnses::Vararg{AbelianNumbers{QN}, N}; signs::NTuple{N, Int}=ntuple(i->1, Val(N))) where {N, QN<:AbelianNumber} -> AbelianNumbers{QN}

Get the direct product of some AbelianNumbers or AbelianNumberses.

Note

Physically, the direct product of a couple of AbelianNumbers or AbelianNumberses are defined by the direct product of the bases of the Hilbert spaces they represent. Therefore, AbelianNumbers with differenct types or AbelianNumberses with differenct eltypes are allowed to be direct producted in principle. However, for simplicity, we only implement a method which handle the situation of two AbelianNumbers with differenct types. The type of the result should be provided as the first parameter. Note that in this situation, the fieldnames and periods of the result type must be exactly equal to the flattened fieldnames and periods of the two input AbelianNumbers, which means, even the order of the input AbelianNumbers matters.

Apparently, the dimension of the result equals the product of those of the inputs. Therefore, the direct product of AbelianNumbers is also an AbelianNumber since its dimension is still one.

For other situations except the one mentioned in Note.1, the input AbelianNumbers or AbelianNumberses must be homogenous. Meanwhile, signs can also be provided for these situations. Note that each quantum number in the contents of the result is obtained by a summation of the corresponding quanum numbers out of the inputs with the correct signs. This is a direct observation of the Abelian nature of our quantum numbers.

Base.length — Method

length(qns::AbelianNumbers) -> Int

Get the number of unduplicate qunatum numbers in the AbelianNumbers.

Base.pairs — Method

pairs(qns::AbelianNumbers, choice::Union{QNIndptr, QNCounts})

Iterate over the AbelianNumber=>slice or AbelianNumber=>count pairs.

Base.show — Method

show(io::IO, qns::AbelianNumbers)

Show an AbelianNumbers.

Base.sort — Method

sort(qns::AbelianNumbers) -> Tuple{AbelianNumbers, Vector{Int}}

Sort the quantum numbers of an AbelianNumber, return the sorted AbelianNumber and the permutation array that sorts the expansion of the original AbelianNumbers.

Base.string — Method

string(qns::AbelianNumbers) -> String

Convert an AbelianNumbers to string.

Base.union — Method

union(qns::Vararg{<:AbelianNumber, N}; signs::NTuple{N, Int}=ntuple(i->1, Val(N))) where N -> AbelianNumbers
union(qnses::Vararg{AbelianNumbers{QN}, N}; signs::NTuple{N, Int}=ntuple(i->1, Val(N))) where {N, QN<:AbelianNumber} -> AbelianNumbers{QN}

Get the direct sum of some AbelianNumbers or AbelianNumberses.

Note

Physically, the direct sum of a couple of AbelianNumbers or AbelianNumberses is defined by the direct sum of the bases of the Hilbert spaces they represent. Therefore, the input AbelianNumbers or AbelianNumberses must be homogenous. Inhomogenous 'AbelianNumber's must be direct producted first to ensure homogenity before the direct sum.

Apparently, the dimension of the result equals the summation of those of the inputs, which means, even for AbelianNumbers, the result will be naturally an AbelianNumbers because the dimension of the result is larger than 1.

Signs of AbelianNumbers or AbelianNumberses can be provided when getting their direct sums.

Base.values — Method

values(qns::AbelianNumbers, ::QNIndptr)
values(qns::AbelianNumbers, ::QNCounts)

Iterate over the slices/counts of the AbelianNumbers.

QuantumLattices.Interfaces.:⊕ — Method

⊕(qns::AbelianNumber...) -> AbelianNumbers{qns|>eltype}
⊕(qnses::AbelianNumbers...) -> qnses|>eltype

Get the direct sum of some AbelianNumbers or AbelianNumberses.

QuantumLattices.Interfaces.:⊗ — Method

⊗(qns::AbelianNumber...) -> eltype(qns)
⊗(qnses::AbelianNumbers...) -> eltype(qnses)

Get the direct product of some AbelianNumbers or AbelianNumberses.

QuantumLattices.Interfaces.decompose — Method

decompose(qnses::NTuple{N, AbelianNumbers{QN}}, target::QN, signs::NTuple{N, Int}, ::QNBruteForce; nmax::Int=20) where {N, QN<:AbelianNumber} -> Vector{NTuple{N, Int}}
decompose(qnses::NTuple{N, AbelianNumbers{QN}}, target::QN, signs::NTuple{N, Int}, ::QNMonteCarlo; nmax::Int=20) where {N, QN<:AbelianNumber} -> Vector{NTuple{N, Int}}

Find a couple of decompositions of target with respect to qnses.

Note

A tuple of integers (i₁, i₂, ...) is called a decomposition of a given target with respect to the given qnses if and only if they satisfy the "decomposition rule":

\[\sum_\text{j} \text{signs}[\text{j}]\times\text{qnses}[\text{j}][\text{i}_{\text{j}}]==\text{target}\]

This equation is in fact a kind of a set of restricted linear Diophantine equations. Indeed, our quantum numbers are always discrete Abelian ones and all instances of a concrete AbelianNumber forms a module over the ring of integers. Therefore, each quantum number can be represented as a integral multiple of the unit element of the Abelian module, which results in the final reduction of the above equation to a set of linear Diophantine equations. Then finding a decomposition is equivalent to find a solution of the reduced linear Diophantine equations, with the restriction that the quantum numbers constructed from the solution should be in the corresponding qnses. Here we provide two methods to find such decompositions, one is by brute force (qnbruteforce case), and the other is by Monte Carlo simultatioins (qnmontecarlo case).

QuantumLattices.Interfaces.dimension — Method

dimension(qns::AbelianNumbers) -> Int

The dimension of the Hilbert space an AbelianNumbers represents.

QuantumLattices.Interfaces.dimension — Method

dimension(::Type{<:AbelianNumber}) -> Int
dimension(::AbelianNumber) -> Int

The dimension of the Hilbert space an AbelianNumber represents. Apparently, this is always 1.

QuantumLattices.Interfaces.expand — Method

expand(qns::AbelianNumbers, ::QNContents) -> Vector{qns|>eltype}
expand(qns::AbelianNumbers, ::QNIndices) -> Vector{Int}

Expand the contents (qncontents case) or indices (qnindices case) of an AbelianNumbers to the uncompressed form.

QuantumLattices.Interfaces.permute — Method

permute(qns::AbelianNumbers, permutation::Vector{Int}, ::QNCompression) -> AbelianNumbers
permute(qns::AbelianNumbers, permutation::Vector{Int}, ::QNExpansion) -> AbelianNumbers

Reorder the quantum numbers contained in an AbelianNumbers with a permutation and return the new one.

For qncompression case, the permutation is for the compressed contents of the original AbelianNumbers while for qnexpansion case, the permutation is for the expanded contents of the original AbelianNumbers.

QuantumLattices.Interfaces.regularize! — Method

regularize!(::Type{QN}, array::AbstractVector{<:Real}) where QN<:AbelianNumber -> typeof(array)
regularize!(::Type{QN}, array::AbstractMatrix{<:Real}) where QN<:AbelianNumber -> typeof(array)

Regularize the elements of an array in place so that it can represent quantum numbers.

QuantumLattices.Interfaces.regularize — Method

regularize(::Type{QN}, array::Union{AbstractVector{<:Real}, AbstractMatrix{<:Real}}) where QN<:AbelianNumber -> typeof(array)

Regularize the elements of an array and return a copy that can represent quantum numbers.

QuantumLattices.Mathematics.QuantumNumbers.PQNS — Method

PQNS(N::Real) -> AbelianNumbers{PQN}

Construct the AbelianNumbers of the Hilbert space of a single-particle state with at most N identical particles.

QuantumLattices.Mathematics.QuantumNumbers.SPQNS — Method

SPQNS(S::Real) -> AbelianNumbers{SPQN}

Construct the AbelianNumbers of the Hilbert space of a single site with internal degrees of freedom that can be ascribed to a spin S.

QuantumLattices.Mathematics.QuantumNumbers.SQNS — Method

SQNS(S::Real) -> AbelianNumbers{SQN}

Construct the AbelianNumbers of the Hilbert space of a signle spin S.

QuantumLattices.Mathematics.QuantumNumbers.SzPQNS — Method

SzPQNS(Sz::Real) -> AbelianNumbers{SPQN}

Construct the AbelianNumbers of the Hilbert space of a single-paritcle state with at most one particle whose spin-z component is Sz.

QuantumLattices.Mathematics.QuantumNumbers.Z2QNS — Method

Z2QNS() -> AbelianNumbers{Z2QN}

Construct the AbelianNumbers of a $Z_2$ Hilbert space.

QuantumLattices.Mathematics.QuantumNumbers.toordereddict — Method

toordereddict(qns::AbelianNumbers, ::QNIndptr) -> OrderedDict{qns|>eltype, UnitRange{Int}}
toordereddict(qns::AbelianNumbers, ::QNCounts) -> OrderedDict{qns|>eltype, Int}

Convert an AbelianNumbers to an ordered dict.

QuantumLattices.Mathematics.QuantumNumbers.ukron — Method

ukron(qnses::Vararg{AbelianNumbers{QN}, N}; signs::NTuple{N, Int}=ntuple(i->1, N)) where {N, QN<:AbelianNumber} -> AbelianNumbers{QN}, Dict{QN, Dict{NTuple{N, QN}, UnitRange{Int}}}

Unitary Kronecker product of several AbelianNumberses. The product result as well as the records of the product will be returned.

Note

All input AbelianNumbers must be 'U' formed or 'C' formed.

Since duplicate quantum number are not allowed in 'U' formed and 'C' formed AbelianNumberses, in general, there exists a merge process of duplicate quantum numbers in the product result. Therefore, records are needed to keep track of this process, which will be returned along with the product result. The records are stored in a Dict{QN, Dict{NTuple{N, QN}, UnitRange{Int}}} typed dict, in which, for each unduplicate quantum number qn in the product result, there exist a record Dict((qn₁, qn₂, ...)=>start:stop, ...) telling what quantum numbers (qn₁, qn₂, ...) a mereged duplicate qn comes from and what slice start:stop this merged duplicate corresponds.