Vector spaces

Vector spaces

A vector space is a linear space, in which the addition of vectors and multiplication of a vector by a scalar are defined.

Vector spaces are frequently encountered in physics, e.g. the Hilbert space in quantum mechanics. In this submodule, we only implement those with finite dimensions. We want to remark that in our implementation, a vector space is a subtype of an abstract vector, therefore, the bases always possess a order, which means, two vector spaces are not considered to be equal to each other even if their corresponding actual mathmatical spaces are the same but the the orders of the bases are different.

VectorSpace

VectorSpace{B} is the abstaction of a vector space, which has only one type parameter:

B<:Any: the type of the bases of the vector space

Basically, a subtype should implement the following 3 methods:

dimension(vs::VectorSpace) -> Int
Get the dimension of a vector space

Base.getindex(vs::VectorSpace{B},i::Int)  where B -> B
Get the ith basis of a vector space

Base.searchsortedfirst(vs::VectorSpace{B},basis::B) where B -> Int
Search the index of a basis in a vector space

However, we provide several interfaces, including type traits and methods to deal with common situations:

A vector space whose bases are stored in a table under the attribute name :table can be ascribed to the HasTable trait and the TableSorted trait. Specifically, the first trait must be implemented as
HasTable(::Type{SubType})=HasTable(true)
While, if the table is unsorted, the second trait should be implemented as
TableSorted(::Type{SubType})=TableSorted(false)
and if the table is sorted, the second trait should be implemented as
TableSorted(::Type{SubType})=TableSorted(true)

A vector space whose bases may be represented by a multiindex (Cartesian index) can be ascribed to the traits IsMultiIndexable and MultiIndexOrderStyle. Specifically, the first trait must be implemented as
IsMultiIndexable(::Type{SubType})=IsMultiIndexable(true)
While, if the order style of the multiindex is C/C++ like, the second trait shoule be implemented as
MultiIndexOrderStyle(::Type{SubType})=MultiIndexOrderStyle('C')
and if the order style is Fortran like, the second trait shoule be implemented as
MultiIndexOrderStyle(::Type{SubType})=MultiIndexOrderStyle('F')
Furthermore, it should implement the following methods

rank(::Type{SubType}) -> Int
Get the rank of a multiindexable vector space.

dims(vs::SubType) -> NTuple{vs|>typeof|>rank,Int}
Get the dimensions along each axes of a multiindexable vector space.

inds(basis,vs::SubType) ->  NTuple{vs|>typeof|>rank,Int}
Get the Cartesian index representation of a basis in a multiindexable vector space.

eltype(SubType).name.wrapper(index::NTuple{N,Int},vs::SubType)
Construct a basis from a tuple of integers and a multiindexable vector space.

Note that a multiindexable vector space can also have a sorted or unsorted table. But then the trait MultiIndexOrderStyle takes no effects and the sequences of its bases will be completely determined by its attribute :table.

If the type taits and methods are defined properly as stated above, the dimension, getindex and searchsortedfirst methods get default implementations. No need to worry about them any more.

Other features include

comparison: == and isequal

iteration: iterate

inquiry: size, findfirst and in

SimpleVectorSpace

SimpleVectorSpace{S,B,N} is the simplest vector space, whose bases are stored in the attribute :table::NTuple{N,B} as an ntuple.

The :table attribute can be sorted or unsorted, which is determined by the type parameter S, with 'T' for sorted and 'F' for unsorted.

OrderedIndices

OrderedIndices{N} defines the simplest abstract class of multiindexable vector spaces, whose bases are just tuples of integers.

This class of vector spaces must have the following attribute: dims::NTuple{N,Int}: the dimesnions of the Cartesian indices along all axes

SimpleIndices

SimpleIndices{M,N} is the simple ordered Cartesian indices.

It is worth noting that

It can be C/C++ ordered or Fortran ordered depending on the first type parameter M, with 'C' for the former and 'F' the latter.

For its bases (Cartesian indices), there is no restriction except that they should be in the proper range defined by its dims.

TabledIndices

TabledIndices{S,N} defines the tabled ordered Cartesian indices.

Compared to SimpleIndices, the bases of this kind of vector spaces are stored in the attribute :table, which must be a vector of tuple of integers. The :table attribute can be sorted or unsorted, which is determined by the type parameter S, with 'T' for sorted and 'F' for unsorted. This type suits the situations when the Cartesian indices are restricted by extra conditions except that propoesed by the attribute :dims.

GradedVectorSpace

GradedVectorSpace{G,B,V,T} defines the abstract type of graded vector spaces, which are vector spaces that have the extra structure of a grading, which is a decomposition of the vector space into a direct sum of vector subspaces.

It has 4 type parameters

G: the type of the grades

B: the eltype of the subspaces

V<:VectorSpace: the type of the subspaces

T<:GradedTables{G,V}: the type of the subspaces' contents

Concrete subtypes must have the following attribute:

:tables::T: the contents of the subspaces, which must be a GradedTables.

Specifically, the dimension, getindex and searchsortedfirst methods are overloaded in support of various purposes. For details, please refer to the manual.

Manul

QuantumLattices.Mathematics.VectorSpaces.GradedTables — Type.

GradedTables(vs::Tuple,ks::Tuple)
GradedTables(::Type{M},n::Int,gs::Val{grades}) where {M<:AbstractCombinatorics,grades}

The tables of a graded vector space.

Alias for Base.Iterators.Pairs{G,V,KS<:Tuple,VS<:Tuple}.

source

QuantumLattices.Mathematics.VectorSpaces.GradedVectorSpace — Type.

GradedVectorSpace{G,B,V<:VectorSpace,T<:GradedTables{G,V}} <: VectorSpace{Tuple{G,B}}

Abstract type of graded vector spaces.

A graded vector space is a vector space that has the extra structure of a grading, which is a decomposition of the vector space into a direct sum of vector subspaces.

Concrete subtypes must have the following attribute:

:tables::GradedTables{G,V} where {G,V<:VectorSpace}: the tables of the subspaces.

source

QuantumLattices.Mathematics.VectorSpaces.HasTable — Type.

HasTable(B::Bool)
HasTable(::Type{<:VectorSpace})

Trait of whether a subtype of VectorSpace has the attribute :table.

Only two instances are allowed, the first of which is the default for a subtype:

HasTable(false): indication of not having the attribute :table

HasTable(true): indication of having the attribute :table

source

QuantumLattices.Mathematics.VectorSpaces.IsMultiIndexable — Type.

IsMultiIndexable(B::Bool)
IsMultiIndexable(::Type{<:VectorSpace})

Trait of whether the bases of a subtype of VectorSpace can be represented by multiindices (Cartesian indices).

Only two instances are allowed, the first of which is the default for a subtype:

IsMultiIndexable(false): indication of irrepresentability by Cartesian indices

IsMultiIndexable(true): indication of representability by Cartesian indices

source

QuantumLattices.Mathematics.VectorSpaces.MultiIndexOrderStyle — Type.

MultiIndexOrderStyle(M::Char)
MultiIndexOrderStyle(::Type{<:VectorSpace})

Trait of the order style of the Cartesian-index representation of the bases of a subtype of VectorSpace.

Only two instances are allowed, the first of which is the default for a subtype:

MultiIndexOrderStyle(C): indication of C/C++ order style

MultiIndexOrderStyle(F): indication of Fortran order style

source

QuantumLattices.Mathematics.VectorSpaces.OrderedIndices — Type.

OrderedIndices{N} <: VectorSpace{NTuple{N,Int}}

The simplest abstract class of multiindexable vector spaces, whose bases are just tuples of integers.

This class of vector spaces must have the following attribute: dims::NTuple{N,Int}: the dimesnions of the Cartesian indices along all axes

source

QuantumLattices.Mathematics.VectorSpaces.SimpleIndices — Type.

SimpleIndices{M}(dims::NTuple{N,Int}) where {M,N}

Simple ordered Cartesian indices.

Note

It can be C/C++ ordered or Fortran ordered depending on the first type parameter M, with 'C' for the former and 'F' the latter.

For its bases (Cartesian indices), there is no restriction except that they should be in the proper range defined by its dims.

source

QuantumLattices.Mathematics.VectorSpaces.SimpleVectorSpace — Type.

SimpleVectorSpace{S}(table::NTuple{N,B}) where {S,B,N}
SimpleVectorSpace{S}(table...) where S

Simplest vector space, whose bases are stored in the attribute :table as an ntuple.

The :table attribute can be sorted or unsorted, which is determined by the type parameter S, with 'T' for sorted and 'F' for unsorted.

source

QuantumLattices.Mathematics.VectorSpaces.TableSorted — Type.

TableSorted(B::Bool)
TableSorted(::Type{<:VectorSpace})

Trait of whether the attribute :table of a subtype of VectorSpace is sorted.

Only two instances are allowed, the first of which is the default for a subtype:

TableSorted(false): indication of unsorted attribute :table

TableSorted(true): indication of sorted attribute :table

source

QuantumLattices.Mathematics.VectorSpaces.TabledIndices — Type.

TabledIndices{S}(dims::NTuple{N,Int},table::Vector{NTuple{N,Int}}) where {S,N}
TabledIndices{N}(::Type{M},len::Int) where {N,M<:AbstractCombinatorics}

Tabled ordered Cartesian indices.

Compared to SimpleIndices, the bases of this kind of vector spaces are stored in the attribute :table, which must be a vector of tuple of integers. The :table attribute can be sorted or unsorted, which is determined by the type parameter S, with 'T' for sorted and 'F' for unsorted. This type suits the situations when the Cartesian indices are restricted by extra conditions except that propoesed by the attribute :dims.

source

QuantumLattices.Mathematics.VectorSpaces.VectorSpace — Type.

VectorSpace{B} <: AbstractVector{B}

Abstract vector space.

source

Base.Sort.searchsortedfirst — Method.

searchsortedfirst(vs::GradedVectorSpace{G,B},pair::Tuple{G,B}) where {G,B} -> Int

Find the index of a grade-basis pair in a graded vector space.

source

Base.Sort.searchsortedfirst — Method.

searchsortedfirst(vs::VectorSpace{B},basis::B) where B -> Int
searchsortedfirst(vs,basis) -> Int

Search the index of a basis in a vector space.

source

Base.eltype — Method.

eltype(vs::GradedVectorSpace,g::Int)
eltype(::Type{V},g::Int) where {V<:GradedVectorSpace}

Get the gth eltype of a graded vector space.

source

Base.findfirst — Method.

findfirst(basis::B,vs::VectorSpace{B}) where B -> Int
findfirst(bases,vs::VectorSpace) -> NTuple{length(bases),Int}

Get the index of a basis or the indices of a couple of bases in a vector space.

source

Base.getindex — Method.

getindex(vs::VectorSpace,i::Int)

Get the ith basis of a vector space by the [] operator.

source

Base.getindex — Method.

getindex(vs::GradedVectorSpace{B},pair::Tuple{B,Int}) where B
getindex(vs::GradedVectorSpace,i::Int)

Get the basis of a graded vector space by a grade-index pair or by an index.

source

Base.in — Method.

in(basis::B,vs::VectorSpace{B}) where B -> Bool

Judge whether a basis belongs to a vector space.

source

Base.iterate — Function.

iterate(vs::GradedVectorSpace,state=(1,1))

Iterate over the whole graded vector space.

source

Base.keys — Method.

keys(vs::GradedVectorSpace) -> Tuple
values(vs::GradedVectorSpace) -> Tuple
pairs(vs::GradedVectorSpace) -> GradedTables

Iterate over the keys, values or pairs of a graded vector space.

source

Base.keytype — Method.

keytype(tables::GradedTables,g::Int)
keytype(::Type{T},g::Int) where {T<:GradedTables}

Get the gth keytype of a graded tables.

source

Base.keytype — Method.

keytype(vs::GradedVectorSpace,g::Int)
keytype(::Type{V},g::Int) where {V<:GradedVectorSpace}

Get the gth keytype of a graded vector space.

source

Base.valtype — Method.

valtype(tables::GradedTables,g::Int)
valtype(::Type{T},g::Int) where {T<:GradedTables}

Get the gth valtype of a graded tables.

source

Base.valtype — Method.

valtype(vs::GradedVectorSpace,g::Int)
valtype(::Type{V},g::Int) where {V<:GradedVectorSpace}

Get the gth valtype of a graded vector space.

source

QuantumLattices.Interfaces.:⊕ — Method.

⊕(basis1::B,basis2::B) where B -> SimpleVectorSpace{'F',B,2}
⊕(basis::B,vs::SimpleVectorSpace{<:Any,B}) where B -> SimpleVectorSpace{'F',B}
⊕(vs::SimpleVectorSpace{<:Any,B},basis::B) where B -> SimpleVectorSpace{'F',B}
⊕(vs1::SimpleVectorSpace{<:Any,B},vs2::SimpleVectorSpace{<:Any,B}) where B -> SimpleVectorSpace{'F',B}

Get the direct sum between bases or simple vector spaces.

source

QuantumLattices.Interfaces.degree — Method.

degree(g::G,vs::GradedVectorSpace{G}) where G -> Int

Get the degree of a vector subspace whose grade are represented by g.

source

QuantumLattices.Interfaces.dimension — Method.

dimension(vs::GradedVectorSpace) -> Int
dimension(vs::GradedVectorSpace{G},g::G) where G -> Int
dimension(vs::GradedVectorSpace{G},gs::NTuple{N,G}) where {G,N} -> Int

Get the dimension of the whole graded vector space or some vector subspaces.

source

QuantumLattices.Interfaces.dimension — Method.

dimension(vs::VectorSpace) -> Int

The dimension of a vector space.

source

QuantumLattices.Interfaces.rank — Method.

rank(tables::GradedTables) -> Int
rank(::Type{T}) where {T<:GradedTables} -> Int

Get the total number of keys or values of a graded tables.

source

QuantumLattices.Interfaces.rank — Method.

rank(vs::GradedVectorSpace) -> Int
rank(::Type{V}) where {V<:GradedVectorSpace} -> Int

Get the rank, i.e. the total number of vector subspaces.

source