Introduction

This module contains the prerequisites of the package.

The constants, types, macros, functions or submodules defined in this module will not be exported by the package. Instead, they serve as the prerequisites. The range of the contents are quite wide, but basically, they fall into two categories:

Global constants and miscellaneous tiny useful functions;

Basic data structures as supplements to the Julia.Base and other common packages.

The first category is contained in the main body of this module, while the other comes in separate submodules.

Constants and functions

All the following constants and functions in this section are defined in the main body and are exported by this module.

QuantumLattices.Prerequisites.atol — Constant

Absolute tolerance for float numbers.

QuantumLattices.Prerequisites.rtol — Constant

Relative tolerance for float numbers.

QuantumLattices.Prerequisites.Float — Type

Default float type.

QuantumLattices.Prerequisites.decimaltostr — Function

decimaltostr(number, ::Int=5)
decimaltostr(number::Integer, n::Int=5)
decimaltostr(number::Rational, n::Int=5)
decimaltostr(number::AbstractFloat, n::Int=5)
decimaltostr(number::Complex, n::Int=5)

Convert a number to a string with at most n decimal places.

QuantumLattices.Prerequisites.ordinal — Function

ordinal(number::Integer)

Convert a positive number to its corresponding ordinal.

QuantumLattices.Prerequisites.delta — Function

delta(i, j) -> Int

Kronecker delta function.

Basic structures

Here lists the table of contents of the basic data structures that are supplements to the Julia.Base and other common packages:

Combinatorics

AbstractCombinatorics

Manual

Traits

Type helpers

Holy traits

Manual

Composite structures

CompositeTuple and CompositeNTuple

CompositeVector

CompositeDict

NamedContainer

Manual

Simple trees

AbstractSimpleTree

SimpleTreeCore and SimpleTree

Manual

Named vectors

NamedVector

HomoNamedVector

Manual

Vector spaces

VectorSpace

EnumerativeVectorSpace

CartesianVectorSpace

NamedVectorSpace

Manual