Introduction

Introduction

This module contains the prerequisites of the package.

The constants, types, macros, functions or submodules defined in this module will NOT be exported by the package. Instead, they serve as the prerequisites and fundamentals. The range of the contents are quite wide, but basically, they fall into 2 categories:

Global constants and miscellaneous tiny useful functions;

Basic data structures as supplements to the Julia.Base and other common packages.

The first category is contained in the main body of this module, while the other comes in separate submodules.

Constants and functions

All the following constants and functions in this section are defined in the main body and are exported by this module.

QuantumLattices.Prerequisites.atol — Constant.

Absolute tolerance for float numbers.

source

QuantumLattices.Prerequisites.rtol — Constant.

Relative tolerance for float numbers.

source

QuantumLattices.Prerequisites.Float — Type.

Default float type.

source

QuantumLattices.Prerequisites.decimaltostr — Function.

decimaltostr(number,::Int=5)
decimaltostr(number::Integer,n::Int=5)
decimaltostr(number::Rational,n::Int=5)
decimaltostr(number::AbstractFloat,n::Int=5)
decimaltostr(number::Complex,n::Int=5)

Convert a number to a string with at most n decimal places.

source

QuantumLattices.Prerequisites.ordinal — Function.

ordinal(number::Interger)

Convert a positive number to its corresponding ordinal.

source

QuantumLattices.Prerequisites.delta — Function.

delta(i,j) -> Int

Kronecker delta function.

source

Basic structures

Here lists the table of contents of the basic data structures that are supplements to the Julia.Base and other common packages:

Type traits

EfficientOperations

MemoryOrder

Manual

Factories

Escape mechanisms

Concrete factories

Manual

Composite structures

CompositeTuple and CompositeNTuple

CompositeVector

CompositeDict

NamedContainer

Manul

Simple trees

AbstractSimpleTree

SimpleTreeCore and SimpleTree

Manual

Named vectors

NamedVector

HomoNamedVector

Manual