Quantum Systems

QuantumLattices.QuantumSystems.FockTerm — Type

FockTerm

Type alias for Union{Onsite, Hopping, Pairing, Hubbard, InterOrbitalInterSpin, InterOrbitalIntraSpin, SpinFlip, PairHopping, Coulomb}.

QuantumLattices.QuantumSystems.PhononTerm — Type

PhononTerm

Type alias for Union{Kinetic, Hooke, Elastic}.

QuantumLattices.QuantumSystems.annihilation — Constant

annihilation

Indicate that the nambu index is annihilation.

QuantumLattices.QuantumSystems.block — Constant

*(f₁::Operator{<:Number, <:ID{FID{:f}}}, f₂::Operator{<:Number, <:ID{FID{:f}}}) -> Union{typeof(0), Operator}
*(f₁::Operator{<:Number, <:ID{Index{Int, <:FID{:f}}}}, f₂::Operator{<:Number, <:ID{Index{Int, <:FID{:f}}}}) -> Union{typeof(0), Operator}
*(f₁::Operator{<:Number, <:ID{AbstractCompositeIndex{<:Index{Int, <:FID{:f}}}}}, f₂::Operator{<:Number, <:ID{AbstractCompositeIndex{<:Index{Int, <:FID{:f}}}}}) -> Union{typeof(0), Operator}

Get the multiplication of two fermionic Fock operators.

QuantumLattices.QuantumSystems.creation — Constant

creation

Indicate that the nambu index is creation.

QuantumLattices.QuantumSystems.latexofbosons — Constant

latexofbosons

The default LaTeX format of a bosonic index.

QuantumLattices.QuantumSystems.latexoffermions — Constant

latexoffermions

The default LaTeX format of a fermionic index.

QuantumLattices.QuantumSystems.latexofparticles — Constant

latexofparticles

The default LaTeX format of a wildcard Fock index.

QuantumLattices.QuantumSystems.latexofphonons — Constant

latexofphonons

The default LaTeX format of a phonon index.

QuantumLattices.QuantumSystems.latexofspins — Constant

latexofspins

The default LaTeX format of a spin index.

QuantumLattices.DegreesOfFreedom.MatrixCoupling — Method

MatrixCoupling(sites::Union{NTuple{2, Int}, Colon}, ::Type{PID}, matrix::AbstractMatrix; rows::Union{AbstractVector, Nothing}=nothing, cols::Union{AbstractVector, Nothing}=nothing)

Construct a set of Couplings corresponding to the dynamical matrix of phonons.

QuantumLattices.DegreesOfFreedom.MatrixCoupling — Method

MatrixCoupling(sites::Union{NTuple{2, Int}, Colon}, ::Type{F}, orbital::Union{AbstractMatrix, Colon}, spin::Union{AbstractMatrix, Colon}, nambu::Union{AbstractMatrix, Colon}) where {F<:FID}

Construct a set of Couplings between two Index{<:Union{Int, Colon}, <:FID}s with the coefficients specified by matrices acting on separated internal spaces.

QuantumLattices.DegreesOfFreedom.MatrixCoupling — Method

MatrixCoupling(sites::Union{NTuple{2, Int}, Colon}, ::Type{S}, matrix::AbstractMatrix; rows::AbstractVector=SVector('x', 'y', 'z'), cols::AbstractVector=SVector('x', 'y', 'z')) where {S<:SID}

Construct a set of Couplings between two Index{<:Union{Int, Colon}, <:SID}s with the coefficients specified by a matrix.

QuantumLattices.QuantumSystems.Coulomb — Type

Coulomb(
    id::Symbol, value, bondkind, coupling=Coupling(Index(:, FID(:, :, :)), Index(:, FID(:, :, :)))^2;
    ishermitian::Bool=true,
    amplitude::Union{Function, Nothing}=nothing,
    ismodulatable::Bool=true
)

Coulomb term.

Type alias for Term{:Coulomb, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.DM — Type

DM(
    id::Symbol,
    value,
    bondkind,
    vectors::Pair{<:AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}}, <:Union{Char, AbstractVector{<:Number}}}...;
    unit::Symbol=:degree,
    amplitude::Union{Function, Nothing}=nothing,
    ismodulatable::Bool=true
)

DM term. Since DM term is antisymmetric on every bond, only the positive direction of a bond is needed. The negative direction of a bond can be handled automatically by this function.

Here, vectors specify the unit DM vector on every bond in the form [bond₁, bond₂, ...]=>v, where bondᵢ can be

a Number specifying the azimuth angle of a bond in the 2-dimensional case, or

a Tuple{Number, Number} specifying the polar and azimuth angle pairs of a bond in the 3-dimensional case, or

an AbstractVector{<:Number} specifying the direction of a bond;

and v can be

a Char of 'x', 'y' or 'z', indicating the unit DM vector on the set of bonds is along the x, y or z direction, or

an AbstractVector{<:Number}, specifying the direction of the DM vector on the set of bonds.

Type alias for Term{:DM, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Elastic — Type

Elastic(id::Symbol, value, bondkind, coupling; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Generic elastic energy of phonons.

Type alias for Term{:Elastic, id, V, B, C<:TermCoupling, A<:TermAmplitude}

QuantumLattices.QuantumSystems.FID — Type

FID{T, O<:Union{Int, Symbol, Colon}, S<:Union{Rational{Int}, Symbol, Colon}, N<:Union{Int, Symbol, Colon}} <: SimpleIID

The Fock id.

QuantumLattices.QuantumSystems.Fock — Type

Fock{T} <: SimpleInternal{FID{T, Int, Rational{Int}, Int}}

The Fock internal degrees of freedom.

QuantumLattices.QuantumSystems.Heisenberg — Type

Heisenberg(id::Symbol, value, bondkind; form::Symbol=Symbol("+-z"), amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Heisenberg term.

Type alias for Term{:Heisenberg, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Hooke — Type

Hooke(id::Symbol, value, bondkind; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Potential energy of phonons by the Hooke's law.

Type alias for Term{:Hooke, id, V, B, C<:TermCoupling, A<:TermAmplitude}

QuantumLattices.QuantumSystems.Hopping — Type

Hopping(id::Symbol, value, bondkind, coupling=Coupling(Index(:, FID(:, :, :)), Index(:, FID(:, :, :))); amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Hopping term.

Type alias for Term{:Hopping, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Hubbard — Type

Hubbard(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Hubbard term.

Type alias for Term{:Hubbard, id, V, Int, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.InterOrbitalInterSpin — Type

InterOrbitalInterSpin(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Interorbital-interspin term.

Type alias for Term{:InterOrbitalInterSpin, id, V, Int, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.InterOrbitalIntraSpin — Type

InterOrbitalIntraSpin(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Interorbital-intraspin term.

Type alias for Term{:InterOrbitalIntraSpin, id, V, Int, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Ising — Type

Ising(id::Symbol, value, bondkind, direction::Char; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Ising term.

Type alias for Term{:Ising, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Kinetic — Type

Kinetic(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Kinetic energy of phonons.

Type alias for Term{:Kinetic, id, V, Int, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Kitaev — Type

Kitaev(
    id::Symbol, value, bondkind;
    x::AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}},
    y::AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}},
    z::AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}},
    unit::Symbol=:degree,
    amplitude::Union{Function, Nothing}=nothing,
    ismodulatable::Bool=true
)

Kitaev term. Since Kitaev term is symmetric on every bond, only one direction of a bond is needed. The inverse direction of a bond can be handled automatically by this function.

Here, x, y and z assign the x-bonds, y-bonds, and z-bonds, respectively, with each kind of bond can be

a Number specifying the azimuth angle of a bond in the 2-dimensional case, or

a Tuple{Number, Number} specifying the polar and azimuth angle pairs of a bond in the 3-dimensional case, or

an AbstractVector{<:Number} specifying the direction of a bond.

Type alias for Term{:Kitaev, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Onsite — Type

Onsite(id::Symbol, value, coupling=Coupling(Index(:, FID(:, :, :)), Index(:, FID(:, :, :))); ishermitian::Bool=true, amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Onsite term.

Type alias for Term{:Onsite, id, V, Int, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.PID — Type

PID{D<:Union{Char, Symbol, Colon}} <: SimpleIID

The phonon id.

QuantumLattices.QuantumSystems.PairHopping — Type

PairHopping(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Pair-hopping term.

Type alias for Term{:PairHopping, id, V, Int, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Pairing — Type

Pairing(id::Symbol, value, bondkind, coupling; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Pairing term.

Type alias for Term{:Pairing, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Phonon — Type

Phonon <: SimpleInternal{PID{Char}}

The phonon internal degrees of freedom.

QuantumLattices.QuantumSystems.SID — Type

SID{S, T<:Union{Char, Symbol, Colon}} <: SimpleIID

The spin id.

QuantumLattices.QuantumSystems.SingleIonAnisotropy — Type

SingleIonAnisotropy(id::Symbol, value, direction::Char; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)
SingleIonAnisotropy(id::Symbol, value, matrix::AbstractMatrix{<:Number}; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Single ion anisotropy term.

Type alias for Term{:SingleIonAnisotropy, id, V, Int, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Spin — Type

Spin{S} <: SimpleInternal{SID{S, Char}}

The spin internal degrees of freedom.

QuantumLattices.QuantumSystems.SpinFlip — Type

SpinFlip(id::Symbol, value; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Spin-flip term.

Type alias for Term{:SpinFlip, id, V, Int, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.SpinTerm — Type

SpinTerm(id::Symbol, value, bondkind, coupling; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Generic spin term.

Type alias for Term{:SpinTerm, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Zeeman — Type

Zeeman(id::Symbol, value, direction::Char, g::Number=1; amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)
Zeeman(id::Symbol, value, direction::Union{AbstractVector{<:Number}, Tuple{Number, Number}}, g::Union{Number, AbstractMatrix{<:Number}}=1; unit::Symbol=:degree, amplitude::Union{Function, Nothing}=nothing, ismodulatable::Bool=true)

Zeeman term.

Type alias for Term{:Zeeman, id, V, Int, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Γ — Type

Γ(
    id::Symbol, value, bondkind;
    x::AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}},
    y::AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}},
    z::AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}},
    unit::Symbol=:degree,
    amplitude::Union{Function, Nothing}=nothing,
    ismodulatable::Bool=true
)

Γ Term. Since Γ term is symmetric on every bond, only one direction of a bond is needed. The inverse direction of a bond can be handled automatically by this function.

Here, x, y and z assign the x-bonds, y-bonds, and z-bonds, respectively, with each kind of bond can be

a Number specifying the azimuth angle of a bond in the 2-dimensional case, or

a Tuple{Number, Number} specifying the polar and azimuth angle pairs of a bond in the 3-dimensional case, or

an AbstractVector{<:Number} specifying the direction of a bond.

Type alias for Term{:Γ, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.Γ′ — Type

Γ′(
    id::Symbol, value, bondkind;
    x::AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}},
    y::AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}},
    z::AbstractVector{<:Union{Number, Tuple{Number, Number}, AbstractVector{<:Number}}},
    unit::Symbol=:degree,
    amplitude::Union{Function, Nothing}=nothing,
    ismodulatable::Bool=true
)

Γ′ Term. Since Γ′ term is symmetric on every bond, only one direction of a bond is needed. The inverse direction of a bond can be handled automatically by this function.

Here, x, y and z assign the x-bonds, y-bonds, and z-bonds, respectively, with each bond can be

a Number specifying the azimuth angle of a bond in the 2-dimensional case, or

a Tuple{Number, Number} specifying the polar and azimuth angle pairs of a bond in the 3-dimensional case, or

an AbstractVector{<:Number} specifying the direction of a bond.

Type alias for Term{:Γ′, id, V, B, C<:TermCoupling, A<:TermAmplitude}.

QuantumLattices.QuantumSystems.@DM_str — Macro

DM"x" => SparseMatrixCSC([0 0 0; 0 0 1; 0 -1 0])
DM"y" => SparseMatrixCSC([0 0 -1; 0 0 0; 1 0 0])
DM"z" => SparseMatrixCSC([0 1 0; -1 0 0; 0 0 0])

The DM coupling matrix.

QuantumLattices.QuantumSystems.@Heisenberg_str — Macro

Heisenberg"" => SparseMatrixCSC([1 0 0; 0 1 0; 0 0 1])

The Heisenberg coupling matrix.

QuantumLattices.QuantumSystems.@Ising_str — Macro

Ising"x" => SparseMatrixCSC([1 0 0; 0 0 0; 0 0 0])
Ising"y" => SparseMatrixCSC([0 0 0; 0 1 0; 0 0 0])
Ising"z" => SparseMatrixCSC([0 0 0; 0 0 0; 0 0 1])

The Ising coupling matrix.

QuantumLattices.QuantumSystems.@L_str — Macro

L"x" => SparseMatrixCSC([0 0 0; 0 0 1im; 0 -1im 0])
L"y" => SparseMatrixCSC([0 0 -1im; 0 0 0; 1im 0 0])
L"z" => SparseMatrixCSC([0 1im 0; -1im 0 0; 0 0 0])

The three-dimensional rotation generators.

QuantumLattices.QuantumSystems.@Γ_str — Macro

Γ"x" => SparseMatrixCSC([0 0 0; 0 0 1; 0 1 0])
Γ"y" => SparseMatrixCSC([0 0 1; 0 0 0; 1 0 0])
Γ"z" => SparseMatrixCSC([0 1 0; 1 0 0; 0 0 0])

The Γ coupling matrix.

QuantumLattices.QuantumSystems.@Γ′_str — Macro

Γ′"x" => SparseMatrixCSC([0 1 1; 1 0 0; 1 0 0])
Γ′"y" => SparseMatrixCSC([0 1 0; 1 0 1; 0 1 0])
Γ′"z" => SparseMatrixCSC([0 0 1; 0 0 1; 1 1 0])

The Γ′ coupling matrix.

QuantumLattices.QuantumSystems.@σ_str — Macro

σ"0" => SparseMatrixCSC([1 0; 0 1])
σ"x" => SparseMatrixCSC([0 1; 1 0])
σ"y" => SparseMatrixCSC([0 -1im; 1im 0])
σ"z" => SparseMatrixCSC([1 0; 0 -1])
σ"+" => SparseMatrixCSC([0 1; 0 0])
σ"-" => SparseMatrixCSC([0 0; 1 0])
σ"11" => SparseMatrixCSC([1 0; 0 0])
σ"22" => SparseMatrixCSC([0 0; 0 1])

The Pauli matrix σ⁰, σˣ, σʸ, σᶻ, σ⁺, σ⁻, σ¹¹, σ²².

Base.angle — Method

angle(id::CompositeIndex{<:Index{Int, <:FID}}, vectors::AbstractVector{<:AbstractVector{<:Number}}, values::AbstractVector{<:Number}) -> Complex{<:Number}

Get the twist phase corresponding to a Fock index.

QuantumLattices.QuantumOperators.matrix — Method

matrix(sid::SID{S, Char}, dtype::Type{<:Number}=Complex{Float}) where S -> Matrix{dtype}
matrix(index::Index{<:Union{Int, Colon}, <:SID}, dtype::Type{<:Number}=Complex{Float}) -> Matrix{dtype}
matrix(index::AbstractCompositeIndex{<:Index{<:Union{Int, Colon}, <:SID}}, dtype::Type{<:Number}=Complex{Float}) -> Matrix{dtype}

Get the matrix representation of a sid.

QuantumLattices.QuantumOperators.script — Method

script(::Val{:direction}, pid::PID; kwargs...) -> String

Get the requested script of an pid.

QuantumLattices.QuantumOperators.script — Method

script(::Val{:orbital}, fid::FID; kwargs...) -> String
script(::Val{:spint}, fid::FID; kwargs...) -> String
script(::Val{:spinsym}, fid::FID; kwargs...) -> String
script(::Val{:nambu}, fid::FID; kwargs...) -> String

Get the requested script of an fid.

QuantumLattices.QuantumOperators.script — Method

script(::Val{:tag}, sid::SID; kwargs...) -> String

Get the requested script of an sid.

QuantumLattices.QuantumSystems.isannihilation — Method

isannihilation(fid::FID) -> Bool
isannihilation(index::Index) -> Bool
isannihilation(index::AbstractCompositeIndex) -> Bool

Judge whether the nambu index is annihilation.

QuantumLattices.QuantumSystems.iscreation — Method

iscreation(fid::FID) -> Bool
iscreation(index::Index) -> Bool
iscreation(index::AbstractCompositeIndex) -> Bool

Judge whether the nambu index is creation.

QuantumLattices.QuantumSystems.isnormalordered — Method

isnormalordered(opt::Operator{<:Number, <:ID{AbstractCompositeIndex{<:Index{Int, <:FID}}}}) -> Bool

Judge whether an operator is normal ordered.

QuantumLattices.QuantumSystems.totalspin — Method

totalspin(::SID) -> Rational{Int}/Int/Symbol
totalspin(::Type{<:SID}) -> Rational{Int}/Int/Symbol

totalspin(::Index{<:Union{Int, Colon}, <:SID}) -> Rational{Int}/Int/Symbol
totalspin(::Type{<:Index{<:Union{Int, Colon}, <:SID}}) -> Rational{Int}/Int/Symbol

totalspin(::AbstractCompositeIndex{<:Index{<:Union{Int, Colon}, <:SID}}) -> Rational{Int}/Int/Symbol
totalspin(::Type{<:AbstractCompositeIndex{<:Index{<:Union{Int, Colon}, <:SID}}}) -> Rational{Int}/Int/Symbol

Get the total spin.

QuantumLattices.QuantumSystems.totalspin — Method

totalspin(::Spin) -> Rational{Int}/Int/Symbol
totalspin(::Type{<:Spin}) -> Rational{Int}/Int/Symbol

Get the total spin.

QuantumLattices.expand — Method

expand(::Val{:Hooke}, pnc::Coupling{<:Number, <:NTuple{2, Index{<:Union{Int, Colon}, PID{Colon}}}}, bond::Bond, hilbert::Hilbert) -> PPExpand

Expand the default phonon potential coupling on a given bond.

QuantumLattices.permute — Method

permute(id₁::FID, id₂::FID) -> Tuple{Vararg{Operator}}

Permute two Fock indexes and get the result.

QuantumLattices.permute — Method

permute(id₁::PID, id₂::PID) -> Tuple{Vararg{Operator}}

Permute two phonon indexes and get the result.

QuantumLattices.permute — Method

permute(id₁::SID, id₂::SID) -> Tuple{Vararg{Operator}}

Permute two spin indexes and get the result.