Spatials

QuantumLattices.Spatials.AbstractLattice — Type

AbstractLattice{N, D<:Number}

Abstract type of a unitcell-described lattice.

It should have the following contents:

name::Symbol: the name of the lattice

coordinates::Matrix{D}: the coordinates of the lattice

vectors::Vector{SVector{N, D}}: the translation vectors of the lattice

reciprocals::Vector{SVector{N, D}}: the reciprocals of the lattice

QuantumLattices.Spatials.Bond — Type

Bond{K, P<:Point}

A generic bond, which could contains several points.

QuantumLattices.Spatials.Bond — Method

Bond(point::Point)
Bond(kind, point₁::Point, point₂::Point, points::Point...)

Construct a bond.

QuantumLattices.Spatials.BrillouinZone — Type

BrillouinZone{K, P<:Momentum, S<:SVector} <: ReciprocalSpace{K, P}

The Brillouin zone of a lattice.

QuantumLattices.Spatials.BrillouinZone — Method

BrillouinZone(reciprocals::AbstractVector, nk::Int)
BrillouinZone(::Type{P}, reciprocals::AbstractVector) where {P<:Momentum}
BrillouinZone(reciprocals::AbstractVector, momenta::AbelianNumbers{<:Momentum})

BrillouinZone{K}(reciprocals::AbstractVector, nk::Int) where K
BrillouinZone{K}(::Type{P}, reciprocals::AbstractVector) where {K, P<:Momentum}
BrillouinZone{K}(reciprocals::AbstractVector, momenta::AbelianNumbers{<:Momentum}) where K

Construct a Brillouin zone.

QuantumLattices.Spatials.Lattice — Type

Lattice{N, D<:Number} <: AbstractLattice{N, D}

Simplest lattice.

A simplest lattice can be constructed from its coordinates and translation vectors.

QuantumLattices.Spatials.Lattice — Method

Lattice(lattice::Lattice, translations::Translations)

Construct a lattice from the translations of another.

QuantumLattices.Spatials.Lattice — Method

Lattice(coordinates::NTuple{N, Number}...; name::Symbol=:lattice, vectors::Union{Vector{<:AbstractVector{<:Number}}, Nothing}=nothing) where N
Lattice(coordinates::Vector{<:Number}...; name::Symbol=:lattice, vectors::Union{Vector{<:AbstractVector{<:Number}}, Nothing}=nothing)

Construct a lattice.

QuantumLattices.Spatials.Neighbors — Type

Neighbors{K, V<:Number} <: CompositeDict{K, V}

Neighbor vs. bond length maps.

QuantumLattices.Spatials.Neighbors — Method

Neighbors(lattice::AbstractLattice, nneighbor::Integer; coordination::Int=8)

Get the neighbor vs. bond length map of a lattice up to the nneighborth order.

QuantumLattices.Spatials.Point — Type

Point{N, D<:Number}

A point in a unitcell-described lattice.

QuantumLattices.Spatials.Point — Method

Point(site::Integer, rcoordinate::SVector{N, D}, icoordinate::SVector{N, D}) where {N, D<:Number}
Point(site::Integer, rcoordinate::NTuple{N, <:Number}, icoordinate::NTuple{N, <:Number}=ntuple(i->0, N)) where N
Point(site::Integer, rcoordinate::AbstractVector{<:Number}, icoordinate::AbstractVector{<:Number}=zero(SVector{length(rcoordinate), Int}))

Construct a labeled point.

QuantumLattices.Spatials.ReciprocalPath — Type

ReciprocalPath{K, S<:SVector} <: ReciprocalSpace{K, S}

A path in the reciprocal space.

QuantumLattices.Spatials.ReciprocalPath — Method

ReciprocalPath(momenta::AbstractVector)
ReciprocalPath(reciprocals::AbstractVector, segments::Segment{<:SVector}...)
ReciprocalPath(reciprocals::AbstractVector, segments::Tuple{Vararg{Segment{<:SVector}}})

ReciprocalPath{K}(momenta::AbstractVector) where K
ReciprocalPath{K}(reciprocals::AbstractVector, segments::Segment{<:SVector}...) where K
ReciprocalPath{K}(reciprocals::AbstractVector, segments::Tuple{Vararg{Segment{<:SVector}}}) where K

Construct a path in the reciprocal space.

QuantumLattices.Spatials.ReciprocalPath — Method

ReciprocalPath(
    reciprocals::AbstractVector, segments::Pair{<:NTuple{N, Number}, <:NTuple{N, Number}}...;
    length::Union{Int, NTuple{M, Int}}=100,
    ends::Union{NTuple{2, Bool}, NTuple{M, NTuple{2, Bool}}}=(true, false)
) where {N, M}
ReciprocalPath(
    reciprocals::AbstractVector, segments::Tuple{Vararg{Pair{<:NTuple{N, Number}, <:NTuple{N, Number}}}};
    length::Union{Int, NTuple{M, Int}}=100,
    ends::Union{NTuple{2, Bool}, NTuple{M, NTuple{2, Bool}}}=(true, false)
) where {N, M}

ReciprocalPath{K}(
    reciprocals::AbstractVector, segments::Pair{<:NTuple{N, Number}, <:NTuple{N, Number}}...;
    length::Union{Int, NTuple{M, Int}}=100,
    ends::Union{NTuple{2, Bool}, NTuple{M, NTuple{2, Bool}}}=(true, false)
) where {K, N, M}
ReciprocalPath{K}(
    reciprocals::AbstractVector, segments::Tuple{Vararg{Pair{<:NTuple{N, Number}, <:NTuple{N, Number}}}};
    length::Union{Int, NTuple{M, Int}}=100,
    ends::Union{NTuple{2, Bool}, NTuple{M, NTuple{2, Bool}}}=(true, false)
) where {K, N, M}

Construct a path in the reciprocal space.

QuantumLattices.Spatials.ReciprocalSpace — Type

ReciprocalSpace{K, P} <: SimpleNamedVectorSpace{K, P}

Abstract type of reciprocal spaces.

QuantumLattices.Spatials.ReciprocalZone — Type

ReciprocalZone{K, S<:SVector, V<:Number} <: ReciprocalSpace{K, S}

A zone in the reciprocal space.

QuantumLattices.Spatials.ReciprocalZone — Method

ReciprocalZone(reciprocals::AbstractVector; length::Int=100)
ReciprocalZone(reciprocals::AbstractVector, bounds::Segment...)
ReciprocalZone(reciprocals::AbstractVector, bounds::Union{Tuple{Vararg{Segment}}, Vector{<:Segment}})

ReciprocalZone{K}(reciprocals::AbstractVector; length::Int=100) where K
ReciprocalZone{K}(reciprocals::AbstractVector, bounds::Segment...) where K
ReciprocalZone{K}(reciprocals::AbstractVector, bounds::Union{Tuple{Vararg{Segment}}, Vector{<:Segment}}) where K

Construct a rectangular zone in the reciprocal space.

QuantumLattices.Spatials.ReciprocalZone — Method

ReciprocalZone(bz::BrillouinZone)

Construct a reciprocal zone from a Brillouin zone.

QuantumLattices.Spatials.Segment — Type

Segment{S} <: AbstractVector{S}

A segment.

QuantumLattices.Spatials.Segment — Method

Segment(start::Number, stop::Number, length::Integer; ends::Tuple{Bool, Bool}=(true, false))
Segment(start::AbstractVector, stop::AbstractVector, length::Integer; ends::Tuple{Bool, Bool}=(true, false))
Segment(start::NTuple{N, Number}, stop::NTuple{N, Number}, length::Integer; ends::Tuple{Bool, Bool}=(true, false)) where N

Construct a segment.

QuantumLattices.Spatials.Translations — Type

Translations{N} <: VectorSpace{NTuple{N, Int}}

A set of translations. The boundary conditions along each translation direction are also included.

QuantumLattices.Spatials.@hexagon_str — Macro

hexagon"P₁-P₂-P₃-..."
hexagon"P₁-P₂-P₃-..., 120°"
hexagon"P₁-P₂-P₃-..., 60°"

Construct a tuple of start-stop point pairs for the hexagonal reciprocal space.

QuantumLattices.Spatials.@line_str — Macro

line"P₁-P₂-P₃-..."

Construct a tuple of start-stop point pairs for the one dimensional reciprocal space.

QuantumLattices.Spatials.@rectangle_str — Macro

rectangle"P₁-P₂-P₃-..."

Construct a tuple of start-stop point pairs for the rectangular reciprocal space.

QuantumLattices.Spatials.@translations_str — Macro

translations"n₁P/O-n₁P/O-..." -> Translations

Construct a set of translations from a string literal.

Base.eltype — Method

eltype(bond::Bond)
eltype(::Type{<:Bond{K, P} where K}) where {P<:Point}

Get the point type contained in a generic bond.

Base.getindex — Method

getindex(lattice::AbstractLattice, i::Integer) -> SVector

Get the ith coordinate.

Base.getindex — Method

getindex(bond::Bond, i::Integer) -> Point

Get the ith point contained in a generic bond.

Base.iterate — Function

iterate(bond::Bond, state=1)

Iterate over the points contained in a generic bond.

Base.length — Method

length(lattice::AbstractLattice) -> Int

Get the number of points contained in a lattice.

Base.length — Method

length(bond::Bond) -> Int

Get the number of points contained in a generic bond.

Base.reverse — Method

reverse(bond::Bond) -> Bond

Get the reversed bond.

QuantumLattices.Spatials.azimuth — Method

azimuth(v::AbstractVector{<:Number}) -> Number

Get the azimuth angle in radians of a vector.

QuantumLattices.Spatials.azimuthd — Method

azimuthd(v::AbstractVector{<:Number}) -> Number

Get the azimuth angle in degrees of a vector.

QuantumLattices.Spatials.bonds! — Method

bonds!(bonds::Vector, lattice::AbstractLattice, nneighbor::Int; coordination::Int=8)
bonds!(bonds::Vector, lattice::AbstractLattice, neighbors::Neighbors) -> typeof(bonds)

Get the required bonds of a lattice and append them to the input bonds.

QuantumLattices.Spatials.bonds — Method

bonds(lattice::AbstractLattice, nneighbor::Int; coordination::Int=8) -> Vector{Bond{Int, Point{dimension(lattice), dtype(lattice)}}}
bonds(lattice::AbstractLattice, neighbors::Neighbors) -> Vector{Bond{keytype(neighbors), Point{dimension(lattice), dtype(lattice)}}}

Get the required bonds of a lattice.

QuantumLattices.Spatials.distance — Method

distance(p₁::AbstractVector{<:Number}, p₂::AbstractVector{<:Number}) -> Number

Get the distance between two points.

Note

Compared to norm(p₁-p₂), this function avoids the memory allocation for p₁-p₂, thus is more efficient.

QuantumLattices.Spatials.icoordinate — Method

icoordinate(bond::Bond) -> SVector

Get the icoordinate of the bond.

QuantumLattices.Spatials.interlinks — Method

interlinks(cluster₁::AbstractMatrix{<:Number}, cluster₂::AbstractMatrix{<:Number}, neighbors::Neighbors) -> Vector{Tuple{Int, Int, Int}}

Use kdtree to get the intercluster nearest neighbors.

QuantumLattices.Spatials.isintracell — Method

isintracell(bond::Bond) -> Bool

Judge whether a bond is intra the unit cell of a lattice.

QuantumLattices.Spatials.isintracell — Method

isintracell(point::Point) -> Bool

Judge whether a point is intra the unitcell.

QuantumLattices.Spatials.isintratriangle — Method

isintratriangle(
    p::AbstractVector{<:Number}, p₁::AbstractVector{<:Number}, p₂::AbstractVector{<:Number}, p₃::AbstractVector{<:Number};
    vertexes::NTuple{3, Bool}=(true, true, true), edges::NTuple{3, Bool}=(true, true, true), atol::Real=atol, rtol::Real=rtol
) -> Bool

Judge whether a point belongs to the interior of a triangle whose vertexes are p₁, 'p₂' and p₃ with the give tolerance. vertexes and edges define whether the interior should contain the vertexes or edges, respectively.

Note

The vertexes are in the order (p₁, p₂, p₃) and the edges are in the order (p1p2, p2p3, p3p1).

The edges do not contain the vertexes.

QuantumLattices.Spatials.isonline — Method

isonline(
    p::AbstractVector{<:Number}, p₁::AbstractVector{<:Number}, p₂::AbstractVector{<:Number};
    ends::Tuple{Bool, Bool}=(true, true), atol::Real=atol, rtol::Real=rtol
) -> Bool

Judge whether a point is on a line segment whose end points are p₁ and p₂ with the given tolerance. ends defines whether the line segment should contain its ends.

QuantumLattices.Spatials.isparallel — Method

isparallel(v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}; atol::Real=atol, rtol::Real=rtol) -> Int

Judge whether two vectors are parallel to each other with the given tolerance, 0 for not parallel, 1 for parallel and -1 for antiparallel.

QuantumLattices.Spatials.issubordinate — Method

issubordinate(coordinate::AbstractVector{<:Number}, vectors::AbstractVector{<:AbstractVector{<:Number}}; atol::Real=atol, rtol::Real=rtol) -> Bool

Judge whether a coordinate belongs to a lattice defined by vectors with the given tolerance.

QuantumLattices.Spatials.minimumlengths — Function

minimumlengths(cluster::AbstractMatrix{<:Number}, vectors::AbstractVector{<:AbstractVector{<:Number}}, nneighbor::Int=1; coordination::Int=8) -> Vector{Float}

Use kdtree to search the lowest several minimum bond lengths within a lattice translated by a cluster.

When the translation vectors are not empty, the lattice will be considered periodic in the corresponding directions. Otherwise the lattice will be open in all directions. To search for the bonds across the periodic boundaries, the cluster will be pre-translated to become a supercluster, which has open boundaries but is large enough to contain all the nearest neighbors within the required order. The coordination parameter sets the average number of each order of nearest neighbors. If it is to small, larger bond lengths may not be searched, and the result will contain Inf. This is a sign that you may need a larger coordination. Another situation that Inf appears in the result occurs when the minimum lengths are searched in open lattices. Indeed, the cluster may be too small so that the required order just goes beyond it. In this case the warning message can be safely ignored.

QuantumLattices.Spatials.polar — Method

polar(v::AbstractVector{<:Number}) -> Number

Get the polar angle in radians of a vector.

QuantumLattices.Spatials.polard — Method

polard(v::AbstractVector{<:Number}) -> Number

Get the polar angle in degrees of a vector.

QuantumLattices.Spatials.rcoordinate — Method

rcoordinate(bond::Bond) -> SVector

Get the rcoordinate of the bond.

QuantumLattices.Spatials.reciprocals — Method

reciprocals(vectors::AbstractVector{AbstractVector{<:Number}}) -> Vector{Vector{<:Number}}

Get the reciprocals dual to the input vectors.

QuantumLattices.Spatials.rotate — Method

rotate(cluster::AbstractMatrix{<:Number}, angle::Number; axis::Tuple{Union{AbstractVector{<:Number}, Nothing}, Tuple{<:Number, <:Number}}=(nothing, (0, 0))) -> Matrix{<:Number}

Get a rotated cluster of the original one by a certain angle around an axis.

The axis is determined by a point it gets through (nothing can be used to denote the origin), and its polar as well as azimuth angles in radians. The default axis is the z axis.

Note

The result is given by the Rodrigues' rotation formula.

Only 2 and 3 dimensional vectors can be rotated.

When the input vectors are 2 dimensional, both the polar and azimuth of the axis must be 0.

QuantumLattices.Spatials.tile — Method

tile(cluster::AbstractMatrix{<:Number}, vectors::AbstractVector{<:AbstractVector{<:Number}}, translations) -> Matrix{<:Number}

Tile a supercluster by translations of the input cluster.

Basically, the final supercluster is composed of several parts, each of which is a translation of the original cluster, with the translation vectors specified by vectors and each set of the translation indices contained in translations. When translation vectors are empty, a copy of the original cluster will be returned.

QuantumLattices.Spatials.translate — Method

translate(cluster::AbstractMatrix{<:Number}, vector::AbstractVector{<:Number}) -> Matrix{vector|>eltype}

Get the translated cluster of the original one by a vector.

QuantumLattices.Spatials.volume — Method

volume(vectors::AbstractVector{<:SVector}) -> Number
volume(v::AbstractVector{<:Number}) -> Number
volume(v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}) -> Number
volume(v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}, v₃::AbstractVector{<:Number}) -> Number

Get the volume spanned by the input vectors.

QuantumLattices.decompose — Method

decompose(v₀::AbstractVector{<:Number}, v₁::AbstractVector{<:Number}) -> Tuple{Number}
decompose(v₀::AbstractVector{<:Number}, v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}) -> Tuple{Number, Number}
decompose(v₀::AbstractVector{<:Number}, v₁::AbstractVector{<:Number}, v₂::AbstractVector{<:Number}, v₃::AbstractVector{<:Number}) -> Tuple{Number, Number, Number}

Decompose a vector with respect to input basis vectors.

QuantumLattices.dimension — Method

dimension(lattice::AbstractLattice) -> Int
dimension(::Type{<:AbstractLattice{N}}) where N -> Int

Get the space dimension of the lattice.

QuantumLattices.dimension — Method

dimension(bond::Bond) -> Int
dimension(::Type{<:Bond{K, P} where K}) where {P<:Point} -> Int

Get the space dimension of a concrete bond.

QuantumLattices.dimension — Method

dimension(point::Point) -> Int
dimension(::Type{<:Point{N}}) where N -> Int

Get the spatial dimension of a point.

QuantumLattices.dtype — Method

dtype(lattice::AbstractLattice)
dtype(::Type{<:AbstractLattice{N, D} where N}) where {D<:Number}

Get the data type of the coordinates of a lattice.

QuantumLattices.dtype — Method

dtype(bond::Bond)
dtype(::Type{<:Bond{K, P} where K}) where {P<:Point}

Get the data type of the coordinates of the points contained in a generic bond.

QuantumLattices.dtype — Method

dtype(bz::BrillouinZone)
dtype(::Type{<:BrillouinZone{K, <:Momentum, S} where K}) where {S<:SVector}

Get the data type of a Brillouin zone.

QuantumLattices.dtype — Method

dtype(point::Point)
dtype(::Type{<:Point{N, D} where N}) where {D<:Number}

Get the data type of the coordinates of a point.

RecipesBase.apply_recipe — Function

@recipe plot(lattice::AbstractLattice, neighbors::Union{Int, Neighbors}, filter::Function=bond->true)

Define the recipe for the visualization of a lattice.