Algebra over fields

Algebra over fields

ID

VectorSpace

Element and Elements

Manual

Hamiltonian.Utilities.AlgebraOverField.VectorSpace — Constant.

VectorSpace

The corresponding vector space of an algebra over a field.

Alias for Union{SimpleVectorSpace,CompositeVectorSpace}.

source

Hamiltonian.Utilities.AlgebraOverField.VectorSpace — Method.

VectorSpace(ids::SimpleID...) -> SimpleVectorSpace
VectorSpace(svses::SimpleVectorSpace...) -> CompositeVectorSpace

Get the corresponding vector space of an algebra over a field.

source

Hamiltonian.Utilities.AlgebraOverField.CompositeVectorSpace — Type.

CompositeVectorSpace(svses::SimpleVectorSpace...)

The vector space spanned by the direct product of simple vector spaces.

source

Hamiltonian.Utilities.AlgebraOverField.Element — Type.

Element{V<:Number,I<:ID}

An element of an algebra over a field.

The first and second attributes of an element must be

value::Nuber: the coefficient of the element

id::ID: the id of the element

source

Hamiltonian.Utilities.AlgebraOverField.Elements — Type.

Elements{I<:ID,M<:Element} <: AbstractDict{I,M}

An set of elements of an algebra over a field.

Alias for Dict{I<:ID,M<:Element}. Similar iterms are automatically merged thanks to the id system.

source

Hamiltonian.Utilities.AlgebraOverField.Elements — Method.

Elements(ms)
Elements(ms::Pair{I,M}...) where {I<:ID,M<:Element}
Elements(ms::Element...)

Get the set of elements with similar items merged.

source

Hamiltonian.Utilities.AlgebraOverField.ID — Type.

ID(ids::NTuple{N,SimpleID}) where N
ID(ids::SimpleID...)
ID(::Type{SID},attrs::Vararg{NTuple{N},M}) where {SID<:SimpleID,N,M}

The id system of an algebra over a field.

Usually, a simple id corresponds to a single generator of the algebra while an id corresponds to an element of the algebra.

source

Hamiltonian.Utilities.AlgebraOverField.SimpleID — Type.

SimpleID <: AbstractNamedVector

A simple id is the building block of the id system of an algebra over a field.

source

Hamiltonian.Utilities.AlgebraOverField.SimpleVectorSpace — Type.

SimpleVectorSpace(ids::SimpleID...)

The vector space spanned by a set of bases specified by their ids.

source

Hamiltonian.Utilities.:⊕ — Method.

⊕(id1::I,id2::I) where {I<:SimpleID} -> SimpleVectorSpace{I}
⊕(id::I,svs::SimpleVectorSpace{I}) where {I<:SimpleID} -> SimpleVectorSpace{I}
⊕(svs::SimpleVectorSpace{I},id::I) where {I<:SimpleID} -> SimpleVectorSpace{I}
⊕(svs1::SVS,svs2::SVS) where {SVS<:SimpleVectorSpace} -> SVS

Get the direct sum of bases or simple vector spaces.

source

Hamiltonian.Utilities.:⊗ — Method.

⊗(sid1::SimpleID,sid2::SimpleID) -> ID
⊗(sid::SimpleID,cid::ID) -> ID
⊗(cid::ID,sid::SimpleID) -> ID
⊗(cid1::ID,cid2::ID) -> ID

Get the direct product of the id system.

source

Hamiltonian.Utilities.:⊗ — Method.

⊗(svs1::SimpleVectorSpace,svs2::SimpleVectorSpace) -> CompositeVectorSpace
⊗(svs::SimpleVectorSpace,cvs::CompositeVectorSpace) -> CompositeVectorSpace
⊗(cvs::CompositeVectorSpace,svs::SimpleVectorSpace) -> CompositeVectorSpace
⊗(cvs1::CompositeVectorSpace,cvs2::CompositeVectorSpace) -> CompositeVectorSpace

Get the direct product of simple vector spaces or composite vector spaces.

source

Hamiltonian.Utilities.AlgebraOverField.add! — Method.

add!(ms::Elements) -> typeof(ms)
add!(ms::Elements,m::Element) -> typeof(ms)
add!(ms::Elements,mms::Elements) -> typeof(ms)

Get the inplace addition of elements to a set.

source

Hamiltonian.Utilities.AlgebraOverField.idtype — Method.

idtype(::Type{<:Element{V,I}}) where {V,I}
idtype(m::Element)

The type of the id of an element.

source

Hamiltonian.Utilities.AlgebraOverField.sub! — Method.

sub!(ms::Elements) -> typeof(ms) -> typeof(ms)
sub!(ms::Elements,m::Element) -> typeof(ms)
sub!(ms::Elements,mms::Elements) -> typeof(ms)

Get the inplace subtraction of elements from a set.

source

Hamiltonian.Utilities.dimension — Method.

dimension(svs::SimpleVectorSpace) -> Int
dimension(cvs::CompositeVectorSpace) -> Int

Get the dimension of a vector space.

source

Hamiltonian.Utilities.rank — Method.

rank(::Type{<:Element{V,I,N}}) where {V,I,N} -> Int
rank(m::Element) -> Int

Get the rank of an element.

source

Hamiltonian.Utilities.rank — Method.

rank(::Type{<:ID{N,I}}) where {N,I} -> Int
rank(id::ID) -> Int

Get the rank of a composite id.

source

Base.:* — Method.

*(factor::Number,m::Element) -> Element
*(m::Element,factor::Number) -> Element
*(m1::Element,m2::Element) -> Element
*(factor::Number,ms::Elements) -> Elements
*(ms::Elements,factor::Number) -> Elements
*(m::Element,ms::Elements) -> Elements
*(ms::Elements,m::Element) -> Elements
*(ms1::Elements,ms2::Elements) -> Elements

Overloaded * operator for element-scalar multiplications and element-element multiplications of an algebra over a field.

source

Base.:+ — Method.

+(m::Element) -> typeof(m)
+(ms::Elements) -> typeof(ms)
+(ms::Elements,m::Element) -> Elements
+(m1::Element,m2::Element) -> Elements
+(m::Element,ms::Elements) -> Elements
+(ms1::Elements,ms2::Elements) -> Elements

Overloaded + operator between elements of an algebra over a field.

source

Base.:- — Method.

-(m::Element) -> typeof(m)
-(ms::Elements) -> typeof(ms)
-(m1::Element,m2::Element) -> Elements
-(m::Element,ms::Elements) -> Elements
-(ms::Elements,m::Element) -> Elements
-(ms1::Elements,ms2::Elements) -> Elements

Overloaded - operator between elements of an algebra over a field.

source

Base.:/ — Method.

/(m::Element,factor::Number)
/(ms::Elements,factor::Number)

Overloaded / operator for element-sclar division of an algebra over a field.

source

Base.:== — Method.

==(m1::M,m2::M) where M<:Element -> Bool
isequal(m1::M,m2::M) where M<:Element -> Bool

Compare two elements and judge whether they are equal to each other.

source

Base.getproperty — Method.

getproperty(cid::ID,name::Symbol)

Get the property of a composite id.

source

Base.hash — Method.

hash(cid::ID,h::UInt)

Hash a composite id.

source

Base.isless — Method.

isless(cid1::ID,cid2::ID) -> Bool
<(cid1::ID,cid2::ID) -> Bool

Compare two ids and judge whether the first is less than the second.

We assume that ids with smaller ranks are always less than those with higher ranks. If two ids are of the same rank, the comparison goes just like that between tuples.

source

Base.propertynames — Method.

propertynames(::Type{I},private::Bool=false) where I<:ID -> Tuple

Get the property names of a composite id.

source

Base.replace — Method.

replace(m::Element;kwargs...) -> typeof(m)

Return a copy of a concrete Element with some of the field values replaced by the keyword arguments.

source

Base.show — Method.

show(io::IO,cid::ID)

Show a composite id.

source

Base.valtype — Method.

valtype(::Type{<:Element{V}}) where {V}
valtype(m::Element)

Get the type of the value of an element.

The result is also the type of the field over which the algebra is defined.

source

Base.zero — Method.

zero(ms::Elements) -> typeof(ms)
zero(::Type{Elements{I,M}}) where {I,M} -> Elements{I,M}

Get a zero set of elements.

A zero set of elements is defined to be the empty one.

source