Factory

Factory

The aim of Factory is to provide tools to hack into Julia codes without knowing the details of their abstract syntax trees and regularize the mechanism to "escape" variables in Expr expressions, so that users can manipulate the existing codes, modify them and generate new ones in macros. In particular, Factory in this module means the representation of certain blocks of Julia codes by a usual Julia struct. This representation is much easier to comprehend than the canonical Expr representation and makes it far more convenient to define macros. In general, we propose the following requirements that any factory must satisfy:

DECOMPOSITION - An Expr expression can be decomposed into its corresponding factory by the factory's constructor.

COMPOSITION - A factory can compose its corresponding Expr expression by calling itself.

ESCAPE - A variable should be escaped or not in the composed Expr expression by a factory depends on keyword arguments escaped or unescaped passed to the factory call, which are tuple of Symbols. Specifically, when escaped is provided, a variable should be escaped if its name is in escaped, whereas, when unescaped is provided, a variable should be escaped if its name is not in unescaped. A factory can choose which keyword arguments should be used for convenience, or can choose both with different keyword arguments for different parts of it.

The first two requirements define the basic interfaces to interact with factories, and the third requirement proposes the escape mechanism of variables.

Out of practical purposes, we implemente 7 kinds of factories, i.e. Inference, Argument, Parameter, Field, Block, FunctionFactory and TypeFactory, which represent a type inference, a function argument, a method or type parameter, a struct field, a begin ... end block, a function itself and a struct itself, respectively. Some of the basic methods making the above requirements fulfilled with these types are based on the powerful functions defined in MacroTools.

Inference

An Inference has 3 attributes:

head::Union{Symbol,Nothing}: the head of the type inference, which must be one of (nothing,:(::),:(<:),:curly)

name::Union{Symbol,Nothing}: the name of the type inference

params::Union{Inference,Vector{Inference},Nothing}: the parameters of the type inference

All valid expressions representing type inferences can be passed to the constructor:

julia> Inference(:T)
Inference(
  head:   nothing
  name:   T
  params: nothing
)

julia> Inference(:(::T))
Inference(
  head:   ::
  name:   nothing
  params: T
)

julia> Inference(:(<:Number))
Inference(
  head:   <:
  name:   nothing
  params: Number
)

julia> Inference(:(Vector{T}))
Inference(
  head:   curly
  name:   Vector
  params: Inference[T]
)

julia> Inference(:(Vector{Tuple{String,Int}}))
Inference(
  head:   curly
  name:   Vector
  params: Inference[Tuple{String, Int}]
)

julia> Inference(:(Type{<:Number}))
Inference(
  head:   curly
  name:   Type
  params: Inference[<:Number]
)

On the other hand, you can use the macro @inference to construct an Inference directly from a type inference:

julia> @inference Vector{Tuple{String,Int}}
Inference(
  head:   curly
  name:   Vector
  params: Inference[Tuple{String, Int}]
)

Note

Inference is a recursive struct, i.e. it recursively decomposes a type inference until the final type inference is just a Symbol.

When the input expression is a Symbol, the head and params attributes of the resulting Inference is nothing. Otherwise, its head is the same with that of the input expression, and the args of the input expression will be further decomposed, whose result will be stored in params.

When the head of the input expression is :(::) or :(<:), the params is an Inference whereas when the head of the input expression is :curly, the params is a Vector{Inference}.

Inference uses the keyword argument unescaped to escape variables, e.g.

julia> Inference(:(Vector{T}))() |> println
($(Expr(:escape, :Vector))){$(Expr(:escape, :T))}

julia> Inference(:(Vector{T}))(unescaped=(:T,)) |> println
($(Expr(:escape, :Vector))){T}

julia> Inference(:(Vector{T}))(unescaped=(:Vector,:T)) |> println
Vector{T}

Argument

An Argument has 4 attributes:

name::Union{Symbol,Nothing}: the name of the argument

type::Inference: the type inference of the argument

slurp::Bool: whether the argument should be expanded by ...

default::Any: the default value of the argument, nothing for those with no default values

All valid expressions representing the arguments of functions can be passed to the constructor:

julia> Argument(:arg)
Argument(
  name:    arg
  type:    Any
  slurp:   false
  default: nothing
)

julia> Argument(:(arg::ArgType))
Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: nothing
)

julia> Argument(:(arg::ArgType...))
Argument(
  name:    arg
  type:    ArgType
  slurp:   true
  default: nothing
)

julia> Argument(:(arg::ArgType=default))
Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: default
)

Or you can use the macro @argument for a direct construction from an argument declaration:

julia> @argument arg::ArgType=default
Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: default
)

The construction from such expressions is based on the the MacroTools.splitarg function.

Argument also uses the keyword argument unescaped to escape variables, e.g.

julia> Argument(:(arg::ArgType=default))(unescaped=(:ArgType,:default)) |> println
arg::ArgType=default

It can be seen the name of an argument will never be escaped even though it is not in the unescaped tuple. This is obvious since the name of a function argument is always local. By the way, the composition of an Argument expression is based on the MacroTools.combinearg function.

Parameter

A Parameter has 2 attributes:

name::Union{Symbol,Nothing}: the name of the parameter

type::Union{Inference,Nothing}: the type inference of the parameter

All expressions that represent type parameters or method parameters are allowed to be passed to the constructor:

julia> Parameter(:T)
Parameter(
  name: T
  type: nothing
)

julia> Parameter(:(<:Number))
Parameter(
  name: nothing
  type: Number
)

julia> Parameter(:(T<:Number))
Parameter(
  name: T
  type: Number
)

julia> Parameter(:(::Int))
Parameter(
  name: nothing
  type: Int
)

The macro @parameter completes the construction directly from a parameter declaration:

julia> @parameter T<:Number
Parameter(
  name: T
  type: Number
)

Note

We use nothing to denote a missing name or type.

Two subtle situations of type/method parameters, e.g. MyType{T} and MyType{Int}, are distinguished by Parameter(:T) and Parameter(:(::Int)). The name and type attributes of the resulting Parameters are, for the first case, :T and nothing, while for the second case, nothing and :T, respectively. Moreover, the callings of the factories for these two cases are also different, e.g. Parameter(:T)()==:T and Parameter(:(::Int))==:(<:$(esc(Int))).

Parameter uses the keyword argument unescaped to escape variables, too, e.g.

julia> Parameter(:(N<:Vector{T}))(unescaped=(:T,)) |> println
N <: ($(Expr(:escape, :Vector))){T}

As is similar to Argument, the name of a method/type parameter will never be escaped because of its local scope.

Field

A Field has 2 attributes:

name::Symbol: the name of the field

type::Inference: the type inference of the field

Legal expressions can be used to construct a Field instance by its constructor:

julia> Field(:field)
Field(
  name: field
  type: Any
)

julia> Field(:(field::FieldType))
Field(
  name: field
  type: FieldType
)

julia> Field(:(field::ParametricType{T}))
Field(
  name: field
  type: ParametricType{T}
)

The macro @field is also provided to help the construction directly from a field declaration:

julia> @field field::FieldType
Field(
  name: field
  type: FieldType
)

The construction from these expressions is based on the MacroTools.splitarg function.

Field uses the keyword argument unescaped to escape variables as well, e.g.

julia> Field(:(field::Dict{N,D}))(unescaped=(:N,:D)) |> println
field::($(Expr(:escape, :Dict))){N, D}

The name of a struct will never be escaped either because it is a local variable tightly binding to a struct. It is noted that the composition of field expressions is based on the MacroTools.combinefield function.

Block

A Block has only one attribute:

body::Vector{Any}: the body of the begin ... end block

Any expression can be passed to the constructor of Block:

julia> Block(:(x=1))
Block(
  body: Any[:(x = 1)]
)

julia> Block(:(x=1;y=2))
Block(
  body: Any[:(x = 1), :(#= none:1 =#), :(y = 2)]
)

julia> Block(:(begin x=1 end))
Block(
  body: Any[:(#= none:1 =#), :(x = 1)]
)

julia> Block(quote
               x=1
               y=2
           end)
Block(
  body: Any[:(#= none:2 =#), :(x = 1), :(#= none:3 =#), :(y = 2)]
)

Or you can construct a Block instance directly from any code by the macro @block:

julia> @block x=1 y=2
Block(
  body: Any[:(x = 1), :(y = 2)]
)

The body of a block can also be extended by the push! function or the @push! macro.

Note

The body of a Block is somewhat "flattened", i.e. it contains no begin ... end blocks. During the initialization, any such input block will be unblocked and added to the body part by part. So is the push! and @push! procedures.

All LineNumberNodes generated by the input codes will also be included in the block's body. However, you can use rmlines! or @rmlines! to remove them from the body of an existing Block, or use rmlines or @rmlines to get a copy with them removed in the body.

Different from previous factories, Block uses the keyword argument escaped to escape variables. This is because variables in a block are often local ones and should not be escaped. Therefore, only those defined in other modules should be noted and escaped, which usually constitute the minority. For example,

julia> Block(:(x=1;y=2;z=Int[1,2,3]))(escaped=(:Int,)) |> println
begin
    x = 1
    #= none:1 =#
    y = 2
    #= none:1 =#
    z = ($(Expr(:escape, :Int)))[1, 2, 3]
end

FunctionFactory

A FunctionFactory has 7 attributes:

name::Union{Symbol,Expr}: the name of the function

params::Vector{Inference}: the method parameters of the function

args::Vector{Argument}: the positional arguments of the function

kwargs::Vector{Argument}: the keyword arguments of the function

rtype::Inference: the return type of the function

whereparams::Vector{Parameter}: the method parameters specified by the where keyword

body::Block: the body of the function

All expressions that represent functions are allowed to be passed to the constructor:

julia> FunctionFactory(:(f()=nothing))
FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[]
  kwargs:      Argument[]
  rtype:       Any
  whereparams: Parameter[]
  body:        begin
                   nothing
               end
)

julia> FunctionFactory(:(f(x)=x))
FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::Any]
  kwargs:      Argument[]
  rtype:       Any
  whereparams: Parameter[]
  body:        begin
                   x
               end
)

julia> FunctionFactory(:(f(x::Int,y::Int;choice::Function=sum)=choice(x,y)))
FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::Int, y::Int]
  kwargs:      Argument[choice::Function=sum]
  rtype:       Any
  whereparams: Parameter[]
  body:        begin
                   choice(x, y)
               end
)

julia> FunctionFactory(:(f(x::T,y::T;choice::Function=sum) where T<:Number=choice(x,y)))
FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[choice::Function=sum]
  rtype:       Any
  whereparams: Parameter[T <: Number]
  body:        begin
                   choice(x, y)
               end
)

julia> FunctionFactory(:((f(x::T,y::T;choice::Function=sum)::T) where T<:Number=choice(x,y)))
FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[choice::Function=sum]
  rtype:       T
  whereparams: Parameter[T <: Number]
  body:        begin
                   choice(x, y)
               end
)

julia> FunctionFactory(:(
           function (f(x::T,y::T;choice::Function=sum)::T) where T<:Number
               choice(x,y)
           end
       ))
FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[choice::Function=sum]
  rtype:       T
  whereparams: Parameter[T <: Number]
  body:        begin
                   #= none:3 =#
                   choice(x, y)
               end
)

julia> FunctionFactory(
           quote
               function (f(x::T,y::T;choice::Function=sum)::T) where T<:Number
                   choice(x,y)
               end
           end
       )
FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[choice::Function=sum]
  rtype:       T
  whereparams: Parameter[T <: Number]
  body:        begin
                   #= none:4 =#
                   choice(x, y)
               end
)

Similarly, an instance can also be constructed from the macro @functionfactory:

julia> @functionfactory (f(x::T,y::T;choice::Function=sum)::T) where T<:Number=choice(x,y)
FunctionFactory(
  name:        f
  params:      Inference[]
  args:        Argument[x::T, y::T]
  kwargs:      Argument[choice::Function=sum]
  rtype:       T
  whereparams: Parameter[T <: Number]
  body:        begin
                   choice(x, y)
               end
)

The construction from such expressions are based on the MacroTools.splitdef function.

Note

Since Julia 0.7, the form MyType{D}(data::D) where D only appears in struct constructors, therefore, the attribute :params of a function factory is nonempty only when this factory aims to represent a struct constructor.

Usually, the name of a function factory is a Symbol. However, if the factory aims to extend some methods of a function defined in another module, e.g., Base.eltype, the name will be an Expr.

Since FunctionFactory involves not only factories using unescaped but also factories using escaped, it adopts both to escape variables, with unescaped for params, args, kwargs, rtype and whereparams while escaped for name and body. It is worth to emphasize that the name of a function factory is affected by the escaped argument. Specifically, when the name is a Symbol and is in the escaped tuple, it will be escaped. Otherwise it will not, especially when it is an Expr, it will never be escaped because an Expr cannot be a element of a NTuple{N,Symbol} where N. See examples,

julia> FunctionFactory(:((f(x::T,y::T;choice::Function=sum)::T) where T<:Number=max(x,y,choice(x,y))))(unescaped=(:T,),escaped=(:f,:max,)) |> println
function (($(Expr(:escape, :f)))(x::T, y::T; choice::$(Expr(:escape, :Function))=$(Expr(:escape, :sum)))::T) where T <: $(Expr(:escape, :Number))
    #= /home/travis/.julia/packages/MacroTools/4AjBS/src/utils.jl:302 =#
    begin
        ($(Expr(:escape, :max)))(x, y, choice(x, y))
    end
end

julia> FunctionFactory(:((f(x::T,y::T;choice::Function=sum)::T) where T<:Number=max(x,y,choice(x,y))))(unescaped=(:T,),escaped=(:max,)) |> println
function (f(x::T, y::T; choice::$(Expr(:escape, :Function))=$(Expr(:escape, :sum)))::T) where T <: $(Expr(:escape, :Number))
    #= /home/travis/.julia/packages/MacroTools/4AjBS/src/utils.jl:302 =#
    begin
        ($(Expr(:escape, :max)))(x, y, choice(x, y))
    end
end

The compositions of function expressions are based on the MacroTools.combinedef function.

Other features include:

Positional arguments can be added by addargs! or @addargs!

Keyword arguments can be added by addkwargs! or @addkwargs!

Where parameters can be added by addwhereparams! or @addwhereparams!

Body can be extended by extendbody! or @extendbody!

TypeFactory

A TypeFactory has 6 attributes:

name::Symbol: the name of the struct

mutable::Bool: whether or not the struct is mutable

params::Vector{Parameter}: the type parameters of the struct

supertype::Inference: the supertype of the struct

fields::Vector{Field}: the fields of the struct

constructors::Vector{FunctionFactory}: the inner constructors of the struct

Any expression representing valid struct definitions can be passed to the constructor:

julia> TypeFactory(:(struct StructName end))
TypeFactory(
  name:         StructName
  mutable:      false
  params:       Parameter[]
  supertype:    Any
  fields:       Field[]
  constructors: FunctionFactory[]
)

julia> TypeFactory(:(struct StructName{T} end))
TypeFactory(
  name:         StructName
  mutable:      false
  params:       Parameter[T]
  supertype:    Any
  fields:       Field[]
  constructors: FunctionFactory[]
)

julia> TypeFactory(:(struct Child{T} <: Parent{T} end))
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T]
  supertype:    Parent{T}
  fields:       Field[]
  constructors: FunctionFactory[]
)

julia> TypeFactory(:(
           struct Child{T<:Number} <: Parent{T}
               field1::T
               field2::T
           end
       ))
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T <: Number]
  supertype:    Parent{T}
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[]
)

julia> TypeFactory(
           quote
               struct Child{T<:Number} <: Parent{T}
                   field1::T
                   field2::T
               end
           end
       )
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T <: Number]
  supertype:    Parent{T}
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[]
)

Also, the macro @typefactory supports the construction directly from a type definition:

julia> @typefactory struct Child{T<:Number} <: Parent{T}
                       field1::T
                       field2::T
                       Child(field1::T,field2::T=zero(T)) where T=new{T}(field1,field2)
                   end
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T <: Number]
  supertype:    Parent{T}
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[
                  function (Child(field1::T, field2::T=zero(T); )::Any) where T
                      #= /home/travis/.julia/packages/MacroTools/4AjBS/src/utils.jl:302 =#
                      begin
                          new{T}(field1, field2)
                      end
                  end
                  ]
)

The construction from these expressions is based on the MacroTools.splitstructdef function.

TypeFactory also uses both keyword arguments, i.e. unescaped and escaped, to escape variables, with the former for params, supertype and fields, the latter for name, and both for constructors. For example,

julia> (@typefactory struct Child{T<:Number} <: Parent{T} field::T; Child(field::T) where T=new{T}(field) end)(unescaped=(:T,),escaped=(:Child,)) |>println
struct ($(Expr(:escape, :Child))){T <: $(Expr(:escape, :Number))} <: ($(Expr(:escape, :Parent))){T}
    #= /home/travis/.julia/packages/MacroTools/4AjBS/src/structdef.jl:57 =#
    field::T
    #= /home/travis/.julia/packages/MacroTools/4AjBS/src/structdef.jl:58 =#
    function (($(Expr(:escape, :Child)))(field::T; )::$(Expr(:escape, :Any))) where T
        #= /home/travis/.julia/packages/MacroTools/4AjBS/src/utils.jl:302 =#
        begin
            new{T}(field)
        end
    end
end

The composition of a type expression is based on the MacroTools.combinestructdef function.

Other features include:

Fields can be added by addfields! or @addfields!

Type parameters can be added by addparams! or @addparams!

Inner constructors can be added by addconstructors! or @addconstructors!

Manual

Hamiltonian.Utilities.Factory.FExpr — Constant.

Factory expression types, which is defined as Union{Symbol,Expr}.

source

Hamiltonian.Utilities.Factory.RawExpr — Constant.

Whether or not to show raw expressions of factoies.

source

Hamiltonian.Utilities.Factory.Argument — Type.

Argument(name::Union{Symbol,Nothing},type::Inference,slurp::Bool,default::Any)
Argument(;name::Union{Symbol,Nothing}=nothing,type::Inference=Inference(:Any),slurp::Bool=false,default::Any=nothing)
Argument(expr::FExpr)

The struct to describe a argument of a function.

source

Hamiltonian.Utilities.Factory.Argument — Method.

(a::Argument)(::typeof(RawExpr))
(a::Argument)(;unescaped::NTuple{N,Symbol}=()) where N

Convert an Argument to the Expr representation of the argument it describes.

source

Hamiltonian.Utilities.Factory.Block — Type.

Block(parts::FExpr...)

The struct to describe a begin ... end block.

source

Hamiltonian.Utilities.Factory.Block — Method.

(b::Block)(::typeof(RawExpr))
(b::Block)(;escaped::NTuple{N,Symbol}=()) where N

Convert a Block to the Expr representation of the begin ... end block it describes.

source

Hamiltonian.Utilities.Factory.Field — Type.

Field(name::Symbol,type::Inference)
Field(;name::Symbol,type::FExpr=Inference(:Any))
Field(expr::Expr)

The struct to describe a field of a struct.

source

Hamiltonian.Utilities.Factory.Field — Method.

(f::Field)(::typeof(RawExpr))
(f::Field)(;unescaped::NTuple{N,Symbol}=()) where N

Convert a Field to the Expr representation of the field it describes.

source

Hamiltonian.Utilities.Factory.FunctionFactory — Type.

FunctionFactory(name::FExpr,params::Vector{Inference},args::Vector{Argument},kwargs::Vector{Argument},rtype::Inference,whereparams::Vector{Parameter},body::Block)
FunctionFactory(    ;name::FExpr,
                    params::Vector{Inference}=Inference[],
                    args::Vector{Argument}=Argument[],
                    kwargs::Vector{Argument}=Argument[],
                    rtype::Inference=Inference(:Any),
                    whereparams::Vector{Parameter}=Parameter[],
                    body::Block=Block()
                    )
FunctionFactory(expr::Expr)

The struct to describe a function.

source

Hamiltonian.Utilities.Factory.FunctionFactory — Method.

(ff::FunctionFactory)(::typeof(RawExpr))
(ff::FunctionFactory)(;unescaped::NTuple{N,Symbol}=(),escaped::NTuple{M,Symbol}=()) where {N,M}

Convert a FunctionFactory to the Expr representation of the function it describes.

source

Hamiltonian.Utilities.Factory.Inference — Type.

Inference(head::Union{Symbol,Nothing},name::Union{Symbol,Nothing},params::Union{Inference,Vector{Inference},Nothing})
Inference(;
        head::Union{Symbol,Nothing}=nothing,
        name::Union{Symbol,Nothing}=nothing,
        params::Union{Inference,Vector{Inference},Nothing}=nothing,
        )
Inference(expr::FExpr)

The struct to describe a type inference.

source

Hamiltonian.Utilities.Factory.Inference — Method.

(i::Inference)(::typeof(RawExpr))
(i::Inference)(;unescaped::NTuple{N,Symbol}=()) where N

Convert a Inference to the Expr representation of the type inference it describes.

source

Hamiltonian.Utilities.Factory.Parameter — Type.

Parameter(name::Union{Symbol,Nothing},type::Union{Inference,Nothing})
Parameter(;name::Union{Symbol,Nothing}=nothing,type::Union{Inference,Nothing}=nothing)
Parameter(expr::FExpr)

The struct to describe a parameter of a function or a type.

source

Hamiltonian.Utilities.Factory.Parameter — Method.

(p::Parameter)(::typeof(RawExpr))
(p::Parameter)(;unescaped::NTuple{N,Symbol}=()) where N

Convert a Parameter to the Expr representation of the parameter it describes.

source

Hamiltonian.Utilities.Factory.TypeFactory — Type.

TypeFactory(name::Symbol,mutable::Bool,params::Vector{Parameter},supertype::Inference,fields::Vector{Field},constructors::Vector{FunctionFactory})
TypeFactory(    ;name::Symbol,
                mutable::Bool=false,
                params::Vector{Parameter}=Parameter[],
                supertype::Inference=Inference(:Any),
                fields::Vector{Field}=Field[],
                constructors::Vector{FunctionFactory}=FunctionFactory[],
                )
TypeFactory(expr::Expr)

The struct to describe a struct.

source

Hamiltonian.Utilities.Factory.TypeFactory — Method.

(tf::TypeFactory)(::typeof(RawExpr))
(tf::TypeFactory)(;unescaped::NTuple{N,Symbol}=()) where N

Convert a TypeFactory to the Expr representation of the struct it describes.

source

Hamiltonian.Utilities.Factory.@addargs! — Macro.

@addargs! ff args::FExpr...

Add a couple of positional arguments to a function factory.

source

Hamiltonian.Utilities.Factory.@addconstructors! — Macro.

@addconstructors! tf constructors::Expr...

Add a couple of constructors to a type factory.

source

Hamiltonian.Utilities.Factory.@addfields! — Macro.

@addfields! tf fields::FExpr...

Add a couple of fields to a type factory.

source

Hamiltonian.Utilities.Factory.@addkwargs! — Macro.

@addkwargs! ff kwargs::FExpr...

Add a couple of keyword arguments to a function factory.

source

Hamiltonian.Utilities.Factory.@addparams! — Macro.

@addparams! f params::FExpr...

Add a couple of method parameters to a function factory or a type factory.

source

Hamiltonian.Utilities.Factory.@addwhereparams! — Macro.

@addwhereparams! f whereparams::FExpr...

Add a couple of method parameters to a function factory or a type factory.

source

Hamiltonian.Utilities.Factory.@argument — Macro.

@argument expr::FExpr

Construct an Argument directly from an argument statement.

source

Hamiltonian.Utilities.Factory.@block — Macro.

@block parts::FExpr...

Construct a Block directly from a begin ... end block definition.

source

Hamiltonian.Utilities.Factory.@extendbody! — Macro.

@extendbody! ff parts::FExpr...

Extend the body of a function factory.

source

Hamiltonian.Utilities.Factory.@field — Macro.

@field expr::FExpr

Construct a Field directly from a field statement.

source

Hamiltonian.Utilities.Factory.@functionfactory — Macro.

@functionfactory expr::FExpr

Construct a FunctionFactory directly from a function definition.

source

Hamiltonian.Utilities.Factory.@inference — Macro.

@inference expr::FExpr

Construct an Inference directly from a type inference.

source

Hamiltonian.Utilities.Factory.@parameter — Macro.

@parameter expr::FExpr

Construct a Parameter directly from an parameter statement.

source

Hamiltonian.Utilities.Factory.@push! — Macro.

@push! b parts::FExpr...

Push other parts into the body of a block.

source

Hamiltonian.Utilities.Factory.@rmlines! — Macro.

@rmlines! b::Expr

Remove line number nodes in the body of a block.

source

Hamiltonian.Utilities.Factory.@rmlines — Macro.

@rmlines b::Expr

Return a copy of a block with the line number nodes removed.

source

Hamiltonian.Utilities.Factory.@typefactory — Macro.

@typefactory expr::Expr

Construct a TypeFactory directly from a type definition.

source

Hamiltonian.Utilities.Factory.addargs! — Method.

addargs!(ff::FunctionFactory,args::Argument...)
addargs!(ff::FunctionFactory,args::FExpr...)

Add a couple of positional arguments to a function factory.

source

Hamiltonian.Utilities.Factory.addconstructors! — Method.

addconstructors!(tf::TypeFactory,constructors::FunctionFactory...)
addconstructors!(tf::TypeFactory,constructors::Expr...)

Add a couple of constructors to a type factory.

source

Hamiltonian.Utilities.Factory.addfields! — Method.

addfields!(tf::TypeFactory,fields::Field...)
addfields!(tf::TypeFactory,fields::FExpr...)

Add a couple of fields to a type factory.

source

Hamiltonian.Utilities.Factory.addkwargs! — Method.

addkwargs!(ff::FunctionFactory,kwargs::Argument...)
addkwargs!(ff::FunctionFactory,kwargs::FExpr...)

Add a couple of keyword arguments to a function factory.

source

Hamiltonian.Utilities.Factory.addparams! — Method.

addparams!(f::FunctionFactory,params::Inference...)
addparams!(f::TypeFactory,params::Parameter...)
addparams!(f::Union{FunctionFactory,TypeFactory},params::FExpr...)

Add a couple of parameters to a function factory or a type factory.

source

Hamiltonian.Utilities.Factory.addwhereparams! — Method.

addwhereparams!(f::FunctionFactory,whereparams::Parameter...)
addwhereparams!(f::FunctionFactory,whereparams::FExpr...)

Add a couple of method where parameters to a function factory or a type factory.

source

Hamiltonian.Utilities.Factory.escape — Method.

escape(expr,::NTuple{N,Symbol}=()) where N
escape(expr::Symbol,escaped::NTuple{N,Symbol}=()) where N
escape(expr::Expr,escaped::NTuple{N,Symbol}=()) where N

Escape the variables sepecified by escaped in the input expression.

source

Hamiltonian.Utilities.Factory.extendbody! — Method.

extendbody!(ff::FunctionFactory,parts::FExpr...)
extendbody!(ff::FunctionFactory,parts::Block...)

Extend the body of a function factory.

source

Hamiltonian.Utilities.Factory.rmlines! — Method.

rmlines!(b::Block)

Remove line number nodes in the body of a block.

source

MacroTools.rmlines — Method.

rmlines(b::Block)

Return a copy of a block with the line number nodes removed.

source

Hamiltonian.Utilities.Factory.AbstractFactory — Type.

AbstractFactory

Abstract type for all concrete factories.

source

Base.:== — Method.

==(f1::F,f2::F) where F<:AbstractFactory

Overloaded == operator.

source

Base.names — Method.

names(expr::Union{Symbol,Expr})

Get all the variable names in an Expr.

source

Base.push! — Method.

push!(b::Block,parts::FExpr...)
push!(b::Block,parts::Block...)

Push other parts into the body of a block.

source

Base.replace — Method.

replace(f::AbstractFactory;kwargs...)

Return a copy of a concrete AbstractFactory with some of the field values replaced by the keyword arguments.

source

Base.show — Method.

show(io::IO,f::AbstractFactory)

Show a concrete AbstractFactory.

source