Factory

Factory

The aim of Factory is to provide tools to hack into Julia codes without knowing the details of their abstract syntax trees, so that users can somewhat manipulate the existing codes, modify them and generate new ones. In particular, Factory in this module means the representation of certain blocks of Julia codes by a usual Julia struct. This representation is much easier to comprehend than the canonical Expr representation. In general, we propose two basic requirements that any factory must satisfy:

Besides the default constructor, a concrete factory can be constructed from any legal Expr expression that represents the block of codes it aims to represent;

The canonical Expr expression of the block of codes that a concrete factory represents can be obtained by "calling" the factory itself.

These two requirements also defines the basic interfaces to interact with factories.

Out of practical purposes, we only implemente 5 kinds of factories, i.e. Block, Argument, FunctionFactory, Field and TypeFactory, which represent a begin ... end block, an argument of a function, a function itself, a field of a struct and a struct itself, respectively. Some of the basic methods making the above requirements fulfilled with these types are based on the powerful functions defined in MacroTools.

Block

A Block has only one attribute:

body::Vector{Any}: the body of the begin ... end block

Any expression can be passed to the constructor of Block, such as:

:(x=1)

:(x=1;y=2)

:(begin x=1 end)

quote
    x=1
    y=2
end

Or you can construct a Block instance directly from any code by the macro @block:

@block x=1 y=2

whose result is equivalent to

Block(Expr(:block,:(x=1),:(y=2)))

The body of a block can also be extended by the push! function or the @push! macro.

Note

The body of a Block is somewhat "flattened", i.e. it contains no begin ... end blocks. During the initialization, any such input block will be unblocked and added to the body part by part. So is the push! and @push! processes.

Argument

An Argument has 4 attributes:

name::Symbol: the name of the argument

type::Union{Symbol,Expr}: the type of the argument

slurp::Bool: whether the argument should be expanded by ...

default::Any: the default value of the argument, nothing for those with no default values

Valid expressions that can be passed to the constructor include:

:(arg)

:(arg::ArgType)

:(arg::ArgType...)

:(arg::ArgType=default)

Or you can use the macro @argument for construction directly from an argument declaration:

@argument arg::ArgType=default

The construction from such expressions is based on the the MacroTools.splitarg function. On the other hand, calling an instance of Argument will get the corresponding Expr expression, e.g.,

Argument(:arg,:ArgType,false,default)()

will return

:(arg::ArgType=default)

This feature is based on the MacroTools.combinearg function.

FunctionFactory

A FunctionFactory has 6 attributes:

name::Symbol: the name of the function

args::Vector{Union{Symbol,Expr}}: the positional arguments of the function

kwargs::Vector{Union{Symbol,Expr}}: the keyword arguments of the function

rtype::Union{Symbol,Expr}: the return type of the function

params::Vector{Union{Symbol,Expr}}: the method parameters specified by the where keyword

body::Union{Symbol,Expr}: the body of the function

All expressions that represent functions are allowed to be passed to the constructor, such as:

:(f()=nothing)

:(f(x)=x)

:(f(x::Int,y::Int;choice::Function=sum)=choice(x,y))

:(f(x::T,y::T;choice::Function=sum) where T<:Number=choice(x,y))

:((f(x::T,y::T;choice::Function=sum)::T) where T<:Number=choice(x,y))

:(function (f(x::T,y::T;choice::Function=sum)::T) where T<:Numbe
      choice(x,y)
  end
)

quote
    function (f(x::T,y::T;choice::Function=sum)::T) where T<:Numbe
        choice(x,y)
    end
end

Similarly, an instance can also be constructed from the macro @functionfactory:

@functionfactory (f(x::T,y::T;choice::Function=sum)::T) where T<:Number=choice(x,y)

The construction from and the convertion to such expressions are based on the MacroTools.splitdef and MacroTools.combinedef functions, respectively.

Note

Because the form f{T}(x::T,y::T;choice::Function=sum) has no longer been supported since Julia 0.7, the entry :params in the returned dict by MacroTools.splitarg is always missing. Therefore, we abandon its corresponding field in FunctionFactory but use the attribute :params to denote the :whereparams entry.

Other features include:

Positional arguments can be added by addargs! or @addargs!

Keyword arguments can be added by addkwargs! or @addkwargs!

Body can be extended by extendbody! or @extendbody!

Field

A Field has 2 attributes:

name::Symbol: the name of the field

type::Union{Symbol,Expr}: the type of the field

Legal expressions that can be passed to the constructor includes the following forms:

:(field)

:(field::FieldType)

:(field::ParametricType{Parameter})

The macro @field is also provided to help the construction directly from a field declaration:

@field field::FieldType

The construction from these expressions is based on the MacroTools.splitarg function and the convertion to these expressions is based on the MacroTools.combinefield function.

Type Factory

A TypeFactory has 6 attributes:

name::Symbol: the name of the struct

mutable::Bool: whether or not the struct is mutable

params::Vector{Union{Symbol,Expr}}: the type parameters of the struct

supertype::Union{Symbol,Expr}: the supertype of the struct

fields::Vector{Tuple{Symbol,Union{Symbol,Expr}}}: the fields of the struct

constructors::Vector{Expr}: the inner constructors of the struct

Any expression representing valid struct definitions can be passed to the constructor, whose common forms include the followings:

:(struct StructName end)

:(struct StructName{T} end)

:(struct Child{T} <: Parent{T} where T end)

:(struct Child{T} <: Parent{T} where T<:TypeParameter
      field1::FiledType1
      filed2::FieldType2
  end
)

quote
    struct Child{T} <: Parent{T} where T<:TypeParameter
        field1::FiledType1
        filed2::FieldType2
    end
end

Also, the macro @typefactory simplifies the construction directly from a type definition:

@typefactory struct Child{T} <: Parent{T} where T<:TypeParameter
                field1::FiledType1
                filed2::FieldType2
            end

The construction from these expressions is based on the MacroTools.splitstructdef function. Meanwhile, the convertion to the corresponding expression from a TypeFactory is based on the MacroTools.combinestructdef function.

Other features include:

Fields can be added by addfields! or @addfields!

Inner constructors can be added by addconstructors! or @addconstructors!

Manual

Hamiltonian.Utilities.Factory.Argument — Type.

Argument(name::Symbol,type::FExpr,slurp::Bool,default::Any)
Argument(name::Symbol;type::FExpr=:Any,slurp::Bool=false,default::Any=nothing)
Argument(expr::Expr)

The struct to describe a argument of a function.

source

Hamiltonian.Utilities.Factory.Argument — Method.

(a::Argument)()

Convert a Argument to the Expr representation of the argument it describes.

source

Hamiltonian.Utilities.Factory.Block — Type.

Block(parts::FExpr...)

The struct to describe a begin ... end block.

source

Hamiltonian.Utilities.Factory.Block — Method.

(b::Block)()

Convert a Block to the Expr representation of the begin ... end block it describes.

source

Hamiltonian.Utilities.Factory.Field — Type.

Field(name::Symbol,type::FExpr)
Field(name::Symbol;type::FExpr=:Any)
Field(expr::Expr)

The struct to describe a field of a struct.

source

Hamiltonian.Utilities.Factory.Field — Method.

(f::Field)()

Convert a Field to the Expr representation of the field it describes.

source

Hamiltonian.Utilities.Factory.FunctionFactory — Type.

FunctionFactory(name::Symbol,args::Vector,kwargs::Vector,rtype::FExpr,params::Vector,body::FExpr)
FunctionFactory(name::Symbol;args::Vector=Any[],kwargs::Vector=Any[],rtype::FExpr=:Any,params::Vector=Any[],body::FExpr)
FunctionFactory(expr::Expr)

The struct to describe a function.

source

Hamiltonian.Utilities.Factory.FunctionFactory — Method.

(ff::FunctionFactory)()

Convert a FunctionFactory to the Expr representation of the function it describes.

source

Hamiltonian.Utilities.Factory.TypeFactory — Type.

TypeFactory(name::Symbol,mutable::Bool,params::Vector,supertype::FExpr,fields::Vector,constructors::Vector)
TypeFactory(name::Symbol;mutable::Bool=false,params::Vector=Any[],supertype::FExpr=:Any,fields::Vector=Any[],constructors::Vector=Any[])
TypeFactory(expr::Expr)

The struct to describe a struct.

source

Hamiltonian.Utilities.Factory.TypeFactory — Method.

(tf::TypeFactory)()

Convert a TypeFactory to the Expr representation of the struct it describes.

source

Hamiltonian.Utilities.Factory.@addargs! — Macro.

@addargs! ff args::FExpr...

Add a couple of positional arguments to a function factory.

source

Hamiltonian.Utilities.Factory.@addconstructors! — Macro.

@addconstructors! rf constructors::Expr...

Add a couple of constructors to a type factory.

source

Hamiltonian.Utilities.Factory.@addfields! — Macro.

@addfields! rf fields::FExpr...

Add a couple of fields to a type factory.

source

Hamiltonian.Utilities.Factory.@addkwargs! — Macro.

@addkwargs! ff kwargs::FExpr...

Add a couple of keyword arguments to a function factory.

source

Hamiltonian.Utilities.Factory.@argument — Macro.

@argument expr::FExpr

Construct an Argument directly from an argument statement.

source

Hamiltonian.Utilities.Factory.@block — Macro.

@block parts::FExpr...

Construct a Block directly from a begin ... end block definition.

source

Hamiltonian.Utilities.Factory.@extendbody! — Macro.

@extendbody! ff parts::FExpr...

Extend the body of a function factory.

source

Hamiltonian.Utilities.Factory.@field — Macro.

@field expr::FExpr

Construct a Field directly from a field statement.

source

Hamiltonian.Utilities.Factory.@functionfactory — Macro.

@functionfactory expr::Expr

Construct a FunctionFactory directly from a function definition.

source

Hamiltonian.Utilities.Factory.@push! — Macro.

@push! b parts::FExpr...

Push other parts into the body of a block.

source

Hamiltonian.Utilities.Factory.@typefactory — Macro.

@typefactory expr::Expr

Construct a TypeFactory directly from a type definition.

source

Hamiltonian.Utilities.Factory.addargs! — Method.

addargs!(ff::FunctionFactory,args::FExpr...)
addargs!(ff::FunctionFactory,args::Argument...)

Add a couple of positional arguments to a function factory.

source

Hamiltonian.Utilities.Factory.addconstructors! — Method.

addconstructors!(rf::TypeFactory,constructors::Expr...)
addconstructors!(rf::TypeFactory,constructors::FunctionFactory...)

Add a couple of constructors to a type factory.

source

Hamiltonian.Utilities.Factory.addfields! — Method.

addfields!(rf::TypeFactory,fields::FExpr...)
addfields!(rf::TypeFactory,fields::Field...)

Add a couple of fields to a type factory.

source

Hamiltonian.Utilities.Factory.addkwargs! — Method.

addkwargs!(ff::FunctionFactory,kwargs::FExpr...)
addkwargs!(ff::FunctionFactory,kwargs::Argument...)

Add a couple of keyword arguments to a function factory.

source

Hamiltonian.Utilities.Factory.extendbody! — Method.

extendbody!(ff::FunctionFactory,parts::FExpr...)
extendbody!(ff::FunctionFactory,parts::Block...)

Extend the body of a function factory.

source

Hamiltonian.Utilities.Factory.FExpr — Constant.

Factory expression types.

source

Hamiltonian.Utilities.Factory.AbstractFactory — Type.

AbstractFactory

Abstract type for all concrete factories.

source

Base.:== — Method.

==(f1::F,f2::F) where F<:AbstractFactory

Overloaded == operator.

source

Base.push! — Method.

push!(b::Block,parts::FExpr...)
push!(b::Block,parts::Block...)

Push other parts into the body of a block.

source

Base.replace — Method.

replace(f::AbstractFactory;kwargs...)

Return a copy of a concrete AbstractFactory with some of the field values replaced by the keyword arguments.

source

Base.show — Method.

show(io::IO,f::AbstractFactory)

Show a concrete AbstractFactory.

source