Factory

Factory

The aim of Factory is to provide tools to hack into Julia codes without knowing the details of their abstract syntax trees and regularize the mechanism to "escape" variables in Expr expressions, so that users can manipulate the existing codes, modify them and generate new ones in macros. In particular, Factory in this module means the representation of certain blocks of Julia codes by a usual Julia struct. This representation is much easier to comprehend than the canonical Expr representation and makes it far more convenient to define macros. In general, we propose two basic requirements that any factory must satisfy:

FEATURE-1: A concrete factory can be constructed from a legal Expr expression that represents the block of codes it aims to represent;

FEATURE-2: The canonical Expr expression of a block of codes that a concrete factory represents can be obtained by "calling" the factory itself.

Also, we propose a simple mechanism to implement the escape of variables in factories:

FEATURE-3: During the construction from Expr expressions, a keyword argument unescaped, which is a tuple of Symbols, can be passed to the constructor, and all variables whose names are not in this keyword argument will be escaped.

Different from the former two requirements, this mechanism is just an optional feature of concrete factories. Meanwhile, all of these features also defines the basic interfaces to interact with factories.

Out of practical purposes, we implemente 7 kinds of factories, i.e. Inference, Argument, Parameter, Field, Block, FunctionFactory and TypeFactory, which represent a type inference, a function argument, a method or type parameter, a struct field, a begin ... end block, a function itself and a struct itself, respectively. Some of the basic methods making the above requirements fulfilled with these types are based on the powerful functions defined in MacroTools.

Inference

An Inference has 4 attributes:

head::Union{Symbol,Nothing}: the head of the type inference, which must be one of (nothing,:(::),:(<:),:curly)

name::Union{Symbol,Nothing}: the name of the type inference

params::Union{Inference,Vector{Inference},Nothing}: the parameters of the type inference

escape::Bool: whether the name of the type inference should be escaped when calling the factory

All valid expressions representing type inferences can be passed to the constructor:

julia> Inference(:T,unescaped=(:T,))
Inference(
  head:   nothing
  name:   T
  params: nothing
  escape: false
)

julia> Inference(:(::T),unescaped=(:T,))
Inference(
  head:   ::
  name:   nothing
  params: T
  escape: false
)

julia> Inference(:(<:Number))
Inference(
  head:   <:
  name:   nothing
  params: $(Expr(:escape, :Number))
  escape: false
)

julia> Inference(:(Vector{T}),unescaped=(:T,))
Inference(
  head:   curly
  name:   Vector
  params: Inference[T]
  escape: true
)

julia> Inference(:(Vector{Tuple{String,Int}}))
Inference(
  head:   curly
  name:   Vector
  params: Inference[($(Expr(:escape, :Tuple))){$(Expr(:escape, :String)), $(Expr(:escape, :Int))}]
  escape: true
)

julia> Inference(:(Type{<:Number}))
Inference(
  head:   curly
  name:   Type
  params: Inference[<:$(Expr(:escape, :Number))]
  escape: true
)

On the other hand, you can use the macro @inference to construct an Inference directly from a type inference:

julia> @inference Vector{Tuple{String,Int}}
Inference(
  head:   curly
  name:   Vector
  params: Inference[($(Expr(:escape, :Tuple))){$(Expr(:escape, :String)), $(Expr(:escape, :Int))}]
  escape: true
)

Note

Inference is a recursive struct, i.e. it recursively decomposes a type inference until the final type inference is just a Symbol.

When the input expression representing a type inference is a Symbol, the head and params attributes of the resulting Inference is nothing. Otherwise, its head is the same with that of the input expression, and the args of the input expression will be further decomposed, whose result will be stored in params.

When the head of the input expression is :(::) or :(<:), the params is an Inference whereas when the head of the input expression is :curly, the params is a Vector{Inference}.

Argument

An Argument has 4 attributes:

name::Union{Symbol,Nothing}: the name of the argument

type::Inference: the type inference of the argument

slurp::Bool: whether the argument should be expanded by ...

default::Any: the default value of the argument, nothing for those with no default values

All valid expressions representing the arguments of functions can be passed to the constructor:

julia> Argument(:arg)
Argument(
  name:    arg
  type:    $(Expr(:escape, :Any))
  slurp:   false
  default: nothing
)

julia> Argument(:(arg::ArgType),unescaped=(:ArgType,))
Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: nothing
)

julia> Argument(:(arg::ArgType...),unescaped=(:ArgType,))
Argument(
  name:    arg
  type:    ArgType
  slurp:   true
  default: nothing
)

julia> Argument(:(arg::ArgType=default),unescaped=(:ArgType,:default))
Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: default
)

Or you can use the macro @argument for construction directly from an argument declaration:

julia> @argument arg::ArgType=default (:ArgType,:default)
Argument(
  name:    arg
  type:    ArgType
  slurp:   false
  default: default
)

The construction from such expressions is based on the the MacroTools.splitarg function. On the other hand, calling an instance of Argument will get the corresponding Expr expression, e.g.,

julia> Argument(:(arg::ArgType=default),unescaped=(:ArgType,:default))()
:(arg::ArgType=default)

This feature is based on the MacroTools.combinearg function.

Parameter

A Parameter has 2 attributes:

name::Union{Symbol,Nothing}: the name of the parameter

type::Union{Inference,Nothing}: the type inference of the parameter

All expressions that represent a type parameter or a method parameter are allowed to be passed to the constructor:

julia> Parameter(:T)
Parameter(
  name: T
  type: nothing
)

julia> Parameter(:(<:Number))
Parameter(
  name: nothing
  type: $(Expr(:escape, :Number))
)

julia> Parameter(:(T<:Number))
Parameter(
  name: T
  type: $(Expr(:escape, :Number))
)

julia> Parameter(:(::Int))
Parameter(
  name: nothing
  type: $(Expr(:escape, :Int))
)

The macro @parameter completes the construction directly from a parameter declaration:

julia> @parameter T<:Number
Parameter(
  name: T
  type: $(Expr(:escape, :Number))
)

Note

We use nothing to denote a missing name or type.

Two subtle situations of type/method parameters, e.g. MyType{T} and MyType{Int}, are distinguished by Parameter(:T) and Parameter(:(::Int)). For the first case, Parameter(:T).name==:T and Parameter(:T).type==:nothing while for the second case, Parameter(:(::Int)).name==:nothing and Parameter(:(::Int)).name==:Int. Moreover, the callings of them return different forms, e.g. Parameter(:T)()==:T and Parameter(:(::Int))==:(<:$(esc(Int))).

Field

A Field has 2 attributes:

name::Symbol: the name of the field

type::Inference: the type inference of the field

Legal expressions can be used to construct a Field instance by its constructor:

julia> Field(:field)
Field(
  name: field
  type: $(Expr(:escape, :Any))
)

julia> Field(:(field::FieldType),unescaped=(:FieldType,))
Field(
  name: field
  type: FieldType
)

julia> Field(:(field::ParametricType{T}),unescaped=(:ParametricType,:T))
Field(
  name: field
  type: ParametricType{T}
)

The macro @field is also provided to help the construction directly from a field declaration:

julia> @field field::FieldType (:FieldType,)
Field(
  name: field
  type: FieldType
)

The construction from these expressions is based on the MacroTools.splitarg function and the convertion to these expressions is based on the MacroTools.combinefield function.

Block

A Block has only one attribute:

body::Vector{Any}: the body of the begin ... end block

Any expression can be passed to the constructor of Block:

julia> Block(:(x=1))
Block(
  body: Any[:(x = 1)]
)

julia> Block(:(x=1;y=2))
Block(
  body: Any[:(x = 1), :(#= none:1 =#), :(y = 2)]
)

julia> Block(:(begin x=1 end))
Block(
  body: Any[:(#= none:1 =#), :(x = 1)]
)

julia> Block(quote
               x=1
               y=2
           end)
Block(
  body: Any[:(#= none:2 =#), :(x = 1), :(#= none:3 =#), :(y = 2)]
)

Or you can construct a Block instance directly from any code by the macro @block:

julia> @block x=1 y=2
Block(
  body: Any[:(x = 1), :(y = 2)]
)

The body of a block can also be extended by the push! function or the @push! macro.

Note

The body of a Block is somewhat "flattened", i.e. it contains no begin ... end blocks. During the initialization, any such input block will be unblocked and added to the body part by part. So is the push! and @push! processes.

All LineNumberNodes generated by the input codes will also be included in the block's body. However, you can use rmlines! or @rmlines! to remove them from the body of an existing Block, or use rmlines or @rmlines to get a copy with them removed in the body.

Block is the only factory that DOES NOT cope with the "escape" mechanism we propose, because usually variables in a block are local ones and should not be escaped. However, the function escape is provided to escape the variables in a Expr expression with a reverse mechanism, where a tuple of variables to be escaped should be transmitted.

FunctionFactory

A FunctionFactory has 6 attributes:

name::Symbol: the name of the function

args::Vector{Argument}: the positional arguments of the function

kwargs::Vector{Argument}: the keyword arguments of the function

rtype::Inference: the return type of the function

params::Vector{Parameter}: the method parameters specified by the where keyword

body::Block: the body of the function

All expressions that represent functions are allowed to be passed to the constructor:

julia> FunctionFactory(:(f()=nothing))
FunctionFactory(
  name:   f
  args:   Argument[]
  kwargs: Argument[]
  rtype:  $(Expr(:escape, :Any))
  params: Parameter[]
  body:   begin
              nothing
          end
)

julia> FunctionFactory(:(f(x)=x))
FunctionFactory(
  name:   f
  args:   Argument[x::$(Expr(:escape, :Any))]
  kwargs: Argument[]
  rtype:  $(Expr(:escape, :Any))
  params: Parameter[]
  body:   begin
              x
          end
)

julia> FunctionFactory(:(f(x::Int,y::Int;choice::Function=sum)=choice(x,y)))
FunctionFactory(
  name:   f
  args:   Argument[x::$(Expr(:escape, :Int)), y::$(Expr(:escape, :Int))]
  kwargs: Argument[choice::$(Expr(:escape, :Function))=$(Expr(:escape, :sum))]
  rtype:  $(Expr(:escape, :Any))
  params: Parameter[]
  body:   begin
              choice(x, y)
          end
)

julia> FunctionFactory(:(f(x::T,y::T;choice::Function=sum) where T<:Number=choice(x,y)),unescaped=(:T,))
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::$(Expr(:escape, :Function))=$(Expr(:escape, :sum))]
  rtype:  $(Expr(:escape, :Any))
  params: Parameter[T <: $(Expr(:escape, :Number))]
  body:   begin
              choice(x, y)
          end
)

julia> FunctionFactory(:((f(x::T,y::T;choice::Function=sum)::T) where T<:Number=choice(x,y)),unescaped=(:T,))
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::$(Expr(:escape, :Function))=$(Expr(:escape, :sum))]
  rtype:  T
  params: Parameter[T <: $(Expr(:escape, :Number))]
  body:   begin
              choice(x, y)
          end
)

julia> FunctionFactory(:(
           function (f(x::T,y::T;choice::Function=sum)::T) where T<:Number
               choice(x,y)
           end
           ),
           unescaped=(:T,)
       )
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::$(Expr(:escape, :Function))=$(Expr(:escape, :sum))]
  rtype:  T
  params: Parameter[T <: $(Expr(:escape, :Number))]
  body:   begin
              #= none:3 =#
              choice(x, y)
          end
)

julia> FunctionFactory(
           quote
               function (f(x::T,y::T;choice::Function=sum)::T) where T<:Number
                   choice(x,y)
               end
           end,
           unescaped=(:T,)
       )
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::$(Expr(:escape, :Function))=$(Expr(:escape, :sum))]
  rtype:  T
  params: Parameter[T <: $(Expr(:escape, :Number))]
  body:   begin
              #= none:4 =#
              choice(x, y)
          end
)

Similarly, an instance can also be constructed from the macro @functionfactory:

julia> @functionfactory (f(x::T,y::T;choice::Function=sum)::T) where T<:Number=choice(x,y) (:T,)
FunctionFactory(
  name:   f
  args:   Argument[x::T, y::T]
  kwargs: Argument[choice::$(Expr(:escape, :Function))=$(Expr(:escape, :sum))]
  rtype:  T
  params: Parameter[T <: $(Expr(:escape, :Number))]
  body:   begin
              choice(x, y)
          end
)

The construction from and the convertion to such expressions are based on the MacroTools.splitdef and MacroTools.combinedef functions, respectively.

Note

Because the form f{T}(x::T,y::T;choice::Function=sum) has no longer been supported since Julia 0.7, the entry :params in the returned dict by MacroTools.splitarg is always missing. Therefore, we abandon its corresponding field in FunctionFactory but use the attribute :params to denote the :whereparams entry.

Other features include:

Positional arguments can be added by addargs! or @addargs!

Keyword arguments can be added by addkwargs! or @addkwargs!

Method parameters can be added by addparams! or @addparams!

Body can be extended by extendbody! or @extendbody!

TypeFactory

A TypeFactory has 6 attributes:

name::Symbol: the name of the struct

mutable::Bool: whether or not the struct is mutable

params::Vector{Parameter}: the type parameters of the struct

supertype::Inference: the supertype of the struct

fields::Vector{Field}: the fields of the struct

constructors::Vector{FunctionFactory}: the inner constructors of the struct

Any expression representing valid struct definitions can be passed to the constructor:

julia> TypeFactory(:(struct StructName end))
TypeFactory(
  name:         StructName
  mutable:      false
  params:       Parameter[]
  supertype:    $(Expr(:escape, :Any))
  fields:       Field[]
  constructors: FunctionFactory[]
)

julia> TypeFactory(:(struct StructName{T} end),unescaped=(:T,))
TypeFactory(
  name:         StructName
  mutable:      false
  params:       Parameter[T]
  supertype:    $(Expr(:escape, :Any))
  fields:       Field[]
  constructors: FunctionFactory[]
)

julia> TypeFactory(:(struct Child{T} <: Parent{T} end),unescaped=(:T,))
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T]
  supertype:    ($(Expr(:escape, :Parent))){T}
  fields:       Field[]
  constructors: FunctionFactory[]
)

julia> TypeFactory(:(
           struct Child{T<:Number} <: Parent{T}
               field1::T
               field2::T
           end
           ),
           unescaped=(:T,)
       )
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T <: $(Expr(:escape, :Number))]
  supertype:    ($(Expr(:escape, :Parent))){T}
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[]
)

julia> TypeFactory(
           quote
               struct Child{T<:Number} <: Parent{T}
                   field1::T
                   field2::T
               end
           end,
           unescaped=(:T,)
       )
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T <: $(Expr(:escape, :Number))]
  supertype:    ($(Expr(:escape, :Parent))){T}
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[]
)

Also, the macro @typefactory supports the construction directly from a type definition:

julia> @typefactory struct Child{T<:Number} <: Parent{T} field1::T; field2::T; Child(field1::T,field2::T=zero(T)) where T=new{T}(field1,field2) end (:T,)
TypeFactory(
  name:         Child
  mutable:      false
  params:       Parameter[T <: $(Expr(:escape, :Number))]
  supertype:    ($(Expr(:escape, :Parent))){T}
  fields:       Field[field1::T, field2::T]
  constructors: FunctionFactory[
                  function (Child(field1::T, field2::T=($(Expr(:escape, :zero)))(T); )::$(Expr(:escape, :Any))) where T
                      #= /home/travis/.julia/packages/MacroTools/4AjBS/src/utils.jl:302 =#
                      begin
                          new{T}(field1, field2)
                      end
                  end
                  ]
)

The construction from these expressions is based on the MacroTools.splitstructdef function. Meanwhile, the convertion to the corresponding expression from a TypeFactory is based on the MacroTools.combinestructdef function.

Other features include:

Fields can be added by addfields! or @addfields!

Type parameters can be added by addparams! or @addparams!

Inner constructors can be added by addconstructors! or @addconstructors!

More about escape

As you may have noticed, during the construction of factoies, we often provide a keyword argument unescaped to tell what variables should not be escaped. The criterion to determine such variables is quite direct and simple: if a variable is local to current module, or to be defined, or to be declared, it should not be escaped, while if it has been defined or declared in other modules, it should be escaped. Therefore, the name of a function argument, or a type/method parameter, or a struct field, should not be escaped because the first is a local variable and the latter two are to-be-declared ones. In fact, by design, these names will NEVER be escaped even when they are not provided in the unescaped keyword argument. Another common situation where unescaped variables exist occurs when methods or structs have parameters. In this case, the type inferences specified by these method/type parameters should not be escaped because they are also local ones. Note that this situation is not automatically handled by the above factories, because the method/type parameters can be modified after the factoies are constructed, which means, you have to offer the unescaped tuple by hand. To help decorate existing methods or structs, we define an inquiry function paramnames to obtain the method/type parameter names. By the way, the factory Block does not support variable escape by design, because it is usually used as the body of a function, which is always a local environment. This is in sharp contrast to the method/type declaration, where type inferences are usually involved with variables that are defined or declared in other modules. This also explains the philosophy behind our design that for such factories, variables that are not escaped should be assigned while for the function escape, variables to be escaped should be assigned.

Manual

Hamiltonian.Utilities.Factory.FExpr — Constant.

Factory expression types, which is defined as Union{Symbol,Expr}.

source

Hamiltonian.Utilities.Factory.Argument — Type.

Argument(name::Union{Symbol,Nothing},type::Inference,slurp::Bool,default::Any)
Argument(;name::Union{Symbol,Nothing}=nothing,type::Inference=Inference(:Any),slurp::Bool=false,default::Any=nothing)
Argument(expr::FExpr;unescaped::NTuple{N,Symbol}=()) where N)

The struct to describe a argument of a function.

source

Hamiltonian.Utilities.Factory.Argument — Method.

(a::Argument)()

Convert an Argument to the Expr representation of the argument it describes.

source

Hamiltonian.Utilities.Factory.Block — Type.

Block(parts::FExpr...)

The struct to describe a begin ... end block.

source

Hamiltonian.Utilities.Factory.Block — Method.

(b::Block)()

Convert a Block to the Expr representation of the begin ... end block it describes.

source

Hamiltonian.Utilities.Factory.Field — Type.

Field(name::Symbol,type::Inference)
Field(name::Symbol;type::FExpr=Inference(:Any))
Field(expr::Expr;unescaped::NTuple{N,Symbol}=()) where N

The struct to describe a field of a struct.

source

Hamiltonian.Utilities.Factory.Field — Method.

(f::Field)()

Convert a Field to the Expr representation of the field it describes.

source

Hamiltonian.Utilities.Factory.FunctionFactory — Type.

FunctionFactory(name::Symbol,args::Vector{Argument},kwargs::Vector{Argument},rtype::Inference,params::Vector{Parameter},body::Block)
FunctionFactory(    name::Symbol;
                    args::Vector{Argument}=Argument[],
                    kwargs::Vector{Argument}=Argument[],
                    rtype::Inference=Inference(:Any),
                    params::Vector{Parameter}=Parameter[],
                    body::Block=Block()
                    )
FunctionFactory(expr::Expr;unescaped::NTuple{N,Symbol}=()) where N

The struct to describe a function.

source

Hamiltonian.Utilities.Factory.FunctionFactory — Method.

(ff::FunctionFactory)()

Convert a FunctionFactory to the Expr representation of the function it describes.

source

Hamiltonian.Utilities.Factory.Inference — Type.

Inference(head::Union{Symbol,Nothing},name::Union{Symbol,Nothing},params::Union{Inference,Vector{Inference},Nothing},escape::Bool)
Inference(;
        head::Union{Symbol,Nothing}=nothing,
        name::Union{Symbol,Nothing}=nothing,
        params::Union{Inference,Vector{Inference},Nothing}=nothing,
        escape::Bool=true
        )
Inference(expr::FExpr;unescaped::NTuple{N,Symbol}=()) where N

The struct to describe a type inference.

source

Hamiltonian.Utilities.Factory.Inference — Method.

(i::Inference)()

Convert a Inference to the Expr representation of the type inference it describes.

source

Hamiltonian.Utilities.Factory.Parameter — Type.

Parameter(name::Union{Symbol,Nothing},type::Union{Inference,Nothing})
Parameter(;name::Union{Symbol,Nothing}=nothing,type::Union{Inference,Nothing}=nothing)
Parameter(expr::FExpr;unescaped::NTuple{N,Symbol}=()) where N

The struct to describe a parameter of a function or a type.

source

Hamiltonian.Utilities.Factory.Parameter — Method.

(p::Parameter)()

Convert a Parameter to the Expr representation of the parameter it describes.

source

Hamiltonian.Utilities.Factory.TypeFactory — Type.

TypeFactory(name::Symbol,mutable::Bool,params::Vector{Parameter},supertype::Inference,fields::Vector{Field},constructors::Vector{FunctionFactory})
TypeFactory(    name::Symbol;
                mutable::Bool=false,
                params::Vector{Parameter}=Parameter[],
                supertype::Inference=Inference(:Any),
                fields::Vector{Field}=Field[],
                constructors::Vector{FunctionFactory}=FunctionFactory[],
                )
TypeFactory(expr::Expr;unescaped::NTuple{N,Symbol}=()) where N

The struct to describe a struct.

source

Hamiltonian.Utilities.Factory.TypeFactory — Method.

(tf::TypeFactory)()

Convert a TypeFactory to the Expr representation of the struct it describes.

source

Hamiltonian.Utilities.Factory.@addargs! — Macro.

@addargs! ff args::FExpr...

Add a couple of positional arguments to a function factory.

source

Hamiltonian.Utilities.Factory.@addconstructors! — Macro.

@addconstructors! tf constructors::Expr...

Add a couple of constructors to a type factory.

source

Hamiltonian.Utilities.Factory.@addfields! — Macro.

@addfields! tf fields::FExpr...

Add a couple of fields to a type factory.

source

Hamiltonian.Utilities.Factory.@addkwargs! — Macro.

@addkwargs! ff kwargs::FExpr...

Add a couple of keyword arguments to a function factory.

source

Hamiltonian.Utilities.Factory.@addparams! — Macro.

@addparams! f params::FExpr...

Add a couple of method parameters to a function factory or a type factory.

source

Hamiltonian.Utilities.Factory.@argument — Macro.

@argument expr::FExpr unescaped::FExpr=:()

Construct an Argument directly from an argument statement.

source

Hamiltonian.Utilities.Factory.@block — Macro.

@block parts::FExpr...

Construct a Block directly from a begin ... end block definition.

source

Hamiltonian.Utilities.Factory.@extendbody! — Macro.

@extendbody! ff parts::FExpr...

Extend the body of a function factory.

source

Hamiltonian.Utilities.Factory.@field — Macro.

@field expr::FExpr unescaped::FExpr=:()

Construct a Field directly from a field statement.

source

Hamiltonian.Utilities.Factory.@functionfactory — Macro.

@functionfactory expr::FExpr unescaped::FExpr=:()

Construct a FunctionFactory directly from a function definition.

source

Hamiltonian.Utilities.Factory.@inference — Macro.

@inference expr::FExpr unescaped::FExpr=:()

Construct an Inference directly from a type inference.

source

Hamiltonian.Utilities.Factory.@parameter — Macro.

@parameter expr::FExpr unescaped::FExpr=:()

Construct a Parameter directly from an parameter statement.

source

Hamiltonian.Utilities.Factory.@push! — Macro.

@push! b parts::FExpr...

Push other parts into the body of a block.

source

Hamiltonian.Utilities.Factory.@rmlines! — Macro.

@rmlines! b::Expr

Remove line number nodes in the body of a block.

source

Hamiltonian.Utilities.Factory.@rmlines — Macro.

@rmlines b::Expr

Return a copy of a block with the line number nodes removed.

source

Hamiltonian.Utilities.Factory.@typefactory — Macro.

@typefactory expr::Expr unescaped::FExpr=:()

Construct a TypeFactory directly from a type definition.

source

Hamiltonian.Utilities.Factory.addargs! — Method.

addargs!(ff::FunctionFactory,args::Argument...)
addargs!(ff::FunctionFactory,args::FExpr...;unescaped::NTuple{N,Symbol}=()) where N

Add a couple of positional arguments to a function factory.

source

Hamiltonian.Utilities.Factory.addconstructors! — Method.

addconstructors!(tf::TypeFactory,constructors::FunctionFactory...)
addconstructors!(tf::TypeFactory,constructors::Expr...;unescaped::NTuple{N,Symbol}=()) where N

Add a couple of constructors to a type factory.

source

Hamiltonian.Utilities.Factory.addfields! — Method.

addfields!(tf::TypeFactory,fields::Field...)
addfields!(tf::TypeFactory,fields::FExpr...;unescaped::NTuple{N,Symbol}=()) where N

Add a couple of fields to a type factory.

source

Hamiltonian.Utilities.Factory.addkwargs! — Method.

addkwargs!(ff::FunctionFactory,kwargs::Argument...)
addkwargs!(ff::FunctionFactory,kwargs::FExpr...;unescaped::NTuple{N,Symbol}=()) where N

Add a couple of keyword arguments to a function factory.

source

Hamiltonian.Utilities.Factory.addparams! — Method.

addparams!(f::Union{FunctionFactory,TypeFactory},params::Parameter...)
addparams!(f::Union{FunctionFactory,TypeFactory},params::FExpr...;unescaped::NTuple{N,Symbol}=()) where N

Add a couple of method parameters to a function factory or a type factory.

source

Hamiltonian.Utilities.Factory.escape — Method.

escape(expr,::NTuple{N,Symbol}=()) where N
escape(expr::Symbol,escaped::NTuple{N,Symbol}=()) where N
escape(expr::Expr,escaped::NTuple{N,Symbol}=()) where N

Escape the symbols sepecified by escaped in the input expression.

source

Hamiltonian.Utilities.Factory.extendbody! — Method.

extendbody!(ff::FunctionFactory,parts::FExpr...)
extendbody!(ff::FunctionFactory,parts::Block...)

Extend the body of a function factory.

source

Hamiltonian.Utilities.Factory.paramnames — Method.

paramnames(expr::Union{Expr,FunctionFactory,TypeFactory}) -> NTuple{N,Symbol} where N

Get the type/method parameters names.

source

Hamiltonian.Utilities.Factory.rmlines! — Method.

rmlines!(b::Block)

Remove line number nodes in the body of a block.

source

Hamiltonian.Utilities.Factory.symbols — Method.

symbols(expr)

source

MacroTools.rmlines — Method.

rmlines(b::Block)

Return a copy of a block with the line number nodes removed.

source

Hamiltonian.Utilities.Factory.AbstractFactory — Type.

AbstractFactory

Abstract type for all concrete factories.

source

Base.:== — Method.

==(f1::F,f2::F) where F<:AbstractFactory

Overloaded == operator.

source

Base.Broadcast.broadcasted — Method.

SubTypeofAbstractFactory.(exprs::Vector{FExpr};unescaped::NTuple{N,Symbol}=()) where N
SubTypeofAbstractFactory.(exprs::NTuple{N,FExpr};unescaped::NTuple{M,Symbol}=()) where {N,M}

Broadcast construction of concrete subtypes of AbstractFactory.

source

Base.convert — Method.

convert(::Type{Tuple},f::Field)
convert(::Type{Tuple{Symbol,FExpr}},f::Field)

Convert a Field to tuple.

source

Base.push! — Method.

push!(b::Block,parts::FExpr...)
push!(b::Block,parts::Block...)

Push other parts into the body of a block.

source

Base.replace — Method.

replace(f::AbstractFactory;kwargs...)

Return a copy of a concrete AbstractFactory with some of the field values replaced by the keyword arguments.

source

Base.show — Method.

show(io::IO,f::AbstractFactory)

Show a concrete AbstractFactory.

source