Good quantum numbers

Good quantum numbers

Good qunatum numbers can be considered as the conserved labels for the bases of a Hilbert space when a quantum system hosts some symmetries. Here we only implement Abelian good quantum numbers because non-Abelian ones are far more complicated yet much less used. In practice, good quantum numbers can be integers or half integers, therefore, we use real numbers to denote them in this module for simplicity. Independent quantum numbers, such as the particle number and the spin z-component, can coexist at the same time. We use type QuantumNumber to represent the complete set of independent ones for a single basis of a Hilbert space, and type QuantumNumbers to represent the whole quantum numbers for the total bases.

QuantumNumber

The abstract type for the complete set of independent good quantum numbers for a single basis.

Main features include:

function fieldnames: get the names of the quantum numbers

function periods: get the periods of the quantum numbers

arithmetic operations: +, -, *, ^, ⊕, ⊗

hashable: concrete instances can be used as keys for a dict or a set

iterable: concrete instances are iterable over their values

comparable: two concrete instances can be compared

In particular, QuantumNumber <: AbstractNamedVector{Float64}, all features supported by AbstractNamedVector are also available for QuantumNumber. See also AbstractNamedVector.

For convenience, 4 kinds of good quantum numbers are predefined in this module, i.e.

SQN: for spin z-component reserved systems

PQN: for particle number reserved systems

SPQN: for both particle number and spin-z component reserved systems

Z2QN: for systems with a $Z_2$ conservation quantum number

Users who want to define their own $Z_N$-like quantum numbers must handle the periodicities in the construction function, otherwise, wrong results will be get when arithmetic operations, such as + or -, are involved. It is highly recommended to use the macro @quantumnumber to define your own concrete QuantumNumbers.

QuantumNumbers

The whole quantum numbers for the total bases, which has three forms

'G' form: the general form, which has no restriction for the contents of the QuantumNumbers

'U' form: the unitary form, which requires no duplicates in the contents of the QuantumNumbers

'C' form: the canonical form, which requires not only no duplicates but also accending-order storage in the contents of the QuantumNumber

Usually, 'G'-formed and 'U'-formed QuantumNumberses can be transformed to the corresponding 'C'-formed ones by the sort function.

To achieve high efficiency:

The contents of a QuantumNumbers are an homogenous array of a certain kind of concrete QuantumNumbers.

The quantum numbers are stored in a compressed form similiar to that of a CSC/CSR sparse matrix.

Main features include:

function eltype: get the concrete type of the quantum numbers it contains

arithmetic operations: +, -, *, ^, ⊗, ⊕

iterable: various iteration supports, including functions such as iterate, keys, values and pairs

...

For a complete summation of its features, please refer to the manual.

Manual

Hamiltonian.Utilities.GoodQuantumNumber.qnsbruteforce — Constant.

Choice associated with QuantumNumbers, meaning 'by brute force'.

source

Hamiltonian.Utilities.GoodQuantumNumber.qnscontents — Constant.

Choice associated with QuantumNumbers, meaning 'for contents'.

source

Hamiltonian.Utilities.GoodQuantumNumber.qnscounts — Constant.

Choice associated with QuantumNumbers, meaning 'by counts'.

source

Hamiltonian.Utilities.GoodQuantumNumber.qnsexpansion — Constant.

Choice associated with QuantumNumbers, meaning 'for expansion'.

source

Hamiltonian.Utilities.GoodQuantumNumber.qnsindices — Constant.

Choice associated with QuantumNumbers, meaning 'for indices'.

source

Hamiltonian.Utilities.GoodQuantumNumber.qnsindptr — Constant.

Choice associated with QuantumNumbers, meaning 'by indptr'.

source

Hamiltonian.Utilities.GoodQuantumNumber.qnsmontecarlo — Constant.

Choice associated with quantumnumbers, meaning 'by Monte Carlo'.

source

Hamiltonian.Utilities.GoodQuantumNumber.PQN — Type.

PQN(N::Real)

The concrete QuantumNumber of a quantum system with particle number N conserved.

source

Hamiltonian.Utilities.GoodQuantumNumber.QuantumNumber — Type.

Abstract type for all concrete quantum numbers for a single basis.

source

Hamiltonian.Utilities.GoodQuantumNumber.QuantumNumbers — Type.

QuantumNumbers(form::Char,contents::Vector{<:QuantumNumber},counts::Vector{Int},::typeof(qnscounts))
QuantumNumbers(form::Char,contents::Vector{<:QuantumNumber},indptr::Vector{Int},::typeof(qnsindptr))

The whole quantum numbers of the total bases of a Hilbert space. The default constructors construct a QuantumNumbers from a vector of concrete quantum numbers and an vector containing their counts or indptr.

source

Hamiltonian.Utilities.GoodQuantumNumber.QuantumNumbers — Type.

QuantumNumbers(qn::QuantumNumber,count::Int=1)

Construct a QuantumNumbers with one unique quantum number which occurs count times.

source

Hamiltonian.Utilities.GoodQuantumNumber.QuantumNumbers — Method.

QuantumNumbers(od::OrderedDict{<:QuantumNumber,Int})

Construct a QuantumNumbers from an ordered dict containing concrete quantum numbers and their counts.

source

Hamiltonian.Utilities.GoodQuantumNumber.QuantumNumbers — Method.

QuantumNumbers(od::OrderedDict{<:QuantumNumber,UnitRange{Int}})

Construct a QuantumNumbers from an ordered dict containing concrete quantum numbers and their slices.

source

Hamiltonian.Utilities.GoodQuantumNumber.SPQN — Type.

SPQN(N::Real,Sz::Real)

The concrete QuantumNumber of a quantum system with both particle number N and spin z-component Sz conserved.

source

Hamiltonian.Utilities.GoodQuantumNumber.SQN — Type.

SQN(Sz::Real)

The concrete QuantumNumber of a quantum system with spin z-component Sz conserved.

source

Hamiltonian.Utilities.GoodQuantumNumber.Z2QN — Type.

Z2QN(N::Real)

The concrete QuantumNumber of a quantum system with a Z₂-like conserved quantity.

source

Hamiltonian.Utilities.GoodQuantumNumber.@quantumnumber — Macro.

@quantumnumber typename fieldnames fieldperiods

Construct a concrete QuantumNumber with the type name being typename, fieldnames specified by fieldnames and periods specified by fieldperiods.

source

Hamiltonian.Utilities.GoodQuantumNumber.:⊕ — Method.

⊕(qns::NTuple{N,<:QuantumNumber},signs::NTuple{N,Int}=ntuple(i->1,N)) where N -> QuantumNumbers
⊕(qnses::NTuple{N,QuantumNumbers{QN}},signs::NTuple{N,Int}=ntuple(i->1,N)) where {N,QN<:QuantumNumber} -> QuantumNumbers{QN}

Get the direct sum of some QuantumNumbers or QuantumNumberses.

Note

Physically, the direct sum of a couple of QuantumNumbers or QuantumNumberses is defined by the direct sum of the bases of the Hilbert spaces they represent. Therefore, the input QuantumNumbers or QuantumNumberses must be homogenous. Inhomogenous 'QuantumNumber's must be direct producted first to ensure homogenity before the direct sum.

Apparently, the dimension of the result equals the summation of those of the inputs, which means, even for QuantumNumbers, the result will be naturally a QuantumNumbers because the dimension of the result is largeer than 1.

Signs of QuantumNumbers or QuantumNumberses can be provided when getting their direct sums.

source

Hamiltonian.Utilities.GoodQuantumNumber.:⊗ — Method.

⊗(::Type{QN},qn1::QuantumNumber,qn2::QuantumNumber) where QN<:QuantumNumber -> QN
⊗(qns::NTuple{N,<:QuantumNumber},signs::NTuple{N,Int}=ntuple(i->1,N)) where N -> QuantumNumber
⊗(qnses::NTuple{N,QuantumNumbers{QN}},signs::NTuple{N,Int}=ntuple(i->1,N)) where {N,QN<:QuantumNumber} -> QuantumNumbers{QN}

Get the direct product of some QuantumNumbers or QuantumNumberses.

Note

Physically, the direct product of a couple of QuantumNumbers or QuantumNumberses are defined by the direct product of the bases of the Hilbert spaces they represent. Therefore, QuantumNumbers with differenct types or QuantumNumberses with differenct eltypes are allowed to be direct producted in principle. However, for simplicity, we only implement a method which handle the situation of two QuantumNumbers with differenct types. The type of the result should be provided as the first parameter. Note that in this situation, the fieldnames and periods of the result type must be exactly equal to the flattened fieldnames and periods of the two input QuantumNumbers, which means, even the order of the input QuantumNumbers matters.

Apparently, the dimension of the result equals the product of those of the inputs. Therefore, the direct product of QuantumNumbers is also a QuantumNumber since its dimension is still one.

For other situations except the one mentioned in Note.1, the input QuantumNumbers or QuantumNumberses must be homogenous. Meanwhile, signs can also be provided for these situations. Note that each quantum number in the contents of the result is obtained by a summation of the corresponding quanum numbers out of the inputs with the correct signs. This is a direct observation of the Abelian nature of our quantum numbers.

source

Hamiltonian.Utilities.GoodQuantumNumber.PQNS — Method.

PQNS(N::Real)

Construct the QuantumNumbers of the Hilbert space of a single-particle state with at most N identical particles.

source

Hamiltonian.Utilities.GoodQuantumNumber.SPQNS — Method.

SPQNS(S::Real)

Construct the QuantumNumbers of the Hilbert space of a single site with internal degrees of freedom that can be ascribed to a spin S.

source

Hamiltonian.Utilities.GoodQuantumNumber.SQNS — Method.

SQNS(S::Real)

Construct the QuantumNumbers of the Hilbert space of a signle spin S.

source

Hamiltonian.Utilities.GoodQuantumNumber.SzPQNS — Method.

SzPQNS(Sz::Real)

Construct the QuantumNumbers of the Hilbert space of a single-paritcle state with at most one particle whose spin-z component is Sz.

source

Hamiltonian.Utilities.GoodQuantumNumber.Z2QNS — Method.

Z2QNS()

Construct the QuantumNumbers of a $Z_2$ Hilbert space.

source

Hamiltonian.Utilities.GoodQuantumNumber.decompose — Method.

decompose(qnses::NTuple{N,QuantumNumbers{QN}},target::QN,signs::NTuple{N,Int},::typeof(qnsbruteforce);nmax::Int=20) where {N,QN<:QuantumNumber} -> Vector{NTuple{N,Int}}
decompose(qnses::NTuple{N,QuantumNumbers{QN}},target::QN,signs::NTuple{N,Int},::typeof(qnsmontecarlo);nmax::Int=20) where {N,QN<:QuantumNumber} -> Vector{NTuple{N,Int}}

Find a couple of decompositions of target with respect to qnses.

Note

A tuple of integers (i₁,i₂,...) is called a decomposition of a given target with respect to the given qnses if and only if they satisfy the "decomposition rule":

\[\sum_\text{j} \text{signs}[\text{j}]\times\text{qnses}[\text{j}][\text{i}_{\text{j}}]==\text{target}\]

This equation is in fact a kind of a set of restricted linear Diophantine equations. Indeed, our quantum numbers are always discrete Abelian ones and all instances of a concrete QuantumNumber forms a module over the ring of integers. Therefore, each quantum number can be represented as a integral multiple of the unit element of the Abelian module, which results in the final reduction of the above equation to a set of linear Diophantine equations. Then finding a decomposition is equivalent to find a solution of the reduced linear Diophantine equations, with the restriction that the quantum numbers constructed from the solution should be in the corresponding qnses. Here we provide two methods to find such decompositions, one is by brute force (qnsbruteforce case), and the other is by Monte Carlo simultatioins (qnsmontecarlo case).

source

Hamiltonian.Utilities.GoodQuantumNumber.dimension — Method.

dimension(qns::QuantumNumbers) -> Int

The dimension of the Hilbert space a QuantumNumbers represents.

source

Hamiltonian.Utilities.GoodQuantumNumber.dimension — Method.

dimension(::Type{<:QuantumNumber}) -> Int
dimension(::QuantumNumber) -> Int

The dimension of the Hilbert space a QuantumNumber represents. Apparently, this is always 1.

source

Hamiltonian.Utilities.GoodQuantumNumber.expand — Method.

expand(qns::QuantumNumbers,::typeof(qnscontents)) -> Vector{qns|>eltype}
expand(qns::QuantumNumbers,::typeof(qnsindices)) -> Vector{Int}

Expand the contents (qnscontents case) or indices (qnsindices case) of a QuantumNumbers to the uncompressed form.

source

Hamiltonian.Utilities.GoodQuantumNumber.regularize! — Method.

regularize!(::Type{QN},array::AbstractVector{Float64}) where QN<:QuantumNumber
regularize!(::Type{QN},array::AbstractMatrix{Float64}) where QN<:QuantumNumber

Regularize the elements of an array in place so that it can represent quantum numbers.

source

Hamiltonian.Utilities.GoodQuantumNumber.regularize — Method.

regularize(::Type{QN},array::Union{AbstractVector{Float64},AbstractMatrix{Float64}}) where {QN<:QuantumNumber}

Regularize the elements of an array and return a copy that can represent quantum numbers.

source

Hamiltonian.Utilities.GoodQuantumNumber.reorder — Method.

reorder(qns::QuantumNumbers,permutation::Vector{Int},::typeof(qnscontents)) -> QuantumNumbers
reorder(qns::QuantumNumbers,permutation::Vector{Int},::typeof(qnsexpansion)) -> QuantumNumbers

Reorder the quantum numbers contained in a QuantumNumbers with a permutation and return the new one. For qnscontents case, the permutation is for the contents of the original QuantumNumbers while for qnsexpansion case, the permutation is for the expansion of the original QuantumNumbers.

source

Hamiltonian.Utilities.GoodQuantumNumber.subset — Method.

subset(qns::QuantumNumbers{QN},target::QN) where QN<:QuantumNumber -> QuantumNumbers{QN}
subset(qns::QuantumNumbers{QN},targets::NTuple{N,QN}) where {N,QN<:QuantumNumber} -> QuantumNumbers{QN}

Find a subset of a QuantumNumbers by picking out the quantum numbers in targets.

source

Hamiltonian.Utilities.GoodQuantumNumber.toordereddict — Method.

toordereddict(qns::QuantumNumbers,::typeof(qnsindptr)) -> OrderedDict{qns|>eltype,UnitRange{Int}}
toordereddict(qns::QuantumNumbers,::typeof(qnscounts)) -> OrderedDict{qns|>eltype,Int}

Convert a QuantumNumbers to an ordered dict.

source

Hamiltonian.Utilities.GoodQuantumNumber.ukron — Method.

ukron(qnses::NTuple{N,QuantumNumbers{QN}},signs::NTuple{N,Int}=ntuple(i->1,N)) where {N,QN<:QuantumNumber} -> QuantumNumbers{QN},Dict{QN,Dict{NTuple{N,QN},UnitRange{Int}}}

Unitary Kronecker product of several QuantumNumberses. The product result as well as the records of the product will be returned.

Note

All input QuantumNumbers must be 'U' formed or 'C' formed.

Since duplicate quantum number are not allowed in 'U' formed and 'C' formed QuantumNumberses, in general, there exists a merge process of duplicate quantum numbers in the product result. Therefore, records are needed to keep track of this process, which will be returned along with the product result. The records are stored in a Dict{QN,Dict{NTuple{N,QN},UnitRange{Int}}} typed dict, in which, for each unduplicate quantum number qn in the product result, there exist a record Dict((qn₁,qn₂,...)=>start:stop,...) telling what quantum numbers (qn₁,qn₂,...) a mereged duplicate qn comes from and what slice start:stop this merged duplicate corresponds.

source

Base.:* — Method.

*(qn::QuantumNumber,factor::Integer) -> QuantumNumber
*(factor::Integer,qn::QuantumNumber) -> QuantumNumber
*(qns::QuantumNumbers,factor::Integer) -> QuantumNumbers
*(factor::Integer,qns::QuantumNumbers) -> QuantumNumbers

Overloaded * operator for the multiplication between an integer and a QuantumNumber or a QuantumNumbers.

source

Base.:+ — Method.

+(qn::QuantumNumber) -> QuantumNumber
+(qn::QN,qns::QN...) where QN<:QuantumNumber -> QN
+(qns::QuantumNumbers) -> QuantumNumbers
+(qn::QN,qns::QuantumNumbers{QN}) where QN<:QuantumNumber -> QuantumNumbers{QN}
+(qns::QuantumNumbers{QN},qn::QN) where QN<:QuantumNumber -> QuantumNumbers{QN}

Overloaded + operator for QuantumNumber and QuantumNumbers.

Note

The addition between a QuantumNumbers and a QuantumNumber is just a global shift of the contents of the QuantumNumbers by the QuantumNumber, therefore, the result is a QuantumNumbers.

+ cannot be used between two QuantumNumbers because the result is ambiguous. Instead, use ⊕ for direct sum and ⊗ for direct product.

To ensure type stability, two QuantumNumber can be added together if and only if they are of the same type.

Similarly, a QuantumNumber and a QuantumNumbers can be added together if and only if the former's type is the same with the latter's eltype.

source

Base.:- — Method.

-(qn::QuantumNumber) -> QuantumNumber
-(qn1::QN,qn2::QN) where QN<:QuantumNumber -> QN
-(qns::QuantumNumbers) -> QuantumNumbers
-(qn::QN,qns::QuantumNumbers{QN}) where QN<:QuantumNumber -> QuantumNumbers{QN}
-(qns::QuantumNumbers{QN},qn::QN) where QN<:QuantumNumber -> QuantumNumbers{QN}

Overloaded - operator for QuantumNumber and QuantumNumbers.

Note

The subtraction between a QuantumNumbers and a QuantumNumber is just a global shift of the contents of the QuantumNumbers by the QuantumNumber, therefore, the result is a QuantumNumbers.

- cannot be used between two QuantumNumbers because the result is ambiguous. Instead, use ⊕ with signs for direct sum and ⊗ with signs for direct product.

To ensure type stability, a QuantumNumber can be subtracted by another QuantumNumber if and only if they are of the same type.

Similarly, a QuantumNumber can be subtracted by a QuantumNumbers or vice versa if and only if the former's type is the same with the latter's eltype.

source

Base.:== — Method.

==(qns1::QuantumNumbers,qns2::QuantumNumbers) -> Bool

Overloaded == operator. Two QuantumNumberses are equal to each other if and only if both their contentses and indptrs are elementwise equal to each other.

Note

It is not necessary for two QuantumNumberses to have the same eltype nor the same form to be equal to each other.

source

Base.:^ — Method.

^(qn::QuantumNumber,factor::Integer) -> QuantumNumber
^(qns::QuantumNumbers,factor::Integer) -> QuantumNumbers

Overloaded ^ operator for QuantumNumber and QuantumNumbers. This operation translates into the direct product ⊗ of factor copies of qn or qns.

source

Base.eltype — Method.

eltype(::Type{<:QuantumNumbers{QN}}) where QN
eltype(qns::QuantumNumbers)

Get the type of the concrete QuantumNumber contained in a QuantumNumbers.

source

Base.findall — Method.

findall(qns::QuantumNumbers{QN},target::QN,choice::Union{typeof(qnscontents),typeof(qnsexpansion)}) where QN<:QuantumNumber -> Vector{Int}
findall(qns::QuantumNumbers{QN},targets::NTuple{N,QN},::typeof(qnscontents)) where {N,QN<:QuantumNumber} -> Vector{Int}
findall(qns::QuantumNumbers{QN},targets::NTuple{N,QN},::typeof(qnsexpansion)) where {N,QN<:QuantumNumber} -> Vector{Int}

Find all the indices of the target quantum numbers in the contents (qnscontents case) or the expansion (qnsexpansion case) of a QuantumNumbers.

source

Base.getindex — Method.

getindex(qns::QuantumNumbers,index::Int) -> QuantumNumber
getindex(qns::QuantumNumbers,slice::UnitRange{Int}) -> QuantumNumbers
getindex(qns::QuantumNumbers,indices::Vector{Int}) -> QuantumNumbers

Overloaded [] operator.

Note

For a QuantumNumbers, all these getindex functions act on its contents, i.e. its compressed data, but not on its expansion, i.e. the uncompressed data. This definition is consistent with the length function.

When the index is an integer, the result is a QuantumNumber, while when the index is a unit range or a vector of intgers, the result is a QuantumNumbers. The logic is quite reasonable because such behaviors are much alike to those of a vector container.

source

Base.iterate — Function.

iterate(qns::QuantumNumbers,state::Int=1)
iterate(rv::Iterators.Reverse{<:QuantumNumbers},state::Int=length(rv.itr,false))

Iterate or reversely iterate over the concrete QuantumNumbers contained in a QuantumNumbers.

source

Base.keys — Method.

keys(qns::QuantumNumbers) -> Vector{qns|>eltype}

Iterate over the concrete QuantumNumbers contained in a QuantumNumbers.

source

Base.kron — Method.

kron(::Type{QN},qn1::QuantumNumber,qn2::QuantumNumber) where QN<:QuantumNumber -> QN
kron(qns::NTuple{N,<:QuantumNumber},signs::NTuple{N,Int}=ntuple(i->1,N)) where N -> QuantumNumber
kron(qnses::NTuple{N,QuantumNumbers{QN}},signs::NTuple{N,Int}=ntuple(i->1,N)) where {N,QN<:QuantumNumber} -> QuantumNumbers{QN}

Kronecker product of some QuantumNumbers or QuantumNumberses. This is defined to be equivalent to the direct product ⊗.

source

Base.length — Method.

length(qns::QuantumNumbers) -> Int

Get the number of unduplicate qunatum numbers in the QuantumNumbers.

source

Base.pairs — Method.

pairs(qns::QuantumNumbers,choice::Union{typeof(qnsindptr),typeof(qnscounts)})

Iterate over the QuantumNumber=>slice or QuantumNumber=>count pairs.

source

Base.show — Method.

show(io::IO,qns::QuantumNumbers)

Show a QuantumNumbers.

source

Base.sort — Method.

sort(qns::QuantumNumbers) -> QuantumNumbers,Vector{Int}

Sort the quantum numbers of a QuantumNumber, return the sorted QuantumNumber and the permutation array that sorts the expansion of the original QuantumNumbers.

source

Base.string — Method.

string(qns::QuantumNumbers) -> String

Convert a QuantumNumbers to string.

source

Base.values — Method.

values(qns::QuantumNumbers,::typeof(qnsindptr))
values(qns::QuantumNumbers,::typeof(qnscounts))

Iterate over the slices/counts of the QuantumNumbers.

source