Introduction

Introduction

This module contains the utilities of the Hamiltonian package, all of whose variables will NOT be exported by Hamiltonian.

Useful constants and functions

Hamiltonian.Utilities.atol — Constant.

Absolute tolerance for float numbers.

source

Hamiltonian.Utilities.rtol — Constant.

Relative tolerance for float numbers.

source

Hamiltonian.Utilities.Float — Type.

Default float type.

source

Hamiltonian.Utilities.forder — Constant.

forder

Indicate that the convertion between Cartesian index and linear index is using the Fortran order.

source

Hamiltonian.Utilities.corder — Constant.

corder

Indicate that the convertion between Cartesian index and linear index is using the C/C++ order.

source

Hamiltonian.Utilities.ind2sub — Function.

ind2sub(dims::Tuple,ind::Int,order::FOrder) -> Tuple
ind2sub(dims::Tuple,ind::Int,order::COrder) -> Tuple

Convert an linear index to Cartesian index. Fortran-order or C-order can be assigned.

source

Hamiltonian.Utilities.sub2ind — Function.

sub2ind(dims::NTuple{N,Int},inds::NTuple{N,Int},order::FOrder) where N -> Int
sub2ind(dims::NTuple{N,Int},inds::NTuple{N,Int},order::COrder) where N -> Int

Convert an Cartesian index to linear index. Fortran-order or C-order can be assigned.

source

Hamiltonian.Utilities.decimaltostr — Function.

decimaltostr(number::Integer,n::Int=5)
decimaltostr(number::Rational,n::Int=5)
decimaltostr(number::AbstractFloat,n::Int=5)
decimaltostr(number::Complex,n::Int=5)

Convert a number to a string with at most n decimal places.

source

Hamiltonian.Utilities.ordinal — Function.

ordinal(number::Interger)

Convert a positive number to its corresponding ordinal.

source

Generic functions for overloading

Hamiltonian.Utilities.:⊕ — Function.

Generic interface of the direct sum of some types.

source

Hamiltonian.Utilities.:⊗ — Function.

Generic interface of the direct product of some types.

source

Hamiltonian.Utilities.rank — Function.

Generic interface of the rank of some types.

source

Hamiltonian.Utilities.dimension — Function.

Generic interface of the dimension of some types.

source

Hamiltonian.Utilities.permute — Function.

Generic interface of permuting of some types.

source

Prerequisites for Essentials

Factory

Escape mechanisms

Concrete factories

Manual

Composite structure

CompositeNTuple

CompositeVector

CompositeDict

Manul

Tree

AbstractTree

TreeCore and SimpleTree

Manual

Named vector

AbstractNamedVector

HomoNamedVector

Manual

Algebra over fields

ID

VectorSpace

Element and Elements

Manual

Necessities for Algorithms

Quantum numbers

AbstractQuantumNumber

QuantumNumbers

Manual