NamedVector

NamedVector

A named vector is similiar to a named tuple, which associate each of its values with a name, represented with a Symbol. Unlike named tuples, the values of a named vector can be modified. Yet the names of a named vector cannot be changed. With experiences of namedtuple in Python, we treat the names of a named vector as type traits, by overloading the Base.filednames function, so that all instances of a certain concrete named vector share the same names, which is different from the predefined NamedTuple in Julia. For users who want to define their own concrete named vector, we recommend the use of the macro @namedvector.

Hamiltonian.Utilities.NamedVector.AbstractNamedVector — Type.

AbstractNamedVector{T,A<:AbstractVector{T}}

Abstract type for all concrete named vectors. To subtype it, please note:

The concrete types must have the field values::A, which is used to store the values by design. A recommended template for the subtype is struct YourNamedVector{T,A} <: AbstractNamedVector{T,A} values::A end

The concrete types must implement their own Base.fieldnames, which defines the names of the type. A recommended template for this method is Base.fieldnames(::Type{YourNamedVector},private=false)=private ? tuple(YourNames...,:values) : YourNames

Arithmetic operations, such as +, -, *, \, %, ÷, etc. are NOT supported. Thus users have to overload these operations themselves if needed.

source

Hamiltonian.Utilities.NamedVector.@namedvector — Macro.

@namedvector typename fieldnames dtype::Union{Expr,Symbol}=:nothing atype::Union{Expr,Symbol}=:nothing supertypename=:AbstractNamedVector

Construct a concrete named vector with the type name being typename, fieldnames specified by fieldnames, and optionally, the type parameters specified by dtype and atype, and the supertype specified by supertypename.

source

Base.:== — Method.

==(nv1::AbstractNamedVector,nv2::AbstractNamedVector)

Overloaded == operator.

source

Base.eltype — Method.

eltype(::Type{NV},choice::Int=1) where NV<:AbstractNamedVector{T,A} where {T,A}

Get the type parameter of a concrete AbstractNamedVector.

source

Base.getindex — Method.

getindex(nv::AbstractNamedVector,index::Int)

Get the value by the [] syntax.

source

Base.getproperty — Method.

getproperty(nv::AbstractNamedVector,key::Symbol)

Get the value by the . syntax.

source

Base.hash — Method.

hash(nv::AbstractNamedVector,h::UInt)

Hash a concrete AbstractNamedVector.

source

Base.iterate — Function.

iterate(nv::AbstractNamedVector)
iterate(nv::AbstractNamedVector,state)

Iterate over the values of a concrete AbstractNamedVector.

source

Base.keys — Method.

keys(nv::AbstractNamedVector)

Iterate over the names.

source

Base.length — Method.

length(nv::AbstractNamedVector)

Get the length of a concrete AbstractNamedVector.

source

Base.pairs — Method.

pairs(nv::AbstractNamedVector)

Iterate over the name=>value pairs.

source

Base.replace — Method.

replace(nv::AbstractNamedVector;kwargs...)

Return a copy of a concrete AbstractNamedVector with some of the filed values replaced by the keyword arguments.

source

Base.setindex! — Method.

setindex!(nv::AbstractNamedVector,value,index::Int)

Set the value by the [] syntax.

source

Base.setproperty! — Method.

setproperty!(nv::AbstractNamedVector,key::Symbol,value)

Set the value by the . syntax. Note the value of the field :values as a whole can not be changed directly by design. Instead, the elements of :values can be changed.

source

Base.show — Method.

show(io::IO,nv::AbstractNamedVector)

Show a concrete AbstractNamedVector.

source

Base.values — Method.

values(nv::AbstractNamedVector)

Iterate over the values.

source

Base.zero — Method.

zero(nv::AbstractNamedVector)

Get a concrete AbstractNamedVector with all values being zero.

source